### 2026-01-18 #9850 @editorialbot
[PRE REVIEW]: pharmsol: A high-performance Rust library for pharmacokinetic/pharmacodynamic modeling and simulation
**Submitting author:** @mhovd (<a href="http://orcid.org/0000-0002-6077-0934">Markus Hovd</a>)
**Repository:** https://github.com/LAPKB/pharmsol
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.22
