### 2026-01-18 eee53fc
Update joss/paper.md
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
