### 2026-01-19 17d4bd4
paper: Minor changes
