### 2026-01-19 0e52970
paper: Pre-review changes (#201)
* paper: Pre-review changes

* Update joss/paper.md

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* paper: Comments by Julian (#202)

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
Co-authored-by: Julian Otalvaro <juliandavid347@gmail.com>
