+++ joss/paper.bib
@@ -76,4 +76,24 @@
   publisher = {MDPI AG},
 }
 
+@Book{npod,
+  author    = {Hovd, Markus and Kryshchenko, Alona and Neely, Michael N. and Otalvaro, Julian and Schumitzky, Alan and Yamada, Walter M.},
+  editor    = {Cristina Garcia-Cardona and Harlin Lee},
+  publisher = {Springer Nature Switzerland},
+  title     = {A Non-parametric Optimal Design Algorithm for Population Pharmacokinetics},
+  year      = {2025},
+  address   = {Cham},
+  isbn      = {9783031878046},
+  number    = {37},
+  series    = {Association for Women in Mathematics Series},
+  booktitle = {Advances in Data Science: Women in Data Science and Mathematics (WiSDM) 2023},
+  doi       = {10.1007/978-3-031-87804-6_12},
+  issn      = {2364-5741},
+  pages     = {259--274},
+  pagetotal = {1357},
+  ppn_gvk   = {1937384802},
+  subtitle  = {Women in data science and mathematics (WiSDM) 2023},
+  url       = {https://doi.org/10.1007/978-3-031-87804-6_12},
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
