### 2026-01-19 #9850 @mhovd
Dear @arfon, thank you for handling our submission!

We have updated our submission to adhere to the new submission requirements.
The relevant changes are gathered in this commit: https://github.com/LAPKB/pharmsol/commit/0e52970dc8dcc596acb828c039683f2cc1b587df to the `joss` branch of the repository.

Despite the relatively small _current_ research impact, we believe that the potential research impact of `pharmsol` is great. At least for our research group, it has been very useful for pharmacokinetic modeling and simulation.

Please let us know if any additional editorial changes are warranted, otherwise we look forward to the review process.

Unfortunately, I could not find any reviewers with a background in pharmacology/pharmacokinetics AND Rust. Some relevant reviewers with a relevant background in pharmacology could be: [ashahide](https://github.com/ashahide), [metelkin](https://github.com/metelkin), or [ajbaird](https://github.com/ajbaird).  Of the reviewers with a background in Rust, and a background related to pharmacology (e.g. biology) could be [maxall41](https://github.com/maxall41), [luciorq](https://github.com/luciorq), and [axiomcura](https://github.com/axiomcura).
