### 2026-01-21 f267be1
suggestions: Suggestions for renamed items (#196)
* chore: Rename modules and structures

* Update src/error/mod.rs

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Update src/data/parser/mod.rs

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

* Name changes

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
