+++ src/data/parser/pmetrics.rs
@@ -4,45 +4,11 @@ use serde::de::{MapAccess, Visitor};
 use serde::{de, Deserialize, Deserializer, Serialize};
 use std::collections::HashMap;
 
+use crate::data::row::build_data;
+use crate::data::row::DataError;
+use crate::data::row::DataRow;
 use std::fmt;
 use std::str::FromStr;
-use thiserror::Error;
-
-/// Custom error type for the module
-#[allow(private_interfaces)]
-#[derive(Error, Debug, Clone)]
-pub enum PmetricsError {
-    /// Error encountered when reading CSV data
-    #[error("CSV error: {0}")]
-    CSVError(String),
-    /// Error during data deserialization
-    #[error("Parse error: {0}")]
-    SerdeError(String),
-    /// Encountered an unknown EVID value
-    #[error("Unknown EVID: {evid} for ID {id} at time {time}")]
-    UnknownEvid { evid: isize, id: String, time: f64 },
-    /// Required observation value (OUT) is missing
-    #[error("Observation OUT is missing for {id} at time {time}")]
-    MissingObservationOut { id: String, time: f64 },
-    /// Required observation output equation (OUTEQ) is missing
-    #[error("Observation OUTEQ is missing in for {id} at time {time}")]
-    MissingObservationOuteq { id: String, time: f64 },
-    /// Required infusion dose amount is missing
-    #[error("Infusion amount (DOSE) is missing for {id} at time {time}")]
-    MissingInfusionDose { id: String, time: f64 },
-    /// Required infusion input compartment is missing
-    #[error("Infusion compartment (INPUT) is missing for {id} at time {time}")]
-    MissingInfusionInput { id: String, time: f64 },
-    /// Required infusion duration is missing
-    #[error("Infusion duration (DUR) is missing for {id} at time {time}")]
-    MissingInfusionDur { id: String, time: f64 },
-    /// Required bolus dose amount is missing
-    #[error("Bolus amount (DOSE) is missing for {id} at time {time}")]
-    MissingBolusDose { id: String, time: f64 },
-    /// Required bolus input compartment is missing
-    #[error("Bolus compartment (INPUT) is missing for {id} at time {time}")]
-    MissingBolusInput { id: String, time: f64 },
-}
 
 /// Read a Pmetrics datafile and convert it to a [Data] object
 ///
@@ -56,7 +22,7 @@ pub enum PmetricsError {
 ///
 /// # Returns
 ///
-/// * `Result<Data, PmetricsError>` - A result containing either the parsed [Data] object or an error
+/// * `Result<Data, DataError>` - A result containing either the parsed [Data] object or an error
 ///
 /// # Example
 ///
@@ -78,32 +44,32 @@ pub enum PmetricsError {
 ///
 /// For specific column definitions, see the `Row` struct.
 #[allow(dead_code)]
-pub fn read_pmetrics(path: impl Into<String>) -> Result<Data, PmetricsError> {
+pub fn read_pmetrics(path: impl Into<String>) -> Result<Data, DataError> {
     let path = path.into();
 
     let mut reader = csv::ReaderBuilder::new()
         .comment(Some(b'#'))
         .has_headers(true)
         .from_path(&path)
-        .map_err(|e| PmetricsError::CSVError(e.to_string()))?;
+        .map_err(|e| DataError::CSVError(e.to_string()))?;
     // Convert headers to lowercase
     let headers = reader
         .headers()
-        .map_err(|e| PmetricsError::CSVError(e.to_string()))?
+        .map_err(|e| DataError::CSVError(e.to_string()))?
         .iter()
         .map(|h| h.to_lowercase())
         .collect::<Vec<_>>();
[…]
