+++ src/data/row.rs
@@ -1,26 +1,12 @@
-//! Normalized row representation for flexible data parsing
-//!
-//! This module provides a format-agnostic intermediate representation that decouples
-//! column naming/mapping from event creation logic. Any data source (CSV with custom
-//! columns, Excel, DataFrames) can construct [`NormalizedRow`] instances, then use
-//! [`NormalizedRow::into_events()`] to get properly parsed pharmsol Events.
-//!
-//! # Design Philosophy
-//!
-//! The key insight is separating two concerns:
-//! 1. **Row Normalization** - Transform arbitrary input formats into a standard representation
-//! 2. **Event Creation** - Convert normalized rows into pharmsol Events (with ADDL expansion, etc.)
-//!
-//! This allows any consumer (GUI applications, scripts, other tools) to bring their own
-//! "column mapping" while reusing parsing logic.
+//! Row representation of [Data] for flexible parsing
 //!
 //! # Example
 //!
 //! ```rust
-//! use pharmsol::data::parser::NormalizedRow;
+//! use pharmsol::data::parser::DataRow;
 //!
 //! // Create a dosing row with ADDL expansion
-//! let row = NormalizedRow::builder("subject_1", 0.0)
+//! let row = DataRow::builder("subject_1", 0.0)
 //!     .evid(1)
 //!     .dose(100.0)
 //!     .input(1)
@@ -33,15 +19,15 @@
 //! ```
 //!
 
-use super::PmetricsError;
 use crate::data::*;
 use std::collections::HashMap;
+use thiserror::Error;
 
 /// A format-agnostic representation of a single data row
 ///
 /// This struct represents the canonical fields needed to create pharmsol Events.
 /// Consumers construct this from their source data (regardless of column names),
-/// then call [`into_events()`](NormalizedRow::into_events) to get properly parsed
+/// then call [`into_events()`](DataRow::into_events) to get properly parsed
 /// Events with full ADDL expansion, EVID handling, censoring, etc.
 ///
 /// # Fields
@@ -54,17 +40,17 @@ use std::collections::HashMap;
 /// # Example
 ///
 /// ```rust
-/// use pharmsol::data::parser::NormalizedRow;
+/// use pharmsol::data::parser::DataRow;
 ///
 /// // Observation row
-/// let obs = NormalizedRow::builder("pt1", 1.0)
+/// let obs = DataRow::builder("pt1", 1.0)
 ///     .evid(0)
 ///     .out(25.5)
 ///     .outeq(1)
 ///     .build();
 ///
 /// // Dosing row with negative ADDL (doses before time 0)
-/// let dose = NormalizedRow::builder("pt1", 0.0)
+/// let dose = DataRow::builder("pt1", 0.0)
 ///     .evid(1)
 ///     .dose(100.0)
 ///     .input(1)
@@ -77,7 +63,7 @@ use std::collections::HashMap;
 /// assert_eq!(events.len(), 11);
 /// ```
 #[derive(Debug, Clone, Default)]
-pub struct NormalizedRow {
+pub struct DataRow {
     /// Subject identifier (required)
     pub id: String,
     /// Event time (required)
@@ -112,8 +98,8 @@ pub struct NormalizedRow {
     pub covariates: HashMap<String, f64>,
 }
 
-impl NormalizedRow {
-    /// Create a new builder for constructing a NormalizedRow
+impl DataRow {
+    /// Create a new builder for constructing a DataRow
     ///
     /// # Arguments
     ///
@@ -123,16 +109,16 @@ impl NormalizedRow {
     /// # Example
     ///
     /// ```rust
-    /// use pharmsol::data::parser::NormalizedRow;
+    /// use pharmsol::data::parser::DataRow;
     ///
-    /// let row = NormalizedRow::builder("patient_001", 0.0)
[…]
