### 2026-01-28 d9c2730
paper: Draft state of the field
As requested by editor of JOSS
