+++ joss/paper.bib
@@ -96,4 +96,12 @@
   url       = {https://doi.org/10.1007/978-3-031-87804-6_12},
 }
 
+@Manual{mrgsolve,
+  title = {mrgsolve: Simulate from ODE-Based Models},
+  author = {Kyle T Baron},
+  year = {2025},
+  note = {R package version 1.6.1},
+  url = {https://mrgsolve.org/docs/},
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
