### 2026-01-30 845bff1
paper: Some refinement
