### 2026-01-30 bc1044c
paper: Refinement
