### 2026-01-30 1f2d695
paper: State of the field (#203)
* paper: Draft state of the field

---------

Co-authored-by: Julián D. Otálvaro <juliandavid347@gmail.com>
