+++ joss/paper.bib
@@ -96,4 +96,23 @@
   url       = {https://doi.org/10.1007/978-3-031-87804-6_12},
 }
 
+@Manual{mrgsolve,
+  title = {mrgsolve: Simulate from ODE-Based Models},
+  author = {Kyle T Baron},
+  year = {2025},
+  note = {R package version 1.6.1},
+  url = {https://mrgsolve.org/docs/},
+}
+
+@Article{rackauckas2020pumas,
+  author  = {Rackauckas, Chris and Ma, Yingbo and Noack, Andreas and Dixit, Vaibhav and Mogensen, Patrick Kofod and Byrne, Simon and Maddhashiya, Shubham and Santiago-Calder\'on, Jos\'e Bayo\'an and Nyberg, Joakim and Gobburu, Jogarao V. S. and others},
+  title   = {Accelerated predictive healthcare analytics with {Pumas}, a high performance pharmaceutical modeling and simulation platform},
+  journal = {bioRxiv},
+  year    = {2020},
+  doi     = {10.1101/2020.11.28.402297},
+  url     = {https://www.biorxiv.org/content/10.1101/2020.11.28.402297v1},
+}
+
+
+
 @Comment{jabref-meta: databaseType:bibtex;}
