### 2026-01-30 #9850 @mhovd
Dear @arfon, thank you again for making sure the submission is up to par.

We have complied with the requested additions, the latest changes including the state of the field was added in this commit: https://github.com/LAPKB/pharmsol/commit/1f2d6953ddfb717171ffdc61c43f1cf3460a2b89 to the `joss` branch containing the paper.

Please let us know if any additional changes are needed.
