### 2026-02-05 2fe3327
making sure both 0-index and 1-index data are supported
