### 2026-02-06 26246a8
fix: vector size missmatch
