### 2026-02-06 2efb0f0
fix: vector size missmatch
