### 2026-02-06 42b48f2
chore: Remove NCA implementation
