+++ src/data/structs.rs
@@ -285,30 +285,6 @@ impl Data {
         outeq_values.dedup();
         outeq_values
     }
-
-    /// Perform Non-Compartmental Analysis (NCA) on all subjects in the dataset
-    ///
-    /// This method iterates through all subjects and performs NCA analysis
-    /// for each subject's data, returning a collection of results.
-    ///
-    /// # Arguments
-    ///
-    /// * `options` - NCA calculation options
-    /// * `outeq` - Output equation index to analyze (0-indexed)
-    ///
-    /// # Returns
-    ///
-    /// Vector of `Result<NCAResult, NCAError>` for each subject-occasion combination
-    pub fn nca(
-        &self,
-        options: &crate::nca::NCAOptions,
-        outeq: usize,
-    ) -> Vec<Result<crate::nca::NCAResult, crate::nca::NCAError>> {
-        self.subjects
-            .iter()
-            .flat_map(|subject| subject.nca(options, outeq))
-            .collect()
-    }
 }
 
 impl IntoIterator for Data {
@@ -492,114 +468,6 @@ impl Subject {
         let mut hasher = std::hash::DefaultHasher::new();
         self.id.hash(&mut hasher);
         hasher.finish()
-    }
-
-    /// Perform Non-Compartmental Analysis (NCA) on this subject's data
-    ///
-    /// Calculates standard NCA parameters (Cmax, Tmax, AUC, half-life, etc.)
-    /// from the subject's observed concentration-time data.
-    ///
-    /// # Arguments
-    ///
-    /// * `options` - NCA calculation options
-    /// * `outeq` - Output equation index to analyze (default: 0)
-    ///
-    /// # Returns
-    ///
-    /// Vector of `NCAResult`, one per occasion
-    ///
-    /// # Examples
-    ///
-    /// ```rust,ignore
-    /// use pharmsol::prelude::*;
-    /// use pharmsol::nca::NCAOptions;
-    ///
-    /// let subject = Subject::builder("patient_001")
-    ///     .bolus(0.0, 100.0, 0)
-    ///     .observation(1.0, 10.0, 0)
-    ///     .observation(2.0, 8.0, 0)
-    ///     .observation(4.0, 4.0, 0)
-    ///     .build();
-    ///
-    /// let results = subject.nca(&NCAOptions::default(), 0);
-    /// if let Ok(res) = &results[0] {
-    ///     println!("Cmax: {:.2}", res.exposure.cmax);
-    /// }
-    /// ```
-    pub fn nca(
-        &self,
-        options: &crate::nca::NCAOptions,
-        outeq: usize,
-    ) -> Vec<Result<crate::nca::NCAResult, crate::nca::NCAError>> {
-        self.occasions
-            .iter()
-            .map(|occasion| occasion.nca(options, outeq, Some(self.id.clone())))
-            .collect()
-    }
-
-    /// Extract time-concentration data for a specific output equation
-    ///
-    /// Returns vectors of (times, concentrations, censoring) for the specified outeq.
-    /// This is useful for NCA calculations or other analysis.
-    ///
-    /// # Arguments
-    ///
-    /// * `outeq` - Output equation index to extract
-    ///
-    /// # Returns
-    ///
-    /// Tuple of (times, concentrations, censoring) vectors
-    pub fn get_observations(&self, outeq: usize) -> (Vec<f64>, Vec<f64>, Vec<Censor>) {
-        let mut times = Vec::new();
-        let mut concs = Vec::new();
-        let mut censoring = Vec::new();
-
-        for occasion in &self.occasions {
-            for event in occasion.events() {
-                if let Event::Observation(obs) = event {
-                    if obs.outeq() == outeq {
-                        if let Some(value) = obs.value() {
-                            times.push(obs.time());
[…]
