+++ src/nca/analyze.rs
@@ -1,517 +1,0 @@
-//! Main NCA analysis orchestrator
-//!
-//! This module contains the core analysis function that computes all NCA parameters
-//! from a validated profile and options.
-
-use super::calc;
-use super::error::NCAError;
-use super::profile::Profile;
-use super::types::*;
-
-// ============================================================================
-// Dose Context (internal - auto-detected from data structures)
-// ============================================================================
-
-/// Dose and route information detected from data
-///
-/// This is constructed internally by `Occasion::nca()` from the dose events in the data.
-#[derive(Debug, Clone)]
-pub(crate) struct DoseContext {
-    /// Total dose amount
-    pub amount: f64,
-    /// Infusion duration (None for bolus)
-    pub duration: Option<f64>,
-    /// Administration route
-    pub route: Route,
-}
-
-impl DoseContext {
-    /// Create a new dose context
-    pub fn new(amount: f64, duration: Option<f64>, route: Route) -> Self {
-        Self {
-            amount,
-            duration,
-            route,
-        }
-    }
-}
-
-// ============================================================================
-// Main Analysis Function
-// ============================================================================
-
-/// Perform complete NCA analysis on a profile
-///
-/// This is an internal function. External users should use `analyze_arrays`
-/// or the `.nca()` method on data structures.
-///
-/// # Arguments
-/// * `profile` - Validated concentration-time profile
-/// * `dose` - Dose context (detected from data, None if no dosing info)
-/// * `options` - Analysis configuration
-#[allow(dead_code)] // Used only in tests; main entry point is analyze_arrays
-pub(crate) fn analyze(
-    profile: &Profile,
-    dose: Option<&DoseContext>,
-    options: &NCAOptions,
-) -> Result<NCAResult, NCAError> {
-    // When called without raw data, calculate tlag from the (filtered) profile
-    #[allow(deprecated)]
-    let raw_tlag = calc::tlag(profile);
-    analyze_with_raw_tlag(profile, dose, options, raw_tlag)
-}
-
-/// Internal analysis with pre-computed raw tlag
-fn analyze_with_raw_tlag(
-    profile: &Profile,
-    dose: Option<&DoseContext>,
-    options: &NCAOptions,
-    raw_tlag: Option<f64>,
-) -> Result<NCAResult, NCAError> {
-    if profile.times.is_empty() {
-        return Err(NCAError::InsufficientData { n: 0, required: 2 });
-    }
-
-    // Core exposure parameters (always calculated)
-    let mut exposure = compute_exposure(profile, options, raw_tlag)?;
-
-    // Terminal phase parameters (if lambda-z can be estimated)
-    let (terminal, lambda_z_result) = compute_terminal(profile, options);
-
-    // Update exposure with AUCinf if we have terminal phase
-    if let Some(ref lz) = lambda_z_result {
-        update_exposure_with_terminal(&mut exposure, profile, lz, options);
-    }
-
-    // Clearance parameters (if we have dose and terminal phase)
-    let clearance = dose
-        .and_then(|d| lambda_z_result.as_ref().map(|lz| (d, lz)))
-        .map(|(d, lz)| compute_clearance(d.amount, exposure.auc_inf, lz.lambda_z));
-
-    // Route-specific parameters
-    let (iv_bolus, iv_infusion) =
-        compute_route_specific(profile, dose, lambda_z_result.as_ref(), options);
-
[…]
