+++ tests/nca/mod.rs
@@ -1,10 +1,0 @@
-// NCA Integration Tests Module
-// Tests using the public NCA API via Subject::builder().nca()
-//
-// Note: Most NCA tests are in src/nca/tests.rs (internal unit tests).
-// These integration tests verify the public API works correctly.
-
-pub mod test_auc;
-pub mod test_params;
-pub mod test_quality;
-pub mod test_terminal;
