+++ src/lib.rs
@@ -2,7 +2,6 @@ pub mod data;
 pub mod error;
 #[cfg(feature = "exa")]
 pub mod exa;
-pub mod nca;
 pub mod optimize;
 pub mod simulator;
 
