### 2026-02-06 2ad2176
feat: Normalized rows  (#205)
* feat: add NormalizedRow API and ResidualErrorModel

This commit adds two related features:

## NormalizedRow API (parser/)
- New  struct for format-agnostic data parsing
- Decouples column mapping from event creation logic
- Full ADDL/II expansion support (both positive and negative directions)
- Refactors pmetrics.rs to use NormalizedRow internally
- Enables external tools (like vial) to reuse parsing logic without
  reimplementing ADDL expansion

## ResidualErrorModel (data/)
- New  for parametric algorithms (SAEM, FOCE)
- Uses prediction-based sigma (vs obser
[…]
