+++ src/simulator/likelihood/mod.rs
@@ -1,592 +1,212 @@
-use crate::simulator::likelihood::progress::ProgressTracker;
-use crate::Censor;
-use crate::ErrorModelError;
-use crate::{
-    data::error_model::ErrorModels, Data, Equation, ErrorPoly, Observation, PharmsolError,
-    Predictions,
-};
-use ndarray::{Array2, Axis, ShapeBuilder};
-use rayon::prelude::*;
-use statrs::distribution::ContinuousCDF;
-use statrs::distribution::Normal;
+//! Likelihood calculation module for pharmacometric analyses.
+//!
+//! This module provides functions and types for computing log-likelihoods
+//! in pharmacometric population modeling. It supports both:
+//!
+//! - **Non-parametric algorithms** (NPAG, NPOD): Use [`ErrorModels`] with observation-based sigma
+//! - **Parametric algorithms** (SAEM, FOCE): Use [`ResidualErrorModels`] with prediction-based sigma
+//!
+//! # Module Organization
+//!
+//! - [`distributions`]: Statistical distribution functions (log-normal PDF, CDF)
+//! - [`prediction`]: Single observation-prediction pairs
+//! - [`subject`]: Subject-level prediction collections
+//! - [`matrix`]: Population-level log-likelihood matrix computation
+//!
+//! # Key Functions
+//!
+//! ## For Non-Parametric Algorithms
+//!
+//! Use [`log_likelihood_matrix`] to compute a matrix of log-likelihoods across
+//! all subjects and support points:
+//!
+//! ```ignore
+//! use pharmsol::prelude::simulator::{log_likelihood_matrix, LikelihoodMatrixOptions};
+//!
+//! let log_liks = log_likelihood_matrix(
+//!     &equation,
+//!     &data,
+//!     &support_points,
+//!     &error_models,
+//!     LikelihoodMatrixOptions::new().with_progress(),
+//! )?;
+//! ```
+//!
+//! ## For Parametric Algorithms
+//!
+//! Use [`log_likelihood_batch`] when each subject has individual parameters:
+//!
+//! ```ignore
+//! use pharmsol::prelude::simulator::log_likelihood_batch;
+//!
+//! let log_liks = log_likelihood_batch(
+//!     &equation,
+//!     &data,
+//!     &parameters,
+//!     &residual_error_models,
+//! )?;
+//! ```
+//!
+//! # Numerical Stability
+//!
+//! All likelihood functions operate in log-space for numerical stability.
+//! The deprecated `likelihood()` and `psi()` functions are provided for
+//! backward compatibility but should be avoided in new code.
 
+mod distributions;
+mod matrix;
+mod prediction;
 mod progress;
+mod subject;
 
-const FRAC_1_SQRT_2PI: f64 =
-    std::f64::consts::FRAC_2_SQRT_PI * std::f64::consts::FRAC_1_SQRT_2 / 2.0;
+// Re-export main types
+pub use matrix::{log_likelihood_matrix, LikelihoodMatrixOptions};
+pub use prediction::Prediction;
+pub use subject::{PopulationPredictions, SubjectPredictions};
 
-// ln(2π) = ln(2) + ln(π) ≈ 1.8378770664093453
-const LOG_2PI: f64 = 1.8378770664093453_f64;
+// Deprecated re-exports for backward compatibility
+#[allow(deprecated)]
+pub use matrix::{log_psi, psi};
 
-/// Container for predictions associated with a single subject.
+use ndarray::Array2;
+use rayon::prelude::*;
+
+use crate::{Data, Equation, PharmsolError, Predictions, Subject};
+
+/// Compute log-likelihoods for all subjects in parallel, where each subject
+/// has its own parameter vector.
 ///
[…]
