### 2026-02-06 f5894e0
feat: Error handling during simulation
