+++ examples/covariates.rs
@@ -34,20 +34,22 @@ fn main() {
             //Struct
             dx[0] = -ka * x[0] + b[0];
             dx[1] = ka * x[0] - ke * x[1];
+            Ok(())
         },
         // This blocks defines the lag-time of the bolus dose
         |p, _t, _cov| {
             fetch_params!(p, _ka, _ke, tlag, _v);
             // Macro used to define the lag-time for the input of the bolus dose
-            lag! {0=>tlag}
+            Ok(lag! {0=>tlag})
         },
-        |_p, _t, _cov| fa! {},
-        |_p, _t, _cov, _x| {},
+        |_p, _t, _cov| Ok(fa! {}),
+        |_p, _t, _cov, _x| Ok(()),
         |x, p, _t, _cov, y| {
             fetch_params!(p, _ka, _ke, _tlag, v);
 
             // Define the predicted concentration as the amount in the central compartment divided by volume
             y[0] = x[1] / v;
+            Ok(())
         },
         (2, 1),
     );
