### 2026-02-12 2415f1c
wip: adding missing functionality and re-defining base structs
