+++ src/nca/summary.rs
@@ -1,0 +1,490 @@
+//! Population summary statistics for NCA results
+//!
+//! Computes descriptive statistics across multiple [`NCAResult`]s,
+//! including geometric mean, CV%, and percentiles — standard PK reporting metrics.
+//!
+//! # Example
+//!
+//! ```rust,ignore
+//! use pharmsol::nca::{summarize, NCAOptions, NCA};
+//!
+//! let results: Vec<NCAResult> = subjects.iter()
+//!     .flat_map(|s| s.nca(&NCAOptions::default(), 0))
+//!     .filter_map(|r| r.ok())
+//!     .collect();
+//!
+//! let summary = summarize(&results);
+//! println!("N subjects: {}", summary.n_subjects);
+//! for p in &summary.parameters {
+//!     println!("{}: mean={:.2} CV%={:.1}", p.name, p.mean, p.cv_pct);
+//! }
+//! ```
+
+use super::types::NCAResult;
+use serde::{Deserialize, Serialize};
+
+// ============================================================================
+// Types
+// ============================================================================
+
+/// Descriptive statistics for a single NCA parameter across subjects
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct ParameterSummary {
+    /// Parameter name (matches keys from `NCAResult::to_params()`)
+    pub name: String,
+    /// Number of subjects with this parameter
+    pub n: usize,
+    /// Arithmetic mean
+    pub mean: f64,
+    /// Standard deviation
+    pub sd: f64,
+    /// Coefficient of variation (%)
+    pub cv_pct: f64,
+    /// Median
+    pub median: f64,
+    /// Minimum
+    pub min: f64,
+    /// Maximum
+    pub max: f64,
+    /// Geometric mean (NaN if any values ≤ 0)
+    pub geo_mean: f64,
+    /// Geometric CV% (NaN if any values ≤ 0)
+    pub geo_cv_pct: f64,
+    /// 5th percentile
+    pub p5: f64,
+    /// 25th percentile (Q1)
+    pub p25: f64,
+    /// 75th percentile (Q3)
+    pub p75: f64,
+    /// 95th percentile
+    pub p95: f64,
+}
+
+/// Summary of NCA results across a population
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct PopulationSummary {
+    /// Total number of NCA results summarized
+    pub n_subjects: usize,
+    /// Per-parameter descriptive statistics
+    pub parameters: Vec<ParameterSummary>,
+}
+
+// ============================================================================
+// Public API
+// ============================================================================
+
+/// Compute population summary statistics from a collection of NCA results
+///
+/// Extracts each named parameter via [`NCAResult::to_params()`], then computes
+/// descriptive statistics across all results that have that parameter.
+///
+/// Parameters are returned in a stable alphabetical order.
+pub fn summarize(results: &[NCAResult]) -> PopulationSummary {
+    if results.is_empty() {
+        return PopulationSummary {
+            n_subjects: 0,
+            parameters: Vec::new(),
+        };
+    }
+
+    // Collect all parameter names across all results
+    let mut all_params: std::collections::BTreeMap<&'static str, Vec<f64>> =
+        std::collections::BTreeMap::new();
+
+    for result in results {
+        let params = result.to_params();
+        for (name, value) in params {
+            all_params.entry(name).or_default().push(value);
+        }
+    }
+
+    // Compute summary for each parameter
+    let parameters: Vec<ParameterSummary> = all_params
+        .into_iter()
[…]
