+++ src/nca/profile.rs
@@ -1,389 +1,0 @@
-//! Internal profile representation for NCA analysis
-//!
-//! The Profile struct is a validated, analysis-ready concentration-time dataset.
-//! It handles BLQ processing and caches key indices for efficiency.
-
-use super::error::NCAError;
-use super::types::BLQRule;
-use crate::Censor;
-
-/// A validated concentration-time profile ready for NCA analysis
-///
-/// This is an internal structure that normalizes data from various sources
-/// (raw arrays, Occasion) into a consistent format with BLQ handling applied.
-#[derive(Debug, Clone)]
-pub(crate) struct Profile {
-    /// Time points (sorted, ascending)
-    pub times: Vec<f64>,
-    /// Concentration values (parallel to times)
-    pub concentrations: Vec<f64>,
-    /// Index of Cmax in the arrays
-    pub cmax_idx: usize,
-    /// Index of Clast (last positive concentration)
-    pub tlast_idx: usize,
-}
-
-impl Profile {
-    /// Create a profile from time/concentration/censoring arrays
-    ///
-    /// BLQ/ALQ status is determined by the `Censor` marking:
-    /// - `Censor::BLOQ`: Below limit of quantification - value is the lower limit
-    /// - `Censor::ALOQ`: Above limit of quantification - value is the upper limit
-    /// - `Censor::None`: Quantifiable observation - value is the measured concentration
-    ///
-    /// # Arguments
-    /// * `times` - Time points
-    /// * `concentrations` - Concentration values (for censored samples, this is the LOQ/ULQ)
-    /// * `censoring` - Censoring status for each observation
-    /// * `blq_rule` - How to handle BLQ values
-    ///
-    /// # Errors
-    /// Returns error if data is insufficient or invalid
-    pub fn from_arrays(
-        times: &[f64],
-        concentrations: &[f64],
-        censoring: &[Censor],
-        blq_rule: BLQRule,
-    ) -> Result<Self, NCAError> {
-        if times.len() != concentrations.len() || times.len() != censoring.len() {
-            return Err(NCAError::InvalidParameter {
-                param: "arrays".to_string(),
-                value: format!(
-                    "array lengths mismatch: times={}, concentrations={}, censoring={}",
-                    times.len(),
-                    concentrations.len(),
-                    censoring.len()
-                ),
-            });
-        }
-
-        if times.is_empty() {
-            return Err(NCAError::InsufficientData { n: 0, required: 2 });
-        }
-
-        // Check time sequence is valid
-        for i in 1..times.len() {
-            if times[i] < times[i - 1] {
-                return Err(NCAError::InvalidTimeSequence);
-            }
-        }
-
-        // For Positional rule, we need tfirst and tlast first
-        // For TmaxRelative, we need tmax
-        // Do a preliminary pass to find these indices
-        let (tfirst_idx, tlast_idx) = if matches!(blq_rule, BLQRule::Positional) {
-            Self::find_tfirst_tlast(concentrations, censoring)
-        } else {
-            (None, None)
-        };
-
-        let tmax_idx = if matches!(blq_rule, BLQRule::TmaxRelative { .. }) {
-            Self::find_tmax_idx(concentrations, censoring)
-        } else {
-            None
-        };
-
-        let mut proc_times = Vec::with_capacity(times.len());
-        let mut proc_concs = Vec::with_capacity(concentrations.len());
-
-        for i in 0..times.len() {
-            let time = times[i];
-            let conc = concentrations[i];
-            let censor = censoring[i];
-
-            // BLQ is determined by the Censor marking
[…]
