### 2026-02-13 4528d8d
total rewrite of the NCA module
