+++ src/nca/tests.rs
@@ -6,6 +6,7 @@
 use crate::data::Subject;
 use crate::nca::*;
 use crate::SubjectBuilderExt;
+use crate::Data;
 
 // ============================================================================
 // Test subject builders
@@ -104,7 +105,7 @@ fn no_dose_subject() -> Subject {
 fn test_nca_basic_exposure() {
     let subject = single_dose_oral();
     let options = NCAOptions::default();
-    let results = subject.nca(&options, 0);
+    let results = subject.nca_all(&options);
     let result = results[0].as_ref().unwrap();
 
     // Check Cmax/Tmax
@@ -123,7 +124,7 @@ fn test_nca_basic_exposure() {
 fn test_nca_with_dose() {
     let subject = single_dose_oral();
     let options = NCAOptions::default();
-    let results = subject.nca(&options, 0);
+    let results = subject.nca_all(&options);
     let result = results[0].as_ref().unwrap();
 
     // Should have clearance parameters if lambda-z was estimated
@@ -137,7 +138,7 @@ fn test_nca_with_dose() {
 fn test_nca_without_dose() {
     let subject = no_dose_subject();
     let options = NCAOptions::default();
-    let results = subject.nca(&options, 0);
+    let results = subject.nca_all(&options);
     let result = results[0].as_ref().unwrap();
 
     // Exposure should still be computed
@@ -150,7 +151,7 @@ fn test_nca_without_dose() {
 fn test_nca_terminal_phase() {
     let subject = single_dose_oral();
     let options = NCAOptions::default();
-    let results = subject.nca(&options, 0);
+    let results = subject.nca_all(&options);
     let result = results[0].as_ref().unwrap();
 
     // Check terminal phase was estimated
@@ -180,7 +181,7 @@ fn test_nca_terminal_phase() {
 fn test_auc_linear_method() {
     let subject = single_dose_oral();
     let options = NCAOptions::default().with_auc_method(AUCMethod::Linear);
-    let results = subject.nca(&options, 0);
+    let results = subject.nca_all(&options);
     let result = results[0].as_ref().unwrap();
 
     assert!(result.exposure.auc_last > 0.0);
@@ -190,7 +191,7 @@ fn test_auc_linear_method() {
 fn test_auc_linuplogdown_method() {
     let subject = single_dose_oral();
     let options = NCAOptions::default().with_auc_method(AUCMethod::LinUpLogDown);
-    let results = subject.nca(&options, 0);
+    let results = subject.nca_all(&options);
     let result = results[0].as_ref().unwrap();
 
     assert!(result.exposure.auc_last > 0.0);
@@ -203,12 +204,12 @@ fn test_auc_methods_differ() {
     let linear = NCAOptions::default().with_auc_method(AUCMethod::Linear);
     let logdown = NCAOptions::default().with_auc_method(AUCMethod::LinUpLogDown);
 
-    let result_linear = subject.nca(&linear, 0)[0]
+    let result_linear = subject.nca_all(&linear)[0]
         .as_ref()
         .unwrap()
         .exposure
         .auc_last;
-    let result_logdown = subject.nca(&logdown, 0)[0]
+    let result_logdown = subject.nca_all(&logdown)[0]
         .as_ref()
         .unwrap()
         .exposure
@@ -229,7 +230,7 @@ fn test_auc_methods_differ() {
 fn test_iv_bolus_route() {
     let subject = iv_bolus_subject();
     let options = NCAOptions::default();
-    let results = subject.nca(&options, 0);
+    let results = subject.nca_all(&options);
     let result = results[0].as_ref().unwrap();
 
     // Should have IV bolus parameters
@@ -248,7 +249,7 @@ fn test_iv_bolus_route() {
 fn test_iv_infusion_route() {
[…]

[…]
