+++ src/lib.rs
@@ -59,9 +59,9 @@ pub mod prelude {
     };
 
     // NCA extension traits (provides .nca(), .nca_all(), etc. on data types)
+    pub use crate::data::traits::{MetricsError, ObservationMetrics};
     pub use crate::nca::NCA;
     pub use crate::nca::{NCAOptions, NCAPopulation, SubjectNCAResult};
-    pub use crate::data::traits::{ObservationMetrics, MetricsError};
 
     // AUC primitives for direct use on raw arrays
     pub use crate::data::auc::{auc, auc_interval, aumc, interpolate_linear};
