### 2026-02-15 e7376b2
Remove JSON from this PR
A copy before this change is at https://github.com/LAPKB/pharmsol/pull/209
