+++ src/json/codegen/analytical.rs
@@ -1,11 +1,0 @@
-//! Analytical model code generation
-//!
-//! This module contains specialized code generation logic for analytical models.
-//! Most of the heavy lifting is done by the ClosureGenerator in closures.rs.
-
-// Currently, all analytical-specific generation is handled in mod.rs
-// and closures.rs. This module is reserved for future specialized logic
-// such as:
-// - Analytical function parameter validation
-// - Secondary equation optimization
-// - Symbolic differentiation for sensitivity analysis
