+++ src/json/mod.rs
@@ -1,219 +1,0 @@
-//! JSON Model Definition and Code Generation
-//!
-//! This module provides functionality for defining pharmacometric models using JSON
-//! and generating Rust code that can be compiled by the `exa` module.
-//!
-//! # Overview
-//!
-//! The JSON model system provides a declarative way to define pharmacometric models
-//! without writing Rust code directly. Models are defined in JSON following a
-//! structured schema, then validated and compiled to native code.
-//!
-//! The system supports three equation types:
-//! - **Analytical**: Built-in closed-form solutions (fastest execution)
-//! - **ODE**: Custom ordinary differential equations
-//! - **SDE**: Stochastic differential equations with particle filtering
-//!
-//! # Quick Start
-//!
-//! ```ignore
-//! use pharmsol::json::{parse_json, validate_json, generate_code};
-//!
-//! // Define a model in JSON
-//! let json = r#"{
-//!     "schema": "1.0",
-//!     "id": "pk_1cmt_oral",
-//!     "type": "analytical",
-//!     "analytical": "one_compartment_with_absorption",
-//!     "parameters": ["ka", "ke", "V"],
-//!     "output": "x[1] / V"
-//! }"#;
-//!
-//! // Parse and validate
-//! let validated = validate_json(json)?;
-//!
-//! // Generate Rust code
-//! let code = generate_code(json)?;
-//! println!("Generated: {}", code.equation_code);
-//! ```
-//!
-//! # Using the Model Library
-//!
-//! The library provides pre-built standard PK models:
-//!
-//! ```ignore
-//! use pharmsol::json::ModelLibrary;
-//!
-//! let library = ModelLibrary::builtin();
-//!
-//! // List available models
-//! for id in library.list() {
-//!     println!("Available: {}", id);
-//! }
-//!
-//! // Get a specific model
-//! let model = library.get("pk/1cmt-oral").unwrap();
-//!
-//! // Search by keyword
-//! let oral_models = library.search("oral");
-//!
-//! // Filter by type
-//! let ode_models = library.filter_by_type(ModelType::Ode);
-//! ```
-//!
-//! # Model Inheritance
-//!
-//! Models can extend base models to add customizations:
-//!
-//! ```ignore
-//! use pharmsol::json::{JsonModel, ModelLibrary};
-//!
-//! let mut library = ModelLibrary::builtin();
-//!
-//! // Define a model that extends a library model
-//! let derived = JsonModel::from_str(r#"{
-//!     "schema": "1.0",
-//!     "id": "pk_1cmt_wt",
-//!     "extends": "pk/1cmt-oral",
-//!     "type": "analytical",
-//!     "analytical": "one_compartment_with_absorption",
-//!     "parameters": ["ka", "ke", "V"],
-//!     "covariates": [{ "id": "WT", "reference": 70.0 }],
-//!     "covariateEffects": [{
-//!         "on": "V",
-//!         "covariate": "WT",
-//!         "type": "allometric",
-//!         "exponent": 1.0,
-//!         "reference": 70.0
-//!     }]
-//! }"#)?;
-//!
-//! // Resolve inherits base model's output expression
-//! let resolved = library.resolve(&derived)?;
-//! ```
-//!
-//! # JSON Schema
-//!
-//! ## Required Fields
-//!
-//! | Field | Description |
-//! |-------|-------------|
-//! | `schema` | Schema version (currently `"1.0"`) |
-//! | `id` | Unique model identifier |
-//! | `type` | Equation type: `"analytical"`, `"ode"`, or `"sde"` |
-//!
-//! ## Model Type Specific Fields
-//!
-//! ### Analytical Models
[…]
