+++ src/json/library/mod.rs
@@ -1,517 +1,0 @@
-//! Model Library
-//!
-//! Provides a registry of built-in pharmacometric models that can be:
-//! - Used directly via their ID
-//! - Extended via the `extends` field for customization
-//!
-//! # Example
-//!
-//! ```rust,ignore
-//! use pharmsol::json::library::ModelLibrary;
-//!
-//! let library = ModelLibrary::builtin();
-//!
-//! // List available models
-//! for id in library.list() {
-//!     println!("Available: {}", id);
-//! }
-//!
-//! // Get a model
-//! if let Some(model) = library.get("pk/1cmt-iv") {
-//!     println!("Found model: {}", model.id);
-//! }
-//! ```
-
-use crate::json::errors::JsonModelError;
-use crate::json::model::JsonModel;
-use crate::json::types::{DisplayInfo, Documentation, ModelType};
-use std::collections::HashMap;
-use std::path::Path;
-
-/// A registry of JSON model definitions
-#[derive(Debug, Clone)]
-pub struct ModelLibrary {
-    models: HashMap<String, JsonModel>,
-}
-
-// Embed built-in models at compile time
-mod embedded {
-    // PK Analytical Models
-    pub const PK_1CMT_IV: &str = include_str!("models/pk_1cmt_iv.json");
-    pub const PK_1CMT_ORAL: &str = include_str!("models/pk_1cmt_oral.json");
-    pub const PK_2CMT_IV: &str = include_str!("models/pk_2cmt_iv.json");
-    pub const PK_2CMT_ORAL: &str = include_str!("models/pk_2cmt_oral.json");
-    pub const PK_3CMT_IV: &str = include_str!("models/pk_3cmt_iv.json");
-    pub const PK_3CMT_ORAL: &str = include_str!("models/pk_3cmt_oral.json");
-
-    // PK ODE Models
-    pub const PK_1CMT_IV_ODE: &str = include_str!("models/pk_1cmt_iv_ode.json");
-    pub const PK_1CMT_ORAL_ODE: &str = include_str!("models/pk_1cmt_oral_ode.json");
-    pub const PK_2CMT_IV_ODE: &str = include_str!("models/pk_2cmt_iv_ode.json");
-    pub const PK_2CMT_ORAL_ODE: &str = include_str!("models/pk_2cmt_oral_ode.json");
-}
-
-impl ModelLibrary {
-    /// Create a new empty library
-    pub fn new() -> Self {
-        Self {
-            models: HashMap::new(),
-        }
-    }
-
-    /// Create a library with all built-in models
-    pub fn builtin() -> Self {
-        let mut library = Self::new();
-
-        // Load embedded models
-        let embedded_models = [
-            embedded::PK_1CMT_IV,
-            embedded::PK_1CMT_ORAL,
-            embedded::PK_2CMT_IV,
-            embedded::PK_2CMT_ORAL,
-            embedded::PK_3CMT_IV,
-            embedded::PK_3CMT_ORAL,
-            embedded::PK_1CMT_IV_ODE,
-            embedded::PK_1CMT_ORAL_ODE,
-            embedded::PK_2CMT_IV_ODE,
-            embedded::PK_2CMT_ORAL_ODE,
-        ];
-
-        for json in embedded_models {
-            if let Ok(model) = JsonModel::from_str(json) {
-                library.models.insert(model.id.clone(), model);
-            }
-        }
-
-        library
-    }
-
-    /// Load models from a directory (recursively searches for .json files)
-    pub fn from_dir(path: &Path) -> Result<Self, JsonModelError> {
-        let mut library = Self::new();
-        library.load_dir(path)?;
-        Ok(library)
-    }
-
-    /// Load models from a directory into this library
[…]

[…]
