### 2026-02-16 06b32f0
chore: Update rand_distr requirement from 0.5.0 to 0.6.0
Updates the requirements on [rand_distr](https://github.com/rust-random/rand_distr) to permit the latest version.
- [Release notes](https://github.com/rust-random/rand_distr/releases)
- [Changelog](https://github.com/rust-random/rand_distr/blob/master/CHANGELOG.md)
- [Commits](https://github.com/rust-random/rand_distr/compare/0.5.0...0.6.0)

---
updated-dependencies:
- dependency-name: rand_distr
  dependency-version: 0.6.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
