+++ Cargo.toml
@@ -18,8 +18,8 @@ diffsol = "=0.10.4"
 libloading = { version = "0.9.0", optional = true, features = [] }
 nalgebra = "0.34.1"
 ndarray = { version = "0.16.1", features = ["rayon"] }
-rand = "0.9.0"
-rand_distr = "0.5.0"
+rand = "0.10.0"
+rand_distr = "0.6.0"
 rayon = "1.10.0"
 serde = { version = "1.0.201", features = ["derive"] }
 serde_json = "1.0.117"
