+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "pharmsol"
-version = "0.22.1"
+version = "0.23.0"
 edition = "2021"
 authors = ["Julián D. Otálvaro <juliandavid347@gmail.com>", "Markus Hovd"]
 description = "Rust library for solving analytic and ode-defined pharmacometric models."
