### 2026-02-18 6900ed5
refactor: Use faer instead of nalgebra
