+++ src/lib.rs
@@ -16,7 +16,6 @@ pub use crate::simulator::equation::{self, ODE};
 pub use error::PharmsolError;
 #[cfg(feature = "exa")]
 pub use exa::*;
-pub use nalgebra::dmatrix;
 pub use std::collections::HashMap;
 
 /// Prelude module that re-exports all commonly used types and traits.
