+++ Cargo.toml
@@ -16,7 +16,6 @@ cached = { version = "0.56.0" }
 csv = "1.3.0"
 diffsol = "=0.10.4"
 libloading = { version = "0.9.0", optional = true, features = [] }
-nalgebra = "0.34.1"
 ndarray = { version = "0.17.2", features = ["rayon"] }
 rand = "0.10.0"
 rand_distr = "0.6.0"
@@ -28,6 +27,7 @@ thiserror = "2.0.11"
 argmin = "0.11.0"
 argmin-math = "0.5.1"
 tracing = "0.1.41"
+faer = "0.23"
 
 [dev-dependencies]
 criterion = { version = "0.8.0", features = ["html_reports"] }

[…]
