+++ src/simulator/equation/ode/closure.rs
@@ -3,9 +3,7 @@ use diffsol::{
     ConstantOp, FaerContext, FaerMat, LinearOp, MatrixCommon, NonLinearOp, NonLinearOpJacobian,
     OdeEquations, OdeEquationsRef, Op, Vector,
 };
-use std::{cell::RefCell, cmp::Ordering, sync::LazyLock};
-
-static FAER_CTX: LazyLock<FaerContext> = LazyLock::new(FaerContext::default);
+use std::{cell::RefCell, cmp::Ordering};
 
 type M = FaerMat<f64>;
 type V = <M as MatrixCommon>::V;
@@ -130,6 +128,7 @@ where
     func: &'a F,
     rateiv_buffer: &'a RefCell<V>,
     zero_bolus: &'a V,
+    ctx: FaerContext,
 }
 
 impl<F> Op for PmRhs<'_, F>
@@ -150,7 +149,7 @@ where
         self.nparams
     }
     fn context(&self) -> &Self::C {
-        &FAER_CTX
+        &self.ctx
     }
 }
 
@@ -158,6 +157,7 @@ pub struct PmMass {
     nstates: usize,
     nout: usize,
     nparams: usize,
+    ctx: FaerContext,
 }
 
 impl Op for PmMass {
@@ -175,7 +175,7 @@ impl Op for PmMass {
         self.nparams
     }
     fn context(&self) -> &Self::C {
-        &FAER_CTX
+        &self.ctx
     }
 }
 
@@ -184,6 +184,7 @@ pub struct PmInit<'a> {
     nout: usize,
     nparams: usize,
     init: &'a V,
+    ctx: FaerContext,
 }
 
 impl Op for PmInit<'_> {
@@ -201,7 +202,7 @@ impl Op for PmInit<'_> {
         self.nparams
     }
     fn context(&self) -> &Self::C {
-        &FAER_CTX
+        &self.ctx
     }
 }
 
@@ -215,6 +216,7 @@ pub struct PmRoot {
     nstates: usize,
     nout: usize,
     nparams: usize,
+    ctx: FaerContext,
 }
 
 impl Op for PmRoot {
@@ -232,7 +234,7 @@ impl Op for PmRoot {
         self.nparams
     }
     fn context(&self) -> &Self::C {
-        &FAER_CTX
+        &self.ctx
     }
 }
 
@@ -240,6 +242,7 @@ pub struct PmOut {
     nstates: usize,
     nout: usize,
     nparams: usize,
+    ctx: FaerContext,
 }
 
 impl Op for PmOut {
@@ -257,7 +260,7 @@ impl Op for PmOut {
         self.nparams
     }
     fn context(&self) -> &Self::C {
-        &FAER_CTX
+        &self.ctx
     }
 }
 
@@ -325,6 +328,7 @@ where
     covariates: &'a Covariates,
     infusion_schedule: InfusionSchedule,
     rateiv_buffer: RefCell<V>,
+    ctx: FaerContext,
 }
 
 impl<'a, F> PMProblem<'a, F>
@@ -343,10 +347,11 @@ where
         init: V,
     ) -> Self {
         let nparams = p.len();
-        let rateiv_buffer = RefCell::new(V::zeros(nstates, FaerContext::default()));
+        let ctx = FaerContext::default();
+        let rateiv_buffer = RefCell::new(V::zeros(nstates, ctx));
         let infusion_schedule = InfusionSchedule::new(nstates, infusions);
         // Pre-allocate zero bolus vector
-        let zero_bolus = V::zeros(nstates, FaerContext::default());
+        let zero_bolus = V::zeros(nstates, ctx);
 
         Self {
             func,
@@ -359,6 +364,7 @@ where
             covariates,
             infusion_schedule,
             rateiv_buffer,
+            ctx,
         }
[…]
