### 2026-02-19 d6c7e78
feat: Use a sized cache with new subject hashing
