+++ src/simulator/equation/ode/mod.rs
@@ -55,11 +55,10 @@ impl State for V {
 }
 
 /// Hash support points to a u64 for cache key generation.
-/// Uses DefaultHasher for good distribution and collision resistance.
 #[inline(always)]
 fn spphash(spp: &[f64]) -> u64 {
     use std::hash::{Hash, Hasher};
-    let mut hasher = std::hash::DefaultHasher::new();
+    let mut hasher = ahash::AHasher::default();
     for &value in spp {
         // Normalize -0.0 to 0.0 for consistent hashing
         let bits = if value == 0.0 { 0u64 } else { value.to_bits() };
