### 2026-02-19 59a8335
Add unit tests for the new hash methods
