### 2026-02-24 776fb6b
chore: making private some structs that are not part of the API, removing unused code
