### 2026-02-24 cb084ac
fix: propagate errors from auc functions instead of silently returning 0
- auc_segment/auc_interval/aumc now return Result<f64, ObservationError>
- dt <= 0 yields Err(InvalidTimeSequence) instead of Ok(0.0)
- Cascade through ObservationProfile, traits, analyze, steady-state
- Rewrite trapezoidal_auc in superposition to avoid dependency on data::auc
- Fix README .nca() API example
- Fix {:.2} format specifier for usize outeq in prediction.rs
- Fix typos in CHANGELOG.md
