+++ src/data/auc.rs
@@ -20,13 +20,14 @@
 //! let times = [0.0, 1.0, 2.0, 4.0, 8.0];
 //! let concs = [0.0, 10.0, 8.0, 4.0, 2.0];
 //!
-//! let total = auc(&times, &concs, &AUCMethod::Linear);
-//! let partial = auc_interval(&times, &concs, 1.0, 4.0, &AUCMethod::Linear);
-//! let moment = aumc(&times, &concs, &AUCMethod::Linear);
+//! let total = auc(&times, &concs, &AUCMethod::Linear).unwrap();
+//! let partial = auc_interval(&times, &concs, 1.0, 4.0, &AUCMethod::Linear).unwrap();
+//! let moment = aumc(&times, &concs, &AUCMethod::Linear).unwrap();
 //! let c_at_3 = interpolate_linear(&times, &concs, 3.0);
 //! ```
 
 use crate::data::event::AUCMethod;
+use crate::data::observation_error::ObservationError;
 
 // ============================================================================
 // Segment-level helpers (private)
@@ -73,13 +74,19 @@ fn aumc_log(t1: f64, c1: f64, t2: f64, c2: f64, dt: f64) -> f64 {
 /// calculation cannot know Tmax context. Use [`auc`] or
 /// [`auc_segment_with_tmax`] for proper LinLog handling.
 #[inline]
-pub fn auc_segment(t1: f64, c1: f64, t2: f64, c2: f64, method: &AUCMethod) -> f64 {
+pub fn auc_segment(
+    t1: f64,
+    c1: f64,
+    t2: f64,
+    c2: f64,
+    method: &AUCMethod,
+) -> Result<f64, ObservationError> {
     let dt = t2 - t1;
     if dt <= 0.0 {
-        return 0.0;
+        return Err(ObservationError::InvalidTimeSequence);
     }
 
-    match method {
+    Ok(match method {
         AUCMethod::Linear | AUCMethod::LinLog => auc_linear(c1, c2, dt),
         AUCMethod::LinUpLogDown => {
             if use_log_linear(c1, c2) {
@@ -88,7 +95,7 @@ pub fn auc_segment(t1: f64, c1: f64, t2: f64, c2: f64, method: &AUCMethod) -> f6
                 auc_linear(c1, c2, dt)
             }
         }
-    }
+    })
 }
 
 /// Calculate AUC for a segment with Tmax context (for LinLog method)
@@ -103,13 +110,13 @@ pub fn auc_segment_with_tmax(
     c2: f64,
     tmax: f64,
     method: &AUCMethod,
-) -> f64 {
+) -> Result<f64, ObservationError> {
     let dt = t2 - t1;
     if dt <= 0.0 {
-        return 0.0;
+        return Err(ObservationError::InvalidTimeSequence);
     }
 
-    match method {
+    Ok(match method {
         AUCMethod::Linear => auc_linear(c1, c2, dt),
         AUCMethod::LinUpLogDown => {
             if use_log_linear(c1, c2) {
@@ -125,7 +132,7 @@ pub fn auc_segment_with_tmax(
                 auc_log(c1, c2, dt)
             }
         }
-    }
+    })
 }
 
 /// Calculate AUMC for a segment with Tmax context (for LinLog method)
@@ -137,13 +144,13 @@ pub fn aumc_segment_with_tmax(
     c2: f64,
     tmax: f64,
     method: &AUCMethod,
-) -> f64 {
+) -> Result<f64, ObservationError> {
     let dt = t2 - t1;
     if dt <= 0.0 {
-        return 0.0;
+        return Err(ObservationError::InvalidTimeSequence);
     }
 
-    match method {
[…]
