### 2026-02-24 759baa7
Update documentation in auc.rs
Removed design section from documentation comments.
