### 2026-02-25 6bd3255
Add a function to clear caches
