### 2026-02-25 61668f9
Proof of concept vibe coded DSL
