+++ src/lib.rs
@@ -1,9 +1,15 @@
+// Allow the crate to reference itself as `::pharmsol` (used by proc macro output)
+extern crate self as pharmsol;
+
 pub mod data;
 pub mod error;
 #[cfg(feature = "exa")]
 pub mod exa;
 pub mod optimize;
 pub mod simulator;
+
+// Re-export the pk_model proc macro
+pub use pharmsol_macros::pk_model;
 
 //extension traits
 pub use crate::data::builder::SubjectBuilderExt;
@@ -99,6 +105,8 @@ pub mod prelude {
     pub use crate::fetch_params;
     #[doc(inline)]
     pub use crate::lag;
+    #[doc(inline)]
+    pub use crate::pk_model;
 }
 
 #[macro_export]
@@ -154,5 +162,258 @@ mod tests {
         assert_eq!(ka, 1.0);
         assert_eq!(ke, 2.5);
         assert_eq!(v, 3.7);
+    }
+
+    #[test]
+    fn test_pk_model_basic() {
+        use crate::prelude::*;
+
+        let subject = Subject::builder("id1")
+            .bolus(0.0, 100.0, 0)
+            .observation(1.0, 0.0, 0)
+            .observation(2.0, 0.0, 0)
+            .build();
+
+        // Clean declarative DSL — no closures needed
+        let ode_dsl = pk_model! {
+            params: (ke, v),
+            diffeq: {
+                dx[0] = -ke * x[0];
+            },
+            out: {
+                y[0] = x[0] / v;
+            },
+            neqs: (1, 1),
+        };
+
+        // Manual version for comparison
+        let ode_manual = equation::ODE::new(
+            |x, p, _t, dx, _b, _rateiv, _cov| {
+                fetch_params!(p, ke, _v);
+                dx[0] = -ke * x[0];
+            },
+            |_p, _t, _cov| lag! {},
+            |_p, _t, _cov| fa! {},
+            |_p, _t, _cov, _x| {},
+            |x, p, _t, _cov, y| {
+                fetch_params!(p, _ke, v);
+                y[0] = x[0] / v;
+            },
+            (1, 1),
+        );
+
+        let params = vec![0.1, 50.0];
+        let pred_dsl = ode_dsl.estimate_predictions(&subject, &params).unwrap();
+        let pred_manual = ode_manual.estimate_predictions(&subject, &params).unwrap();
+
+        let dsl_preds = pred_dsl.flat_predictions();
+        let manual_preds = pred_manual.flat_predictions();
+
+        assert_eq!(dsl_preds.len(), manual_preds.len());
+        for (d, m) in dsl_preds.iter().zip(manual_preds.iter()) {
+            assert!(
+                (d - m).abs() < 1e-10,
+                "Predictions differ: DSL={}, Manual={}",
+                d,
+                m
+            );
+        }
+    }
+
+    #[test]
+    fn test_pk_model_with_lag() {
+        use crate::prelude::*;
+
+        let subject = Subject::builder("id1")
+            .bolus(0.0, 100.0, 0)
+            .observation(0.5, 0.0, 0)
+            .observation(1.0, 0.0, 0)
+            .observation(2.0, 0.0, 0)
+            .build();
+
+        let ode = pk_model! {
+            params: (ka, ke, tlag, v),
+            lag: { 0 => tlag },
+            diffeq: {
+                dx[0] = -ka * x[0];
+                dx[1] = ka * x[0] - ke * x[1];
+            },
+            out: {
+                y[0] = x[1] / v;
+            },
+            neqs: (2, 1),
+        };
+
+        let params = vec![1.0, 0.1, 0.5, 50.0];
[…]
