+++ pharmsol-macros/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "pharmsol-macros"
+version = "0.1.0"
+edition = "2021"
+description = "Procedural macros for the pharmsol pharmacometric modeling library"
+license = "GPL-3.0"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "2", features = ["full"] }
+quote = "1"
+proc-macro2 = "1"

[…]
