+++ src/simulator/equation/sde/mod.rs
@@ -17,6 +17,9 @@ use crate::{
     Subject,
 };
 
+use super::id_hash;
+use super::spphash;
+
 use diffsol::VectorCommon;
 
 use crate::PharmsolError;
@@ -408,26 +411,12 @@ impl Equation for SDE {
     }
 }
 
-//TODO: Add hash impl on dedicated structure!
-/// Hash support points to a u64 for cache key generation.
-#[inline(always)]
-fn spphash(spp: &[f64]) -> u64 {
-    use std::hash::{Hash, Hasher};
-    let mut hasher = ahash::AHasher::default();
-    for &value in spp {
-        // Normalize -0.0 to 0.0 for consistent hashing
-        let bits = if value == 0.0 { 0u64 } else { value.to_bits() };
-        bits.hash(&mut hasher);
-    }
-    hasher.finish()
-}
-
 #[inline(always)]
 #[cached(
     name = "SDE_PREDICTIONS_CACHE",
     ty = "SizedCache<(u64, u64, u64), f64>",
     create = "{ SizedCache::with_size(100_000) }",
-    convert = r#"{ ((subject.hash()), spphash(support_point), error_models.hash()) }"#,
+    convert = r#"{ (id_hash(subject.id()), spphash(support_point), error_models.hash()) }"#,
     result = "true"
 )]
 fn _estimate_likelihood(
