+++ src/simulator/equation/analytical/mod.rs
@@ -7,6 +7,9 @@ pub use one_compartment_models::*;
 pub use three_compartment_models::*;
 pub use two_compartment_models::*;
 
+use super::id_hash;
+use super::spphash;
+
 use crate::data::error_model::AssayErrorModels;
 use crate::PharmsolError;
 use crate::{
@@ -314,25 +317,12 @@ impl Equation for Analytical {
     }
 }
 
-/// Hash support points to a u64 for cache key generation.
-#[inline(always)]
-fn spphash(spp: &[f64]) -> u64 {
-    use std::hash::{Hash, Hasher};
-    let mut hasher = ahash::AHasher::default();
-    for &value in spp {
-        // Normalize -0.0 to 0.0 for consistent hashing
-        let bits = if value == 0.0 { 0u64 } else { value.to_bits() };
-        bits.hash(&mut hasher);
-    }
-    hasher.finish()
-}
-
 #[inline(always)]
 #[cached(
     name = "ANA_PREDICTIONS_CACHE",
     ty = "SizedCache<(u64, u64), SubjectPredictions>",
     create = "{ SizedCache::with_size(100_000) }",
-    convert = r#"{ (subject.hash(), spphash(support_point)) }"#,
+    convert = r#"{ (id_hash(subject.id()), spphash(support_point)) }"#,
     result = "true"
 )]
 fn _subject_predictions(
