### 2026-03-03 7672cab
paper: Update reference for diffsol
Congratulations to Martin and Alex!
