+++ joss/paper.bib
@@ -35,13 +35,15 @@
   url     = {https://joss.theoj.org/papers/10.21105/joss.06394},
 }
 
-@Misc{diffsol,
-  author    = {Robinson, Martin},
-  title     = {diffsol: A Rust library for solving ordinary differential equations},
-  year      = {2024},
-  publisher = {GitHub},
-  url       = {https://github.com/martinjrobins/diffsol},
-}
+@article{diffsol, 
+  doi = {10.21105/joss.09384}, 
+  url = {https://doi.org/10.21105/joss.09384}, 
+  year = {2026}, 
+  publisher = {The Open Journal}, 
+  volume = {11}, number = {117}, 
+  pages = {9384}, author = {Robinson, Martin and Allmont, Alex}, 
+  title = {diffsol: Rust crate for solving differential equations}, 
+  journal = {Journal of Open Source Software} }
 
 @Article{nonmem,
   author    = {Beal, Stuart L. and Sheiner, Lewis B.},
