+++ joss/paper.bib
@@ -40,8 +40,10 @@
   url = {https://doi.org/10.21105/joss.09384}, 
   year = {2026}, 
   publisher = {The Open Journal}, 
-  volume = {11}, number = {117}, 
-  pages = {9384}, author = {Robinson, Martin and Allmont, Alex}, 
+  volume = {11}, 
+  number = {117}, 
+  pages = {9384}, 
+  author = {Robinson, Martin and Allmont, Alex}, 
   title = {diffsol: Rust crate for solving differential equations}, 
   journal = {Journal of Open Source Software} }
 
