### 2026-03-03 6d8c8e0
paper: Update reference for diffsol (#216)
* paper: Update reference for diffsol

Congratulations to Martin and Alex!

* Formatting
