### 2026-03-04 6bdcd33
Configure cache globally
Also removes deprecated functions
