+++ src/simulator/equation/analytical/mod.rs
@@ -11,12 +11,11 @@ use super::id_hash;
 use super::spphash;
 
 use crate::data::error_model::AssayErrorModels;
+use crate::simulator::cache::{ana_cache_lock_read, cache_enabled};
 use crate::PharmsolError;
 use crate::{
     data::Covariates, simulator::*, Equation, EquationPriv, EquationTypes, Observation, Subject,
 };
-use cached::proc_macro::cached;
-use cached::{Cached, SizedCache};
 
 /// Model equation using analytical solutions.
 ///
@@ -292,9 +291,8 @@ impl Equation for Analytical {
         subject: &Subject,
         support_point: &Vec<f64>,
         error_models: &AssayErrorModels,
-        cache: bool,
     ) -> Result<f64, PharmsolError> {
-        _estimate_likelihood(self, subject, support_point, error_models, cache)
+        _estimate_likelihood(self, subject, support_point, error_models)
     }
 
     fn estimate_log_likelihood(
@@ -302,13 +300,8 @@ impl Equation for Analytical {
         subject: &Subject,
         support_point: &Vec<f64>,
         error_models: &AssayErrorModels,
-        cache: bool,
     ) -> Result<f64, PharmsolError> {
-        let ypred = if cache {
-            _subject_predictions(self, subject, support_point)
-        } else {
-            _subject_predictions_no_cache(self, subject, support_point)
-        }?;
+        let ypred = _subject_predictions(self, subject, support_point)?;
         ypred.log_likelihood(error_models)
     }
 
@@ -318,24 +311,26 @@ impl Equation for Analytical {
 }
 
 #[inline(always)]
-#[cached(
-    name = "ANA_PREDICTIONS_CACHE",
-    ty = "SizedCache<(u64, u64), SubjectPredictions>",
-    create = "{ SizedCache::with_size(100_000) }",
-    convert = r#"{ (id_hash(subject.id()), spphash(support_point)) }"#,
-    result = "true"
-)]
 fn _subject_predictions(
     ode: &Analytical,
     subject: &Subject,
     support_point: &Vec<f64>,
 ) -> Result<SubjectPredictions, PharmsolError> {
-    Ok(ode.simulate_subject(subject, support_point, None)?.0)
-}
+    if cache_enabled() {
+        let key = (id_hash(subject.id()), spphash(support_point));
+        let cache_guard = ana_cache_lock_read()?;
+        if let Some(cached) = cache_guard.get(&key) {
+            return Ok(cached);
+        }
+        drop(cache_guard);
 
-/// Clear the analytical predictions cache.
-pub(crate) fn clear_cache() {
-    ANA_PREDICTIONS_CACHE.lock().unwrap().cache_clear();
+        let result = ode.simulate_subject(subject, support_point, None)?.0;
+        let cache_guard = ana_cache_lock_read()?;
+        cache_guard.insert(key, result.clone());
+        Ok(result)
+    } else {
+        Ok(ode.simulate_subject(subject, support_point, None)?.0)
+    }
 }
 
 fn _estimate_likelihood(
@@ -343,12 +338,7 @@ fn _estimate_likelihood(
     subject: &Subject,
     support_point: &Vec<f64>,
     error_models: &AssayErrorModels,
-    cache: bool,
 ) -> Result<f64, PharmsolError> {
-    let ypred = if cache {
-        _subject_predictions(ode, subject, support_point)
-    } else {
-        _subject_predictions_no_cache(ode, subject, support_point)
-    }?;
+    let ypred = _subject_predictions(ode, subject, support_point)?;
     Ok(ypred.log_likelihood(error_models)?.exp())
 }
