+++ Cargo.toml
@@ -13,7 +13,6 @@ exa = ["libloading"]
 
 [dependencies]
 ahash = "0.8"
-cached = { version = "0.56.0" }
 csv = "1.3.0"
 diffsol = "=0.10.4"
 libloading = { version = "0.9.0", optional = true, features = [] }
@@ -29,6 +28,7 @@ thiserror = "2.0.11"
 argmin = "0.11.0"
 argmin-math = "0.5.1"
 tracing = "0.1.41"
+moka = { version = "0.12.14", features = ["sync"] }
 
 [dev-dependencies]
 criterion = { version = "0.8.0", features = ["html_reports"] }

[…]
