### 2026-03-04 b734fdc
Use full subject for hash
