### 2026-03-06 c7bbe92
wip: saving progress so far
