+++ examples/json_comprehensive_test.rs
@@ -1,0 +1,1216 @@
+// Run with: cargo run --example json_comprehensive_test --features exa
+//
+// Comprehensive test of every JSON model feature against hand-written Rust equivalents.
+// Tests:
+//   1. Simple ODE (1-compartment IV, named compartment)
+//   2. ODE with absorption + rateiv (2-compartment)
+//   3. Analytical model (1-compartment IV)
+//   4. Analytical model with absorption
+//   5. ODE with lag time
+//   6. ODE with bioavailability (fa)
+//   7. ODE with initial conditions
+//   8. ODE with derived parameters (CL/V → ke)
+//   9. ODE with multiple output equations
+//  10. ODE with covariates (allometric scaling)
+//  11. Built-in library model compilation
+
+#[cfg(feature = "exa")]
+fn main() {
+    use pharmsol::prelude::*;
+    use pharmsol::{exa, json, Analytical, ODE};
+    use std::path::PathBuf;
+
+    let template_path = std::env::temp_dir().join("json_comprehensive_test");
+    let test_dir = std::env::temp_dir().join("json_comprehensive_output");
+    std::fs::create_dir_all(&test_dir).ok();
+
+    let mut all_passed = true;
+    let mut test_count = 0;
+    let mut pass_count = 0;
+
+    // Helper macro to run a test
+    macro_rules! run_test {
+        ($name:expr, $body:expr) => {{
+            test_count += 1;
+            print!("  Test {}: {} ... ", test_count, $name);
+            match std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| $body)) {
+                Ok(true) => {
+                    println!("PASS");
+                    pass_count += 1;
+                }
+                Ok(false) => {
+                    println!("FAIL");
+                    all_passed = false;
+                }
+                Err(e) => {
+                    println!("PANIC: {:?}", e.downcast_ref::<String>().unwrap_or(&"unknown".to_string()));
+                    all_passed = false;
+                }
+            }
+        }};
+    }
+
+    println!("═══════════════════════════════════════════════════════════════════════");
+    println!("  JSON Model Comprehensive Test Suite");
+    println!("═══════════════════════════════════════════════════════════════════════");
+
+    // ═══════════════════════════════════════════════════════════════════════════
+    // PART A: Code generation tests (no compilation needed)
+    // ═══════════════════════════════════════════════════════════════════════════
+    println!("\n── Part A: Code Generation (parse → validate → codegen) ──────────────\n");
+
+    // Test 1: Simple ODE with named compartment
+    run_test!("Simple ODE codegen (named compartment)", {
+        let json_str = r#"{
+            "schema": "1.0",
+            "id": "test_1cmt_iv_ode",
+            "type": "ode",
+            "parameters": ["ke", "V"],
+            "compartments": ["central"],
+            "diffeq": {
+                "central": "-ke * central + rateiv[0]"
+            },
+            "output": "central / V"
+        }"#;
+
+        let generated = json::generate_code(json_str).expect("codegen failed");
+        let code = &generated.equation_code;
+
+        // Must contain compartment binding
+        assert!(code.contains("let central = x[0];"), "Missing compartment binding: {}", code);
+        // Must contain the diffeq expression
+        assert!(code.contains("dx[0] = -ke * central + rateiv[0];"), "Wrong diffeq: {}", code);
+        // Must contain the output expression
+        assert!(code.contains("y[0] = central / V;"), "Wrong output: {}", code);
+        // Must contain fetch_params
+        assert!(code.contains("fetch_params!(p, ke, V);"), "Missing fetch_params: {}", code);
+        true
+    });
+
+    // Test 2: ODE with 2 compartments + absorption
[…]
