### 2026-03-06 23d5db2
fix(exa): invalidate compilation cache when pharmsol version changes
Also recover from incomplete template directories.
