+++ src/exa/build.rs
@@ -270,7 +270,15 @@ fn create_template(temp_dir: PathBuf) -> Result<PathBuf, io::Error> {
         "#,
     );
 
-    if !template_dir.exists() {
+    let src_dir = template_dir.join("src");
+    let needs_scaffold = !template_dir.exists() || !src_dir.exists();
+
+    if needs_scaffold {
+        // Remove incomplete template directory if it exists but is malformed
+        if template_dir.exists() {
+            fs::remove_dir_all(&template_dir)?;
+        }
+
         let output = Command::new(find_cargo())
             .arg("new")
             .arg("template")
@@ -282,9 +290,23 @@ fn create_template(temp_dir: PathBuf) -> Result<PathBuf, io::Error> {
         io::stderr().write_all(&output.stderr)?;
         io::stdout().write_all(&output.stdout)?;
 
-        fs::write(cargo_toml_path, cargo_toml_content)?;
+        fs::write(&cargo_toml_path, &cargo_toml_content)?;
     } else if !cargo_toml_path.exists() {
-        fs::write(cargo_toml_path, cargo_toml_content)?;
+        fs::write(&cargo_toml_path, &cargo_toml_content)?;
+    } else {
+        // Check if the pharmsol dependency has changed (e.g. after a version upgrade).
+        // If so, rewrite Cargo.toml and remove build artifacts to force recompilation.
+        // Without this, a stale dylib compiled against an older pharmsol version would be
+        // loaded at runtime, causing ABI mismatches and crashes.
+        let existing_content = fs::read_to_string(&cargo_toml_path).unwrap_or_default();
+        if existing_content.trim() != cargo_toml_content.trim() {
+            tracing::info!("pharmsol dependency changed, invalidating exa compilation cache");
+            fs::write(&cargo_toml_path, &cargo_toml_content)?;
+            let target_dir = template_dir.join("target");
+            if target_dir.exists() {
+                fs::remove_dir_all(&target_dir)?;
+            }
+        }
     };
     Ok(template_dir)
 }
