### 2026-03-09 0c6ea67
chore: Update cached requirement from 0.56.0 to 0.58.0
Updates the requirements on [cached](https://github.com/jaemk/cached) to permit the latest version.
- [Changelog](https://github.com/jaemk/cached/blob/master/CHANGELOG.md)
- [Commits](https://github.com/jaemk/cached/commits)

---
updated-dependencies:
- dependency-name: cached
  dependency-version: 0.58.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
