+++ Cargo.toml
@@ -12,7 +12,7 @@ default = []
 exa = ["libloading"]
 
 [dependencies]
-cached = { version = "0.56.0" }
+cached = { version = "0.58.0" }
 csv = "1.3.0"
 diffsol = "=0.10.4"
 libloading = { version = "0.9.0", optional = true, features = [] }
