### 2026-03-10 6ff8717
fix: when ordering events by priority, observations should have the highest one
