+++ src/data/structs.rs
@@ -589,16 +589,18 @@ impl Occasion {
         }
     }
 
-    /// Sort events by time, then by [Event] type so that [Bolus] and [Infusion] come before [Observation]
+    /// Sort events by time, then by [Event] type so that [Observation] comes before [Bolus] and [Infusion]
     pub(crate) fn sort(&mut self) {
         self.events.sort_by(|a, b| {
             // Helper function to get event type order
+            // Observations come first so that at the same time point,
+            // the pre-dose state is observed before the dose is applied.
             #[inline]
             fn event_type_order(event: &Event) -> u8 {
                 match event {
-                    Event::Bolus(_) => 1,
-                    Event::Infusion(_) => 2,
-                    Event::Observation(_) => 3,
+                    Event::Observation(_) => 1,
+                    Event::Bolus(_) => 2,
+                    Event::Infusion(_) => 3,
                 }
             }
 
@@ -953,7 +955,7 @@ mod tests {
     }
 
     #[test]
-    fn test_occasion_sort() {
+    fn test_occasion_sort_by_time() {
         let mut occasion = Occasion::new(0);
         occasion.add_observation(2.0, 1.0, 1, None, Censor::None);
         occasion.add_bolus(1.0, 100.0, 1);
@@ -961,12 +963,87 @@ mod tests {
         let events = occasion.process_events(None, false);
         match &events[0] {
             Event::Bolus(b) => assert_eq!(b.time(), 1.0),
-            _ => panic!("First event should be a Bolus"),
+            _ => panic!("First event should be a Bolus (earlier time)"),
         }
         match &events[1] {
             Event::Observation(o) => assert_eq!(o.time(), 2.0),
-            _ => panic!("Second event should be an Observation"),
+            _ => panic!("Second event should be an Observation (later time)"),
         }
+    }
+
+    #[test]
+    fn test_observation_before_bolus_at_same_time() {
+        let mut occasion = Occasion::new(0);
+        occasion.add_bolus(1.0, 100.0, 1);
+        occasion.add_observation(1.0, 5.0, 1, None, Censor::None);
+        occasion.sort();
+        let events = occasion.process_events(None, false);
+        assert_eq!(events.len(), 2);
+        match &events[0] {
+            Event::Observation(o) => assert_eq!(o.time(), 1.0),
+            _ => panic!("Observation should come before Bolus at the same time"),
+        }
+        match &events[1] {
+            Event::Bolus(b) => assert_eq!(b.time(), 1.0),
+            _ => panic!("Bolus should come after Observation at the same time"),
+        }
+    }
+
+    #[test]
+    fn test_observation_before_infusion_at_same_time() {
+        let mut occasion = Occasion::new(0);
+        occasion.add_infusion(1.0, 100.0, 1, 0.5);
+        occasion.add_observation(1.0, 5.0, 1, None, Censor::None);
+        occasion.sort();
+        let events = occasion.process_events(None, false);
+        assert_eq!(events.len(), 2);
+        match &events[0] {
+            Event::Observation(o) => assert_eq!(o.time(), 1.0),
+            _ => panic!("Observation should come before Infusion at the same time"),
+        }
+        match &events[1] {
+            Event::Infusion(i) => assert_eq!(i.time(), 1.0),
+            _ => panic!("Infusion should come after Observation at the same time"),
+        }
+    }
+
+    #[test]
+    fn test_sort_mixed_events_same_time() {
+        // All three event types at the same time
+        let mut occasion = Occasion::new(0);
+        occasion.add_infusion(0.0, 500.0, 1, 1.0);
[…]
