### 2026-03-13 569512e
docs: Updated documentation for add_event
