### 2026-03-16 a2fb2d2
fix: Update sort order of events (#219)
* fix: when ordering events by priority, observations should have the highest one
