### 2026-03-16 dadad26
fix(exa): invalidate compilation cache when pharmsol version changes (#217)
* fix(exa): invalidate compilation cache when pharmsol version changes

Also recover from incomplete template directories.

* Apply suggestion from @Copilot

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
