### 2026-03-16 51f69fc
Enable cache by default
