### 2026-03-16 30e87d7
feat: Use a sized cache with new subject hashing (#214)
* feat: Use a sized cache with new subject hashing

* Use ahash for faster hashing

* Add unit tests for the new hash methods

* Dont hash covariates

* Add a function to clear caches

* Documentation

* Update the hash key and refactor

* Configure cache globally

Also removes deprecated functions

* Enable cache by default
