+++ src/simulator/equation/sde/mod.rs
@@ -6,9 +6,6 @@ use ndarray::{concatenate, Array2, Axis};
 use rand::{rng, RngExt};
 use rayon::prelude::*;
 
-use cached::proc_macro::cached;
-use cached::UnboundCache;
-
 use crate::{
     data::{Covariates, Infusion},
     error_model::AssayErrorModels,
@@ -16,6 +13,10 @@ use crate::{
     simulator::{Diffusion, Drift, Fa, Init, Lag, Neqs, Out, V},
     Subject,
 };
+
+use super::id_hash;
+use super::spphash;
+use crate::simulator::cache::{cache_enabled, sde_cache_lock_read};
 
 use diffsol::VectorCommon;
 
@@ -358,7 +359,6 @@ impl Equation for SDE {
     /// * `subject` - Subject data containing observations
     /// * `support_point` - Parameter vector for the model
     /// * `error_model` - Error model to use for likelihood calculations
-    /// * `cache` - Whether to cache likelihood results for reuse
     ///
     /// # Returns
     ///
@@ -368,15 +368,8 @@ impl Equation for SDE {
         subject: &Subject,
         support_point: &Vec<f64>,
         error_models: &AssayErrorModels,
-        cache: bool,
     ) -> Result<f64, PharmsolError> {
-        if cache {
-            _estimate_likelihood(self, subject, support_point, error_models)
-        } else {
-            // No cache version: directly simulate
-            let ypred = self.simulate_subject(subject, support_point, Some(error_models))?;
-            Ok(ypred.1.unwrap())
-        }
+        _estimate_likelihood(self, subject, support_point, error_models)
     }
 
     fn estimate_log_likelihood(
@@ -384,17 +377,10 @@ impl Equation for SDE {
         subject: &Subject,
         support_point: &Vec<f64>,
         error_models: &AssayErrorModels,
-        cache: bool,
     ) -> Result<f64, PharmsolError> {
         // For SDE, the particle filter computes likelihood in regular space.
         // We compute it directly and then take the log.
-        let lik = if cache {
-            _estimate_likelihood(self, subject, support_point, error_models)?
-        } else {
-            // No cache version: directly simulate
-            let ypred = self.simulate_subject(subject, support_point, Some(error_models))?;
-            ypred.1.unwrap()
-        };
+        let lik = _estimate_likelihood(self, subject, support_point, error_models)?;
 
         if lik > 0.0 {
             Ok(lik.ln())
@@ -408,37 +394,34 @@ impl Equation for SDE {
     }
 }
 
-//TODO: Add hash impl on dedicated structure!
-/// Hash support points to a u64 for cache key generation.
-/// Uses DefaultHasher for good distribution and collision resistance.
 #[inline(always)]
-
-fn spphash(spp: &[f64]) -> u64 {
-    use std::hash::{Hash, Hasher};
-    let mut hasher = std::hash::DefaultHasher::new();
-    for &value in spp {
-        // Normalize -0.0 to 0.0 for consistent hashing
-        let bits = if value == 0.0 { 0u64 } else { value.to_bits() };
-        bits.hash(&mut hasher);
-    }
-    hasher.finish()
-}
-
-#[inline(always)]
-#[cached(
-    ty = "UnboundCache<(u64, u64, u64), f64>",
-    create = "{ UnboundCache::with_capacity(100_000) }",
-    convert = r#"{ ((subject.hash()), spphash(support_point), error_models.hash()) }"#,
-    result = "true"
-)]
 fn _estimate_likelihood(
     sde: &SDE,
     subject: &Subject,
     support_point: &Vec<f64>,
     error_models: &AssayErrorModels,
 ) -> Result<f64, PharmsolError> {
[…]
