### 2026-03-16 b10c3ca
fix: Prevents creating a terminal on windows (#179)
* Prevents creating a terminal on windows

* Update build.rs

---------

Co-authored-by: Markus Hovd <markushh@uio.no>
