### 2026-03-16 41c4131
fix: Fix test and capture error to drop
