### 2026-03-16 002bac4
feat: Remove LikelihoodMatrixOptions
Since it only contained one field
