### 2026-03-16 f268d2f
feat: Remove LikelihoodMatrixOptions (#222)
Since it only contained one field
