+++ src/simulator/likelihood/mod.rs
@@ -60,9 +60,14 @@ mod progress;
 mod subject;
 
 // Re-export main types
-pub use matrix::{log_likelihood_matrix, LikelihoodMatrixOptions};
+pub use matrix::log_likelihood_matrix;
 pub use prediction::Prediction;
 pub use subject::{PopulationPredictions, SubjectPredictions};
+
+#[allow(deprecated)]
+pub use matrix::log_psi;
+#[allow(deprecated)]
+pub use matrix::psi;
 
 use ndarray::Array2;
 use rayon::prelude::*;
