+++ src/lib.rs
@@ -54,6 +54,7 @@ pub mod prelude {
         Covariates, Data, Event, Interpolation, Occasion, Subject,
     };
 
+    #[allow(deprecated)]
     // Simulator submodule for internal use and advanced users
     pub mod simulator {
         pub use crate::simulator::{
@@ -61,8 +62,8 @@ pub mod prelude {
             equation,
             equation::Equation,
             likelihood::{
-                log_likelihood_batch, log_likelihood_matrix, log_likelihood_subject,
-                LikelihoodMatrixOptions, PopulationPredictions, Prediction, SubjectPredictions,
+                log_likelihood_batch, log_likelihood_matrix, log_likelihood_subject, log_psi, psi,
+                PopulationPredictions, Prediction, SubjectPredictions,
             },
         };
     }
