### 2026-03-16 966ff39
chore: release v0.23.0 (#223)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
