+++ src/data/errormodel/residual.rs
@@ -1,0 +1,283 @@
+use super::{ErrorModel, ErrorModelError, ErrorModels};
+use serde::{Deserialize, Serialize};
+
+/// Residual error model for parametric estimation algorithms.
+///
+/// Unlike [`super::AssayErrorModel`] which uses observations, this uses
+/// the model **prediction** to compute the standard deviation.
+#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq)]
+pub enum ResidualErrorModel {
+    /// Constant (additive) error model: σ = a
+    Constant {
+        /// Additive error standard deviation
+        a: f64,
+    },
+
+    /// Proportional error model: σ = b * |f|
+    Proportional {
+        /// Proportional coefficient (e.g., 0.1 = 10% CV)
+        b: f64,
+    },
+
+    /// Combined (additive + proportional) error model: σ = sqrt(a² + b² * f²)
+    Combined {
+        /// Additive component (a)
+        a: f64,
+        /// Proportional component (b)
+        b: f64,
+    },
+
+    /// Exponential error model (for log-transformed data): σ = sigma (constant on log scale)
+    Exponential {
+        /// Error standard deviation on log scale
+        sigma: f64,
+    },
+}
+
+impl Default for ResidualErrorModel {
+    fn default() -> Self {
+        ResidualErrorModel::Constant { a: 1.0 }
+    }
+}
+
+impl ErrorModel for ResidualErrorModel {
+    /// Compute sigma from a prediction value.
+    ///
+    /// Returns a cutoff minimum to avoid numerical issues with very small σ.
+    fn sigma(&self, prediction: f64) -> Result<f64, ErrorModelError> {
+        let raw_sigma = match self {
+            ResidualErrorModel::Constant { a } => *a,
+            ResidualErrorModel::Proportional { b } => b * prediction.abs(),
+            ResidualErrorModel::Combined { a, b } => {
+                (a.powi(2) + b.powi(2) * prediction.powi(2)).sqrt()
+            }
+            ResidualErrorModel::Exponential { sigma } => *sigma,
+        };
+
+        // Apply cutoff to prevent division by zero in likelihood
+        Ok(raw_sigma.max(f64::EPSILON.sqrt()))
+    }
+}
+
+impl ResidualErrorModel {
+    /// Create a constant (additive) error model
+    pub fn constant(a: f64) -> Self {
+        ResidualErrorModel::Constant { a }
+    }
+
+    /// Create a proportional error model
+    pub fn proportional(b: f64) -> Self {
+        ResidualErrorModel::Proportional { b }
+    }
+
+    /// Create a combined (additive + proportional) error model
+    pub fn combined(a: f64, b: f64) -> Self {
+        ResidualErrorModel::Combined { a, b }
+    }
+
+    /// Create an exponential error model
+    pub fn exponential(sigma: f64) -> Self {
+        ResidualErrorModel::Exponential { sigma }
+    }
+
+    /// Compute the weighted residual for M-step sigma updates
+    pub fn weighted_squared_residual(&self, observation: f64, prediction: f64) -> f64 {
+        let residual = observation - prediction;
+        let residual_sq = residual * residual;
+
+        match self {
+            ResidualErrorModel::Constant { .. } => residual_sq,
+            ResidualErrorModel::Proportional { .. } => {
+                let pred_sq = prediction.powi(2).max(f64::EPSILON);
+                residual_sq / pred_sq
+            }
+            ResidualErrorModel::Combined { a, b } => {
[…]
