### 2026-03-18 134084a
feat: tests are now hardcoed and not json-generated files
