+++ tests/nca/test_pknca.rs
@@ -1,96 +1,22 @@
 //! PKNCA Cross-Validation Tests
 //!
 //! This module validates pharmsol's NCA implementation against expected values
-//! generated by PKNCA (the gold-standard R package for NCA).
+//! computed by PKNCA v0.12.1 (the gold-standard R package for NCA).
 //!
-//! The validation uses a clean-room approach:
-//! 1. Test scenarios are independently designed based on PK principles
-//! 2. PKNCA computes expected values (run `Rscript generate_expected.R`)
-//! 3. This module compares pharmsol's results against those expected values
+//! Each test is a self-contained scenario with hardcoded inputs and expected outputs.
+//! The reference values were generated using PKNCA via `Rscript generate_expected.R`.
 //!
-//! Run with: `cargo test pknca_validation`
+//! Tolerance: 0.1% relative, 1e-10 absolute.
 
 use pharmsol::nca::{AUCMethod, BLQRule, NCAOptions, Route, RouteParams};
 use pharmsol::{prelude::*, Censor};
-use serde::Deserialize;
-use std::collections::HashMap;
-use std::fs;
-use std::path::Path;
 
-/// Tolerance for floating-point comparisons
-/// NCA calculations should match within 0.1% for most parameters
 const RELATIVE_TOLERANCE: f64 = 0.001;
-
-/// Absolute tolerance for very small values (near zero)
 const ABSOLUTE_TOLERANCE: f64 = 1e-10;
 
-// =============================================================================
-// JSON Structures for Test Data
-// =============================================================================
-
-#[derive(Debug, Deserialize)]
-struct TestScenarios {
-    scenarios: Vec<Scenario>,
-}
-
-#[derive(Debug, Deserialize)]
-#[allow(dead_code)]
-struct Scenario {
-    id: String,
-    name: String,
-    route: String,
-    dose: DoseInfo,
-    times: Vec<f64>,
-    concentrations: Vec<f64>,
-    #[serde(default)]
-    auc_method: Option<String>,
-    #[serde(default)]
-    blq_rule: Option<String>,
-    #[serde(default)]
-    blq_indices: Option<Vec<usize>>,
-    #[serde(default)]
-    loq: Option<f64>,
-    #[serde(default)]
-    partial_auc_interval: Option<Vec<f64>>,
-    #[serde(default)]
-    tau: Option<f64>,
-    test_params: Vec<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct DoseInfo {
-    amount: f64,
-    time: f64,
-    #[serde(default)]
-    duration: Option<f64>,
-}
-
-#[derive(Debug, Deserialize)]
-struct ExpectedValues {
-    generated_at: String,
-    pknca_version: String,
-    results: HashMap<String, ScenarioResult>,
-}
-
-#[derive(Debug, Deserialize)]
-#[allow(dead_code)]
-struct ScenarioResult {
-    id: String,
-    name: String,
-    #[serde(default)]
-    parameters: HashMap<String, f64>,
-    #[serde(default)]
-    error: Option<String>,
-}
-
-// =============================================================================
-// Test Utilities
-// =============================================================================
-
-/// Check if two floating-point values are approximately equal
-fn approx_eq(a: f64, b: f64, rel_tol: f64, abs_tol: f64) -> bool {
+fn approx_eq(a: f64, b: f64) -> bool {
     if a.is_nan() && b.is_nan() {
-        return true; // Both NaN is considered equal for our purposes
+        return true;
     }
     if a.is_nan() || b.is_nan() {
         return false;
@@ -104,400 +30,1009 @@ fn approx_eq(a: f64, b: f64, rel_tol: f64, abs_tol: f64) -> bool {
 
     let diff = (a - b).abs();
[…]
