+++ tests/nca/pknca/generate_expected.R
@@ -1,250 +1,0 @@
-#!/usr/bin/env Rscript
-# =============================================================================
-# PKNCA Cross-Validation: Generate Expected Values
-# =============================================================================
-#
-# This script reads test scenarios from test_scenarios.json, runs PKNCA to
-# compute NCA parameters, and saves the expected values to expected_values.json.
-#
-# Usage: Rscript generate_expected.R
-#
-# Requirements: R with PKNCA, jsonlite packages installed
-# =============================================================================
-
-library(PKNCA)
-library(jsonlite)
-
-cat("PKNCA Cross-Validation - Generating Expected Values\n")
-cat("====================================================\n\n")
-
-# Read test scenarios
-scenarios_raw <- fromJSON("test_scenarios.json", simplifyVector = FALSE)
-scenarios <- scenarios_raw$scenarios
-cat(sprintf("Loaded %d test scenarios\n\n", length(scenarios)))
-
-# Helper function to map our route names to PKNCA expectations
-get_route_for_pknca <- function(route) {
-    switch(route,
-        "extravascular" = "extravascular",
-        "iv_bolus" = "intravascular",
-        "iv_infusion" = "intravascular",
-        route
-    )
-}
-
-# Helper function to get AUC method name for PKNCA
-get_auc_method <- function(method) {
-    if (is.null(method)) {
-        return("lin up/log down")
-    }
-    method
-}
-
-# Process each scenario
-results <- list()
-
-for (scenario in scenarios) {
-    cat(sprintf("Processing: %s (%s)\n", scenario$name, scenario$id))
-
-    tryCatch(
-        {
-            # Build concentration data frame - unlist JSON arrays
-            times <- unlist(scenario$times)
-            concs <- unlist(scenario$concentrations)
-
-            conc_data <- data.frame(
-                ID = 1,
-                time = times,
-                conc = concs
-            )
-
-            # Handle BLQ if specified
-            if (!is.null(scenario$blq_indices)) {
-                # Mark BLQ as 0 (PKNCA convention)
-                # Note: blq_indices are 0-based from JSON
-                blq_idx <- unlist(scenario$blq_indices)
-                for (idx in blq_idx) {
-                    conc_data$conc[idx + 1] <- 0
-                }
-            }
-
-            # Build dose data frame
-            dose_data <- data.frame(
-                ID = 1,
-                time = scenario$dose$time,
-                dose = scenario$dose$amount
-            )
-
-            # Add duration for infusions
-            if (scenario$route == "iv_infusion" && !is.null(scenario$dose$duration)) {
-                dose_data$duration <- scenario$dose$duration
-            }
-
-            # Create PKNCA objects
-            conc_obj <- PKNCAconc(conc_data, conc ~ time | ID)
-
-            if (scenario$route == "iv_infusion" && !is.null(scenario$dose$duration)) {
-                dose_obj <- PKNCAdose(dose_data, dose ~ time | ID,
-                    route = "intravascular",
-                    duration = "duration"
-                )
-            } else {
-                dose_obj <- PKNCAdose(dose_data, dose ~ time | ID,
-                    route = get_route_for_pknca(scenario$route)
-                )
-            }
-
-            # Set up intervals - request all parameters up to infinity
-            intervals <- data.frame(
-                start = 0,
-                end = Inf,
-                cmax = TRUE,
-                tmax = TRUE,
-                tlast = TRUE,
-                clast.obs = TRUE,
-                auclast = TRUE,
-                aucall = TRUE,
-                aumclast = TRUE,
-                half.life = TRUE,
-                lambda.z = TRUE,
-                r.squared = TRUE,
-                adj.r.squared = TRUE,
-                lambda.z.n.points = TRUE,
[…]
