### 2026-03-18 ed51745
feat: seconday equations now can use expressions
