+++ examples/json_exa.rs
@@ -1,11 +1,7 @@
-// Run with: cargo run --example json_exa --features exa
+// cargo run --example json_exa --features exa
 //
-// This example demonstrates JSON model compilation using the `exa` feature.
-// It compares predictions from:
-// 1. A statically defined ODE model (Rust code)
-// 2. A dynamically compiled ODE model (via exa, raw Rust string)
-// 3. A JSON-defined ODE model (via compile_json)
-// 4. A JSON-defined Analytical model (via compile_json)
+// Compiles a 1-compartment IV model four different ways and checks that they
+// all agree: static Rust, exa (raw Rust string), JSON ODE, JSON analytical.
 
 #[cfg(feature = "exa")]
 fn main() {
@@ -13,7 +9,6 @@ fn main() {
     use pharmsol::{exa, json, Analytical, ODE};
     use std::path::PathBuf;
 
-    // Create test subject with infusion and observations
     let subject = Subject::builder("1")
         .infusion(0.0, 500.0, 0, 0.5)
         .observation(0.5, 1.645776, 0)
@@ -25,18 +20,13 @@ fn main() {
         .observation(8.0, 0.001017932, 0)
         .build();
 
-    // Parameters: ke (elimination rate constant), V (volume of distribution)
-    let params = vec![1.2, 50.0];
+    let params = vec![1.2, 50.0]; // ke, V
 
-    let test_dir = std::env::current_dir().expect("Failed to get current directory");
-
-    // Shared template path for all compilations (they run sequentially)
+    let cwd = std::env::current_dir().unwrap();
     let template_path = std::env::temp_dir().join("exa_json_example");
 
-    // =========================================================================
-    // 1. Create ODE model directly (static Rust code)
-    // =========================================================================
-    println!("1. Creating static ODE model...");
+    // -- Static ODE (plain Rust) ----------------------------------------------
+
     let static_ode = equation::ODE::new(
         |x, p, _t, dx, _bolus, rateiv, _cov| {
             fetch_params!(p, ke, _v);
@@ -51,14 +41,10 @@ fn main() {
         },
         (1, 1),
     );
-    println!("   ✓ Static ODE model created\n");
 
-    // =========================================================================
-    // 2. Compile ODE model dynamically using exa (raw Rust string)
-    // =========================================================================
-    println!("2. Compiling ODE model via exa (raw Rust)...");
-    let exa_ode_path = test_dir.join("exa_ode_model.pkm");
+    // -- Exa ODE (raw Rust string, compiled at runtime) -----------------------
 
+    let exa_ode_path = cwd.join("exa_ode_model.pkm");
     let exa_ode_compiled = exa::build::compile::<ODE>(
         r#"
             equation::ODE::new(
@@ -82,231 +68,115 @@ fn main() {
         template_path.clone(),
         |_, _| {},
     )
-    .expect("Failed to compile ODE model via exa");
+    .unwrap();
 
     let exa_ode_path = PathBuf::from(&exa_ode_compiled);
-    let (_lib_exa_ode, (dynamic_exa_ode, _)) =
-        unsafe { exa::load::load::<ODE>(exa_ode_path.clone()) };
-    println!("   ✓ Compiled to: {}\n", exa_ode_compiled);
+    let (_lib_exa, (dyn_exa_ode, _)) = unsafe { exa::load::load::<ODE>(exa_ode_path.clone()) };
 
-    // =========================================================================
-    // 3. Compile ODE model from JSON using compile_json
-    // =========================================================================
-    println!("3. Compiling ODE model from JSON...");
[…]
