+++ src/json/expression.rs
@@ -1,0 +1,1198 @@
+//! Language-agnostic math expression parser and Rust code emitter.
+//!
+//! This module parses a simple math expression language used in JSON model
+//! definitions and transpiles it to Rust code. The expression language is
+//! intentionally simple and language-agnostic — it uses standard mathematical
+//! notation (e.g. `^` for power, `exp()`, `ln()`) rather than Rust-specific
+//! syntax (e.g. `.powf()`, `.ln()`).
+//!
+//! # Expression Grammar
+//!
+//! ```text
+//! expr     := ternary
+//! ternary  := logic ("?" logic ":" logic)?
+//! logic    := compare (("and" | "or") compare)*
+//! compare  := addition (("==" | "!=" | "<=" | ">=" | "<" | ">") addition)?
+//! addition := multiply (("+" | "-") multiply)*
+//! multiply := unary (("*" | "/") unary)*
+//! unary    := ("-" | "not") unary | power
+//! power    := call ("^" unary)?
+//! call     := IDENT "(" args ")" | indexing
+//! indexing := atom ("[" expr "]")?
+//! atom     := NUMBER | IDENT | "(" expr ")"
+//! args     := expr ("," expr)*
+//! ```
+//!
+//! # Supported Functions
+//!
+//! `exp`, `ln`, `log`, `log2`, `log10`, `sqrt`, `abs`, `min`, `max`,
+//! `floor`, `ceil`, `sin`, `cos`, `if`
+//!
+//! # Examples
+//!
+//! | Expression | Generated Rust |
+//! |---|---|
+//! | `CL * (wt / 70)^0.75` | `CL * (wt / 70.0).powf(0.75)` |
+//! | `exp(-ke * t)` | `(-ke * t).exp()` |
+//! | `if(sex == 1, V * 0.8, V)` | `if sex == 1.0 { V * 0.8 } else { V }` |
+//! | `ln(x)` | `(x).ln()` |
+
+use std::collections::HashMap;
+use std::fmt;
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// AST
+// ═══════════════════════════════════════════════════════════════════════════════
+
+/// AST node for parsed expressions
+#[derive(Debug, Clone, PartialEq)]
+pub enum Expr {
+    /// Numeric literal
+    Number(f64),
+    /// Variable or parameter reference
+    Ident(String),
+    /// Binary operation
+    BinOp {
+        op: BinOp,
+        left: Box<Expr>,
+        right: Box<Expr>,
+    },
+    /// Unary negation
+    Neg(Box<Expr>),
+    /// Logical not
+    Not(Box<Expr>),
+    /// Function call (e.g., exp(x), ln(x), sqrt(x))
+    FuncCall { name: String, args: Vec<Expr> },
+    /// Conditional: if(cond, then, else)
+    If {
+        condition: Box<Expr>,
+        then_expr: Box<Expr>,
+        else_expr: Box<Expr>,
+    },
+    /// Array/vector indexing: name[index]
+    Index { name: String, index: Box<Expr> },
+}
+
+/// Binary operators
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum BinOp {
+    Add,
+    Sub,
+    Mul,
+    Div,
+    Pow,
+    Eq,
+    Neq,
+    Lt,
+    Gt,
+    Lte,
+    Gte,
+    And,
+    Or,
+}
+
+impl fmt::Display for BinOp {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            BinOp::Add => write!(f, "+"),
+            BinOp::Sub => write!(f, "-"),
+            BinOp::Mul => write!(f, "*"),
[…]
