+++ src/simulator/equation/ode/mod.rs
@@ -410,18 +410,16 @@ impl Equation for ODE {
         let nstates = self.get_nstates();
         let ndrugs = self.get_ndrugs();
 
-        let state_buffer_size = nstates;
-        let output_buffer_size = self.get_nouteqs();
-
-        // Preallocate reusable vectors for bolus computation
-        let mut state_with_bolus = V::zeros(state_buffer_size, NalgebraContext);
-        let mut state_without_bolus = V::zeros(state_buffer_size, NalgebraContext);
-        let zero_vector = V::zeros(state_buffer_size, NalgebraContext);
-        let mut bolus_v = V::zeros(state_buffer_size, NalgebraContext);
+        // Preallocate reusable vectors for bolus computation (sized by ndrugs)
+        let mut state_with_bolus = V::zeros(nstates, NalgebraContext);
+        let mut state_without_bolus = V::zeros(nstates, NalgebraContext);
+        let zero_bolus = V::zeros(ndrugs, NalgebraContext);
+        let zero_rateiv = V::zeros(ndrugs, NalgebraContext);
+        let mut bolus_v = V::zeros(ndrugs, NalgebraContext);
         let spp_v: V = DVector::from_vec(support_point.clone()).into();
 
         // Pre-allocate output vector for observations
-        let mut y_out = V::zeros(output_buffer_size, NalgebraContext);
+        let mut y_out = V::zeros(self.get_nouteqs(), NalgebraContext);
 
         // Iterate over occasions
         for occasion in subject.occasions() {
