### 2026-03-18 ede21b7
feat: new constructor for Equation, adding with_nstates and with_nout methods; there is also a new array size definition ndrugs, this one defines the size of bolus[] and rateiv[] (previously we were using nstates); We are now also exporting the ODE solver, so users can decide which one to use
