+++ src/simulator/equation/analytical/two_compartment_models.rs
@@ -9,7 +9,7 @@ use nalgebra::{DVector, Matrix2, Vector2};
 /// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
 /// - `x` is a vector of length 2
 /// - covariates are not used
-pub fn two_compartments(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
+pub fn two_compartments(x: &V, p: &V, t: T, rateiv: &V, _cov: &Covariates) -> V {
     let ke = p[0];
     let kcp = p[1];
     let kpc = p[2];
@@ -52,7 +52,7 @@ pub fn two_compartments(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
 /// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
 /// - `x` is a vector of length 3
 /// - covariates are not used
-pub fn two_compartments_with_absorption(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
+pub fn two_compartments_with_absorption(x: &V, p: &V, t: T, rateiv: &V, _cov: &Covariates) -> V {
     let ke = p[0];
     let ka = p[1];
     let kcp = p[2];
@@ -175,7 +175,7 @@ mod tests {
                 fetch_params!(p, ke, kcp, kpc, _v);
 
                 dx[0] = rateiv[0] - ke * x[0] - kcp * x[0] + kpc * x[1] + b[0];
-                dx[1] = kcp * x[0] - kpc * x[1] + b[1];
+                dx[1] = kcp * x[0] - kpc * x[1];
             },
             |_p, _t, _cov| lag! {},
             |_p, _t, _cov| fa! {},
@@ -184,8 +184,10 @@ mod tests {
                 fetch_params!(p, _ke, _kcp, _kpc, v);
                 y[0] = x[0] / v;
             },
-            (2, 1),
-        );
+        )
+        .with_nstates(2)
+        .with_ndrugs(1)
+        .with_nout(1);
 
         let analytical = equation::Analytical::new(
             two_compartments,
@@ -197,8 +199,10 @@ mod tests {
                 fetch_params!(p, _ke, _kcp, _kpc, v);
                 y[0] = x[0] / v;
             },
-            (2, 1),
-        );
+        )
+        .with_nstates(2)
+        .with_ndrugs(1)
+        .with_nout(1);
 
         let op_ode = ode
             .estimate_predictions(&subject, &vec![0.1, 3.0, 1.0, 1.0])
@@ -229,7 +233,7 @@ mod tests {
 
                 dx[0] = -ka * x[0] + b[0];
                 dx[1] = rateiv[0] - ke * x[1] + ka * x[0] - kcp * x[1] + kpc * x[2] + b[1];
-                dx[2] = kcp * x[1] - kpc * x[2] + b[2];
+                dx[2] = kcp * x[1] - kpc * x[2];
             },
             |_p, _t, _cov| lag! {},
             |_p, _t, _cov| fa! {},
@@ -238,8 +242,10 @@ mod tests {
                 fetch_params!(p, _ke, _ka, _kcp, _kpc, v);
                 y[0] = x[1] / v;
             },
-            (3, 1),
-        );
+        )
+        .with_nstates(3)
+        .with_ndrugs(2)
+        .with_nout(1);
 
         let analytical = equation::Analytical::new(
             two_compartments_with_absorption,
@@ -251,8 +257,10 @@ mod tests {
                 fetch_params!(p, _ke, _ka, _kcp, _kpc, v);
                 y[0] = x[1] / v;
             },
-            (3, 1),
-        );
+        )
[…]
