+++ tests/ode_optimizations.rs
@@ -92,8 +92,9 @@ fn single_iv_bolus_matches_analytical() {
             fetch_params!(p, _ke, v);
             y[0] = x[0] / v;
         },
-        (1, 1),
-    );
+    )
+    .with_nstates(1)
+    .with_nout(1);
 
     let ode = equation::ODE::new(
         |x, p, _t, dx, b, _rateiv, _cov| {
@@ -108,8 +109,9 @@ fn single_iv_bolus_matches_analytical() {
             fetch_params!(p, _ke, v);
             y[0] = x[0] / v;
         },
-        (1, 1),
-    );
+    )
+    .with_nstates(1)
+    .with_nout(1);
 
     // ke = 0.1, v = 50
     assert_ode_matches_analytical("single_iv_bolus", &analytical, &ode, &subject, &[0.1, 50.0]);
@@ -143,8 +145,9 @@ fn multiple_iv_boluses_match_analytical() {
             fetch_params!(p, _ke, v);
             y[0] = x[0] / v;
         },
-        (1, 1),
-    );
+    )
+    .with_nstates(1)
+    .with_nout(1);
 
     let ode = equation::ODE::new(
         |x, p, _t, dx, b, _rateiv, _cov| {
@@ -158,8 +161,9 @@ fn multiple_iv_boluses_match_analytical() {
             fetch_params!(p, _ke, v);
             y[0] = x[0] / v;
         },
-        (1, 1),
-    );
+    )
+    .with_nstates(1)
+    .with_nout(1);
 
     assert_ode_matches_analytical(
         "multiple_iv_boluses",
@@ -194,8 +198,9 @@ fn oral_bolus_with_absorption_matches_analytical() {
             fetch_params!(p, _ka, _ke, v);
             y[0] = x[1] / v; // Central compartment is x[1]
         },
-        (2, 1),
-    );
+    )
+    .with_nstates(2)
+    .with_nout(1);
 
     let ode = equation::ODE::new(
         |x, p, _t, dx, b, _rateiv, _cov| {
@@ -210,8 +215,9 @@ fn oral_bolus_with_absorption_matches_analytical() {
             fetch_params!(p, _ka, _ke, v);
             y[0] = x[1] / v;
         },
-        (2, 1),
-    );
+    )
+    .with_nstates(2)
+    .with_nout(1);
 
     // ka = 1.0, ke = 0.1, v = 50
     assert_ode_matches_analytical(
@@ -253,8 +259,9 @@ fn multiple_oral_doses_match_analytical() {
             fetch_params!(p, _ka, _ke, v);
             y[0] = x[1] / v;
         },
-        (2, 1),
-    );
+    )
+    .with_nstates(2)
+    .with_nout(1);
 
     let ode = equation::ODE::new(
         |x, p, _t, dx, b, _rateiv, _cov| {
@@ -269,8 +276,9 @@ fn multiple_oral_doses_match_analytical() {
             fetch_params!(p, _ka, _ke, v);
             y[0] = x[1] / v;
         },
-        (2, 1),
-    );
+    )
+    .with_nstates(2)
+    .with_nout(1);
 
     assert_ode_matches_analytical(
         "multiple_oral_doses",
@@ -309,8 +317,9 @@ fn single_infusion_matches_analytical() {
             fetch_params!(p, _ke, v);
             y[0] = x[0] / v;
         },
-        (1, 1),
-    );
+    )
+    .with_nstates(1)
+    .with_nout(1);
 
     let ode = equation::ODE::new(
         |x, p, _t, dx, _b, rateiv, _cov| {
@@ -324,8 +333,9 @@ fn single_infusion_matches_analytical() {
             fetch_params!(p, _ke, v);
             y[0] = x[0] / v;
         },
[…]

[…]
