### 2026-03-18 08f7ee4
feat: the solver selection now allows to pass different Diffsol's tablueaus to each solver (and potentially define their own). Added more tests, and error checks
