+++ src/simulator/equation/analytical/two_compartment_models.rs
@@ -259,7 +259,7 @@ mod tests {
             },
         )
         .with_nstates(3)
-        .with_ndrugs(1)
+        .with_ndrugs(2)
         .with_nout(1);
 
         let op_ode = ode
