### 2026-03-18 79a711f
feat: ode! macro that automatically sets the number of states,drugs and outeqs
