+++ pharmsol-macros/src/lib.rs
@@ -1,0 +1,284 @@
+use proc_macro::TokenStream;
+use proc_macro2::TokenTree;
+use quote::{quote, ToTokens};
+use syn::{
+    parse::{Parse, ParseStream},
+    ExprClosure, Ident, Pat, Token,
+};
+
+// ---------------------------------------------------------------------------
+// Macro input parsing
+// ---------------------------------------------------------------------------
+
+struct OdeInput {
+    diffeq: ExprClosure,
+    lag: Option<ExprClosure>,
+    fa: Option<ExprClosure>,
+    init: Option<ExprClosure>,
+    out: ExprClosure,
+}
+
+impl Parse for OdeInput {
+    fn parse(input: ParseStream) -> syn::Result<Self> {
+        let mut diffeq = None;
+        let mut lag = None;
+        let mut fa_val = None;
+        let mut init = None;
+        let mut out = None;
+
+        while !input.is_empty() {
+            let key: Ident = input.parse()?;
+            input.parse::<Token![:]>()?;
+            let closure: ExprClosure = input.parse()?;
+
+            match key.to_string().as_str() {
+                "diffeq" => diffeq = Some(closure),
+                "lag" => lag = Some(closure),
+                "fa" => fa_val = Some(closure),
+                "init" => init = Some(closure),
+                "out" => out = Some(closure),
+                other => {
+                    return Err(syn::Error::new_spanned(
+                        &key,
+                        format!("unknown field `{other}`, expected: diffeq, lag, fa, init, out"),
+                    ));
+                }
+            }
+
+            // optional trailing comma
+            if !input.is_empty() {
+                input.parse::<Token![,]>()?;
+            }
+        }
+
+        Ok(OdeInput {
+            diffeq: diffeq.ok_or_else(|| input.error("missing required field `diffeq`"))?,
+            lag,
+            fa: fa_val,
+            init,
+            out: out.ok_or_else(|| input.error("missing required field `out`"))?,
+        })
+    }
+}
+
+// ---------------------------------------------------------------------------
+// Helpers
+// ---------------------------------------------------------------------------
+
+/// Return the identifier string for a closure parameter (empty for wildcards).
+fn param_name(pat: &Pat) -> String {
+    match pat {
+        Pat::Ident(p) => p.ident.to_string(),
+        _ => String::new(),
+    }
+}
+
+fn closure_param_names(c: &ExprClosure) -> Vec<String> {
+    c.inputs.iter().map(param_name).collect()
+}
+
+/// Recursively scan `tokens` for `ident[literal_int]` patterns where the
+/// ident matches one of `names`.  Returns the maximum literal integer found.
+fn max_literal_index(tokens: proc_macro2::TokenStream, names: &[&str]) -> Option<usize> {
+    let tts: Vec<TokenTree> = tokens.into_iter().collect();
+    let mut best: Option<usize> = None;
+
+    for (i, tt) in tts.iter().enumerate() {
+        match tt {
+            TokenTree::Ident(ident) => {
+                let s = ident.to_string();
+                if names.contains(&s.as_str()) {
+                    if let Some(TokenTree::Group(g)) = tts.get(i + 1) {
+                        if g.delimiter() == proc_macro2::Delimiter::Bracket {
+                            let inner: Vec<TokenTree> = g.stream().into_iter().collect();
+                            if inner.len() == 1 {
+                                if let TokenTree::Literal(lit) = &inner[0] {
+                                    if let Ok(n) = lit.to_string().parse::<usize>() {
[…]
