+++ Cargo.toml
@@ -15,6 +15,7 @@ exa = ["libloading"]
 ahash = "0.8"
 csv = "1.3.0"
 diffsol = "=0.10.4"
+pharmsol-macros = { path = "pharmsol-macros" }
 libloading = { version = "0.9.0", optional = true, features = [] }
 nalgebra = "0.34.1"
 ndarray = { version = "0.17.2", features = ["rayon"] }
