+++ src/simulator/equation/analytical/mod.rs
@@ -1,8 +1,10 @@
+pub mod one_compartment_cl_models;
 pub mod one_compartment_models;
 pub mod three_compartment_models;
 pub mod two_compartment_models;
 
 use diffsol::{NalgebraContext, Vector, VectorHost};
+pub use one_compartment_cl_models::*;
 pub use one_compartment_models::*;
 pub use three_compartment_models::*;
 pub use two_compartment_models::*;
