+++ src/lib.rs
@@ -77,14 +77,16 @@ pub mod prelude {
 
     // Analytical model functions
     pub use crate::simulator::equation::analytical::{
-        one_compartment, one_compartment_with_absorption, three_compartments,
-        three_compartments_with_absorption, two_compartments, two_compartments_with_absorption,
+        one_compartment, one_compartment_cl, one_compartment_cl_with_absorption,
+        one_compartment_with_absorption, three_compartments, three_compartments_with_absorption,
+        two_compartments, two_compartments_with_absorption,
     };
 
     /// Models submodule for organized access to analytical model functions
     pub mod models {
         pub use crate::simulator::equation::analytical::{
-            one_compartment, one_compartment_with_absorption, three_compartments,
+            one_compartment, one_compartment_cl, one_compartment_cl_with_absorption,
+            one_compartment_with_absorption, three_compartments,
             three_compartments_with_absorption, two_compartments, two_compartments_with_absorption,
         };
     }
