+++ src/simulator/equation/analytical/one_compartment_cl_models.rs
@@ -1,0 +1,202 @@
+use crate::{data::Covariates, simulator::*};
+
+/// Analytical solution for one compartment model parameterized by clearance.
+///
+/// # Assumptions
+/// - `p` is a vector of length 2 with CL and V in that order
+/// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
+/// - `x` is a vector of length 1
+/// - covariates are not used
+pub fn one_compartment_cl(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
+    let mut xout = x.clone();
+    let cl = p[0];
+    let v = p[1];
+    let ke = cl / v;
+
+    xout[0] = x[0] * (-ke * t).exp() + rateiv[0] / ke * (1.0 - (-ke * t).exp());
+    xout
+}
+
+/// Analytical solution for one compartment model with first-order absorption,
+/// parameterized by clearance.
+///
+/// # Assumptions
+/// - `p` is a vector of length 3 with ka, CL and V in that order
+/// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
+/// - `x` is a vector of length 2
+/// - covariates are not used
+pub fn one_compartment_cl_with_absorption(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
+    let mut xout = x.clone();
+    let ka = p[0];
+    let cl = p[1];
+    let v = p[2];
+    let ke = cl / v;
+
+    xout[0] = x[0] * (-ka * t).exp();
+
+    xout[1] = x[1] * (-ke * t).exp()
+        + rateiv[0] / ke * (1.0 - (-ke * t).exp())
+        + ((ka * x[0]) / (ka - ke)) * ((-ke * t).exp() - (-ka * t).exp());
+
+    xout
+}
+
+#[cfg(test)]
+mod tests {
+    use super::{one_compartment_cl, one_compartment_cl_with_absorption};
+    use crate::*;
+    use approx::assert_relative_eq;
+
+    enum SubjectInfo {
+        InfusionDosing,
+        OralInfusionDosage,
+    }
+
+    impl SubjectInfo {
+        fn get_subject(&self) -> Subject {
+            match self {
+                SubjectInfo::InfusionDosing => Subject::builder("id1")
+                    .bolus(0.0, 100.0, 0)
+                    .infusion(24.0, 150.0, 0, 3.0)
+                    .missing_observation(0.0, 0)
+                    .missing_observation(1.0, 0)
+                    .missing_observation(2.0, 0)
+                    .missing_observation(4.0, 0)
+                    .missing_observation(8.0, 0)
+                    .missing_observation(12.0, 0)
+                    .missing_observation(24.0, 0)
+                    .missing_observation(25.0, 0)
+                    .missing_observation(26.0, 0)
+                    .missing_observation(27.0, 0)
+                    .missing_observation(28.0, 0)
+                    .missing_observation(32.0, 0)
+                    .missing_observation(36.0, 0)
+                    .build(),
+
+                SubjectInfo::OralInfusionDosage => Subject::builder("id1")
+                    .bolus(0.0, 100.0, 1)
+                    .infusion(24.0, 150.0, 0, 3.0)
+                    .bolus(48.0, 100.0, 0)
+                    .missing_observation(0.0, 0)
+                    .missing_observation(1.0, 0)
+                    .missing_observation(2.0, 0)
+                    .missing_observation(4.0, 0)
+                    .missing_observation(8.0, 0)
+                    .missing_observation(12.0, 0)
+                    .missing_observation(24.0, 0)
+                    .missing_observation(25.0, 0)
[…]
