### 2026-03-18 1c30865
wip: add two comp cl model
