+++ src/simulator/equation/analytical/two_compartment_cl_models.rs
@@ -1,0 +1,286 @@
+use crate::{data::Covariates, simulator::*};
+use diffsol::VectorCommon;
+use nalgebra::{DVector, Matrix2, Vector2};
+
+/// Analytical solution for two compartment model parameterized by clearances.
+///
+/// # Assumptions
+/// - `p` is a vector of length 4 with CL, Q, Vc and Vp in that order
+/// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
+/// - `x` is a vector of length 2
+/// - covariates are not used
+pub fn two_compartments_cl(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
+    let cl = p[0];
+    let q = p[1];
+    let vc = p[2];
+    let vp = p[3];
+
+    let ke = cl / vc;
+    let kcp = q / vc;
+    let kpc = q / vp;
+
+    let sqrt = (ke + kcp + kpc).powi(2) - 4.0 * ke * kpc;
+    if sqrt < 0.0 {
+        panic!("Imaginary solutions, program stopped!");
+    }
+    let sqrt = sqrt.sqrt();
+    let l1 = (ke + kcp + kpc + sqrt) / 2.0;
+    let l2 = (ke + kcp + kpc - sqrt) / 2.0;
+    let exp_l1_t = (-l1 * t).exp();
+    let exp_l2_t = (-l2 * t).exp();
+    let non_zero_matrix = Matrix2::new(
+        (l1 - kpc) * exp_l1_t + (kpc - l2) * exp_l2_t,
+        -kpc * exp_l1_t + kpc * exp_l2_t,
+        -kcp * exp_l1_t + kcp * exp_l2_t,
+        (l1 - ke - kcp) * exp_l1_t + (ke + kcp - l2) * exp_l2_t,
+    );
+
+    let non_zero = (non_zero_matrix * x.inner()) / (l1 - l2);
+
+    let infusion_vector = Vector2::new(
+        ((l1 - kpc) / l1) * (1.0 - exp_l1_t) + ((kpc - l2) / l2) * (1.0 - exp_l2_t),
+        (-kcp / l1) * (1.0 - exp_l1_t) + (kcp / l2) * (1.0 - exp_l2_t),
+    );
+
+    let infusion = infusion_vector * (rateiv[0] / (l1 - l2));
+
+    let result_vector = non_zero + infusion;
+
+    DVector::from_vec(vec![result_vector[0], result_vector[1]]).into()
+}
+
+/// Analytical solution for two compartment model with first-order absorption,
+/// parameterized by clearances.
+///
+/// # Assumptions
+/// - `p` is a vector of length 5 with ka, CL, Q, Vc and Vp in that order
+/// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
+/// - `x` is a vector of length 3
+/// - covariates are not used
+pub fn two_compartments_cl_with_absorption(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
+    let ka = p[0];
+    let cl = p[1];
+    let q = p[2];
+    let vc = p[3];
+    let vp = p[4];
+    let mut xout = x.clone();
+
+    let ke = cl / vc;
+    let kcp = q / vc;
+    let kpc = q / vp;
+
+    let sqrt = (ke + kcp + kpc).powi(2) - 4.0 * ke * kpc;
+    if sqrt < 0.0 {
+        panic!("Imaginary solutions, program stopped!");
+    }
+    let sqrt = sqrt.sqrt();
[…]
