+++ src/lib.rs
@@ -79,7 +79,8 @@ pub mod prelude {
     pub use crate::simulator::equation::analytical::{
         one_compartment, one_compartment_cl, one_compartment_cl_with_absorption,
         one_compartment_with_absorption, three_compartments, three_compartments_with_absorption,
-        two_compartments, two_compartments_with_absorption,
+        two_compartments, two_compartments_cl, two_compartments_cl_with_absorption,
+        two_compartments_with_absorption,
     };
 
     /// Models submodule for organized access to analytical model functions
@@ -87,7 +88,8 @@ pub mod prelude {
         pub use crate::simulator::equation::analytical::{
             one_compartment, one_compartment_cl, one_compartment_cl_with_absorption,
             one_compartment_with_absorption, three_compartments,
-            three_compartments_with_absorption, two_compartments, two_compartments_with_absorption,
+            three_compartments_with_absorption, two_compartments, two_compartments_cl,
+            two_compartments_cl_with_absorption, two_compartments_with_absorption,
         };
     }
 
