### 2026-03-19 4a38f59
wip: Define models at runtime
