+++ src/simulator/equation/ode/runtime.rs
@@ -1,0 +1,1031 @@
+use std::cell::RefCell;
+use std::collections::HashMap;
+use std::fmt;
+use std::sync::Arc;
+
+use diffsol::{
+    error::OdeSolverError, ode_solver::method::OdeSolverMethod, Bdf, NalgebraContext,
+    NewtonNonlinearSolver, NoLineSearch, OdeBuilder, OdeSolverStopReason, Vector, VectorHost,
+};
+use nalgebra::DVector;
+use serde::{Deserialize, Serialize};
+
+use crate::data::error_model::AssayErrorModels;
+use crate::prelude::simulator::SubjectPredictions;
+use crate::simulator::{Fa, Lag, Neqs, M, V};
+use crate::{Covariates, Infusion, Observation, PharmsolError, Subject};
+
+use super::closure::PMProblem;
+use super::{Equation, EquationPriv, EquationTypes};
+
+const RTOL: f64 = 1e-4;
+const ATOL: f64 = 1e-4;
+
+fn no_lag(_p: &V, _t: f64, _cov: &Covariates) -> HashMap<usize, f64> {
+    HashMap::new()
+}
+
+fn no_fa(_p: &V, _t: f64, _cov: &Covariates) -> HashMap<usize, f64> {
+    HashMap::new()
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct RuntimeOdeModel {
+    pub states: Vec<String>,
+    pub parameters: Vec<String>,
+    pub outputs: Vec<String>,
+    pub derivatives: Vec<String>,
+    pub output_equations: Vec<String>,
+    #[serde(default)]
+    pub covariates: Vec<String>,
+    #[serde(default)]
+    pub init: Vec<String>,
+}
+
+impl RuntimeOdeModel {
+    pub fn from_json(json: &str) -> Result<Self, PharmsolError> {
+        serde_json::from_str(json)
+            .map_err(|e| PharmsolError::OtherError(format!("invalid runtime ODE JSON: {e}")))
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct RuntimeODE {
+    lag: Lag,
+    fa: Fa,
+    neqs: Neqs,
+    program: Arc<RuntimeProgram>,
+}
+
+impl RuntimeODE {
+    pub fn from_model(model: RuntimeOdeModel) -> Result<Self, PharmsolError> {
+        let program = RuntimeProgram::compile(model)?;
+        Ok(Self {
+            lag: no_lag,
+            fa: no_fa,
+            neqs: (program.nstates(), program.noutputs()),
+            program: Arc::new(program),
+        })
+    }
+
+    pub fn from_json(json: &str) -> Result<Self, PharmsolError> {
+        let model = RuntimeOdeModel::from_json(json)?;
+        Self::from_model(model)
+    }
+}
+
+impl EquationTypes for RuntimeODE {
+    type S = V;
+    type P = SubjectPredictions;
+}
+
+impl EquationPriv for RuntimeODE {
+    #[inline(always)]
+    fn lag(&self) -> &Lag {
+        &self.lag
+    }
+
+    #[inline(always)]
+    fn fa(&self) -> &Fa {
+        &self.fa
+    }
+
+    #[inline(always)]
+    fn get_nstates(&self) -> usize {
+        self.neqs.0
+    }
+
+    #[inline(always)]
+    fn get_nouteqs(&self) -> usize {
+        self.neqs.1
+    }
+
+    fn solve(
+        &self,
+        state: &mut Self::S,
+        support_point: &Vec<f64>,
+        covariates: &Covariates,
+        infusions: &Vec<Infusion>,
+        start_time: f64,
+        end_time: f64,
+    ) -> Result<(), PharmsolError> {
[…]
