+++ src/lib.rs
@@ -12,7 +12,7 @@ pub use crate::data::*;
 pub use crate::equation::*;
 pub use crate::optimize::effect::get_e2;
 pub use crate::optimize::spp::SppOptimizer;
-pub use crate::simulator::equation::{self, ODE};
+pub use crate::simulator::equation::{self, RuntimeODE, RuntimeOdeModel, ODE};
 pub use error::PharmsolError;
 #[cfg(feature = "exa")]
 pub use exa::*;
