### 2026-03-23 a38ce63
Update example to not use ref
