### 2026-03-23 3c4bb57
feat: Added the analytic solution for CL based two three comparment models. I'm also directly using the rate-constant solutions.
