+++ src/simulator/equation/analytical/three_compartment_cl_models.rs
@@ -1,55 +1,63 @@
 use crate::{data::Covariates, simulator::*};
 use diffsol::{NalgebraContext, Vector};
 
-use super::two_compartment_models::{two_compartments, two_compartments_with_absorption};
+use super::three_compartment_models::{three_compartments, three_compartments_with_absorption};
 
-/// Analytical solution for two compartment model parameterized by clearances.
+/// Analytical solution for three compartment model parameterized by clearances.
 ///
-/// Converts CL/Q/Vc/Vp to ke/kcp/kpc and delegates to [`two_compartments`].
+/// Converts CL/Q2/Q3/Vc/V2/V3 to k10/k12/k13/k21/k31 and delegates to [`three_compartments`].
 ///
 /// # Assumptions
-/// - `p` is a vector of length 4 with CL, Q, Vc and Vp in that order
-/// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
-/// - `x` is a vector of length 2
-/// - covariates are not used
-pub fn two_compartments_cl(x: &V, p: &V, t: T, rateiv: V, cov: &Covariates) -> V {
-    let cl = p[0];
-    let q = p[1];
-    let vc = p[2];
-    let vp = p[3];
-    let ke = cl / vc;
-    let kcp = q / vc;
-    let kpc = q / vp;
-    let p_ke = V::from_vec(vec![ke, kcp, kpc], NalgebraContext);
-    two_compartments(x, &p_ke, t, rateiv, cov)
-}
-
-/// Analytical solution for two compartment model with first-order absorption,
-/// parameterized by clearances.
-///
-/// Converts CL/Q/Vc/Vp to ke/kcp/kpc and delegates to [`two_compartments_with_absorption`].
-///
-/// # Assumptions
-/// - `p` is a vector of length 5 with ka, CL, Q, Vc and Vp in that order
+/// - `p` is a vector of length 6 with CL, Q2, Q3, Vc, V2 and V3 in that order
 /// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
 /// - `x` is a vector of length 3
 /// - covariates are not used
-pub fn two_compartments_cl_with_absorption(x: &V, p: &V, t: T, rateiv: V, cov: &Covariates) -> V {
+pub fn three_compartments_cl(x: &V, p: &V, t: T, rateiv: V, cov: &Covariates) -> V {
+    let cl = p[0];
+    let q2 = p[1];
+    let q3 = p[2];
+    let vc = p[3];
+    let v2 = p[4];
+    let v3 = p[5];
+    let k10 = cl / vc;
+    let k12 = q2 / vc;
+    let k13 = q3 / vc;
+    let k21 = q2 / v2;
+    let k31 = q3 / v3;
+    let p_ke = V::from_vec(vec![k10, k12, k13, k21, k31], NalgebraContext);
+    three_compartments(x, &p_ke, t, rateiv, cov)
+}
+
+/// Analytical solution for three compartment model with first-order absorption,
+/// parameterized by clearances.
+///
+/// Converts CL/Q2/Q3/Vc/V2/V3 to k10/k12/k13/k21/k31 and delegates to [`three_compartments_with_absorption`].
+///
+/// # Assumptions
+/// - `p` is a vector of length 7 with ka, CL, Q2, Q3, Vc, V2 and V3 in that order
+/// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
+/// - `x` is a vector of length 4
+/// - covariates are not used
+pub fn three_compartments_cl_with_absorption(x: &V, p: &V, t: T, rateiv: V, cov: &Covariates) -> V {
     let ka = p[0];
[…]
