### 2026-03-23 7041ac1
Remove extraneous functions in SubjectBuilder
