### 2026-03-23 02813c1
Chore: remove duplicated SubjetInfo code in tests.
