### 2026-03-23 c5f505b
feat: Add CL-based one and two compartment models (#227)
* feat: Added the analytic solution for CL based two three compartment models.

* chore: remove duplicated SubjetInfo code in tests.

---------

Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>
Co-authored-by: Julián D. Otálvaro <juliandavid347@gmail.com>
