+++ src/simulator/equation/analytical/mod.rs
@@ -1,10 +1,16 @@
+pub mod one_compartment_cl_models;
 pub mod one_compartment_models;
+pub mod three_compartment_cl_models;
 pub mod three_compartment_models;
+pub mod two_compartment_cl_models;
 pub mod two_compartment_models;
 
 use diffsol::{NalgebraContext, Vector, VectorHost};
+pub use one_compartment_cl_models::*;
 pub use one_compartment_models::*;
+pub use three_compartment_cl_models::*;
 pub use three_compartment_models::*;
+pub use two_compartment_cl_models::*;
 pub use two_compartment_models::*;
 
 use super::id_hash;
@@ -216,10 +222,63 @@ impl EquationPriv for Analytical {
 }
 
 #[cfg(test)]
-mod tests {
+pub(crate) mod tests {
     use super::*;
     use crate::SubjectBuilderExt;
     use std::collections::HashMap;
+
+    pub(crate) enum SubjectInfo {
+        InfusionDosing,
+        OralInfusionDosage,
+    }
+    impl SubjectInfo {
+        pub(crate) fn get_subject(&self) -> Subject {
+            match self {
+                SubjectInfo::InfusionDosing => Subject::builder("id1")
+                    .bolus(0.0, 100.0, 0)
+                    .infusion(24.0, 150.0, 0, 3.0)
+                    .missing_observation(0.0, 0)
+                    .missing_observation(1.0, 0)
+                    .missing_observation(2.0, 0)
+                    .missing_observation(4.0, 0)
+                    .missing_observation(8.0, 0)
+                    .missing_observation(12.0, 0)
+                    .missing_observation(24.0, 0)
+                    .missing_observation(25.0, 0)
+                    .missing_observation(26.0, 0)
+                    .missing_observation(27.0, 0)
+                    .missing_observation(28.0, 0)
+                    .missing_observation(32.0, 0)
+                    .missing_observation(36.0, 0)
+                    .build(),
+
+                SubjectInfo::OralInfusionDosage => Subject::builder("id1")
+                    .bolus(0.0, 100.0, 1)
+                    .infusion(24.0, 150.0, 0, 3.0)
+                    .bolus(48.0, 100.0, 0)
+                    .missing_observation(0.0, 0)
+                    .missing_observation(1.0, 0)
+                    .missing_observation(2.0, 0)
+                    .missing_observation(4.0, 0)
+                    .missing_observation(8.0, 0)
+                    .missing_observation(12.0, 0)
+                    .missing_observation(24.0, 0)
+                    .missing_observation(25.0, 0)
+                    .missing_observation(26.0, 0)
+                    .missing_observation(27.0, 0)
+                    .missing_observation(28.0, 0)
+                    .missing_observation(32.0, 0)
+                    .missing_observation(36.0, 0)
+                    .missing_observation(48.0, 0)
+                    .missing_observation(49.0, 0)
+                    .missing_observation(50.0, 0)
+                    .missing_observation(52.0, 0)
+                    .missing_observation(56.0, 0)
+                    .missing_observation(60.0, 0)
+                    .build(),
+            }
+        }
+    }
 
     #[test]
     fn secondary_equations_accumulate_within_single_solve() {
