### 2026-03-23 3a959f5
feat: Improved functionality for defining equations (#226)
* feat: new constructor for Equation, adding with_nstates and with_nout methods; there is also a new array size definition ndrugs, this one defines the size of bolus[] and rateiv[] (previously we were using nstates); We are now also exporting the ODE solver, so users can decide which one to use

* feat: the solver selection now allows to pass different Diffsol's tablueaus to each solver (and potentially define their own). Added more tests, and error checks

* feat: ode! macro that automatically sets the number of states,drugs and outeq
[…]
