+++ examples/exa.rs
@@ -42,8 +42,9 @@ fn main() {
             fetch_params!(p, _ke, v);
             y[0] = x[0] / v;
         },
-        (1, 1),
-    );
+    )
+    .with_nstates(1)
+    .with_nout(1);
 
     // =========================================================================
     // 2. Compile and load ODE model dynamically using exa
@@ -66,8 +67,9 @@ fn main() {
                     fetch_params!(p, _ke, v);
                     y[0] = x[0] / v;
                 },
-                (1, 1),
             )
+            .with_nstates(1)
+            .with_nout(1)
         "#
         .to_string(),
         Some(ode_output_path),
@@ -100,8 +102,9 @@ fn main() {
                     fetch_params!(p, _ke, v);
                     y[0] = x[0] / v;
                 },
-                (1, 1),
             )
+            .with_nstates(1)
+            .with_nout(1)
         "#
         .to_string(),
         Some(analytical_output_path),
