+++ pharmsol-macros/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "pharmsol-macros"
+version = "0.1.0"
+edition = "2021"
+description = "Useful macros for pharmsol."
+license = "GPL-3.0"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+proc-macro2 = "1"
+quote = "1"
+syn = { version = "2", features = ["full"] }

[…]
