### 2026-03-24 7e6be81
fix: Use workspace
To allow easier publishing of pharmsol-macros to crates.io
