+++ Cargo.toml
@@ -1,3 +1,6 @@
+[workspace]
+members = [".", "pharmsol-macros"]
+
 [package]
 name = "pharmsol"
 version = "0.24.0"
@@ -15,7 +18,7 @@ exa = ["libloading"]
 ahash = "0.8"
 csv = "1.3.0"
 diffsol = "=0.10.4"
-pharmsol-macros = { path = "pharmsol-macros" }
+pharmsol-macros = { path = "pharmsol-macros", version = "0.1.0" }
 libloading = { version = "0.9.0", optional = true, features = [] }
 nalgebra = "0.34.1"
 ndarray = { version = "0.17.2", features = ["rayon"] }
