### 2026-03-24 47dd3ef
Update resolver
So it doesn't fall back to pre-2021 behaviour
