### 2026-03-24 d894de8
fix: Use workspace (#229)
* fix: Use workspace, and add documentation for pharmsol-macros
