### 2026-03-24 1219c13
fix: Handle missing observations in log-likelihood
Should return zero, i.e. no contribution, instead of returning an error.
