### 2026-03-24 086861e
Chore: cleaning documentation and aswering to some of @mhovd comments
