### 2026-03-24 70fa1f4
Chore: cleaning documentation and aswering to some of @mhovd comments
