+++ src/data/traits.rs
@@ -1,13 +1,5 @@
 //! Extension traits for observation-level pharmacokinetic metrics
 //!
-//! These traits provide convenient access to AUC, Cmax, Tmax, and other
-//! observation-derived metrics on [`Data`], [`Subject`], and [`Occasion`].
-//! These are generic observation-level computations, not NCA-specific —
-//! they belong in the data layer because they operate on raw observed data
-//! and are useful for any downstream analysis (NCA, BestDose, model diagnostics, etc.).
-//!
-//! # Example
-//!
 //! ```rust,ignore
 //! use pharmsol::prelude::*;
 //!
@@ -18,14 +10,16 @@
 //!     .observation(4.0, 4.0, 0)
 //!     .build();
 //!
-//! let auc = subject.auc(0, &AUCMethod::Linear, &BLQRule::Exclude);
-//! let cmax = subject.cmax(0, &BLQRule::Exclude);
-//! let cmax_val = subject.cmax_first(0, &BLQRule::Exclude).unwrap();
+//! let auc = subject.auc(0, &AUCMethod::Linear);
+//! let cmax = subject.cmax(0);
+//!
+//! // With BLQ handling
+//! let auc_blq = subject.auc_blq(0, &AUCMethod::Linear, &BLQRule::Zero);
 //! ```
 
 use crate::data::event::{AUCMethod, BLQRule};
-use crate::data::observation::ObservationProfile;
 use crate::data::observation_error::ObservationError;
+use crate::nca::observation::ObservationProfile;
 use crate::{Data, Occasion, Subject};
 use rayon::prelude::*;
 
@@ -33,60 +27,45 @@ use rayon::prelude::*;
 ///
 /// Wraps [`ObservationError`] with optional context about which subject,
 /// occasion, or output equation failed. This provides better error messages
-/// than bare `ObservationError`.
+/// Wraps [`ObservationError`] with context about which subject or output equation failed.
 #[derive(Debug, Clone, thiserror::Error)]
 pub enum MetricsError {
-    /// An error from observation data processing
     #[error(transparent)]
     Observation(#[from] ObservationError),
 
-    /// Output equation not found in subject data
     #[error("Output equation {outeq} not found in subject{}", subject_id.as_ref().map(|id| format!(" '{}'", id)).unwrap_or_default())]
     OutputEquationNotFound {
-        /// The requested output equation index
         outeq: usize,
-        /// Optional subject identifier for context
         subject_id: Option<String>,
     },
 }
 
-/// Extension trait for observation-level pharmacokinetic metrics
+/// Observation-level pharmacokinetic metrics (AUC, Cmax, Tmax, etc.)
 ///
-/// Provides convenient access to AUC, Cmax, Tmax, etc. without running
-/// full NCA analysis. Each method returns one result per occasion.
-///
-/// For single-occasion convenience, use the `_first()` variants which
-/// return a single `Result` instead of `Vec<Result<...>>`.
-///
-/// # Example
+/// Methods without `_blq` default to [`BLQRule::Exclude`].
+/// The `_first` variants return a single result for the first occasion.
 ///
 /// ```rust,ignore
 /// use pharmsol::prelude::*;
 ///
-/// let subject = Subject::builder("pt1")
-///     .bolus(0.0, 100.0, 0)
-///     .observation(1.0, 10.0, 0)
-///     .observation(2.0, 8.0, 0)
-///     .observation(4.0, 4.0, 0)
-///     .build();
-///
-/// // Per-occasion results
-/// let auc = subject.auc(0, &AUCMethod::Linear, &BLQRule::Exclude);
[…]
