+++ src/nca/traits.rs
@@ -1,17 +1,20 @@
-//! Extension traits for NCA analysis on pharmsol data types
-//!
-//! The [`NCA`] trait adds full non-compartmental analysis to [`Data`], [`Subject`],
-//! and [`Occasion`] without creating a dependency from `data` → `nca`.
-//!
+//! Extension traits for NCA analysis and observation-level pharmacokinetic metrics
 //!
 //! ```rust,ignore
 //! use pharmsol::prelude::*;
 //!
+//! // Full NCA
 //! let result = subject.nca(&NCAOptions::default())?;
+//!
+//! // Quick metrics
+//! let auc = subject.auc(0, &AUCMethod::Linear);
+//! let cmax = subject.cmax(0);
 //! ```
 
 use super::observation::ObservationProfile;
-use crate::nca::analyze::analyze;
+use crate::data::event::{AUCMethod, BLQRule};
+use crate::data::observation_error::ObservationError;
+use crate::nca::analyze::{analyze, AnalysisContext};
 use crate::nca::calc::tlag_from_raw;
 use crate::nca::error::NCAError;
 use crate::nca::types::{NCAOptions, NCAResult, Warning};
@@ -133,16 +136,16 @@ impl Occasion {
         let profile = ObservationProfile::from_occasion(self, options.outeq, &options.blq_rule)?;
         let (times, concs, censoring) = self.get_observations(options.outeq);
         let raw_tlag = tlag_from_raw(&times, &concs, &censoring);
-        analyze(
-            &profile,
-            Some(dose_amount),
+        analyze(&AnalysisContext {
+            profile: &profile,
+            dose_amount: Some(dose_amount),
             route,
             infusion_duration,
             options,
             raw_tlag,
-            None,
-            Some(self.index()),
-        )
+            subject_id: None,
+            occasion: Some(self.index()),
+        })
     }
 }
 
@@ -171,7 +174,7 @@ impl Subject {
     ) -> Result<NCAResult, NCAError> {
         let occasion = self
             .occasions()
-            .into_iter()
+            .iter()
             .next()
             .ok_or(NCAError::InvalidParameter {
                 param: "occasion".to_string(),
@@ -266,16 +269,16 @@ fn nca_occasion(
     let infusion_duration = occasion.infusion_duration();
 
     // Calculate NCA directly on the profile
-    let mut result = analyze(
-        &profile,
+    let mut result = analyze(&AnalysisContext {
+        profile: &profile,
         dose_amount,
         route,
         infusion_duration,
         options,
         raw_tlag,
         subject_id,
-        Some(occasion.index()),
-    )?;
+        occasion: Some(occasion.index()),
+    })?;
 
     // Warn about mixed routes if no explicit override was given
     let routes = occasion.routes();
@@ -287,4 +290,510 @@ fn nca_occasion(
     }
 
     Ok(result)
+}
+
+// ============================================================================
+// Observation-level metrics (AUC, Cmax, Tmax, etc.)
+// ============================================================================
+
+/// Error type for observation metric computations
+///
+/// Wraps [`ObservationError`] with context about which subject or output equation failed.
+#[derive(Debug, Clone, thiserror::Error)]
+pub enum MetricsError {
+    #[error(transparent)]
+    Observation(#[from] ObservationError),
+
[…]
