+++ src/data/traits.rs
@@ -1,536 +1,0 @@
-//! Extension traits for observation-level pharmacokinetic metrics
-//!
-//! ```rust,ignore
-//! use pharmsol::prelude::*;
-//!
-//! let subject = Subject::builder("pt1")
-//!     .bolus(0.0, 100.0, 0)
-//!     .observation(1.0, 10.0, 0)
-//!     .observation(2.0, 8.0, 0)
-//!     .observation(4.0, 4.0, 0)
-//!     .build();
-//!
-//! let auc = subject.auc(0, &AUCMethod::Linear);
-//! let cmax = subject.cmax(0);
-//!
-//! // With BLQ handling
-//! let auc_blq = subject.auc_blq(0, &AUCMethod::Linear, &BLQRule::Zero);
-//! ```
-
-use crate::data::event::{AUCMethod, BLQRule};
-use crate::data::observation_error::ObservationError;
-use crate::nca::observation::ObservationProfile;
-use crate::{Data, Occasion, Subject};
-use rayon::prelude::*;
-
-/// Error type for observation metric computations
-///
-/// Wraps [`ObservationError`] with optional context about which subject,
-/// occasion, or output equation failed. This provides better error messages
-/// Wraps [`ObservationError`] with context about which subject or output equation failed.
-#[derive(Debug, Clone, thiserror::Error)]
-pub enum MetricsError {
-    #[error(transparent)]
-    Observation(#[from] ObservationError),
-
-    #[error("Output equation {outeq} not found in subject{}", subject_id.as_ref().map(|id| format!(" '{}'", id)).unwrap_or_default())]
-    OutputEquationNotFound {
-        outeq: usize,
-        subject_id: Option<String>,
-    },
-}
-
-/// Observation-level pharmacokinetic metrics (AUC, Cmax, Tmax, etc.)
-///
-/// Methods without `_blq` default to [`BLQRule::Exclude`].
-/// The `_first` variants return a single result for the first occasion.
-///
-/// ```rust,ignore
-/// use pharmsol::prelude::*;
-///
-/// let auc = subject.auc(0, &AUCMethod::Linear);
-/// let cmax_val = subject.cmax_first(0).unwrap();
-/// ```
-pub trait ObservationMetrics {
-    // ========================================================================
-    // Required methods — with explicit BLQ rule
-    // ========================================================================
-
-    /// Calculate AUC from time 0 to Tlast with explicit BLQ handling
-    fn auc_blq(
-        &self,
-        outeq: usize,
-        method: &AUCMethod,
-        blq_rule: &BLQRule,
-    ) -> Vec<Result<f64, MetricsError>>;
-
-    /// Calculate partial AUC over a time interval with explicit BLQ handling
-    fn auc_interval_blq(
-        &self,
-        outeq: usize,
-        start: f64,
-        end: f64,
-        method: &AUCMethod,
-        blq_rule: &BLQRule,
-    ) -> Vec<Result<f64, MetricsError>>;
-
-    /// Get Cmax with explicit BLQ handling
-    fn cmax_blq(&self, outeq: usize, blq_rule: &BLQRule) -> Vec<Result<f64, MetricsError>>;
-
-    /// Get Tmax with explicit BLQ handling
-    fn tmax_blq(&self, outeq: usize, blq_rule: &BLQRule) -> Vec<Result<f64, MetricsError>>;
-
-    /// Get Clast with explicit BLQ handling
-    fn clast_blq(&self, outeq: usize, blq_rule: &BLQRule) -> Vec<Result<f64, MetricsError>>;
-
-    /// Get Tlast with explicit BLQ handling
[…]
