+++ src/lib.rs
@@ -62,7 +62,7 @@ pub mod prelude {
     };
 
     // NCA extension traits (provides .nca(), .nca_all(), etc. on data types)
-    pub use crate::data::traits::{MetricsError, ObservationMetrics};
+    pub use crate::nca::{MetricsError, ObservationMetrics};
     pub use crate::nca::NCA;
     pub use crate::nca::{NCAOptions, NCAPopulation, SubjectNCAResult};
 
