+++ src/data/mod.rs
@@ -39,11 +39,10 @@ pub mod parser;
 pub mod residual_error;
 pub mod row;
 pub mod structs;
-pub mod traits;
+pub use crate::nca::{MetricsError, ObservationMetrics};
 pub use covariate::*;
 pub use error_model::*;
 pub use event::*;
 pub use observation_error::ObservationError;
 pub use residual_error::*;
 pub use structs::{Data, Occasion, Subject};
-pub use traits::{MetricsError, ObservationMetrics};
