### 2026-03-24 4979362
chore: remove unused functions
