+++ src/nca/calc.rs
@@ -393,13 +393,13 @@ fn kahan_sum(iter: impl Iterator<Item = f64>) -> f64 {
 
 /// Calculate terminal half-life
 #[inline]
-pub fn half_life(lambda_z: f64) -> f64 {
+pub(crate) fn half_life(lambda_z: f64) -> f64 {
     std::f64::consts::LN_2 / lambda_z
 }
 
 /// Calculate AUC extrapolated to infinity
 #[inline]
-pub fn auc_inf(auc_last: f64, clast: f64, lambda_z: f64) -> f64 {
+pub(crate) fn auc_inf(auc_last: f64, clast: f64, lambda_z: f64) -> f64 {
     if lambda_z <= 0.0 {
         return f64::NAN;
     }
@@ -408,7 +408,7 @@ pub fn auc_inf(auc_last: f64, clast: f64, lambda_z: f64) -> f64 {
 
 /// Calculate percentage of AUC extrapolated
 #[inline]
-pub fn auc_extrap_pct(auc_last: f64, auc_inf: f64) -> f64 {
+pub(crate) fn auc_extrap_pct(auc_last: f64, auc_inf: f64) -> f64 {
     if auc_inf <= 0.0 || !auc_inf.is_finite() {
         return f64::NAN;
     }
@@ -416,7 +416,7 @@ pub fn auc_extrap_pct(auc_last: f64, auc_inf: f64) -> f64 {
 }
 
 /// Calculate AUMC extrapolated to infinity
-pub fn aumc_inf(aumc_last: f64, clast: f64, tlast: f64, lambda_z: f64) -> f64 {
+pub(crate) fn aumc_inf(aumc_last: f64, clast: f64, tlast: f64, lambda_z: f64) -> f64 {
     if lambda_z <= 0.0 {
         return f64::NAN;
     }
@@ -425,7 +425,7 @@ pub fn aumc_inf(aumc_last: f64, clast: f64, tlast: f64, lambda_z: f64) -> f64 {
 
 /// Calculate mean residence time
 #[inline]
-pub fn mrt(aumc_inf: f64, auc_inf: f64) -> f64 {
+pub(crate) fn mrt(aumc_inf: f64, auc_inf: f64) -> f64 {
     if auc_inf <= 0.0 || !auc_inf.is_finite() {
         return f64::NAN;
     }
@@ -434,7 +434,7 @@ pub fn mrt(aumc_inf: f64, auc_inf: f64) -> f64 {
 
 /// Calculate clearance
 #[inline]
-pub fn clearance(dose: f64, auc_inf: f64) -> f64 {
+pub(crate) fn clearance(dose: f64, auc_inf: f64) -> f64 {
     if auc_inf <= 0.0 || !auc_inf.is_finite() {
         return f64::NAN;
     }
@@ -443,7 +443,7 @@ pub fn clearance(dose: f64, auc_inf: f64) -> f64 {
 
 /// Calculate volume of distribution
 #[inline]
-pub fn vz(dose: f64, lambda_z: f64, auc_inf: f64) -> f64 {
+pub(crate) fn vz(dose: f64, lambda_z: f64, auc_inf: f64) -> f64 {
     if lambda_z <= 0.0 || auc_inf <= 0.0 || !auc_inf.is_finite() {
         return f64::NAN;
     }
@@ -540,7 +540,7 @@ fn c0_logslope(profile: &Profile) -> Option<f64> {
 
 /// Calculate Vd for IV bolus
 #[inline]
-pub fn vd_bolus(dose: f64, c0: f64) -> f64 {
+pub(crate) fn vd_bolus(dose: f64, c0: f64) -> f64 {
[…]
