+++ Cargo.toml
@@ -22,7 +22,7 @@ diffsol = "=0.10.4"
 pharmsol-macros = { path = "pharmsol-macros", version = "0.1.0" }
 libloading = { version = "0.9.0", optional = true, features = [] }
 nalgebra = "0.34.1"
-ndarray = { version = "0.16.1", features = ["rayon"] }
+ndarray = { version = "0.17.2", features = ["rayon"] }
 rand = "0.10.0"
 rand_distr = "0.6.0"
 rayon = "1.10.0"
