### 2026-03-24 d94ca90
feat: Non-compartmental analysis (#208)
* feat: Non-compartmental analysis (NCA) (#189)

* nca

* feat: nca

* clenup

* chore: documentation

* chore: cleanup

* chore: cleanup

* chore: deprecating ErrorModel in favor of AssayErrorModel, subdividing the likelihood module and deprecating linear space likelihood calculation functions

* feat: the Data parsing is centraliced to NormalizedRow

* feat: the ErrorModel -> AssayErrorModel

* feat: validation

* chore: cleanup

* chore: cleanup

* chore: cleanup

* chore: fmt

* feat: more validation scenarios

* feat: Json

* chore: wire up the test
[…]
