+++ src/data/auc.rs
@@ -1,0 +1,627 @@
+//! Trapezoidal AUC/AUMC calculation on `&[f64]` slices.
+//!
+//! Building blocks for [`ObservationProfile`](crate::data::observation::ObservationProfile)
+//! and NCA analysis. Supports linear, lin-up/log-down, and lin-log methods.
+//!
+//! ```rust
+//! use pharmsol::data::auc::{auc, auc_interval, aumc, interpolate_linear};
+//! use pharmsol::prelude::AUCMethod;
+//!
+//! let times = [0.0, 1.0, 2.0, 4.0, 8.0];
+//! let concs = [0.0, 10.0, 8.0, 4.0, 2.0];
+//!
+//! let total = auc(&times, &concs, &AUCMethod::Linear).unwrap();
+//! let partial = auc_interval(&times, &concs, 1.0, 4.0, &AUCMethod::Linear).unwrap();
+//! let moment = aumc(&times, &concs, &AUCMethod::Linear).unwrap();
+//! let c_at_3 = interpolate_linear(&times, &concs, 3.0).unwrap();
+//! ```
+
+use crate::data::event::AUCMethod;
+use crate::data::observation_error::ObservationError;
+
+// ============================================================================
+// Segment-level helpers (private)
+// ============================================================================
+
+/// Check if log-linear method should be used for this segment
+#[inline]
+fn use_log_linear(c1: f64, c2: f64) -> bool {
+    c2 < c1 && c1 > 0.0 && c2 > 0.0 && ((c1 / c2) - 1.0).abs() >= 1e-10
+}
+
+/// Linear trapezoidal AUC for a single segment
+#[inline]
+fn auc_linear(c1: f64, c2: f64, dt: f64) -> f64 {
+    (c1 + c2) / 2.0 * dt
+}
+
+/// Log-linear AUC for a single segment (assumes c1 > c2 > 0)
+#[inline]
+fn auc_log(c1: f64, c2: f64, dt: f64) -> f64 {
+    (c1 - c2) * dt / (c1 / c2).ln()
+}
+
+/// Linear trapezoidal AUMC for a single segment
+#[inline]
+fn aumc_linear(t1: f64, c1: f64, t2: f64, c2: f64, dt: f64) -> f64 {
+    (t1 * c1 + t2 * c2) / 2.0 * dt
+}
+
+/// Log-linear AUMC for a single segment (PKNCA formula)
+#[inline]
+fn aumc_log(t1: f64, c1: f64, t2: f64, c2: f64, dt: f64) -> f64 {
+    let k = (c1 / c2).ln() / dt;
+    (t1 * c1 - t2 * c2) / k + (c1 - c2) / (k * k)
+}
+
+// ============================================================================
+// Public segment functions
+// ============================================================================
+
+/// Calculate AUC for a single segment between two time points
+///
+/// For [`AUCMethod::LinLog`], this falls back to linear because segment-level
+/// calculation cannot know Tmax context. Use [`auc`] or
+/// [`auc_segment_with_tmax`] for proper LinLog handling.
+#[inline]
+pub fn auc_segment(
+    t1: f64,
+    c1: f64,
+    t2: f64,
+    c2: f64,
+    method: &AUCMethod,
+) -> Result<f64, ObservationError> {
+    let dt = t2 - t1;
+    if dt <= 0.0 {
+        return Err(ObservationError::InvalidTimeSequence);
+    }
+
+    Ok(match method {
+        AUCMethod::Linear | AUCMethod::LinLog => auc_linear(c1, c2, dt),
+        AUCMethod::LinUpLogDown => {
+            if use_log_linear(c1, c2) {
+                auc_log(c1, c2, dt)
+            } else {
+                auc_linear(c1, c2, dt)
+            }
+        }
+    })
+}
+
[…]
