+++ src/data/mod.rs
@@ -29,16 +29,20 @@
 //!     .build();
 //! ```
 
+pub mod auc;
 pub mod builder;
 pub mod covariate;
 pub mod error_model;
 pub mod event;
+pub mod observation_error;
 pub mod parser;
 pub mod residual_error;
 pub mod row;
 pub mod structs;
+pub use crate::nca::{MetricsError, ObservationMetrics};
 pub use covariate::*;
 pub use error_model::*;
 pub use event::*;
+pub use observation_error::ObservationError;
 pub use residual_error::*;
 pub use structs::{Data, Occasion, Subject};
