### 2026-03-24 bb30fbd
chore(pharmsol): release v0.24.1
