+++ src/json/codegen/closures.rs
@@ -37,6 +37,24 @@ impl<'a> ClosureGenerator<'a> {
         map
     }
 
+    /// Resolve a compartment/state name to its index.
+    /// Accepts either a declared name or a numeric string. Returns an error for unknown names.
+    fn resolve_index(&self, name: &str, context: &str) -> Result<usize, JsonModelError> {
+        if let Some(&idx) = self.compartment_map.get(name) {
+            return Ok(idx);
+        }
+        if let Some(&idx) = self.state_map.get(name) {
+            return Ok(idx);
+        }
+        if let Ok(idx) = name.parse::<usize>() {
+            return Ok(idx);
+        }
+        Err(JsonModelError::CodeGenError(format!(
+            "Unknown compartment or state '{}' in {}",
+            name, context
+        )))
+    }
+
     /// Transpile a math expression to Rust code, resolving named compartment indices
     fn transpile(&self, expr: &str) -> Result<String, JsonModelError> {
         let name_map = self.name_map();
@@ -251,17 +269,18 @@ impl<'a> ClosureGenerator<'a> {
                 s.clone()
             }
             DiffEqSpec::Object(map) => {
-                let mut lines = Vec::new();
+                let mut indexed_lines = Vec::new();
                 for (name, expr) in map {
-                    let idx = self
-                        .compartment_map
-                        .get(name)
-                        .copied()
-                        .unwrap_or_else(|| name.parse::<usize>().unwrap_or(0));
+                    let idx = self.resolve_index(name, "diffeq")?;
                     let rust_expr = self.transpile(expr)?;
-                    lines.push(format!("dx[{}] = {};", idx, rust_expr));
+                    indexed_lines.push((idx, format!("dx[{}] = {};", idx, rust_expr)));
                 }
-                lines.join("\n        ")
+                indexed_lines.sort_by_key(|(idx, _)| *idx);
+                indexed_lines
+                    .into_iter()
+                    .map(|(_, line)| line)
+                    .collect::<Vec<_>>()
+                    .join("\n        ")
             }
         };
 
@@ -298,18 +317,18 @@ impl<'a> ClosureGenerator<'a> {
         let body = match drift {
             DiffEqSpec::String(s) => s.clone(),
             DiffEqSpec::Object(map) => {
-                let mut lines = Vec::new();
+                let mut indexed_lines = Vec::new();
                 for (name, expr) in map {
-                    let idx = self.state_map.get(name).copied().unwrap_or_else(|| {
-                        self.compartment_map
-                            .get(name)
-                            .copied()
-                            .unwrap_or_else(|| name.parse::<usize>().unwrap_or(0))
-                    });
+                    let idx = self.resolve_index(name, "drift")?;
                     let rust_expr = self.transpile(expr)?;
-                    lines.push(format!("dx[{}] = {};", idx, rust_expr));
+                    indexed_lines.push((idx, format!("dx[{}] = {};", idx, rust_expr)));
                 }
-                lines.join("\n        ")
+                indexed_lines.sort_by_key(|(idx, _)| *idx);
+                indexed_lines
+                    .into_iter()
+                    .map(|(_, line)| line)
+                    .collect::<Vec<_>>()
+                    .join("\n        ")
             }
         };
 
@@ -344,21 +363,21 @@ impl<'a> ClosureGenerator<'a> {
         let fetch_params = self.fetch_params();
         let states = self.generate_state_bindings();
 
-        let mut lines = Vec::new();
+        let mut indexed_lines = Vec::new();
         for (name, expr) in diffusion {
[…]
