+++ src/json/validation.rs
@@ -68,7 +68,13 @@ impl Validator {
         // 6. Validate expressions parse correctly
         self.validate_expressions(model)?;
 
-        // 7. Validate analytical function parameters
+        // 7. Validate equation keys match declared compartments/states
+        self.validate_equation_keys(model)?;
+
+        // 8. Validate expression identifiers reference declared names
+        self.validate_expression_identifiers(model)?;
+
+        // 9. Validate analytical function parameters
         if let Some(func) = &model.analytical {
             self.validate_analytical_params(model, func)?;
         }
@@ -219,9 +225,7 @@ impl Validator {
             let mut seen = HashSet::new();
             for cov in covariates {
                 if !seen.insert(cov.clone()) {
-                    return Err(JsonModelError::UndefinedCovariate {
-                        name: format!("duplicate covariate: {}", cov),
-                    });
+                    return Err(JsonModelError::DuplicateCovariate { name: cov.clone() });
                 }
             }
         }
@@ -291,6 +295,173 @@ impl Validator {
                         message: e.to_string(),
                     }
                 })?;
+            }
+        }
+
+        Ok(())
+    }
+
+    /// Validate that diffeq/drift keys match declared compartments/states
+    fn validate_equation_keys(&self, model: &JsonModel) -> Result<(), JsonModelError> {
+        // Validate diffeq keys against compartments
+        if let Some(DiffEqSpec::Object(map)) = &model.diffeq {
+            if let Some(compartments) = &model.compartments {
+                for name in map.keys() {
+                    // Allow numeric indices as well as named compartments
+                    if name.parse::<usize>().is_err() && !compartments.contains(name) {
+                        return Err(JsonModelError::UndefinedCompartment { name: name.clone() });
+                    }
+                }
+            }
+        }
+
+        // Validate drift keys against states
+        if let Some(DiffEqSpec::Object(map)) = &model.drift {
+            if let Some(states) = &model.states {
+                for name in map.keys() {
+                    if name.parse::<usize>().is_err() && !states.contains(name) {
+                        // Also check compartments as fallback
+                        let in_compartments = model
+                            .compartments
+                            .as_ref()
+                            .is_some_and(|c| c.contains(name));
+                        if !in_compartments {
+                            return Err(JsonModelError::UndefinedCompartment {
+                                name: name.clone(),
+                            });
+                        }
+                    }
+                }
+            }
+        }
+
+        // Validate lag keys against compartments
+        if let Some(lag) = &model.lag {
+            if let Some(compartments) = &model.compartments {
+                for name in lag.keys() {
+                    if name.parse::<usize>().is_err() && !compartments.contains(name) {
+                        return Err(JsonModelError::UndefinedCompartment { name: name.clone() });
+                    }
+                }
+            }
+        }
+
+        // Validate fa keys against compartments
+        if let Some(fa) = &model.fa {
+            if let Some(compartments) = &model.compartments {
+                for name in fa.keys() {
+                    if name.parse::<usize>().is_err() && !compartments.contains(name) {
+                        return Err(JsonModelError::UndefinedCompartment { name: name.clone() });
+                    }
+                }
+            }
+        }
+
+        Ok(())
+    }
+
+    /// Validate that expression identifiers reference declared names
+    fn validate_expression_identifiers(&self, model: &JsonModel) -> Result<(), JsonModelError> {
[…]
