+++ tests/test_json.rs
@@ -963,3 +963,301 @@ mod exa_integration {
         std::fs::remove_dir_all(template_path).ok();
     }
 }
+
+// ═══════════════════════════════════════════════════════════════════════════════
+// Library Model Correctness Tests
+// ═══════════════════════════════════════════════════════════════════════════════
+
+mod library_models {
+    use super::*;
+
+    /// All 10 built-in library models should parse and validate without errors
+    #[test]
+    fn test_all_library_models_validate() {
+        let library = ModelLibrary::builtin();
+        let validator = Validator::new();
+
+        let ids = [
+            "pk/1cmt-iv",
+            "pk/1cmt-oral",
+            "pk/2cmt-iv",
+            "pk/2cmt-oral",
+            "pk/3cmt-iv",
+            "pk/3cmt-oral",
+            "pk/1cmt-iv-ode",
+            "pk/1cmt-oral-ode",
+            "pk/2cmt-iv-ode",
+            "pk/2cmt-oral-ode",
+        ];
+
+        for id in &ids {
+            let model = library.get(id).unwrap_or_else(|| panic!("Missing model: {}", id));
+            validator
+                .validate(model)
+                .unwrap_or_else(|e| panic!("Validation failed for {}: {:?}", id, e));
+        }
+    }
+
+    /// All library models should generate code successfully
+    #[test]
+    fn test_all_library_models_generate_code() {
+        let library = ModelLibrary::builtin();
+
+        let ids = [
+            "pk/1cmt-iv",
+            "pk/1cmt-oral",
+            "pk/2cmt-iv",
+            "pk/2cmt-oral",
+            "pk/3cmt-iv",
+            "pk/3cmt-oral",
+            "pk/1cmt-iv-ode",
+            "pk/1cmt-oral-ode",
+            "pk/2cmt-iv-ode",
+            "pk/2cmt-oral-ode",
+        ];
+
+        for id in &ids {
+            let model = library.get(id).unwrap();
+            let gen = CodeGenerator::new(model);
+            gen.generate()
+                .unwrap_or_else(|e| panic!("Code generation failed for {}: {:?}", id, e));
+        }
+    }
+
+    // ── Analytical parameter order tests ─────────────────────────────────
+
+    #[test]
+    fn test_1cmt_iv_parameter_order() {
+        let library = ModelLibrary::builtin();
+        let model = library.get("pk/1cmt-iv").unwrap();
+
+        // one_compartment expects: p[0]=ke
+        let params = model.get_parameters();
+        assert_eq!(params[0], "ke", "p[0] must be ke for one_compartment");
+    }
+
+    #[test]
+    fn test_1cmt_oral_parameter_order() {
+        let library = ModelLibrary::builtin();
+        let model = library.get("pk/1cmt-oral").unwrap();
+
+        // one_compartment_with_absorption expects: p[0]=ka, p[1]=ke
+        let params = model.get_parameters();
+        assert_eq!(params[0], "ka", "p[0] must be ka");
+        assert_eq!(params[1], "ke", "p[1] must be ke");
+    }
+
+    #[test]
+    fn test_2cmt_iv_parameter_order() {
+        let library = ModelLibrary::builtin();
+        let model = library.get("pk/2cmt-iv").unwrap();
+
+        // two_compartments expects: p[0]=ke, p[1]=kcp, p[2]=kpc
+        let params = model.get_parameters();
+        assert_eq!(params[0], "ke");
+        assert_eq!(params[1], "kcp");
+        assert_eq!(params[2], "kpc");
+    }
+
+    #[test]
+    fn test_2cmt_oral_parameter_order() {
+        let library = ModelLibrary::builtin();
+        let model = library.get("pk/2cmt-oral").unwrap();
+
[…]
