### 2026-03-24 8d764c5
automatic calculation of nstates,ndrugs and nouts
