### 2026-03-24 bcfb224
json feature
