### 2026-03-24 2fca843
feat: Json (#199)
* chore: wire up the tests, and update them to the newer API

* feat: Json

* making sure both 0-index and 1-index data are supported

* fix: vector size missmatch

* feat: seconday equations now can use expressions

* chore: fmt

* chore: tests pass

* chore: testing the .json files

* automatic calculation of nstates,ndrugs and nouts

* json feature
