+++ examples/json_exa.rs
@@ -1,0 +1,199 @@
+// cargo run --example json_exa --features exa
+//
+// Compiles a 1-compartment IV model four different ways and checks that they
+// all agree: static Rust, exa (raw Rust string), JSON ODE, JSON analytical.
+
+#[cfg(feature = "exa")]
+fn main() {
+    use pharmsol::prelude::*;
+    use pharmsol::{exa, json, Analytical, ODE};
+    use std::path::PathBuf;
+
+    let subject = Subject::builder("1")
+        .infusion(0.0, 500.0, 0, 0.5)
+        .observation(0.5, 1.645776, 0)
+        .observation(1.0, 1.216442, 0)
+        .observation(2.0, 0.4622729, 0)
+        .observation(3.0, 0.1697458, 0)
+        .observation(4.0, 0.06382178, 0)
+        .observation(6.0, 0.009099384, 0)
+        .observation(8.0, 0.001017932, 0)
+        .build();
+
+    let params = vec![1.2, 50.0]; // ke, V
+
+    let cwd = std::env::current_dir().unwrap();
+    let template_path = std::env::temp_dir().join("exa_json_example");
+
+    // -- Static ODE (plain Rust) ----------------------------------------------
+
+    let static_ode = equation::ODE::new(
+        |x, p, _t, dx, _bolus, rateiv, _cov| {
+            fetch_params!(p, ke, _v);
+            dx[0] = -ke * x[0] + rateiv[0];
+        },
+        |_p, _t, _cov| lag! {},
+        |_p, _t, _cov| fa! {},
+        |_p, _t, _cov, _x| {},
+        |x, p, _t, _cov, y| {
+            fetch_params!(p, _ke, v);
+            y[0] = x[0] / v;
+        },
+    )
+    .with_nstates(1)
+    .with_ndrugs(1)
+    .with_nout(1);
+
+    // -- Exa ODE (raw Rust string, compiled at runtime) -----------------------
+
+    let exa_ode_path = cwd.join("exa_ode_model.pkm");
+    let exa_ode_compiled = exa::build::compile::<ODE>(
+        r#"
+            equation::ODE::new(
+                |x, p, _t, dx, _bolus, rateiv, _cov| {
+                    fetch_params!(p, ke, _V);
+                    dx[0] = -ke * x[0] + rateiv[0];
+                },
+                |_p, _t, _cov| lag! {},
+                |_p, _t, _cov| fa! {},
+                |_p, _t, _cov, _x| {},
+                |x, p, _t, _cov, y| {
+                    fetch_params!(p, _ke, V);
+                    y[0] = x[0] / V;
+                },
+            )
+            .with_nstates(1)
+            .with_ndrugs(1)
+            .with_nout(1)
+        "#
+        .to_string(),
+        Some(exa_ode_path.clone()),
+        vec!["ke".to_string(), "V".to_string()],
+        template_path.clone(),
+        |_, _| {},
+    )
+    .unwrap();
+
+    let exa_ode_path = PathBuf::from(&exa_ode_compiled);
+    let (_lib_exa, (dyn_exa_ode, _)) = unsafe { exa::load::load::<ODE>(exa_ode_path.clone()) };
+
+    // -- JSON ODE -------------------------------------------------------------
+
+    let json_ode_str = r#"{
+        "schema": "1.0",
+        "id": "pk_1cmt_iv_ode",
+        "type": "ode",
+        "parameters": ["ke", "V"],
+        "compartments": ["central"],
+        "diffeq": { "central": "-ke * central + rateiv[0]" },
+        "output": "central / V"
+    }"#;
+
+    // Show the generated Rust before compiling
+    let generated = json::generate_code(json_ode_str).unwrap();
+    println!(
[…]
