### 2026-03-24 87c1b43
chore(pharmsol): release v0.24.1
