+++ src/simulator/equation/analytical/two_compartment_models.rs
@@ -5,36 +5,36 @@ use nalgebra::{DVector, Matrix2, Vector2};
 /// Analytical solution for two compartment model.
 ///
 /// # Assumptions
-/// - `p` is a vector of length 3 with ke, kcp and kpc in that order
+/// - `p` is a vector of length 3 with ke, k12 and k21 in that order
 /// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
 /// - `x` is a vector of length 2
 /// - covariates are not used
 pub fn two_compartments(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
     let ke = p[0];
-    let kcp = p[1];
-    let kpc = p[2];
+    let k12 = p[1];
+    let k21 = p[2];
 
-    let sqrt = (ke + kcp + kpc).powi(2) - 4.0 * ke * kpc;
+    let sqrt = (ke + k12 + k21).powi(2) - 4.0 * ke * k21;
     if sqrt < 0.0 {
         panic!("Imaginary solutions, program stopped!");
     }
     let sqrt = sqrt.sqrt();
-    let l1 = (ke + kcp + kpc + sqrt) / 2.0;
-    let l2 = (ke + kcp + kpc - sqrt) / 2.0;
+    let l1 = (ke + k12 + k21 + sqrt) / 2.0;
+    let l2 = (ke + k12 + k21 - sqrt) / 2.0;
     let exp_l1_t = (-l1 * t).exp();
     let exp_l2_t = (-l2 * t).exp();
     let non_zero_matrix = Matrix2::new(
-        (l1 - kpc) * exp_l1_t + (kpc - l2) * exp_l2_t,
-        -kpc * exp_l1_t + kpc * exp_l2_t,
-        -kcp * exp_l1_t + kcp * exp_l2_t,
-        (l1 - ke - kcp) * exp_l1_t + (ke + kcp - l2) * exp_l2_t,
+        (l1 - k21) * exp_l1_t + (k21 - l2) * exp_l2_t,
+        -k21 * exp_l1_t + k21 * exp_l2_t,
+        -k12 * exp_l1_t + k12 * exp_l2_t,
+        (l1 - ke - k12) * exp_l1_t + (ke + k12 - l2) * exp_l2_t,
     );
 
     let non_zero = (non_zero_matrix * x.inner()) / (l1 - l2);
 
     let infusion_vector = Vector2::new(
-        ((l1 - kpc) / l1) * (1.0 - exp_l1_t) + ((kpc - l2) / l2) * (1.0 - exp_l2_t),
-        (-kcp / l1) * (1.0 - exp_l1_t) + (kcp / l2) * (1.0 - exp_l2_t),
+        ((l1 - k21) / l1) * (1.0 - exp_l1_t) + ((k21 - l2) / l2) * (1.0 - exp_l2_t),
+        (-k12 / l1) * (1.0 - exp_l1_t) + (k12 / l2) * (1.0 - exp_l2_t),
     );
 
     let infusion = infusion_vector * (rateiv[0] / (l1 - l2));
@@ -48,40 +48,40 @@ pub fn two_compartments(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
 /// Analytical solution for two compartment model with first-order absorption.
 ///
 /// # Assumptions
-/// - `p` is a vector of length 4 with ke, ka, kcp and kpc in that order
+/// - `p` is a vector of length 4 with ke, ka, k12 and k21 in that order
 /// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
 /// - `x` is a vector of length 3
 /// - covariates are not used
[…]
