+++ src/simulator/equation/analytical/two_compartment_cl_models.rs
@@ -1,27 +1,54 @@
 use crate::{data::Covariates, simulator::*};
-use diffsol::{NalgebraContext, Vector};
-
-use super::two_compartment_models::{two_compartments, two_compartments_with_absorption};
+use diffsol::VectorCommon;
+use nalgebra::{DVector, Matrix2, Vector2};
 
 /// Analytical solution for two compartment model parameterized by clearances.
 ///
 /// Converts CL/Q/Vc/Vp to ke/kcp/kpc and delegates to [`two_compartments`].
 ///
 /// # Assumptions
-/// - `p` is a vector of length 4 with CL, Q, Vc and Vp in that order
+/// - `p` is a vector of length 4 with CL, Q, V1 and V2 in that order
 /// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
 /// - `x` is a vector of length 2
 /// - covariates are not used
-pub fn two_compartments_cl(x: &V, p: &V, t: T, rateiv: V, cov: &Covariates) -> V {
+pub fn two_compartments_cl(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
     let cl = p[0];
     let q = p[1];
-    let vc = p[2];
-    let vp = p[3];
-    let ke = cl / vc;
-    let kcp = q / vc;
-    let kpc = q / vp;
-    let p_ke = V::from_vec(vec![ke, kcp, kpc], NalgebraContext);
-    two_compartments(x, &p_ke, t, rateiv, cov)
+    let v1 = p[2];
+    let v2 = p[3];
+
+    let ke = cl / v1;
+    let k12 = q / v1;
+    let k21 = q / v2;
+
+    let sqrt = (ke + k12 + k21).powi(2) - 4.0 * ke * k21;
+    if sqrt < 0.0 {
+        panic!("Imaginary solutions, program stopped!");
+    }
+    let sqrt = sqrt.sqrt();
+    let l1 = (ke + k12 + k21 + sqrt) / 2.0;
+    let l2 = (ke + k12 + k21 - sqrt) / 2.0;
+    let exp_l1_t = (-l1 * t).exp();
+    let exp_l2_t = (-l2 * t).exp();
+    let non_zero_matrix = Matrix2::new(
+        (l1 - k21) * exp_l1_t + (k21 - l2) * exp_l2_t,
+        -k21 * exp_l1_t + k21 * exp_l2_t,
+        -k12 * exp_l1_t + k12 * exp_l2_t,
+        (l1 - ke - k12) * exp_l1_t + (ke + k12 - l2) * exp_l2_t,
+    );
+
+    let non_zero = (non_zero_matrix * x.inner()) / (l1 - l2);
+
+    let infusion_vector = Vector2::new(
+        ((l1 - k21) / l1) * (1.0 - exp_l1_t) + ((k21 - l2) / l2) * (1.0 - exp_l2_t),
+        (-k12 / l1) * (1.0 - exp_l1_t) + (k12 / l2) * (1.0 - exp_l2_t),
+    );
+
+    let infusion = infusion_vector * (rateiv[0] / (l1 - l2));
+
+    let result_vector = non_zero + infusion;
+
+    DVector::from_vec(vec![result_vector[0], result_vector[1]]).into()
 }
 
 /// Analytical solution for two compartment model with first-order absorption,
@@ -30,21 +57,66 @@ pub fn two_compartments_cl(x: &V, p: &V, t: T, rateiv: V, cov: &Covariates) -> V
 /// Converts CL/Q/Vc/Vp to ke/kcp/kpc and delegates to [`two_compartments_with_absorption`].
 ///
 /// # Assumptions
-/// - `p` is a vector of length 5 with ka, CL, Q, Vc and Vp in that order
[…]

[…]
