### 2026-03-27 e272e81
Update src/simulator/equation/analytical/three_compartment_cl_models.rs
Co-authored-by: Copilot <175728472+Copilot@users.noreply.github.com>
