### 2026-03-27 cf59e1f
fix: minor copilot findings during PR
