+++ src/simulator/equation/analytical/two_compartment_cl_models.rs
@@ -11,7 +11,7 @@ use nalgebra::{DVector, Matrix2, Vector2};
 /// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
 /// - `x` is a vector of length 2
 /// - covariates are not used
-pub fn two_compartments_cl(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
+pub fn two_compartments_cl(x: &V, p: &V, t: T, rateiv: &V, _cov: &Covariates) -> V {
     let cl = p[0];
     let q = p[1];
     let v1 = p[2];
@@ -61,7 +61,7 @@ pub fn two_compartments_cl(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> 
 /// - `rateiv` is a vector of length 1 with the value of the infusion rate (only one drug)
 /// - `x` is a vector of length 3
 /// - covariates are not used
-pub fn two_compartments_cl_with_absorption(x: &V, p: &V, t: T, rateiv: V, _cov: &Covariates) -> V {
+pub fn two_compartments_cl_with_absorption(x: &V, p: &V, t: T, rateiv: &V, _cov: &Covariates) -> V {
     let ka = p[0];
     let cl = p[1];
     let q = p[2];
@@ -149,8 +149,10 @@ mod tests {
                 fetch_params!(p, _cl, _q, v1, _v2);
                 y[0] = x[0] / v1;
             },
-            (2, 1),
-        );
+        )
+        .with_nstates(2)
+        .with_ndrugs(2)
+        .with_nout(1);
 
         let analytical = equation::Analytical::new(
             two_compartments_cl,
@@ -162,8 +164,10 @@ mod tests {
                 fetch_params!(p, _cl, _q, v1, _v2);
                 y[0] = x[0] / v1;
             },
-            (2, 1),
-        );
+        )
+        .with_nstates(2)
+        .with_ndrugs(2)
+        .with_nout(1);
 
         let op_ode = ode
             .estimate_predictions(&subject, &vec![0.1, 3.0, 1.0, 3.0])
@@ -204,8 +208,10 @@ mod tests {
                 fetch_params!(p, _ka, _cl, _q, v1, _v2);
                 y[0] = x[1] / v1;
             },
-            (3, 1),
-        );
+        )
+        .with_nstates(3)
+        .with_ndrugs(3)
+        .with_nout(1);
 
         let analytical = equation::Analytical::new(
             two_compartments_cl_with_absorption,
@@ -217,8 +223,10 @@ mod tests {
                 fetch_params!(p, _ka, _cl, _q, v1, _v2);
                 y[0] = x[1] / v1;
             },
-            (3, 1),
-        );
+        )
+        .with_nstates(3)
+        .with_ndrugs(3)
+        .with_nout(1);
 
         let op_ode = ode
             .estimate_predictions(&subject, &vec![1.0, 0.1, 3.0, 1.0, 3.0])

[…]
