### 2026-04-08 d47ede1
feat: Use slices instead of vectors when possible
