### 2026-04-08 e5e4f2c
feat: Improved cache performance and API
