+++ src/simulator/cache.rs
@@ -1,199 +1,130 @@
-//! Global cache for predictions
+//! Cache mechanisms for [Equation]s
 //!
-//! This module provides a configurable, concurrent LRU cache for prediction results.
-//! The supports enabling/disabling caching at runtime and adjusting cache size.
-//!
-//! The cache can be cleared between runs, which is useful to avoid stale data and bad hits.
-
+//! This module provides lightweight cache wrappers that can be embedded
+//! directly in equation structs ([`ODE`], [`Analytical`], [`SDE`]).
+//! Each equation instance can optionally own a cache; cloning the equation
+//! produces a shallow clone that shares the same cache data.
 //!
 //! # Example
 //! ```ignore
-//! use pharmsol::simulator::cache::{configure_cache, CacheSettings};
+//! use pharmsol::*;
 //!
-//! // Use a smaller cache
-//! configure_cache(CacheSettings::with_size(10_000));
+//! // No caching (default):
+//! let ode = ODE::new(diffeq, lag, fa, init, out);
 //!
-//! // Disable caching entirely
-//! configure_cache(CacheSettings::disabled());
+//! // Enable caching with default size:
+//! let ode = ODE::new(diffeq, lag, fa, init, out).with_default_cache();
+//!
+//! // Enable caching with custom size:
+//! let ode = ODE::new(diffeq, lag, fa, init, out).with_cache(50_000);
 //! ```
 
-use std::sync::{
-    atomic::{AtomicBool, Ordering},
-    LazyLock, RwLock,
-};
+use std::fmt;
 
 use moka::sync::Cache;
 
-use crate::{simulator::likelihood::SubjectPredictions, PharmsolError};
+use crate::simulator::likelihood::SubjectPredictions;
 
 /// Default maximum number of entries per cache.
 pub const DEFAULT_CACHE_SIZE: u64 = 100_000;
 
-static CACHE_ENABLED: AtomicBool = AtomicBool::new(true);
-
-/// Settings for the prediction cache.
-#[derive(Debug, Clone)]
-pub struct CacheSettings {
-    /// Whether caching is enabled.
-    pub enabled: bool,
-    /// Maximum number of entries per equation type.
-    pub size: u64,
-}
-
-impl Default for CacheSettings {
-    fn default() -> Self {
-        Self {
-            enabled: true,
-            size: DEFAULT_CACHE_SIZE,
-        }
-    }
-}
-
-impl CacheSettings {
-    /// Create settings with caching disabled.
-    pub fn disabled() -> Self {
-        Self {
-            enabled: false,
-            ..Default::default()
-        }
-    }
-
-    /// Create settings with a custom cache size.
-    pub fn with_size(size: u64) -> Self {
-        Self {
-            enabled: true,
-            size,
-        }
-    }
-}
-
-/// Apply new cache settings globally.
-///
-/// This replaces all caches with new instances of the given size
-/// and updates the enabled flag. Any cached entries are discarded.
-pub fn configure_cache(settings: CacheSettings) -> Result<(), PharmsolError> {
-    CACHE_ENABLED.store(settings.enabled, Ordering::Relaxed);
-
-    let size = settings.size;
-
-    // Replace each cache with a fresh instance of the requested size.
-    {
-        let mut c = ana_cache_lock_write()?;
-        *c = Cache::new(size);
-    }
-    {
-        let mut c = ode_cache_lock_write()?;
-        *c = Cache::new(size);
-    }
-    {
[…]
