### 2026-04-08 67e2162
Update structs.rs
