+++ src/simulator/cache.rs
@@ -86,7 +86,9 @@ impl fmt::Debug for PredictionCache {
 /// SDEs do not produce subject predictions that can be cached, but
 /// the likelihood values for a given subject and parameters can still be cached.
 ///
-/// Note that the use of a cache for SDEs
+/// Note that the use of a cache could be counterproductive for SDEs, as this removes the
+/// stochastic nature of the likelihood evaluation. However, it can be useful for
+/// producing a deterministic likelihood for an otherwise stochastic process.
 ///
 /// `Clone` produces a shallow clone that shares the same underlying cache data.
 #[derive(Clone)]
@@ -126,5 +128,146 @@ impl fmt::Debug for SdeLikelihoodCache {
         f.debug_struct("SdeLikelihoodCache")
             .field("entry_count", &self.0.entry_count())
             .finish()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn prediction_cache_miss_returns_none() {
+        let cache = PredictionCache::new(10);
+        assert!(cache.get(&(1, 2)).is_none());
+    }
+
+    #[test]
+    fn prediction_cache_hit_returns_value() {
+        let cache = PredictionCache::new(10);
+        let key: PredictionKey = (42, 99);
+        let preds = SubjectPredictions::default();
+        cache.insert(key, preds.clone());
+        assert!(cache.get(&key).is_some());
+    }
+
+    #[test]
+    fn prediction_cache_entry_count() {
+        let cache = PredictionCache::new(10);
+        assert_eq!(cache.entry_count(), 0);
+        cache.insert((1, 1), SubjectPredictions::default());
+        cache.insert((2, 2), SubjectPredictions::default());
+        // moka may need a short sync before entry_count updates
+        cache.0.run_pending_tasks();
+        assert_eq!(cache.entry_count(), 2);
+    }
+
+    #[test]
+    fn prediction_cache_invalidate_all_clears_entries() {
+        let cache = PredictionCache::new(10);
+        cache.insert((1, 1), SubjectPredictions::default());
+        cache.insert((2, 2), SubjectPredictions::default());
+        cache.0.run_pending_tasks();
+        assert_eq!(cache.entry_count(), 2);
+
+        cache.invalidate_all();
+        cache.0.run_pending_tasks();
+        assert_eq!(cache.entry_count(), 0);
+        assert!(cache.get(&(1, 1)).is_none());
+    }
+
+    #[test]
+    fn prediction_cache_overwrite_same_key() {
+        let cache = PredictionCache::new(10);
+        let key: PredictionKey = (1, 1);
+        cache.insert(key, SubjectPredictions::default());
+        cache.insert(key, SubjectPredictions::default());
+        cache.0.run_pending_tasks();
+        assert_eq!(cache.entry_count(), 1);
+    }
+
+    #[test]
+    fn prediction_cache_clone_shares_data() {
+        let cache = PredictionCache::new(10);
+        cache.insert((1, 1), SubjectPredictions::default());
+        let clone = cache.clone();
+        // Clone sees existing entry
+        assert!(clone.get(&(1, 1)).is_some());
+        // Insert via clone is visible through original
+        clone.insert((2, 2), SubjectPredictions::default());
+        assert!(cache.get(&(2, 2)).is_some());
+    }
+
+    #[test]
+    fn prediction_cache_debug_format() {
+        let cache = PredictionCache::new(10);
[…]
