### 2026-04-08 cdf50bb
feat: Improved cache performance and API (#240)
* feat: Improved cache performance and API

* Also hash covariates

* Update structs.rs

* Add tests
