### 2026-04-08 a8da517
feat: Use slices instead of vectors when possible (#239)
