### 2026-04-09 2e2e1e9
chore: Update diffsol requirement from =0.10.4 to =0.11.0 (#237)
* chore: Update diffsol requirement from =0.10.4 to =0.11.0

Updates the requirements on [diffsol](https://github.com/martinjrobins/diffsol) to permit the latest version.
- [Release notes](https://github.com/martinjrobins/diffsol/releases)
- [Commits](https://github.com/martinjrobins/diffsol/compare/v0.10.4...v0.11.0)

---
updated-dependencies:
- dependency-name: diffsol
  dependency-version: 0.11.0
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>

* fix: Implement missing trait

Reset
[…]
