### 2026-04-09 e8a008c
schema v2: support for multiple outeqs. Robustness, and backwards compatibility
