+++ src/json/library/mod.rs
@@ -24,7 +24,11 @@
 
 use crate::json::errors::JsonModelError;
 use crate::json::model::JsonModel;
-use crate::json::types::{DisplayInfo, Documentation, ModelType};
+use crate::json::types::{
+    CovariateDefinition, DisplayInfo, Documentation, EditorInfo, ExecutableModel, ModelType,
+    NamedEquation,
+};
+use crate::json::Validator;
 use std::collections::HashMap;
 use std::path::Path;
 
@@ -194,7 +198,7 @@ impl ModelLibrary {
                     return true;
                 }
                 // Match by name in display info
-                if let Some(ref display) = model.display {
+                if let Some(display) = model.display_info() {
                     if let Some(ref name) = display.name {
                         if name.to_lowercase().contains(&query_lower) {
                             return true;
@@ -220,7 +224,7 @@ impl ModelLibrary {
         self.models
             .values()
             .filter(|model| {
-                if let Some(ref display) = model.display {
+                if let Some(display) = model.display_info() {
                     if let Some(ref tags) = display.tags {
                         return tags.iter().any(|t| t.to_lowercase() == tag_lower);
                     }
@@ -236,6 +240,13 @@ impl ModelLibrary {
     /// with the derived model's overrides.
     pub fn resolve(&self, model: &JsonModel) -> Result<JsonModel, JsonModelError> {
         self.resolve_with_chain(model, &mut Vec::new())
+    }
+
+    /// Resolve a model's inheritance chain and normalize it into executable form.
+    pub fn resolve_executable(&self, model: &JsonModel) -> Result<ExecutableModel, JsonModelError> {
+        let resolved = self.resolve(model)?;
+        let validated = Validator::new().validate(&resolved)?;
+        validated.executable()
     }
 
     fn resolve_with_chain(
@@ -317,7 +328,6 @@ fn merge_models(base: &JsonModel, derived: &JsonModel) -> JsonModel {
         // ─────────────────────────────────────────────────────────────────────
         // Output
         // ─────────────────────────────────────────────────────────────────────
-        output: derived.output.clone().or_else(|| base.output.clone()),
         outputs: derived.outputs.clone().or_else(|| base.outputs.clone()),
 
         // ─────────────────────────────────────────────────────────────────────
@@ -332,16 +342,13 @@ fn merge_models(base: &JsonModel, derived: &JsonModel) -> JsonModel {
         // ─────────────────────────────────────────────────────────────────────
         // Layer 3: Model Extensions
         // ─────────────────────────────────────────────────────────────────────
-        derived: merge_option_vec(&base.derived, &derived.derived),
         features: merge_option_vec(&base.features, &derived.features),
-        covariates: merge_option_vec_dedup(&base.covariates, &derived.covariates),
+        covariates: merge_covariates(&base.covariates, &derived.covariates),
 
         // ─────────────────────────────────────────────────────────────────────
         // Layer 4: UI Metadata
         // ─────────────────────────────────────────────────────────────────────
-        display: merge_display(&base.display, &derived.display),
-        layout: merge_option_hashmap(&base.layout, &derived.layout),
-        documentation: merge_documentation(&base.documentation, &derived.documentation),
+        editor: merge_editor(&base.editor, &derived.editor),
     }
 }
 
@@ -380,11 +387,34 @@ fn merge_option_vec_dedup(
     }
 }
 
+/// Merge covariates by symbol, preserving order and allowing derived overrides.
+fn merge_covariates(
+    base: &Option<Vec<CovariateDefinition>>,
[…]
