### 2026-04-09 f62f9c8
feat: JSON model representation schema v2 (#241)
* schema v2: support for multiple outeqs. Robustness, and backwards compatibility

* fmt

* renaming function to make it a bit clearer
