+++ tests/test_json.rs
@@ -1,16 +1,10 @@
 #![cfg(feature = "json")]
-//! Integration tests for the JSON model system
-//!
-//! These tests validate the complete pipeline from JSON parsing to code generation.
+//! Integration tests for the JSON model system.
 
 use pharmsol::json::{
-    generate_code, parse_json, validate_json, CodeGenerator, JsonModel, ModelLibrary, ModelType,
-    Validator,
+    generate_code, normalize_json, parse_json, validate_json, CodeGenerator, JsonModel,
+    ModelLibrary, ModelType, Validator,
 };
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// Parsing Tests
-// ═══════════════════════════════════════════════════════════════════════════════
 
 mod parsing {
     use super::*;
@@ -18,103 +12,64 @@ mod parsing {
     #[test]
     fn test_parse_complete_analytical_model() {
         let json = r#"{
-            "schema": "1.0",
-            "id": "pk_2cmt_oral",
+            "schema": "2.0",
+            "id": "pk/2cmt-oral",
             "type": "analytical",
             "version": "1.0.0",
             "analytical": "two_compartments_with_absorption",
             "parameters": ["ke", "ka", "kcp", "kpc", "V"],
-            "output": "x[1] / V",
+            "compartments": ["depot", "central", "peripheral"],
+            "outputs": [
+                { "id": "cp", "equation": "central / V" }
+            ],
             "neqs": [3, 1],
-            "display": {
-                "name": "Two-Compartment Oral",
-                "category": "pk",
-                "tags": ["2-compartment", "oral"]
-            },
-            "documentation": {
-                "summary": "Standard two-compartment oral PK model"
+            "editor": {
+                "display": {
+                    "name": "Two-Compartment Oral",
+                    "category": "pk",
+                    "tags": ["2-compartment", "oral"]
+                },
+                "documentation": {
+                    "summary": "Standard two-compartment oral PK model"
+                }
             }
         }"#;
 
         let model = parse_json(json).expect("Should parse successfully");
-        assert_eq!(model.id, "pk_2cmt_oral");
+        assert_eq!(model.id, "pk/2cmt-oral");
         assert_eq!(model.model_type, ModelType::Analytical);
         assert_eq!(model.parameters.as_ref().unwrap().len(), 5);
+        assert_eq!(model.outputs.as_ref().unwrap()[0].id, "cp");
     }
 
     #[test]
-    fn test_parse_complete_ode_model() {
+    fn test_parse_with_covariates_and_secondary() {
         let json = r#"{
-            "schema": "1.0",
-            "id": "pk_mm_1cmt",
-            "type": "ode",
-            "parameters": ["Vmax", "Km", "V"],
-            "compartments": ["central"],
-            "diffeq": {
-                "central": "-Vmax * (central/V) / (Km + central/V)"
-            },
-            "output": "central / V",
-            "neqs": [1, 1]
-        }"#;
-
-        let model = parse_json(json).expect("Should parse successfully");
-        assert_eq!(model.model_type, ModelType::Ode);
-        assert!(model.diffeq.is_some());
-    }
-
-    #[test]
-    fn test_parse_with_covariates() {
-        let json = r#"{
-            "schema": "1.0",
-            "id": "pk_1cmt_wt",
+            "schema": "2.0",
+            "id": "pk/1cmt-wt",
             "type": "analytical",
             "analytical": "one_compartment",
-            "parameters": ["ke", "V"],
-            "output": "x[0] / V",
-            "covariates": ["WT"],
-            "secondary": {
-                "V": "V * (WT / 70)^0.75"
-            }
+            "parameters": ["CLs", "Vs"],
+            "compartments": ["central"],
[…]

[…]
