### 2026-04-09 d5d1131
chore(pharmsol): release v0.25.0
