### 2026-04-10 4e28131
feat: Update cache API
Updates functions to be more clear and concise in their naming and behaviour
