### 2026-04-10 8bbf090
feat: Cache is now enabled by default
