### 2026-04-10 d3d3ca5
Update trait signature
enable_cache_with_capacity changed to with_cache_capacity
