+++ src/simulator/equation/ode/mod.rs
@@ -129,7 +129,7 @@ impl ODE {
 }
 
 impl super::Cache for ODE {
-    fn enable_cache_with_capacity(mut self, size: u64) -> Self {
+    fn with_cache_capacity(mut self, size: u64) -> Self {
         self.cache = Some(PredictionCache::new(size));
         self
     }
