### 2026-04-10 95f9c1f
feat: Update cache API (#242)
* feat: Update cache API

Updates functions to be more clear and concise in their naming and behaviour

* feat: Cache is now enabled by default

* Update trait signature

enable_cache_with_capacity changed to with_cache_capacity
