+++ src/simulator/equation/analytical/mod.rs
@@ -57,7 +57,7 @@ impl Analytical {
             init,
             out,
             neqs: Neqs::default(),
-            cache: None,
+            cache: Some(PredictionCache::new(DEFAULT_CACHE_SIZE)),
         }
     }
 
@@ -78,27 +78,28 @@ impl Analytical {
         self.neqs.nout = nout;
         self
     }
+}
 
-    /// Enable prediction caching with the given maximum number of entries.
-    ///
-    /// When caching is enabled, predictions for the same (subject, parameters)
-    /// pair are stored and reused. Cloned equations share the same cache.
-    pub fn with_cache(mut self, size: u64) -> Self {
+impl super::Cache for Analytical {
+    fn with_cache_capacity(mut self, size: u64) -> Self {
         self.cache = Some(PredictionCache::new(size));
         self
     }
 
-    /// Enable prediction caching with the default size (100,000 entries).
-    pub fn with_default_cache(mut self) -> Self {
+    fn enable_cache(mut self) -> Self {
         self.cache = Some(PredictionCache::new(DEFAULT_CACHE_SIZE));
         self
     }
 
-    /// Clear all entries from this equation's cache, if caching is enabled.
-    pub fn clear_cache(&self) {
+    fn clear_cache(&self) {
         if let Some(cache) = &self.cache {
             cache.invalidate_all();
         }
+    }
+
+    fn disable_cache(mut self) -> Self {
+        self.cache = None;
+        self
     }
 }
 
