### 2026-04-10 89bf302
chore: Move ODE/analytical comparisons to examples
Instead of running as benchmarks
