### 2026-04-17 5b1e331
Add Pmetrics analytical compatibility wrappers
