+++ src/jit/text.rs
@@ -1,0 +1,210 @@
+//! Plain-text model format for runtime authoring.
+//!
+//! Lets a downstream user (e.g. an R or Python wrapper, or anyone with only
+//! the precompiled `pharmsol` library) define a complete JIT model from a
+//! string and run it without invoking any Rust toolchain.
+//!
+//! # Format
+//!
+//! The format is line-oriented. Blank lines and lines starting with `#` are
+//! ignored. Each non-empty line is one of:
+//!
+//! ```text
+//! name        = my_model
+//! compartments = depot, central
+//! params      = ka, CL, V
+//! covariates  = WT
+//! ndrugs      = 1
+//! dxdt(depot)   = -ka * depot
+//! dxdt(central) =  ka * depot - (CL * pow(WT/70.0, 0.75) / V) * central
+//! out(cp)       =  central / V
+//! ```
+//!
+//! Order of declarations is flexible, but every compartment listed in
+//! `compartments` must have a matching `dxdt(...)` line. `out(...)` lines
+//! define output equations in the order they appear (output `0` is the first
+//! `out` line, etc.).
+//!
+//! Lists are comma-separated; whitespace is ignored. The expression on the
+//! right of `=` follows the syntax documented in [`crate::jit`].
+//!
+//! # Example
+//!
+//! ```ignore
+//! use pharmsol::jit::Model;
+//! let src = r#"
+//!     name = onecmt
+//!     compartments = central
+//!     params = CL, V
+//!     dxdt(central) = rateiv[0] - (CL / V) * central
+//!     out(cp) = central / V
+//! "#;
+//! let ode = Model::from_text(src)?.compile()?;
+//! # Ok::<(), Box<dyn std::error::Error>>(())
+//! ```
+
+use thiserror::Error;
+
+use super::model::Model;
+
+#[derive(Debug, Error)]
+pub enum TextError {
+    #[error("line {line}: expected `key = value` or `dxdt(name) = expr` or `out(name) = expr`")]
+    BadLine { line: usize },
+    #[error("line {line}: unknown key {key:?} (expected name, compartments, params, covariates, ndrugs, dxdt(...), out(...))")]
+    UnknownKey { line: usize, key: String },
+    #[error("line {line}: malformed `{kind}(...)` header: {raw:?}")]
+    BadHeader {
+        line: usize,
+        kind: &'static str,
+        raw: String,
+    },
+    #[error("line {line}: invalid integer for ndrugs: {value:?}")]
+    BadNdrugs { line: usize, value: String },
+    #[error("model has no `compartments` declaration")]
+    NoCompartments,
+}
+
+/// Parse a model from the text format described in [`crate::jit::text`].
+pub(crate) fn parse_text(src: &str) -> Result<Model, TextError> {
+    let mut name = String::from("model");
+    let mut compartments: Option<Vec<String>> = None;
+    let mut params: Vec<String> = Vec::new();
+    let mut covariates: Vec<String> = Vec::new();
+    let mut ndrugs: Option<usize> = None;
+    let mut dxdt: Vec<(String, String)> = Vec::new();
+    let mut outs: Vec<(String, String)> = Vec::new();
+
+    for (lineno, raw_line) in src.lines().enumerate() {
+        let line_num = lineno + 1;
+        let line = strip_comment(raw_line).trim();
+        if line.is_empty() {
+            continue;
+        }
+        let Some(eq_idx) = line.find('=') else {
+            return Err(TextError::BadLine { line: line_num });
+        };
+        let lhs = line[..eq_idx].trim();
+        let rhs = line[eq_idx + 1..].trim();
+
+        if let Some(target) = paren_target(lhs, "dxdt") {
+            dxdt.push((
+                target.ok_or_else(|| TextError::BadHeader {
+                    line: line_num,
[…]
