+++ src/jit/model.rs
@@ -1,0 +1,250 @@
+//! High-level builder DSL for JIT-compiled pharmacometric models.
+//!
+//! Authors a model entirely in Rust — no JSON, no source files, no external
+//! toolchain. After [`Model::compile`], the model is ready to plug into the
+//! existing pharmsol simulator via [`crate::jit::JitOde`].
+//!
+//! # Example
+//!
+//! ```ignore
+//! use pharmsol::jit::Model;
+//! use pharmsol::prelude::*;
+//!
+//! let ode = Model::new("1cmt-iv")
+//!     .compartments(["central"])
+//!     .params(["CL", "V"])
+//!     .covariates(["WT"])
+//!     .dxdt("central", "rateiv[0] - (CL / V) * central")
+//!     .output("cp", "central / V")
+//!     .compile()
+//!     .expect("compile");
+//!
+//! let subject = Subject::builder("p1")
+//!     .infusion(0.0, 100.0, 0, 0.5)
+//!     .observation(1.0, 5.0, 0)
+//!     .covariate("WT", 0.0, 70.0)
+//!     .build();
+//!
+//! let (preds, _) = ode.simulate_subject(&subject, &[10.0, 50.0], None).unwrap();
+//! ```
+
+use std::sync::Arc;
+
+use thiserror::Error;
+
+use super::ast::Expr;
+use super::codegen::compile_artifact;
+use super::equation::JitOde;
+use super::parser::{self, ParseError};
+use super::text::{self, TextError};
+
+/// Errors produced while building or compiling a JIT model.
+#[derive(Error, Debug)]
+pub enum ModelError {
+    #[error("duplicate name {0:?}")]
+    Duplicate(String),
+    #[error("unknown compartment {0:?}")]
+    UnknownCompartment(String),
+    #[error("compartment {0:?} has no dxdt definition")]
+    MissingDxDt(String),
+    #[error("parse error in {context}: {source}")]
+    Parse {
+        context: String,
+        #[source]
+        source: ParseError,
+    },
+    #[error("unresolved identifier {ident:?} in {context}")]
+    UnresolvedIdent { ident: String, context: String },
+    #[error("rateiv index {0} out of range (ndrugs = {1})")]
+    RateIvOutOfRange(usize, usize),
+    #[error("bolus index {0} out of range (ndrugs = {1})")]
+    BolusOutOfRange(usize, usize),
+    #[error("invalid index target {0:?} (only `rateiv` and `bolus` are indexable)")]
+    InvalidIndexTarget(String),
+    #[error("unknown function {name:?} (supported: exp, ln, log, log10, sqrt, abs, pow)")]
+    UnknownFunction { name: String },
+    #[error("function {name:?} expects {expected} args, got {got}")]
+    BadArity {
+        name: String,
+        expected: usize,
+        got: usize,
+    },
+    #[error("codegen error: {0}")]
+    Codegen(String),
+}
+
+/// A symbolic dxdt or output equation, parsed and stored in the model.
+#[derive(Debug, Clone)]
+pub(crate) struct Equation {
+    pub(crate) target: String,
+    pub(crate) expr: Expr,
+}
+
+/// Mutable builder for a JIT model.
+#[derive(Debug, Clone)]
+pub struct Model {
+    pub(crate) name: String,
+    pub(crate) compartments: Vec<String>,
+    pub(crate) params: Vec<String>,
+    pub(crate) covariates: Vec<String>,
+    pub(crate) ndrugs: usize,
+    pub(crate) dxdt: Vec<Equation>,
+    pub(crate) outputs: Vec<Equation>,
+}
+
+impl Model {
+    /// Start a new model with the given identifier (used for diagnostics only).
[…]
