### 2026-04-20 d238ea0
wip: Error handling
