+++ src/jit/model.rs
@@ -30,63 +30,210 @@
 
 use std::sync::Arc;
 
-use thiserror::Error;
-
 use super::ast::Expr;
 use super::codegen::compile_artifact;
 use super::equation::JitOde;
 use super::parser::{self, ParseError};
 use super::text::{self, TextError};
 
+/// All function names recognised by the JIT (used for did-you-mean hints).
+pub(crate) const KNOWN_FUNCTIONS: &[&str] =
+    &["exp", "ln", "log", "log10", "log2", "sqrt", "abs", "pow"];
+
+/// Suggest the closest match from `candidates` to `target` (Levenshtein <= 2).
+pub(crate) fn did_you_mean<'a, I, S>(target: &str, candidates: I) -> Option<String>
+where
+    I: IntoIterator<Item = S>,
+    S: AsRef<str> + 'a,
+{
+    let mut best: Option<(usize, String)> = None;
+    for c in candidates {
+        let s = c.as_ref();
+        if s == target {
+            continue;
+        }
+        let d = levenshtein(target, s);
+        if best.as_ref().map_or(true, |(bd, _)| d < *bd) {
+            best = Some((d, s.to_string()));
+        }
+    }
+    best.and_then(|(d, s)| {
+        let limit = if target.len() <= 3 { 1 } else { 2 };
+        if d <= limit {
+            Some(s)
+        } else {
+            None
+        }
+    })
+}
+
+fn levenshtein(a: &str, b: &str) -> usize {
+    let a: Vec<char> = a.chars().collect();
+    let b: Vec<char> = b.chars().collect();
+    let mut prev: Vec<usize> = (0..=b.len()).collect();
+    let mut curr: Vec<usize> = vec![0; b.len() + 1];
+    for (i, ca) in a.iter().enumerate() {
+        curr[0] = i + 1;
+        for (j, cb) in b.iter().enumerate() {
+            let cost = if ca.eq_ignore_ascii_case(cb) { 0 } else { 1 };
+            curr[j + 1] = (curr[j] + 1).min(prev[j + 1] + 1).min(prev[j] + cost);
+        }
+        std::mem::swap(&mut prev, &mut curr);
+    }
+    prev[b.len()]
+}
+
 /// Errors produced while building or compiling a JIT model.
-#[derive(Error, Debug)]
+#[derive(Debug)]
 pub enum ModelError {
-    #[error("duplicate name {0:?}")]
+    /// Two declarations target the same compartment / output / let name.
     Duplicate(String),
-    #[error("unknown compartment {0:?}")]
+    /// `dxdt(name)` or `init(name)` referenced a compartment not declared in `compartments = ...`.
     UnknownCompartment(String),
-    #[error("compartment {0:?} has no dxdt definition")]
+    /// A compartment was declared but no `dxdt(name) = ...` was provided.
     MissingDxDt(String),
-    #[error("parse error in {context}: {source}")]
+    /// An expression could not be parsed. The original source is included for context.
     Parse {
         context: String,
-        #[source]
+        src: String,
         source: ParseError,
     },
-    #[error("unresolved identifier {ident:?} in {context}")]
-    UnresolvedIdent { ident: String, context: String },
-    #[error("rateiv index {0} out of range (ndrugs = {1})")]
+    /// A name in an expression doesn't match any param, covariate, compartment, let, or `t`.
+    UnresolvedIdent {
+        ident: String,
+        context: String,
+        suggestion: Option<String>,
+        available: Vec<String>,
+    },
+    /// `rateiv[i]` referenced a channel not within `0..ndrugs`.
     RateIvOutOfRange(usize, usize),
-    #[error("bolus index {0} out of range (ndrugs = {1})")]
[…]
