+++ pharmsolr/src/entrypoint.c
@@ -13,9 +13,18 @@ SEXP wrap__simulate_subject(SEXP model, SEXP params, SEXP times, SEXP evids,
                             SEXP amts, SEXP durs, SEXP cmts, SEXP outeqs,
                             SEXP cov_names, SEXP cov_times, SEXP cov_values);
 
+/* Metadata entry points generated by `extendr_module!`. rextendr::document()
+   loads the package and calls `wrap__make_pharmsolr_wrappers` to regenerate
+   the R-side bindings, so it must be registered here even though end users
+   never call it directly. */
+SEXP wrap__make_pharmsolr_wrappers(SEXP use_symbols, SEXP package_name);
+SEXP wrap__get_pharmsolr_metadata(void);
+
 static const R_CallMethodDef CallEntries[] = {
-    {"wrap__compile_model",     (DL_FUNC) &wrap__compile_model,     1},
-    {"wrap__simulate_subject",  (DL_FUNC) &wrap__simulate_subject, 11},
+    {"wrap__compile_model",            (DL_FUNC) &wrap__compile_model,            1},
+    {"wrap__simulate_subject",         (DL_FUNC) &wrap__simulate_subject,        11},
+    {"wrap__make_pharmsolr_wrappers",  (DL_FUNC) &wrap__make_pharmsolr_wrappers,  2},
+    {"wrap__get_pharmsolr_metadata",   (DL_FUNC) &wrap__get_pharmsolr_metadata,   0},
     {NULL, NULL, 0}
 };
 
