+++ pharmsolr/R/extendr-wrappers.R
@@ -1,0 +1,27 @@
+# Generated by extendr: Do not edit by hand
+
+# nolint start
+
+#
+# This file was created with the following call:
+#   .Call("wrap__make_pharmsolr_wrappers", use_symbols = TRUE, package_name = "pharmsolr")
+
+#' @usage NULL
+#' @useDynLib pharmsolr, .registration = TRUE
+NULL
+
+#' Compile a text model definition into a JIT'd ODE model handle.
+#'
+#' Returns an external-pointer SEXP. The pointer is reference-counted via
+#' `Arc<CompiledModel>` so cloning the R object is cheap.
+compile_model <- function(text) .Call(wrap__compile_model, text)
+
+#' Run a single subject through a compiled model.
+#'
+#' All event-shape arguments are parallel vectors of equal length; covariates
+#' are passed as three parallel R objects: a character vector of names, a list
+#' of numeric `time` vectors, and a list of numeric `value` vectors.
+simulate_subject <- function(model, params, times, evids, amts, durs, cmts, outeqs, cov_names, cov_times, cov_values) .Call(wrap__simulate_subject, model, params, times, evids, amts, durs, cmts, outeqs, cov_names, cov_times, cov_values)
+
+
+# nolint end
