+++ pharmsolr/man/simulate_subject.Rd
@@ -1,0 +1,38 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R, R/pharmsolr.R
+\name{simulate_subject}
+\alias{simulate_subject}
+\title{Run a single subject through a compiled model.}
+\usage{
+simulate_subject(model, params, events, covariates = list())
+
+simulate_subject(model, params, events, covariates = list())
+}
+\arguments{
+\item{model}{A model handle returned by [compile_model()].}
+
+\item{params}{Numeric vector of parameter values, in the order declared by
+the `params = ...` line of the model.}
+
+\item{events}{A data.frame with one row per event and columns:
+\describe{
+  \item{time}{numeric — event time}
+  \item{evid}{integer — 0 = observation, 1 = dose (bolus when `dur` is 0, infusion when `dur` > 0)}
+  \item{amt}{numeric — dose amount (ignored for observations)}
+  \item{dur}{numeric — infusion duration; use 0 for a bolus}
+  \item{cmt}{integer — input compartment for bolus/infusion (0-based)}
+  \item{outeq}{integer — output equation index for observation (0-based)}
+}}
+
+\item{covariates}{Optional named list. Each entry is a data.frame with
+columns `time` and `value`, naming the covariate (must match the model's
+`covariates = ...` line).}
+}
+\value{
+A data.frame with columns `time` and `pred`.
+}
+\description{
+All event-shape arguments are parallel vectors of equal length; covariates
+are passed as three parallel R objects: a character vector of names, a list
+of numeric `time` vectors, and a list of numeric `value` vectors.
+}
