+++ pharmsolr/man/compile_model.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/extendr-wrappers.R, R/pharmsolr.R
+\name{compile_model}
+\alias{compile_model}
+\title{Compile a text model definition into a JIT'd ODE model handle.}
+\usage{
+compile_model(text)
+
+compile_model(text)
+}
+\arguments{
+\item{text}{Character scalar containing a pharmsol JIT model definition
+(see the package README for the format).}
+}
+\value{
+An external pointer wrapping the compiled model. Pass to
+  [simulate_subject()].
+}
+\description{
+Returns an external-pointer SEXP. The pointer is reference-counted via
+`Arc<CompiledModel>` so cloning the R object is cheap.
+}
