### 2026-04-20 7f89893
wip: Update definitions
