+++ pharmsolr/R/pharmsolr.R
@@ -10,6 +10,195 @@ compile_model <- function(text) {
   .Call("wrap__compile_model", text, PACKAGE = "pharmsolr")
 }
 
+#' Compartment, output, parameter, and covariate names of a compiled model
+#'
+#' These accessors return the canonical name-to-index map fixed at compile
+#' time. Use them to avoid hard-coding integers for `cmt` and `outeq`.
+#'
+#' `compartments()` index is the value to pass as `cmt` for a dose targeting
+#' that compartment. `outputs()` index is the value to pass as `outeq` for an
+#' observation of that output. `params()` is the order of the `params`
+#' argument to [simulate_subject()].
+#'
+#' @param model A model handle from [compile_model()].
+#' @return Character vector in declaration order.
+#' @name model_metadata
+NULL
+
+#' @rdname model_metadata
+#' @export
+compartments <- function(model) {
+  .Call("wrap__model_compartments", model, PACKAGE = "pharmsolr")
+}
+
+#' @rdname model_metadata
+#' @export
+outputs <- function(model) {
+  .Call("wrap__model_outputs", model, PACKAGE = "pharmsolr")
+}
+
+#' @rdname model_metadata
+#' @export
+params <- function(model) {
+  .Call("wrap__model_params", model, PACKAGE = "pharmsolr")
+}
+
+#' @rdname model_metadata
+#' @export
+covariates <- function(model) {
+  .Call("wrap__model_covariates", model, PACKAGE = "pharmsolr")
+}
+
+#' Look up a compartment's `cmt` index by name (zero-based)
+#'
+#' @param model A model handle from [compile_model()].
+#' @param name Character scalar. Stops with an informative error if not found.
+#' @return Integer scalar (zero-based).
+#' @export
+cmt <- function(model, name) {
+  cmts <- compartments(model)
+  i <- match(name, cmts)
+  if (is.na(i)) {
+    stop(sprintf("unknown compartment %s; declared: [%s]",
+                 dQuote(name), paste(cmts, collapse = ", ")))
+  }
+  as.integer(i - 1L)
+}
+
+#' Look up an output's `outeq` index by name (zero-based)
+#'
+#' @param model A model handle from [compile_model()].
+#' @param name Character scalar. Stops with an informative error if not found.
+#' @return Integer scalar (zero-based).
+#' @export
+outeq <- function(model, name) {
+  outs <- outputs(model)
+  i <- match(name, outs)
+  if (is.na(i)) {
+    stop(sprintf("unknown output %s; declared: [%s]",
+                 dQuote(name), paste(outs, collapse = ", ")))
+  }
+  as.integer(i - 1L)
+}
+
+#' Simulate a single subject through a compiled JIT model
+#'
+#' @param model A model handle returned by [compile_model()].
+#' @param params Numeric vector of parameter values. May be unnamed (uses the
+#'   declared order from the model) or named (will be reordered to match
+#'   [params()]).
+#' @param events A data.frame with one row per event and columns:
+#'   \describe{
+#'     \item{time}{numeric -- event time}
+#'     \item{evid}{integer -- 0 = observation, 1 = dose (bolus when `dur` is 0, infusion when `dur` > 0)}
+#'     \item{amt}{numeric -- dose amount (ignored for observations)}
+#'     \item{dur}{numeric -- infusion duration; use 0 for a bolus}
+#'     \item{cmt}{integer or character -- input compartment for bolus/infusion.
+#'       If character, names are resolved against the compiled model's
+#'       compartments (see [compartments()]). Integer values are zero-based.}
+#'     \item{outeq}{integer or character -- output equation for an observation.
[…]
