### 2026-04-20 4d64ef8
feat: Add Pmetrics analytical compatibility wrappers (#245)
* Add Pmetrics analytical compatibility wrappers

* fmt
