### 2026-04-20 0542be3
chore(pharmsol): release v0.26.1
