+++ Cargo.toml
@@ -4,7 +4,7 @@ resolver = "2"
 
 [package]
 name = "pharmsol"
-version = "0.26.0"
+version = "0.26.1"
 edition = "2021"
 authors = ["Julián D. Otálvaro <juliandavid347@gmail.com>", "Markus Hovd"]
 description = "Rust library for solving analytic and ode-defined pharmacometric models."
