### 2026-04-20 bdf03c4
chore(pharmsol): release v0.26.1 (#247)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
