### 2026-04-20 1bd59e7
chore: Update diffsol requirement from =0.11.0 to =0.12.2
Updates the requirements on [diffsol](https://github.com/martinjrobins/diffsol) to permit the latest version.
- [Release notes](https://github.com/martinjrobins/diffsol/releases)
- [Commits](https://github.com/martinjrobins/diffsol/compare/v0.11.0...v0.12.2)

---
updated-dependencies:
- dependency-name: diffsol
  dependency-version: 0.12.2
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
