+++ Cargo.toml
@@ -19,7 +19,7 @@ exa = ["json", "libloading"]
 [dependencies]
 ahash = "0.8"
 csv = "1.3.0"
-diffsol = "=0.11.0"
+diffsol = "=0.12.2"
 pharmsol-macros = { path = "pharmsol-macros", version = "0.1.0" }
 libloading = { version = "0.9.0", optional = true, features = [] }
 nalgebra = "0.34.1"
