### 2026-04-22 0ef3234
chore(pharmsol): release v0.26.2
