### 2026-04-24 fe7cfa4
refactor: Rename support points to parameters
Simulator is agnostic to NP or P approach, API should reflect that
