+++ src/simulator/equation/mod.rs
@@ -131,7 +131,7 @@ pub(crate) trait EquationPriv: EquationTypes {
     fn solve(
         &self,
         state: &mut Self::S,
-        support_point: &[f64],
+        parameters: &[f64],
         covariates: &Covariates,
         infusions: &[Infusion],
         start_time: f64,
@@ -148,7 +148,7 @@ pub(crate) trait EquationPriv: EquationTypes {
     #[allow(clippy::too_many_arguments)]
     fn process_observation(
         &self,
-        support_point: &[f64],
+        parameters: &[f64],
         observation: &Observation,
         error_models: Option<&AssayErrorModels>,
         time: f64,
@@ -160,7 +160,7 @@ pub(crate) trait EquationPriv: EquationTypes {
 
     fn initial_state(
         &self,
-        support_point: &[f64],
+        parameters: &[f64],
         covariates: &Covariates,
         occasion_index: usize,
     ) -> Self::S;
@@ -168,7 +168,7 @@ pub(crate) trait EquationPriv: EquationTypes {
     #[allow(clippy::too_many_arguments)]
     fn simulate_event(
         &self,
-        support_point: &[f64],
+        parameters: &[f64],
         event: &Event,
         next_event: Option<&Event>,
         error_models: Option<&AssayErrorModels>,
@@ -193,7 +193,7 @@ pub(crate) trait EquationPriv: EquationTypes {
             }
             Event::Observation(observation) => {
                 self.process_observation(
-                    support_point,
+                    parameters,
                     observation,
                     error_models,
                     event.time(),
@@ -208,7 +208,7 @@ pub(crate) trait EquationPriv: EquationTypes {
         if let Some(next_event) = next_event {
             self.solve(
                 x,
-                support_point,
+                parameters,
                 covariates,
                 infusions,
                 event.time(),
@@ -232,7 +232,7 @@ pub(crate) trait EquationPriv: EquationTypes {
 /// is provided for backward compatibility.
 #[allow(private_bounds)]
 pub trait Equation: EquationPriv + 'static + Clone + Sync {
-    /// Estimate the likelihood of the subject given the support point and error model.
+    /// Estimate the likelihood of the subject given the parameter and error model.
     ///
     /// **Deprecated**: Use [`estimate_log_likelihood`](Self::estimate_log_likelihood) instead
     /// for better numerical stability, especially with many observations or extreme parameter values.
@@ -242,7 +242,7 @@ pub trait Equation: EquationPriv + 'static + Clone + Sync {
     ///
     /// # Parameters
     /// - `subject`: The subject data
-    /// - `support_point`: The parameter values
+    /// - `parameters`: The parameter values
     /// - `error_model`: The error model
     ///
     /// # Returns
@@ -254,11 +254,11 @@ pub trait Equation: EquationPriv + 'static + Clone + Sync {
     fn estimate_likelihood(
         &self,
         subject: &Subject,
-        support_point: &[f64],
+        parameters: &[f64],
         error_models: &AssayErrorModels,
     ) -> Result<f64, PharmsolError>;
 
-    /// Estimate the log-likelihood of the subject given the support point and error model.
+    /// Estimate the log-likelihood of the subject given the parameter and error model.
     ///
     /// This function calculates the log of how likely the observed data is given the model
     /// parameters and error model. It is numerically more stable than `estimate_likelihood`
@@ -269,7 +269,7 @@ pub trait Equation: EquationPriv + 'static + Clone + Sync {
     ///
     /// # Parameters
     /// - `subject`: The subject data
-    /// - `support_point`: The parameter values
+    /// - `parameters`: The parameter values
[…]

[…]
