+++ src/dsl/execution.rs
@@ -1,0 +1,1472 @@
+use std::collections::BTreeMap;
+use std::fmt;
+
+use super::{
+    AnalyticalKernel, ConstValue, CovariateInterpolation, Diagnostic, ModelKind, RoutePropertyKind,
+    Span, Symbol, SymbolId, SymbolKind, TypedBinaryOp, TypedCall, TypedExpr, TypedExprKind,
+    TypedModel, TypedModule, TypedRangeExpr, TypedStatePlace, TypedStatementBlock, TypedStmt,
+    TypedStmtKind, TypedUnaryOp, ValueType,
+};
+use super::TypedAssignTargetKind;
+
+pub fn lower_typed_module(module: &TypedModule) -> Result<ExecutionModule, LoweringError> {
+    let mut models = Vec::with_capacity(module.models.len());
+    for model in &module.models {
+        models.push(lower_typed_model(model)?);
+    }
+    Ok(ExecutionModule {
+        models,
+        span: module.span,
+    })
+}
+
+pub fn lower_typed_model(model: &TypedModel) -> Result<ExecutionModel, LoweringError> {
+    ExecutionLowerer::new(model)?.lower()
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ExecutionModule {
+    pub models: Vec<ExecutionModel>,
+    pub span: Span,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ExecutionModel {
+    pub name: String,
+    pub kind: ModelKind,
+    pub metadata: ExecutionMetadata,
+    pub abi: ExecutionAbi,
+    pub kernels: Vec<ExecutionKernel>,
+    pub span: Span,
+}
+
+impl ExecutionModel {
+    pub fn kernel(&self, role: KernelRole) -> Option<&ExecutionKernel> {
+        self.kernels.iter().find(|kernel| kernel.role == role)
+    }
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ExecutionMetadata {
+    pub constants: Vec<ExecutionConstant>,
+    pub parameters: Vec<ExecutionSlot>,
+    pub covariates: Vec<ExecutionCovariate>,
+    pub states: Vec<ExecutionState>,
+    pub routes: Vec<ExecutionRoute>,
+    pub derived: Vec<ExecutionSlot>,
+    pub outputs: Vec<ExecutionSlot>,
+    pub particles: Option<usize>,
+    pub analytical: Option<AnalyticalKernel>,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ExecutionConstant {
+    pub symbol: SymbolId,
+    pub name: String,
+    pub value: ConstValue,
+    pub span: Span,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ExecutionSlot {
+    pub symbol: SymbolId,
+    pub name: String,
+    pub index: usize,
+    pub span: Span,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ExecutionCovariate {
+    pub symbol: SymbolId,
+    pub name: String,
+    pub index: usize,
+    pub interpolation: Option<CovariateInterpolation>,
+    pub span: Span,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ExecutionState {
+    pub symbol: SymbolId,
+    pub name: String,
+    pub offset: usize,
+    pub len: usize,
+    pub span: Span,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct ExecutionRoute {
+    pub symbol: SymbolId,
+    pub name: String,
+    pub index: usize,
[…]
