+++ src/dsl/parser.rs
@@ -4,18 +4,23 @@ use super::diagnostic::{ParseError, Span};
 use super::lexer::{lex, Token, TokenKind};
 
 pub fn parse_module(src: &str) -> Result<Module, ParseError> {
-    let leading = strip_leading_layout(src);
-    if let Some(rest) = leading.strip_prefix("model") {
-        let trimmed = rest.trim_start();
-        if trimmed.starts_with('=') {
-            return authoring::parse_module(src);
+    let parsed = (|| {
+        let leading = strip_leading_layout(src);
+        if let Some(rest) = leading.strip_prefix("model") {
+            let trimmed = rest.trim_start();
+            if trimmed.starts_with('=') {
+                return authoring::parse_module(src);
+            }
+            return Parser::new(src)?.parse_module();
         }
-        return Parser::new(src)?.parse_module();
-    }
-    if leading.is_empty() {
-        return Parser::new(src)?.parse_module();
-    }
-    authoring::parse_module(src)
+        if leading.is_empty() {
+            Parser::new(src)?.parse_module()
+        } else {
+            authoring::parse_module(src)
+        }
+    })();
+
+    parsed.map_err(|error| error.with_source(src))
 }
 
 pub fn parse_model(src: &str) -> Result<Model, ParseError> {
@@ -25,11 +30,13 @@ pub fn parse_model(src: &str) -> Result<Model, ParseError> {
         0 => Err(ParseError::new(
             "expected a `model` declaration",
             Span::empty(src.len()),
-        )),
+        )
+        .with_source(src)),
         _ => Err(ParseError::new(
             "expected exactly one `model` declaration",
             module.models[1].span,
-        )),
+        )
+        .with_source(src)),
     }
 }
 
@@ -54,36 +61,56 @@ fn strip_leading_layout(src: &str) -> &str {
 }
 
 pub(crate) fn parse_expr_fragment(src: &str) -> Result<Expr, ParseError> {
-    let mut parser = Parser::new(src)?;
-    let expr = parser.parse_expr(0)?;
-    if let Some(token) = parser.peek() {
-        return Err(ParseError::new(
-            format!(
-                "expected end of expression, found {}",
-                token.kind.describe()
-            ),
-            token.span,
-        ));
-    }
-    Ok(expr)
+    let parsed = (|| {
+        let mut parser = Parser::new(src)?;
+        let expr = parser.parse_expr(0)?;
+        if let Some(token) = parser.peek() {
+            Err(ParseError::new(
+                format!(
+                    "expected end of expression, found {}",
+                    token.kind.describe()
+                ),
+                token.span,
+            ))
+        } else {
+            Ok(expr)
+        }
+    })();
+
+    parsed.map_err(|error| error.with_source(src))
 }
 
 pub(crate) fn parse_place_fragment(src: &str) -> Result<Place, ParseError> {
-    let mut parser = Parser::new(src)?;
-    let place = parser.parse_place()?;
-    if let Some(token) = parser.peek() {
-        return Err(ParseError::new(
-            format!("expected end of place, found {}", token.kind.describe()),
-            token.span,
-        ));
-    }
-    Ok(place)
+    let parsed = (|| {
+        let mut parser = Parser::new(src)?;
+        let place = parser.parse_place()?;
+        if let Some(token) = parser.peek() {
+            Err(ParseError::new(
+                format!("expected end of place, found {}", token.kind.describe()),
+                token.span,
+            ))
+        } else {
+            Ok(place)
+        }
+    })();
+
+    parsed.map_err(|error| error.with_source(src))
 }
 
 struct Parser {
[…]
