+++ src/dsl/runtime.rs
@@ -1,3 +1,4 @@
+use std::fmt;
 use std::path::Path;
 
 #[cfg(any(
@@ -23,7 +24,10 @@ use super::native::{
 };
 #[cfg(feature = "dsl-wasm")]
 use super::wasm::{export_execution_model_to_wasm, load_wasm_artifact, WasmError};
-use super::{analyze_module, lower_typed_model, parse_module, ExecutionModel, ModelKind};
+use super::{
+    analyze_module, lower_typed_model, parse_module, Diagnostic, DiagnosticReport,
+    ExecutionModel, LoweringError, ModelKind, ParseError, SemanticError,
+};
 use crate::{
     simulator::likelihood::{Prediction, SubjectPredictions},
     PharmsolError, Subject,
@@ -167,14 +171,14 @@ impl CompiledRuntimeModel {
     }
 }
 
-#[derive(Debug, Error)]
+#[derive(Error)]
 pub enum RuntimeError {
     #[error("failed to parse DSL source: {0}")]
-    Parse(String),
+    Parse(#[source] ParseError),
     #[error("failed to analyze DSL source: {0}")]
-    Semantic(String),
+    Semantic(#[source] SemanticError),
     #[error("failed to lower DSL model: {0}")]
-    Lowering(String),
+    Lowering(#[source] LoweringError),
     #[error("{0}")]
     ModelSelection(String),
     #[cfg(feature = "dsl-jit")]
@@ -190,15 +194,57 @@ pub enum RuntimeError {
     Runtime(#[from] PharmsolError),
 }
 
+impl RuntimeError {
+    pub fn diagnostic(&self) -> Option<&Diagnostic> {
+        match self {
+            Self::Parse(error) => Some(error.diagnostic()),
+            Self::Semantic(error) => Some(error.diagnostic()),
+            Self::Lowering(error) => Some(error.diagnostic()),
+            #[cfg(feature = "dsl-jit")]
+            Self::Jit(error) => Some(error.diagnostic()),
+            _ => None,
+        }
+    }
+
+    pub fn render_diagnostic(&self, src: &str) -> Option<String> {
+        self.diagnostic().map(|diagnostic| diagnostic.render(src))
+    }
+
+    pub fn diagnostic_report(&self, source_name: impl Into<String>) -> Option<DiagnosticReport> {
+        let source_name = source_name.into();
+        match self {
+            Self::Parse(error) => Some(error.diagnostic_report(source_name)),
+            Self::Semantic(error) => Some(error.diagnostic_report(source_name)),
+            Self::Lowering(error) => Some(error.diagnostic_report(source_name)),
+            #[cfg(feature = "dsl-jit")]
+            Self::Jit(error) => Some(error.diagnostic_report(source_name)),
+            _ => None,
+        }
+    }
+}
+
+impl fmt::Debug for RuntimeError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Parse(error) => fmt::Display::fmt(error, f),
+            Self::Semantic(error) => fmt::Display::fmt(error, f),
+            Self::Lowering(error) => fmt::Display::fmt(error, f),
+            #[cfg(feature = "dsl-jit")]
+            Self::Jit(error) => fmt::Display::fmt(error, f),
+            _ => fmt::Display::fmt(self, f),
+        }
+    }
+}
+
 pub fn compile_module_source_to_runtime(
     source: &str,
     model_name: Option<&str>,
     target: RuntimeCompilationTarget,
     event_callback: impl Fn(String, String) + Send + Sync + 'static,
 ) -> Result<CompiledRuntimeModel, RuntimeError> {
[…]
