### 2026-04-27 affac01
chore: fix all warnings, clippy, fmt
