### 2026-04-28 07d81b3
chore: extract the frontend dsl into an internal crate
