+++ src/dsl/model_info.rs
@@ -1,8 +1,8 @@
 use serde::{Deserialize, Serialize};
 
 use pharmsol_dsl::execution::{
-    ExecutionExpr, ExecutionExprKind, ExecutionLoad, ExecutionStmt, ExecutionStmtKind,
-    ExecutionModel, KernelImplementation, KernelRole,
+    ExecutionExpr, ExecutionExprKind, ExecutionLoad, ExecutionModel, ExecutionStmt,
+    ExecutionStmtKind, KernelImplementation, KernelRole,
 };
 use pharmsol_dsl::{AnalyticalKernel, ModelKind};
 
