+++ pharmsol-dsl/Cargo.toml
@@ -12,5 +12,5 @@ publish = false
 bench = false
 
 [dependencies]
-serde = { version = "1.0.201", features = ["derive"] }
-serde_json = "1.0.117"
+serde = { version = "1.0.228", features = ["derive"] }
+serde_json = "1.0.149"
