+++ Cargo.toml
@@ -25,39 +25,40 @@ wasm = ["dsl-wasm"]
 pharmsol-dsl = { path = "pharmsol-dsl", version = "0.1.0" }
 pharmsol-macros = { path = "pharmsol-macros", version = "0.1.0" }
 libloading = { version = "0.9.0", optional = true, features = [] }
-cranelift = { version = "0.115", optional = true }
-cranelift-jit = { version = "0.115", optional = true }
-cranelift-module = { version = "0.115", optional = true }
-cranelift-native = { version = "0.115", optional = true }
-serde = { version = "1.0.201", features = ["derive"] }
-serde_json = "1.0.117"
-thiserror = "2.0.11"
-tracing = "0.1.41"
-wasmtime = { version = "28.0.1", optional = true }
-wasm-encoder = { version = "0.246.2", optional = true }
+cranelift = { version = "0.131.0", optional = true }
+cranelift-jit = { version = "0.131.0", optional = true }
+cranelift-module = { version = "0.131.0", optional = true }
+cranelift-native = { version = "0.131.0", optional = true }
+serde = { version = "1.0.228", features = ["derive"] }
+serde_json = "1.0.149"
+thiserror = "2.0.18"
+tracing = "0.1.44"
+wasmtime = { version = "44.0.0", optional = true }
+wasm-encoder = { version = "0.247.0", optional = true }
 
 [target.'cfg(not(all(target_arch = "wasm32", target_os = "unknown")))'.dependencies]
-ahash = "0.8"
-csv = "1.3.0"
-diffsol = "=0.11.0"
-nalgebra = "0.34.1"
+ahash = "0.8.12"
+csv = "1.4.0"
+diffsol = "=0.12.4"
+nalgebra = "0.34.2"
 ndarray = { version = "0.17.2", features = ["rayon"] }
-rand = "0.10.0"
+rand = "0.10.1"
 rand_distr = "0.6.0"
-rayon = "1.10.0"
+rayon = "1.12.0"
 statrs = "0.18.0"
 argmin = "0.11.0"
 argmin-math = "0.5.1"
-moka = { version = "0.12.14", features = ["sync"] }
+moka = { version = "0.12.15", features = ["sync"] }
 
 [dev-dependencies]
-criterion = { version = "0.8.0", features = ["html_reports"] }
+criterion = { version = "0.8.2", features = ["html_reports"] }
 approx = "0.5.1"
-tempfile = "3.10.1"
+tempfile = "3.27.0"
 tiny_http = "0.12.0"
-webbrowser = "1.0.6"
+webbrowser = "1.2.1"
 
 [target.'cfg(all(target_arch = "wasm32", target_os = "unknown"))'.dependencies]
+# Keep these shims on the transitive major lines we still need to feature-activate for browser builds.
 getrandom = { version = "0.2.17", features = ["js"] }
 getrandom_03 = { package = "getrandom", version = "0.3.4", features = ["wasm_js"] }
 
