+++ tests/browser-e2e/compiler-bridge/Cargo.toml
@@ -11,4 +11,4 @@ crate-type = ["cdylib"]
 [dependencies]
 base64 = "0.22.1"
 pharmsol = { path = "../../..", default-features = false, features = ["dsl-wasm-compile"] }
-serde_json = "1.0.117"
+serde_json = "1.0.149"
