### 2026-04-28 4b30a7f
feat: dsl now makes correct recommendations between lowercase and uppercase
