### 2026-04-28 61e151b
chore: remove proposal breadcumbs
