### 2026-04-28 9638b22
chore: extract pharmsol-examples and pharmsolr
