+++ examples/dsl_browser_playground.rs
@@ -1,588 +1,0 @@
-//! Run with:
-//! cargo run --example dsl_browser_playground --features dsl-wasm
-
-#[cfg(feature = "dsl-wasm")]
-mod playground {
-    use std::collections::{hash_map::DefaultHasher, HashMap};
-    use std::fs;
-    use std::hash::{Hash, Hasher};
-    use std::path::PathBuf;
-    use std::process::Command;
-
-    use pharmsol::dsl::{self, CompiledRuntimeModel, RuntimePredictions};
-    use pharmsol::{Subject, SubjectBuilderExt};
-    use serde::de::DeserializeOwned;
-    use serde::{Deserialize, Serialize};
-    use serde_json::{json, Value};
-    use tiny_http::{Header, Method, Request, Response, Server, StatusCode};
-
-    const INDEX_HTML: &str = include_str!("dsl_browser_playground/index.html");
-    const APP_JS: &str = include_str!("dsl_browser_playground/app.js");
-    const COMPILE_WORKER_JS: &str = include_str!("dsl_browser_playground/compile-worker.js");
-    const STYLES_CSS: &str = include_str!("dsl_browser_playground/styles.css");
-
-    const DEFAULT_MODEL_NAME: &str = "bimodal_ke";
-    const DEFAULT_MODEL_SOURCE: &str = r#"
-model = bimodal_ke
-kind = ode
-
-params = ke, v
-states = central
-outputs = cp
-
-infusion(iv) -> central
-
-dx(central) = -ke * central
-
-out(cp) = central / v ~ continuous()
-"#;
-
-    const COMPILER_BRIDGE_MANIFEST: &str = "tests/browser-e2e/compiler-bridge/Cargo.toml";
-    const COMPILER_BRIDGE_TARGET_DIR: &str = "target/dsl-browser-playground-compile-bridge";
-    const COMPILER_BRIDGE_WASM_RELATIVE: &str =
-        "wasm32-unknown-unknown/release/browser_compile_bridge.wasm";
-
-    struct AppState {
-        compiler_bridge_wasm: Vec<u8>,
-        compiled_cache: HashMap<u64, CompiledRuntimeModel>,
-    }
-
-    impl AppState {
-        fn new(compiler_bridge_wasm: Vec<u8>) -> Self {
-            Self {
-                compiler_bridge_wasm,
-                compiled_cache: HashMap::new(),
-            }
-        }
-
-        fn compiled_from_bytes(
-            &mut self,
-            wasm_bytes: &[u8],
-        ) -> Result<&CompiledRuntimeModel, String> {
-            let cache_key = artifact_hash(wasm_bytes);
-            if !self.compiled_cache.contains_key(&cache_key) {
-                let model =
-                    dsl::load_runtime_wasm_bytes(wasm_bytes).map_err(|error| error.to_string())?;
-                self.compiled_cache.clear();
-                self.compiled_cache.insert(cache_key, model);
-            }
-
-            self.compiled_cache.get(&cache_key).ok_or_else(|| {
-                "The compiled artifact could not be cached for simulation.".to_string()
-            })
-        }
-    }
-
-    #[derive(Clone, Debug, Serialize, Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    struct DoseDraft {
-        time: f64,
-        amount: f64,
-        route: String,
-    }
-
-    #[derive(Clone, Debug, Serialize, Deserialize)]
-    #[serde(rename_all = "camelCase")]
-    struct InfusionDraft {
-        time: f64,
-        amount: f64,
-        route: String,
-        duration: f64,
-    }
-
-    #[derive(Clone, Debug, Serialize, Deserialize)]
[…]
