+++ examples/dsl_runtime_jit.rs
@@ -2,83 +2,26 @@
 //! cargo run --example dsl_runtime_jit --features dsl-jit
 
 #[cfg(feature = "dsl-jit")]
-const ODE_SOURCE: &str = r#"
-model = example_ode
-kind = ode
-
-params = ka, cl, v, tlag, f_oral
-covariates = wt@linear
-states = depot, central
-derived = cl_i, ke
-outputs = cp
-
-bolus(oral) -> depot
-infusion(iv) -> central
-
-lag(oral) = tlag
-fa(oral) = f_oral
-
-cl_i = cl * pow(wt / 70.0, 0.75)
-ke = cl_i / v
-
-dx(depot) = -ka * depot
-dx(central) = ka * depot - ke * central
-
-out(cp) = central / v ~ continuous()
-"#;
-
-#[cfg(feature = "dsl-jit")]
-const ANALYTICAL_SOURCE: &str = r#"
-model = example_analytical
-kind = analytical
-
-params = ka, ke, v, tlag, f_oral
-states = depot, central
-outputs = cp
-
-bolus(oral) -> depot
-
-lag(oral) = tlag
-fa(oral) = f_oral
-
-kernel = one_compartment_with_absorption
-
-out(cp) = central / v ~ continuous()
-"#;
-
-#[cfg(feature = "dsl-jit")]
-const SDE_SOURCE: &str = r#"
-model = example_sde
-kind = sde
-
-params = ka, ke0, kcp, kpc, vol, ske
-covariates = wt@locf
-states = depot, central, peripheral, ke_latent
-particles = 16
-outputs = cp
-
-bolus(oral) -> depot
-
-init(ke_latent) = ke0
-
-dx(depot) = -ka * depot
-dx(central) = ka * depot - (ke_latent + kcp) * central + kpc * peripheral
-dx(peripheral) = kcp * central - kpc * peripheral
-dx(ke_latent) = -ke_latent + ke0
-
-noise(ke_latent) = ske
-
-out(cp) = central / (vol * wt) ~ continuous()
-"#;
-
-#[cfg(feature = "dsl-jit")]
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     use std::io;
 
-    use pharmsol::dsl::{self, RuntimeCompilationTarget};
-    use pharmsol::{Subject, SubjectBuilderExt};
+    use pharmsol::prelude::*;
 
-    println!("Sugared DSL models compiled with runtime JIT");
+    let model_source = r#"
+model = bimodal_ke
+kind = ode
+
+params = ke, v
+states = central
+outputs = cp
+
+infusion(iv) -> central
+
+dx(central) = -ke * central
+
+out(cp) = central / v
+"#;
+    let support_point = [1.2, 50.0];
     let show_compile_logs = false;
     let on_compile_event = move |kind: String, message: String| {
         if !show_compile_logs || message.is_empty() {
@@ -92,129 +35,56 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         }
     };
 
-    // 1. Define an ODE model, compile it, simulate one subject, and print predictions.
-    let ode_support_point = [1.2, 5.0, 40.0, 0.25, 0.8];
-    let ode_model = dsl::compile_module_source_to_runtime(
-        ODE_SOURCE,
-        Some("example_ode"),
-        RuntimeCompilationTarget::Jit,
+    // 1. Compile the model to the runtime JIT backend.
+    let model = pharmsol::dsl::compile_module_source_to_runtime(
+        model_source,
+        Some("bimodal_ke"),
+        pharmsol::dsl::RuntimeCompilationTarget::Jit,
         on_compile_event,
     )?;
-    let ode_oral = ode_model
-        .route_index("oral")
[…]
