### 2026-04-28 2ef1056
chore: remove getrandom
