+++ Cargo.toml
@@ -33,7 +33,6 @@ serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.149"
 thiserror = "2.0.18"
 tracing = "0.1.44"
-wasmtime = { version = "44.0.0", optional = true }
 wasm-encoder = { version = "0.247.0", optional = true }
 
 [target.'cfg(not(all(target_arch = "wasm32", target_os = "unknown")))'.dependencies]
@@ -49,6 +48,7 @@ statrs = "0.18.0"
 argmin = "0.11.0"
 argmin-math = "0.5.1"
 moka = { version = "0.12.15", features = ["sync"] }
+wasmtime = { version = "44.0.0", optional = true }
 
 [dev-dependencies]
 criterion = { version = "0.8.2", features = ["html_reports"] }
@@ -56,11 +56,6 @@ approx = "0.5.1"
 tempfile = "3.27.0"
 tiny_http = "0.12.0"
 webbrowser = "1.2.1"
-
-[target.'cfg(all(target_arch = "wasm32", target_os = "unknown"))'.dependencies]
-# Keep these shims on the transitive major lines we still need to feature-activate for browser builds.
-getrandom = { version = "0.2.17", features = ["js"] }
-getrandom_03 = { package = "getrandom", version = "0.3.4", features = ["wasm_js"] }
 
 [lib]
 bench = false
