+++ .github/workflows/build.yml
@@ -89,6 +89,10 @@ jobs:
       - name: Run doc tests
         run: cargo test --doc --all-features
 
+      - name: Run wasm32 dsl-wasm clippy
+        if: ${{ matrix.os == 'ubuntu-latest' }}
+        run: cargo clippy --lib --target wasm32-unknown-unknown --features dsl-wasm -- -D warnings
+
       - name: Run browser E2E
         if: ${{ matrix.os == 'ubuntu-latest' }}
         working-directory: tests/browser-e2e
