### 2026-04-28 2b337e8
feat: Domain Specific Language (DSL) using JIT or AOT (#252)
* wip: Cranelift

* wip: Support Fa and Lag

* wip: Fix EVID

* wip: Error handling

* wip: Bump extendr

* wip: Update definitions

* wip: it seems to work :/

* wip: better

* error report

* DSL-to-WASM

* chore: fix all warnings, clippy, fmt

* chore: extract the frontend dsl into an internal crate

* chore: fmt

* chore: dependencies upgraded to the latest version

* feat: dsl now makes correct recommendations between lowercase and uppercase

* chore: remove proposal breadcumbs

* chore: extract pharmsol-examples and pharmsolr


[…]
