+++ src/json/codegen/mod.rs
@@ -1,341 +1,0 @@
-//! Code generation from JSON models to Rust code
-//!
-//! This module transforms validated JSON models into Rust code strings
-//! that can be compiled by the `exa` module.
-
-mod closures;
-mod ode;
-mod sde;
-
-use crate::json::errors::JsonModelError;
-use crate::json::model::JsonModel;
-use crate::json::types::*;
-use crate::simulator::equation::EqnKind;
-
-pub use closures::ClosureGenerator;
-
-/// Generated Rust code ready for compilation
-#[derive(Debug, Clone)]
-pub struct GeneratedCode {
-    /// The complete equation constructor code
-    pub equation_code: String,
-
-    /// Parameter names in fetch order
-    pub parameters: Vec<String>,
-
-    /// The equation kind (ODE, Analytical, SDE)
-    pub kind: EqnKind,
-}
-
-/// Inferred equation dimensions from generated code, mirroring the
-/// `ode!` proc-macro approach: scan for max literal bracket-indices of
-/// state (`x`, `dx`), drug-input (`rateiv`, `_b`, `b`, `bolus`), and
-/// output (`y`) vectors.
-struct InferredNeqs {
-    nstates: usize,
-    ndrugs: usize,
-    nout: usize,
-}
-
-/// Scan `code` for `name[literal_int]` patterns and return the maximum
-/// literal integer found, or `None` if the name never appears with an index.
-fn max_index_of(code: &str, names: &[&str]) -> Option<usize> {
-    let mut max: Option<usize> = None;
-    for name in names {
-        let needle = format!("{name}[");
-        let mut start = 0;
-        while let Some(pos) = code[start..].find(&needle) {
-            let idx_start = start + pos + needle.len();
-            if let Some(end) = code[idx_start..].find(']') {
-                if let Ok(n) = code[idx_start..idx_start + end].parse::<usize>() {
-                    max = Some(max.map_or(n, |m: usize| m.max(n)));
-                }
-            }
-            start = idx_start;
-        }
-    }
-    max
-}
-
-/// Infer nstates, ndrugs, nout from the generated closure code.
-/// Falls back to 1 for each dimension when no indices are found.
-fn infer_neqs(closures: &[&str]) -> InferredNeqs {
-    let all: String = closures.join("\n");
-
-    let nstates = max_index_of(&all, &["x", "dx"]).map_or(1, |n| n + 1);
-    let ndrugs = max_index_of(&all, &["rateiv", "_b", "b", "bolus"]).map_or(1, |n| n + 1);
-    let nout = max_index_of(&all, &["y"]).map_or(1, |n| n + 1);
-
-    InferredNeqs {
-        nstates,
-        ndrugs,
-        nout,
-    }
-}
-
-/// Code generator for JSON models
-pub struct CodeGenerator<'a> {
-    model: &'a JsonModel,
-    closure_gen: ClosureGenerator<'a>,
-}
-
-impl<'a> CodeGenerator<'a> {
-    /// Create a new code generator for a model
-    pub fn new(model: &'a JsonModel) -> Self {
-        Self {
-            model,
-            closure_gen: ClosureGenerator::new(model),
-        }
-    }
-
-    /// Generate the complete Rust code
-    pub fn generate(&self) -> Result<GeneratedCode, JsonModelError> {
-        let (equation_code, kind) = match self.model.model_type {
-            ModelType::Analytical => {
-                let code = self.generate_analytical()?;
-                (code, EqnKind::Analytical)
-            }
-            ModelType::Ode => {
-                let code = self.generate_ode()?;
-                (code, EqnKind::ODE)
-            }
-            ModelType::Sde => {
-                let code = self.generate_sde()?;
-                (code, EqnKind::SDE)
-            }
-        };
-
[…]
