### 2026-04-28 1f6624b
chore(pharmsol): release v0.26.2
