+++ pharmsol-macros/CHANGELOG.md
@@ -1,0 +1,14 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+## [0.1.1](https://github.com/LAPKB/pharmsol/compare/pharmsol-macros-v0.1.0...pharmsol-macros-v0.1.1) - 2026-04-28
+
+### Added
+
+- Domain Specific Language (DSL) using JIT or AOT ([#252](https://github.com/LAPKB/pharmsol/pull/252))
