### 2026-05-01 75cba57
fix test, add new test to check for full parity. checking also against DSL
