+++ pharmsol-macros/src/lib.rs
@@ -37,10 +37,12 @@ struct OdeInput {
 struct AnalyticalInput {
     name: LitStr,
     params: Vec<Ident>,
+    covariates: Vec<Ident>,
     states: Vec<Ident>,
     outputs: Vec<Ident>,
     routes: Vec<OdeRouteDecl>,
     structure: Ident,
+    sec: Option<ExprClosure>,
     lag: Option<ExprClosure>,
     fa: Option<ExprClosure>,
     init: Option<ExprClosure>,
@@ -50,6 +52,7 @@ struct AnalyticalInput {
 struct SdeInput {
     name: LitStr,
     params: Vec<Ident>,
+    covariates: Vec<Ident>,
     states: Vec<Ident>,
     outputs: Vec<Ident>,
     routes: Vec<OdeRouteDecl>,
@@ -198,7 +201,7 @@ impl Parse for OdeInput {
         let routes = routes.ok_or_else(|| missing_required_ode_field("routes"))?;
         let diffeq = diffeq.ok_or_else(|| missing_required_ode_field("diffeq"))?;
         let out = out.ok_or_else(|| missing_required_ode_field("out"))?;
-        let diffeq_mode = classify_diffeq_mode(&diffeq)?;
+        let diffeq_mode = classify_diffeq_mode(&diffeq, &routes)?;
 
         validate_unique_idents("parameter", &params, "ode!")?;
         validate_unique_idents("covariate", &covariates, "ode!")?;
@@ -206,13 +209,23 @@ impl Parse for OdeInput {
         validate_unique_idents("output", &outputs, "ode!")?;
         validate_routes(&routes, &states, "ode!")?;
         validate_named_binding_compatibility(
-            &params,
-            &states,
-            &outputs,
-            &routes,
-            &diffeq,
-            &out,
-            diffeq_mode,
+            NamedBindingSets {
+                params: &params,
+                covariates: &covariates,
+                states: &states,
+                outputs: &outputs,
+                routes: &routes,
+            },
+            OdeBindingClosures {
+                diffeq: &diffeq,
+                common: CommonBindingClosures {
+                    lag: lag.as_ref(),
+                    fa: fa.as_ref(),
+                    init: init.as_ref(),
+                    out: &out,
+                },
+                diffeq_mode,
+            },
         )?;
 
         Ok(Self {
@@ -245,10 +258,12 @@ impl Parse for AnalyticalInput {
     fn parse(input: ParseStream) -> syn::Result<Self> {
         let mut name = None;
         let mut params = None;
+        let mut covariates = None;
         let mut states = None;
         let mut outputs = None;
         let mut routes = None;
         let mut structure = None;
+        let mut sec = None;
         let mut lag = None;
         let mut fa = None;
         let mut init = None;
@@ -263,6 +278,12 @@ impl Parse for AnalyticalInput {
                 "params" => {
                     set_once_analytical(&mut params, parse_ident_list(input)?, &key, "params")?
                 }
+                "covariates" => set_once_analytical(
+                    &mut covariates,
+                    parse_ident_list(input)?,
+                    &key,
+                    "covariates",
+                )?,
                 "states" => {
                     set_once_analytical(&mut states, parse_ident_list(input)?, &key, "states")?
                 }
@@ -275,6 +296,7 @@ impl Parse for AnalyticalInput {
                 "structure" => {
                     set_once_analytical(&mut structure, input.parse()?, &key, "structure")?
                 }
+                "sec" => set_once_analytical(&mut sec, input.parse()?, &key, "sec")?,
[…]
