### 2026-05-01 9aa1695
chore: fix compilation ubuntu
