### 2026-05-01 0a04f41
feature: INPUT/OUTEQ colums in the Pmetrics format are now read as Strings instead of Integer. This gives flexibility to modelers to define routes in terms of indices or named elements. All three surfaces for Model creation are also updated (macros, SDL, and ::new() constructors)
