+++ pharmsol-dsl/tests/dsl_authoring_edge_cases.rs
@@ -1,4 +1,4 @@
-use pharmsol_dsl::{analyze_model, parse_model, parse_module};
+use pharmsol_dsl::{analyze_model, lower_typed_model, parse_model, parse_module};
 
 #[test]
 fn output_annotation_is_optional() {
@@ -159,6 +159,54 @@ out(cp) = central ~ continous()
         "{}",
         rendered
     );
+}
+
+#[test]
+fn mixed_named_and_numeric_output_labels_lower_and_round_trip() {
+    let src = r#"
+name = mixed_output_labels
+kind = ode
+params = ke, v
+states = central
+outputs = cp, 0, 1
+infusion(iv) -> central
+ddt(central) = -ke * central
+out(cp) = central / v
+out(0) = 2 * central / v
+out(1) = 3 * central / v
+"#;
+
+    let module = parse_module(src).expect("mixed output labels should parse in authoring DSL");
+    let model = module
+        .models
+        .first()
+        .expect("authoring DSL should produce one model");
+    let typed = analyze_model(&model).expect("mixed output labels should analyze");
+    let lowered = lower_typed_model(&typed).expect("mixed output labels should lower");
+
+    assert_eq!(
+        lowered
+            .metadata
+            .outputs
+            .iter()
+            .map(|output| output.name.as_str())
+            .collect::<Vec<_>>(),
+        vec!["cp", "0", "1"]
+    );
+    assert_eq!(
+        lowered
+            .metadata
+            .outputs
+            .iter()
+            .map(|output| output.index)
+            .collect::<Vec<_>>(),
+        vec![0, 1, 2]
+    );
+
+    let rendered = module.to_string();
+    let reparsed = parse_module(&rendered).expect("rendered mixed-output model should reparse");
+
+    assert_eq!(rendered, reparsed.to_string());
 }
 
 #[test]
