+++ src/simulator/equation/analytical/mod.rs
@@ -32,9 +32,7 @@ pub enum AnalyticalMetadataError {
     Validation(#[from] ModelMetadataError),
     #[error("analytical model declares {declared} state metadata entries but model has {expected} states")]
     StateCountMismatch { expected: usize, declared: usize },
-    #[error(
-        "analytical model declares {declared} route metadata entries but model has {expected} input channels"
-    )]
+    #[error("analytical model declares {declared} route metadata entries but model has {expected} inputs")]
     RouteCountMismatch { expected: usize, declared: usize },
     #[error("analytical model declares {declared} output metadata entries but model has {expected} outputs")]
     OutputCountMismatch { expected: usize, declared: usize },
@@ -119,7 +117,7 @@ impl Analytical {
         self
     }
 
-    /// Set the number of drug input channels (size of bolus[] and rateiv[]).
+    /// Set the number of drug inputs (size of bolus[] and rateiv[]).
     pub fn with_ndrugs(mut self, ndrugs: usize) -> Self {
         self.neqs.ndrugs = ndrugs;
         self.invalidate_metadata();
@@ -186,7 +184,7 @@ fn validate_metadata_dimensions(
         });
     }
 
-    let declared_routes = metadata.route_channel_count();
+    let declared_routes = metadata.route_input_count();
     if declared_routes != neqs.ndrugs {
         return Err(AnalyticalMetadataError::RouteCountMismatch {
             expected: neqs.ndrugs,

[…]
