### 2026-05-01 dac1108
fix: The new string-based mapper for inputs/outeq had an issue when ordering indices, lso the implementation was not complete over the DSL frontend
