+++ src/dsl/runtime.rs
@@ -376,11 +376,95 @@ fn runtime_model_from_parts(
 mod tests {
     use super::*;
     use crate::dsl::compile_sde_model_to_jit;
+    use crate::PharmsolError;
     use crate::test_fixtures::STRUCTURED_BLOCK_CORPUS;
     use crate::SubjectBuilderExt;
     use approx::assert_relative_eq;
     use pharmsol_dsl::{DiagnosticPhase, DSL_BACKEND_GENERIC, DSL_PARSE_GENERIC};
     use tempfile::tempdir;
+
+    const MULTI_DIGIT_OUTPUT_ORDER_RUNTIME_DSL: &str = r#"
+name = multi_digit_output_runtime
+kind = ode
+
+params = ke, v
+states = central
+outputs = 2, 10, 11
+
+infusion(iv) -> central
+
+dx(central) = -ke * central
+
+out(10) = central / v ~ continuous()
+out(2) = central / v ~ continuous()
+out(11) = central / v ~ continuous()
+"#;
+
+    const NUMERIC_ROUTE_LABELS_RUNTIME_DSL: &str = r#"
+name = numeric_route_runtime
+kind = ode
+
+params = ke, v
+states = central
+outputs = cp
+
+bolus(10) -> central
+bolus(11) -> central
+
+dx(central) = -ke * central
+
+out(cp) = central / v ~ continuous()
+"#;
+
+    const UNDECLARED_NUMERIC_OUTPUT_LABEL_RUNTIME_DSL: &str = r#"
+name = undeclared_numeric_output_runtime
+kind = ode
+
+params = ke, v
+states = central
+outputs = a0, a1, a2, a3, a4, a5, a6, a7, a8, a9, a10
+
+infusion(iv) -> central
+
+dx(central) = -ke * central
+
+out(a0) = central / v ~ continuous()
+out(a1) = central / v ~ continuous()
+out(a2) = central / v ~ continuous()
+out(a3) = central / v ~ continuous()
+out(a4) = central / v ~ continuous()
+out(a5) = central / v ~ continuous()
+out(a6) = central / v ~ continuous()
+out(a7) = central / v ~ continuous()
+out(a8) = central / v ~ continuous()
+out(a9) = central / v ~ continuous()
+out(a10) = central / v ~ continuous()
+"#;
+
+    const UNDECLARED_NUMERIC_INPUT_LABEL_RUNTIME_DSL: &str = r#"
+name = undeclared_numeric_input_runtime
+kind = ode
+
+params = ke, v
+states = central
+outputs = cp
+
+bolus(r0) -> central
+bolus(r1) -> central
+bolus(r2) -> central
+bolus(r3) -> central
+bolus(r4) -> central
+bolus(r5) -> central
+bolus(r6) -> central
+bolus(r7) -> central
+bolus(r8) -> central
+bolus(r9) -> central
+bolus(r10) -> central
+
+dx(central) = -ke * central
+
+out(cp) = central / v ~ continuous()
+"#;
 
     fn corpus_source() -> &'static str {
         STRUCTURED_BLOCK_CORPUS
@@ -397,17 +481,17 @@ mod tests {
         pharmsol_dsl::lower_typed_model(model).expect("lower corpus model")
     }
 
-    fn ode_subject(output: usize, oral: usize, iv: usize) -> Subject {
+    fn ode_subject() -> Subject {
         Subject::builder("ode")
             .covariate("wt", 0.0, 70.0)
-            .bolus(0.0, 120.0, oral)
-            .infusion(6.0, 60.0, iv, 2.0)
-            .missing_observation(0.5, output)
-            .missing_observation(1.0, output)
-            .missing_observation(2.0, output)
-            .missing_observation(6.0, output)
[…]
