+++ tests/ode_macro_lowering.rs
@@ -56,9 +56,9 @@ fn injected_macro_ode() -> equation::ODE {
         params: [ke, v],
         states: [central],
         outputs: [cp],
-        routes: {
+        routes: [
             infusion(iv) -> central,
-        },
+        ],
         diffeq: |x, _t, dx| {
             dx[central] = -ke * x[central];
         },
@@ -99,66 +99,17 @@ fn injected_handwritten_ode() -> equation::ODE {
     .expect("handwritten injected metadata should validate")
 }
 
-fn explicit_macro_ode() -> equation::ODE {
-    ode! {
-        name: "explicit_one_cpt",
-        params: [ke, v],
-        states: [central],
-        outputs: [cp],
-        routes: {
-            infusion(iv) -> central,
-        },
-        diffeq: |x, _t, dx, _bolus, rateiv| {
-            dx[central] = rateiv[iv] - ke * x[central];
-        },
-        out: |x, _t, y| {
-            y[cp] = x[central] / v;
-        },
-    }
-}
-
-fn explicit_handwritten_ode() -> equation::ODE {
-    equation::ODE::new(
-        |x, p, _t, dx, _bolus, rateiv, _cov| {
-            fetch_params!(p, ke, _v);
-            dx[0] = rateiv[0] - ke * x[0];
-        },
-        |_p, _t, _cov| lag! {},
-        |_p, _t, _cov| fa! {},
-        |_p, _t, _cov, _x| {},
-        |x, p, _t, _cov, y| {
-            fetch_params!(p, _ke, v);
-            y[0] = x[0] / v;
-        },
-    )
-    .with_nstates(1)
-    .with_ndrugs(1)
-    .with_nout(1)
-    .with_metadata(
-        equation::metadata::new("explicit_one_cpt")
-            .parameters(["ke", "v"])
-            .states(["central"])
-            .outputs(["cp"])
-            .route(
-                equation::Route::infusion("iv")
-                    .to_state("central")
-                    .expect_explicit_input(),
-            ),
-    )
-    .expect("handwritten explicit metadata should validate")
-}
-
 fn numeric_label_macro_ode() -> equation::ODE {
     ode! {
         name: "numeric_label_one_cpt",
         params: [ke, v],
         states: [central],
         outputs: [1],
-        routes: {
+        routes: [
             infusion(1) -> central,
-        },
-        diffeq: |x, _t, dx, _bolus, rateiv| {
-            dx[central] = rateiv[1] - ke * x[central];
+        ],
+        diffeq: |x, _t, dx| {
+            dx[central] = -ke * x[central];
         },
         out: |x, _t, y| {
             y[1] = x[central] / v;
@@ -191,7 +142,7 @@ fn numeric_label_handwritten_ode() -> equation::ODE {
             .route(
                 equation::Route::infusion("1")
                     .to_state("central")
-                    .expect_explicit_input(),
+                    .inject_input_to_destination(),
             ),
     )
     .expect("handwritten numeric-label metadata should validate")
@@ -203,13 +154,13 @@ fn shared_input_macro_ode() -> equation::ODE {
         params: [ka, ke, v, tlag, f_oral],
         states: [depot, central],
         outputs: [cp],
-        routes: {
+        routes: [
             bolus(oral) -> depot,
             infusion(iv) -> central,
-        },
-        diffeq: |x, _t, dx, bolus, rateiv| {
-            dx[depot] = bolus[oral] - ka * x[depot];
-            dx[central] = ka * x[depot] + rateiv[iv] - ke * x[central];
+        ],
+        diffeq: |x, _t, dx| {
+            dx[depot] = -ka * x[depot];
+            dx[central] = ka * x[depot] - ke * x[central];
         },
         lag: |_t| {
             lag! { oral => tlag }
[…]
