### 2026-05-01 8cac67c
chore: update examples to use new route API
