+++ examples/ode_readme.rs
@@ -17,15 +17,12 @@ fn main() -> Result<(), pharmsol::PharmsolError> {
         },
     };
 
-    let iv = ode.route_index("iv").expect("iv route exists");
-    let cp = ode.output_index("cp").expect("cp output exists");
-
     let subject = Subject::builder("id1")
-        .infusion(0.0, 100.0, iv, 0.5)
-        .missing_observation(0.5, cp)
-        .missing_observation(1.0, cp)
-        .missing_observation(2.0, cp)
-        .missing_observation(4.0, cp)
+        .infusion(0.0, 100.0, "iv", 0.5)
+        .missing_observation(0.5, "cp")
+        .missing_observation(1.0, "cp")
+        .missing_observation(2.0, "cp")
+        .missing_observation(4.0, "cp")
         .build();
 
     let predictions = ode.estimate_predictions(&subject, &[1.022, 194.0])?;
