### 2026-05-01 7b24e09
feature: DSL supports numeric input/outeq. Implement Equation Trait for runtime environments.
