+++ src/dsl/native.rs
@@ -1,4 +1,5 @@
 use std::cell::RefCell;
+use std::collections::HashMap;
 use std::sync::Arc;
 
 use diffsol::{
@@ -20,14 +21,17 @@ pub use super::model_info::{
     NativeCovariateInfo, NativeModelInfo, NativeOutputInfo, NativeRouteInfo,
 };
 use crate::{
+    data::error_model::AssayErrorModels,
     data::{Covariates, Infusion, InputLabel, OutputLabel},
     simulator::{
+        cache::{PredictionCache, DEFAULT_CACHE_SIZE},
         equation::{
             ode::{closure_helpers::PMProblem, ExplicitRkTableau, OdeSolver, SdirkTableau},
             sde::simulate_sde_event_with,
+            EqnKind, Equation, EquationPriv, EquationTypes,
         },
         likelihood::{Prediction, SubjectPredictions},
-        M, V,
+        Fa, Lag, M, T, V,
     },
     Event, Observation, Occasion, PharmsolError, Subject,
 };
@@ -727,6 +731,7 @@ pub struct NativeOdeModel {
     solver: OdeSolver,
     rtol: f64,
     atol: f64,
+    cache: Option<PredictionCache>,
 }
 
 #[derive(Clone, Debug)]
@@ -754,6 +759,7 @@ impl NativeOdeModel {
             solver: OdeSolver::default(),
             rtol: DEFAULT_ODE_RTOL,
             atol: DEFAULT_ODE_ATOL,
+            cache: Some(PredictionCache::new(DEFAULT_CACHE_SIZE)),
         }
     }
 
@@ -1028,6 +1034,175 @@ impl NativeOdeModel {
         }
 
         Ok(())
+    }
+}
+
+fn runtime_no_lag(_: &V, _: T, _: &Covariates) -> HashMap<usize, T> {
+    HashMap::new()
+}
+
+fn runtime_no_fa(_: &V, _: T, _: &Covariates) -> HashMap<usize, T> {
+    HashMap::new()
+}
+
+#[inline(always)]
+fn runtime_ode_predictions(
+    model: &NativeOdeModel,
+    subject: &Subject,
+    support_point: &[f64],
+) -> Result<SubjectPredictions, PharmsolError> {
+    if let Some(cache) = &model.cache {
+        let key = (
+            subject.hash(),
+            crate::simulator::equation::spphash(support_point),
+        );
+        if let Some(cached) = cache.get(&key) {
+            return Ok(cached);
+        }
+
+        let result = model.estimate_predictions(subject, support_point)?;
+        cache.insert(key, result.clone());
+        Ok(result)
+    } else {
+        model.estimate_predictions(subject, support_point)
+    }
+}
+
+impl crate::simulator::equation::Cache for NativeOdeModel {
+    fn with_cache_capacity(mut self, size: u64) -> Self {
+        self.cache = Some(PredictionCache::new(size));
+        self
+    }
+
+    fn enable_cache(mut self) -> Self {
+        self.cache = Some(PredictionCache::new(DEFAULT_CACHE_SIZE));
+        self
+    }
+
+    fn clear_cache(&self) {
+        if let Some(cache) = &self.cache {
+            cache.invalidate_all();
+        }
+    }
+
+    fn disable_cache(mut self) -> Self {
+        self.cache = None;
+        self
+    }
+}
+
+impl EquationTypes for NativeOdeModel {
+    type S = V;
+    type P = SubjectPredictions;
+}
+
+impl EquationPriv for NativeOdeModel {
+    fn lag(&self) -> &Lag {
+        &(runtime_no_lag as Lag)
+    }
+
+    fn fa(&self) -> &Fa {
[…]

[…]
