+++ src/simulator/equation/metadata.rs
@@ -1,17 +1,40 @@
-//! Shared model metadata for handwritten simulator models.
+//! Metadata builders and validated metadata views for handwritten models.
 //!
-//! This module defines the public metadata contract that handwritten ODE,
-//! analytical, and SDE models can attach to. The field set is intentionally
-//! aligned with the public subset of the DSL/runtime metadata surface.
+//! Use this module when a handwritten [`crate::ODE`], [`crate::Analytical`], or
+//! [`crate::SDE`] model should expose the same public names that appear in data
+//! rows, subject builders, or parsed files.
 //!
-//! Internal runtime layout details such as dense buffer lengths, derived buffer
-//! shape, or ABI-specific offsets remain internal for now.
+//! Metadata gives names to parameters, covariates, states, routes, and outputs.
+//! After validation, the execution layer can resolve public labels such as
+//! `"iv"` and `"cp"` against those declarations before simulation.
+//!
+//! Without metadata, handwritten models fall back to numeric labels. With
+//! metadata, labels are matched by name.
+//!
+//! # Example
+//!
+//! ```rust
+//! use pharmsol::{metadata, ModelKind};
+//!
+//! let metadata = metadata::new("one_cmt")
+//!     .kind(ModelKind::Ode)
+//!     .parameters(["cl", "v"])
+//!     .states(["central"])
+//!     .outputs(["cp"])
+//!     .route(metadata::Route::infusion("iv").to_state("central"))
+//!     .validate()
+//!     .unwrap();
+//!
+//! assert_eq!(metadata.name(), "one_cmt");
+//! assert_eq!(metadata.route("iv").unwrap().destination(), "central");
+//! assert_eq!(metadata.output_index("cp"), Some(0));
+//! ```
 
 use pharmsol_dsl::{AnalyticalKernel, CovariateInterpolation, ModelKind};
 use std::fmt;
 use thiserror::Error;
 
-/// Create a new handwritten-model metadata builder.
+/// Shorthand for [`ModelMetadata::new`].
 pub fn new(name: impl Into<String>) -> ModelMetadata {
     ModelMetadata::new(name)
 }
@@ -71,7 +94,17 @@ impl fmt::Display for NameDomain {
     }
 }
 
-/// Immutable validated metadata view used by later attachment slices.
+/// Validated metadata view used by the execution layer.
+///
+/// This type is what handwritten equation builders store after metadata has
+/// passed validation. It provides stable lookup helpers from public names to the
+/// dense indices used during execution.
+///
+/// Route lookups expose two different indices:
+/// - [`ValidatedModelMetadata::route_declaration_index`] is the route position in
+///   declaration order.
+/// - [`ValidatedModelMetadata::route_index`] is the dense execution input index
+///   for that route kind.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct ValidatedModelMetadata {
     name: String,
@@ -87,10 +120,12 @@ pub struct ValidatedModelMetadata {
 }
 
 impl ValidatedModelMetadata {
+    /// Get the public model name.
     pub fn name(&self) -> &str {
         &self.name
     }
 
+    /// Get the validated model family.
     pub fn kind(&self) -> ModelKind {
         self.kind
     }
@@ -111,6 +146,9 @@ impl ValidatedModelMetadata {
         &self.routes
     }
 
+    /// Get the number of dense execution input slots needed for routes.
+    ///
+    /// This is the maximum of the bolus-route count and infusion-route count.
     pub fn route_input_count(&self) -> usize {
         self.route_input_count
     }
@@ -143,14 +181,17 @@ impl ValidatedModelMetadata {
         self.states.iter().position(|state| state.name() == name)
     }
 
+    /// Look up a route by public name and return its dense execution input index.
[…]
