+++ pharmsol-dsl/tests/dsl_authoring_edge_cases.rs
@@ -162,18 +162,18 @@ out(cp) = central ~ continous()
 }
 
 #[test]
-fn mixed_named_and_numeric_output_labels_lower_and_round_trip() {
+fn mixed_named_and_prefixed_numeric_output_labels_lower_and_round_trip() {
     let src = r#"
 name = mixed_output_labels
 kind = ode
 params = ke, v
 states = central
-outputs = cp, 0, 1
+outputs = cp, outeq_0, outeq_1
 infusion(iv) -> central
 ddt(central) = -ke * central
 out(cp) = central / v
-out(0) = 2 * central / v
-out(1) = 3 * central / v
+out(outeq_0) = 2 * central / v
+out(outeq_1) = 3 * central / v
 "#;
 
     let module = parse_module(src).expect("mixed output labels should parse in authoring DSL");
@@ -191,7 +191,7 @@ out(1) = 3 * central / v
             .iter()
             .map(|output| output.name.as_str())
             .collect::<Vec<_>>(),
-        vec!["cp", "0", "1"]
+        vec!["cp", "outeq_0", "outeq_1"]
     );
     assert_eq!(
         lowered
@@ -210,26 +210,26 @@ out(1) = 3 * central / v
 }
 
 #[test]
-fn shared_numeric_route_and_output_labels_lower_and_round_trip() {
+fn prefixed_numeric_route_and_output_labels_lower_and_round_trip() {
     let src = r#"
-name = shared_numeric_route_output_labels
+name = prefixed_numeric_route_output_labels
 kind = ode
 params = ke, v
 states = central
-outputs = 1
-infusion(1) -> central
+outputs = outeq_1
+infusion(input_1) -> central
 ddt(central) = -ke * central
-out(1) = central / v
+out(outeq_1) = central / v
 "#;
 
-    let module = parse_module(src).expect("shared numeric route/output labels should parse");
+    let module = parse_module(src).expect("prefixed numeric route/output labels should parse");
     let model = module
         .models
         .first()
         .expect("authoring DSL should produce one model");
-    let typed = analyze_model(model).expect("shared numeric route/output labels should analyze");
+    let typed = analyze_model(model).expect("prefixed numeric route/output labels should analyze");
     let lowered =
-        lower_typed_model(&typed).expect("shared numeric route/output labels should lower");
+        lower_typed_model(&typed).expect("prefixed numeric route/output labels should lower");
 
     assert_eq!(
         lowered
@@ -238,7 +238,7 @@ out(1) = central / v
             .iter()
             .map(|route| route.name.as_str())
             .collect::<Vec<_>>(),
-        vec!["1"]
+        vec!["input_1"]
     );
     assert_eq!(
         lowered
@@ -247,13 +247,212 @@ out(1) = central / v
             .iter()
             .map(|output| output.name.as_str())
             .collect::<Vec<_>>(),
-        vec!["1"]
+        vec!["outeq_1"]
     );
 
     let rendered = module.to_string();
     let reparsed = parse_module(&rendered).expect("rendered shared-label model should reparse");
 
     assert_eq!(rendered, reparsed.to_string());
+}
+
+#[test]
+fn rejects_authoring_bare_numeric_output_declarations() {
+    let src = r#"
+name = numeric_outputs
+kind = ode
+states = central
+outputs = 1, 2
+ddt(central) = 0
+out(1) = central
+"#;
+
+    let err = parse_model(src).expect_err("bare numeric output declarations must fail");
+    let rendered = err.render(src);
+
+    assert!(
+        rendered.contains(
+            "bare numeric output labels are not allowed in the DSL; use `outeq_1` instead"
+        ),
+        "{}",
+        rendered
+    );
+    assert!(
[…]
