+++ tests/authoring_parity_corpus.rs
@@ -1,3 +1,4 @@
+#[cfg(feature = "dsl-jit")]
 use approx::assert_relative_eq;
 #[cfg(feature = "dsl-jit")]
 use pharmsol::dsl::{self, RuntimeCompilationTarget, RuntimePredictions};
@@ -59,26 +60,26 @@ kind = ode
 
 params = ke, v
 states = central
-outputs = 2, 10, 11
+outputs = outeq_2, outeq_10, outeq_11
 
 infusion(iv) -> central
 
 dx(central) = -ke * central
 
-out(10) = central / v ~ continuous()
-out(2) = central / v ~ continuous()
-out(11) = central / v ~ continuous()
+out(outeq_10) = central / v ~ continuous()
+out(outeq_2) = central / v ~ continuous()
+out(outeq_11) = central / v ~ continuous()
 "#;
 
 const ODE_NUMERIC_ROUTE_LABELS_AUTHORING_DSL: &str = r#"
-name = authoring_numeric_routes
+name = authoring_prefixed_numeric_routes
 kind = ode
 
 states = first, second
 outputs = cp
 
-bolus(10) -> first
-bolus(11) -> second
+bolus(input_10) -> first
+bolus(input_11) -> second
 
 dx(first) = 0
 dx(second) = 0
@@ -93,8 +94,8 @@ const ODE_NUMERIC_ROUTE_LABELS_STRUCTURED_DSL: &str = r#"model structured_numeri
         second,
     }
     routes {
-        10 -> first
-        11 -> second
+        input_10 -> first
+        input_11 -> second
     }
     dynamics {
         ddt(first) = 0
@@ -149,15 +150,15 @@ kind = ode
 
 params = ke, v
 states = central
-outputs = cp, 0, 1
+outputs = cp, outeq_0, outeq_1
 
 infusion(iv) -> central
 
 dx(central) = -ke * central
 
 out(cp) = central / v ~ continuous()
-out(0) = 2 * central / v ~ continuous()
-out(1) = 3 * central / v ~ continuous()
+out(outeq_0) = 2 * central / v ~ continuous()
+out(outeq_1) = 3 * central / v ~ continuous()
 "#;
 
 #[cfg(feature = "dsl-jit")]
@@ -388,6 +389,26 @@ fn compile_runtime_jit_model(src: &str, model_name: &str) -> dsl::CompiledRuntim
         |_, _| {},
     )
     .expect("DSL runtime model should compile")
+}
+
+#[cfg(feature = "dsl-jit")]
+fn compiled_route_input_index(model: &dsl::CompiledRuntimeModel, name: &str) -> Option<usize> {
+    model
+        .info()
+        .routes
+        .iter()
+        .find(|route| route.name == name)
+        .map(|route| route.index)
+}
+
+#[cfg(feature = "dsl-jit")]
+fn compiled_output_slot_index(model: &dsl::CompiledRuntimeModel, name: &str) -> Option<usize> {
+    model
+        .info()
+        .outputs
+        .iter()
+        .find(|output| output.name == name)
+        .map(|output| output.index)
 }
 
 #[cfg(feature = "dsl-jit")]
@@ -1169,15 +1190,15 @@ fn ode_dsl_declared_output_order_controls_dense_indices_for_multi_digit_labels()
         dsl_view.outputs,
         vec![
             NamedIndex {
-                name: "2".to_string(),
+                name: "outeq_2".to_string(),
                 index: 0,
             },
             NamedIndex {
-                name: "10".to_string(),
+                name: "outeq_10".to_string(),
                 index: 1,
             },
             NamedIndex {
-                name: "11".to_string(),
+                name: "outeq_11".to_string(),
[…]
