+++ pharmsol-macros/Cargo.toml
@@ -13,4 +13,4 @@ proc-macro = true
 [dependencies]
 proc-macro2 = "1.0.106"
 quote = "1.0.45"
-syn = { version = "2.0.117", features = ["full", "visit"] }
+syn = { version = "2.0.117", features = ["full", "visit", "visit-mut"] }
