### 2026-05-10 31f418f
chore: remove duplicated definitions
