### 2026-05-11 8517fa8
feat: applying @mhovd's suggestions
