### 2026-05-11 5a6177e
add missing macros (#253)
* Initial pass

* chore: fmt

* fix test

* fix test, add new test to check for full parity. checking also against DSL

* chore: fmt

* chore: fix compilation ubuntu

* feature: INPUT/OUTEQ colums in the Pmetrics format are now read as Strings instead of Integer. This gives flexibility to modelers to define routes in terms of indices or named elements. All three surfaces for Model creation are also updated (macros, SDL, and ::new() constructors)

* chore: update test

* chore: channeless

* fix: The new string-based mapper for inputs/outeq had an issue when ordering i
[…]
