+++ src/data/builder.rs
@@ -1,6 +1,21 @@
+//! Builder API for constructing [`Subject`] schedules in Rust.
+//!
+//! Use `Subject::builder(...)` when you want to describe a subject directly in
+//! code with a schedule-oriented API. This is the preferred high-level
+//! path for hand-written datasets.
+//!
+//! Builder methods accept public input and output labels. Prefer stable strings
+//! such as `"depot"`, `"iv"`, and `"cp"`. Numeric values are accepted, but
+//! they remain public labels rather than automatically becoming dense internal
+//! indices.
+
 use crate::{data::*, Censor};
 
-/// Extension trait for creating [Subject] instances using the builder pattern
+/// Extension trait that enables `Subject::builder(...)`.
+///
+/// Most users do not need to import [`SubjectBuilder`] directly. Import this
+/// trait from the crate root or [`crate::prelude`] and then start with
+/// `Subject::builder("id")`.
 pub trait SubjectBuilderExt {
     /// Create a new SubjectBuilder with the specified ID
     ///
@@ -14,8 +29,8 @@ pub trait SubjectBuilderExt {
     /// use pharmsol::*;
     ///
     /// let subject = Subject::builder("patient_001")
-    ///     .bolus(0.0, 100.0, 0)
-    ///     .observation(1.0, 10.5, 0)
+    ///     .bolus(0.0, 100.0, "depot")
+    ///     .observation(1.0, 10.5, "cp")
     ///     .build();
     /// ```
     fn builder(id: impl Into<String>) -> SubjectBuilder;
@@ -34,11 +49,37 @@ impl SubjectBuilderExt for Subject {
     }
 }
 
-/// Builder for creating [Subject] instances with a fluent API
+/// Builder for creating [`Subject`] values with a fluent API.
 ///
-/// The [SubjectBuilder] allows for constructing complex subject data with a
-/// chainable, readable syntax. Events like doses and observations can be
-/// added sequentially, and the builder handles organizing them into occasions.
+/// Use [`SubjectBuilder`] when you want to author common dose and observation
+/// schedules directly in Rust without constructing low-level event values by
+/// hand.
+///
+/// A builder instance accumulates events inside the current [`Occasion`].
+/// [`SubjectBuilder::repeat`] duplicates the most recently added event at later
+/// times, and [`SubjectBuilder::reset`] closes the current occasion and starts a
+/// new one with fresh occasion-local state.
+///
+/// Input and output arguments are public labels. Prefer stable model-facing
+/// names such as `"depot"`, `"iv"`, and `"cp"`.
+///
+/// # Example
+///
+/// ```rust
+/// use pharmsol::*;
+///
+/// let subject = Subject::builder("patient_001")
+///     .bolus(0.0, 100.0, "depot")
+///     .repeat(1, 24.0)
+///     .observation(1.0, 12.3, "cp")
+///     .missing_observation(25.0, "cp")
+///     .reset()
+///     .bolus(0.0, 80.0, "depot")
+///     .observation(1.0, 10.1, "cp")
+///     .build();
+///
+/// assert_eq!(subject.occasions().len(), 2);
+/// ```
 #[derive(Debug, Clone)]
 pub struct SubjectBuilder {
     id: String,
@@ -49,52 +90,54 @@ pub struct SubjectBuilder {
 }
 
 impl SubjectBuilder {
-    /// Add an event to the current occasion
+    /// Add a fully constructed event to the current occasion.
     ///
-    /// # Arguments
-    ///
-    /// * `event` - The event to add
+    /// Use this when you want to mix builder convenience methods with direct
+    /// [`Event`] values.
     pub fn event(mut self, event: Event) -> Self {
         self.last_added_event = Some(event.clone());
[…]
