### 2026-05-11 08f5a48
feat: allow analytical parameter projection by name
