+++ src/data/structs.rs
@@ -597,12 +597,12 @@ impl Occasion {
     }
 
     fn add_lagtime(&mut self, reorder: Option<(&Fa, &Lag, &[f64], &Covariates)>) {
-        if let Some((_, fn_lag, spp, covariates)) = reorder {
-            let spp = nalgebra::DVector::from_vec(spp.to_vec());
+        if let Some((_, fn_lag, parameters, covariates)) = reorder {
+            let parameters = nalgebra::DVector::from_vec(parameters.to_vec());
             for event in self.events.iter_mut() {
                 let time = event.time();
                 if let Event::Bolus(bolus) = event {
-                    let lagtime = fn_lag(&spp.clone().into(), time, covariates);
+                    let lagtime = fn_lag(&parameters.clone().into(), time, covariates);
                     if let Some(input) = bolus.input_index() {
                         if let Some(l) = lagtime.get(&input) {
                             *bolus.mut_time() += l;
@@ -616,12 +616,12 @@ impl Occasion {
 
     fn add_bioavailability(&mut self, reorder: Option<(&Fa, &Lag, &[f64], &Covariates)>) {
         // If lagtime is empty, return early
-        if let Some((fn_fa, _, spp, covariates)) = reorder {
-            let spp = nalgebra::DVector::from_vec(spp.to_vec());
+        if let Some((fn_fa, _, parameters, covariates)) = reorder {
+            let parameters = nalgebra::DVector::from_vec(parameters.to_vec());
             for event in self.events.iter_mut() {
                 let time = event.time();
                 if let Event::Bolus(bolus) = event {
-                    let fa = fn_fa(&spp.clone().into(), time, covariates);
+                    let fa = fn_fa(&parameters.clone().into(), time, covariates);
                     if let Some(input) = bolus.input_index() {
                         if let Some(f) = fa.get(&input) {
                             bolus.set_amount(bolus.amount() * f);
@@ -665,7 +665,7 @@ impl Occasion {
     ///
     /// # Arguments
     ///
-    /// * `reorder` - Optional tuple containing references to (Fa, Lag, support point, covariates) for adjustments
+    /// * `reorder` - Optional tuple containing references to (Fa, Lag, parameter values, covariates) for adjustments
     /// # Returns
     ///
     /// Vector of events with times adjusted for lag and bioavailability.
