+++ src/simulator/equation/mod.rs
@@ -183,7 +183,7 @@ pub(crate) trait EquationPriv: EquationTypes {
     fn solve(
         &self,
         state: &mut Self::S,
-        support_point: &[f64],
+        parameters: &[f64],
         covariates: &Covariates,
         infusions: &[Infusion],
         start_time: f64,
@@ -252,7 +252,7 @@ pub(crate) trait EquationPriv: EquationTypes {
     fn resolve_occasion_events(
         &self,
         occasion: &Occasion,
-        support_point: &[f64],
+        parameters: &[f64],
         covariates: &Covariates,
     ) -> Result<Vec<Event>, PharmsolError> {
         let mut resolved = occasion.clone();
@@ -274,10 +274,7 @@ pub(crate) trait EquationPriv: EquationTypes {
             }
         }
 
-        Ok(resolved.process_events(
-            Some((self.fa(), self.lag(), support_point, covariates)),
-            true,
-        ))
+        Ok(resolved.process_events(Some((self.fa(), self.lag(), parameters, covariates)), true))
     }
     #[allow(dead_code)]
     fn is_sde(&self) -> bool {
@@ -287,7 +284,7 @@ pub(crate) trait EquationPriv: EquationTypes {
     #[allow(clippy::too_many_arguments)]
     fn process_observation(
         &self,
-        support_point: &[f64],
+        parameters: &[f64],
         observation: &Observation,
         error_models: Option<&AssayErrorModels>,
         time: f64,
@@ -299,7 +296,7 @@ pub(crate) trait EquationPriv: EquationTypes {
 
     fn initial_state(
         &self,
-        support_point: &[f64],
+        parameters: &[f64],
         covariates: &Covariates,
         occasion_index: usize,
     ) -> Self::S;
@@ -307,7 +304,7 @@ pub(crate) trait EquationPriv: EquationTypes {
     #[allow(clippy::too_many_arguments)]
     fn simulate_event(
         &self,
-        support_point: &[f64],
+        parameters: &[f64],
         event: &Event,
         next_event: Option<&Event>,
         error_models: Option<&AssayErrorModels>,
@@ -339,7 +336,7 @@ pub(crate) trait EquationPriv: EquationTypes {
             }
             Event::Observation(observation) => {
                 self.process_observation(
-                    support_point,
+                    parameters,
                     observation,
                     error_models,
                     event.time(),
@@ -354,7 +351,7 @@ pub(crate) trait EquationPriv: EquationTypes {
         if let Some(next_event) = next_event {
             self.solve(
                 x,
-                support_point,
+                parameters,
                 covariates,
                 infusions,
                 event.time(),
@@ -397,7 +394,7 @@ pub trait Equation: EquationPriv + 'static + Clone + Sync {
         Ok(error_models.bind_to(self)?)
     }
 
-    /// Estimate the likelihood of the subject given the support point and error model.
+    /// Estimate the likelihood of the subject given the parameters and error model.
     ///
     /// **Deprecated**: Use [`estimate_log_likelihood`](Self::estimate_log_likelihood) instead
     /// for better numerical stability, especially with many observations or extreme parameter values.
@@ -407,7 +404,7 @@ pub trait Equation: EquationPriv + 'static + Clone + Sync {
     ///
     /// # Parameters
     /// - `subject`: The subject data
-    /// - `support_point`: The parameter values
+    /// - `parameters`: The parameter values
     /// - `error_model`: The error model
     ///
     /// # Returns
@@ -419,11 +416,11 @@ pub trait Equation: EquationPriv + 'static + Clone + Sync {
     fn estimate_likelihood(
         &self,
         subject: &Subject,
-        support_point: &[f64],
+        parameters: &[f64],
[…]
