+++ src/simulator/equation/sde/mod.rs
@@ -16,7 +16,7 @@ use crate::{
     Subject,
 };
 
-use super::spphash;
+use super::parameters_hash;
 use crate::simulator::cache::{SdeLikelihoodCache, DEFAULT_CACHE_SIZE};
 
 use diffsol::VectorCommon;
@@ -85,7 +85,7 @@ impl InjectedBolusMappings {
 /// * `drift` - Function defining the deterministic component of the SDE
 /// * `difussion` - Function defining the stochastic component of the SDE
 /// * `x` - Current state vector
-/// * `support_point` - Parameter vector for the model
+/// * `parameters` - Parameter vector for the model
 /// * `cov` - Covariates that may influence the system dynamics
 /// * `infusions` - Infusion events to be applied during simulation
 /// * `ti` - Starting time
@@ -100,7 +100,7 @@ fn simulate_sde_event(
     drift: &Drift,
     difussion: &Diffusion,
     x: V,
-    support_point: &[f64],
+    parameters: &[f64],
     cov: &Covariates,
     infusions: &[Infusion],
     ndrugs: usize,
@@ -111,13 +111,13 @@ fn simulate_sde_event(
         return x;
     }
 
-    let params_v = V::from_vec(support_point.to_vec(), NalgebraContext);
+    let parameters_v = V::from_vec(parameters.to_vec(), NalgebraContext);
     let covariates = cov.clone();
     let infusion_events = infusions.to_vec();
     let drift_fn = *drift;
     let diffusion_fn = *difussion;
 
-    let params_for_drift = params_v.clone();
+    let parameters_for_drift = parameters_v.clone();
     let drift_closure = move |time: f64, state: &DVector<f64>, out: &mut DVector<f64>| {
         let mut rateiv = V::zeros(ndrugs, NalgebraContext);
         for infusion in &infusion_events {
@@ -133,7 +133,7 @@ fn simulate_sde_event(
         let mut out_v = V::zeros(state.len(), NalgebraContext);
         drift_fn(
             &state_v,
-            &params_for_drift,
+            &parameters_for_drift,
             time,
             &mut out_v,
             &rateiv,
@@ -144,7 +144,7 @@ fn simulate_sde_event(
 
     let diffusion_closure = move |_time: f64, _state: &DVector<f64>, out: &mut DVector<f64>| {
         let mut out_v = V::zeros(out.len(), NalgebraContext);
-        diffusion_fn(&params_v, &mut out_v);
+        diffusion_fn(&parameters_v, &mut out_v);
         out.copy_from(out_v.inner());
     };
 
@@ -426,13 +426,13 @@ impl EquationPriv for SDE {
     // }
 
     // #[inline(always)]
-    // fn get_lag(&self, spp: &[f64]) -> Option<HashMap<usize, f64>> {
-    //     Some((self.lag)(&V::from_vec(spp.to_owned())))
+    // fn get_lag(&self, parameters: &[f64]) -> Option<HashMap<usize, f64>> {
+    //     Some((self.lag)(&V::from_vec(parameters.to_owned())))
     // }
 
     // #[inline(always)]
-    // fn get_fa(&self, spp: &[f64]) -> Option<HashMap<usize, f64>> {
-    //     Some((self.fa)(&V::from_vec(spp.to_owned())))
+    // fn get_fa(&self, parameters: &[f64]) -> Option<HashMap<usize, f64>> {
+    //     Some((self.fa)(&V::from_vec(parameters.to_owned())))
     // }
 
     #[inline(always)]
@@ -468,7 +468,7 @@ impl EquationPriv for SDE {
     fn solve(
         &self,
         state: &mut Self::S,
-        support_point: &[f64],
+        parameters: &[f64],
         covariates: &Covariates,
         infusions: &[Infusion],
[…]
