### 2026-05-11 9be984b
chore: cargo fmt
