### 2026-05-11 b23e500
chore: Refactor benchmarks
