+++ benches/predictions.rs
@@ -3,10 +3,10 @@ use pharmsol::prelude::*;
 use pharmsol::{Analytical, Cache, ODE};
 use std::hint::black_box;
 
-/// Oral dose with measurements in plasma
-fn oral_subject() -> Subject {
-    Subject::builder("bench_oral")
-        .bolus(0.0, 500.0, "oral")
+/// Oral acetaminophen dose with measurements in plasma
+fn acetaminophen_oral_subject() -> Subject {
+    Subject::builder("bench_acetaminophen_oral")
+        .bolus(0.0, 500.0, "acetaminophen_oral")
         .missing_observation(0.25, "plasma")
         .missing_observation(0.5, "plasma")
         .missing_observation(1.0, "plasma")
@@ -18,10 +18,10 @@ fn oral_subject() -> Subject {
         .build()
 }
 
-/// Intravenous infusion with measurements in plasma
-fn iv_subject() -> Subject {
-    Subject::builder("bench_iv")
-        .infusion(0.0, 500.0, "iv", 0.5)
+/// Intravenous acetaminophen infusion with measurements in plasma
+fn acetaminophen_iv_subject() -> Subject {
+    Subject::builder("bench_acetaminophen_iv")
+        .infusion(0.0, 500.0, "acetaminophen_iv", 0.5)
         .missing_observation(0.5, "plasma")
         .missing_observation(1.0, "plasma")
         .missing_observation(2.0, "plasma")
@@ -34,12 +34,12 @@ fn iv_subject() -> Subject {
 
 fn analytical_one_cpt() -> Analytical {
     analytical! {
-        name: "one_cmt_oral",
+        name: "one_cmt_acetaminophen_oral",
         params: [ka, ke, v],
         states: [gut, central],
         outputs: [plasma],
         routes: [
-            bolus(oral) -> gut,
+            bolus(acetaminophen_oral) -> gut,
         ],
         structure: one_compartment_with_absorption,
         out: |x, _p, _t, _cov, y| {
@@ -51,12 +51,12 @@ fn analytical_one_cpt() -> Analytical {
 /// Two-compartment intravenous model
 fn analytical_two_cpt() -> Analytical {
     analytical! {
-        name: "two_cmt_iv",
+        name: "two_cmt_acetaminophen_iv",
         params: [ke, kcp, kpc, v],
         states: [central, peripheral],
         outputs: [plasma],
         routes: [
-            infusion(iv) -> central,
+            infusion(acetaminophen_iv) -> central,
         ],
         structure: two_compartments,
         out: |x, _p, _t, _cov, y| {
@@ -67,12 +67,12 @@ fn analytical_two_cpt() -> Analytical {
 
 fn ode_one_cpt() -> ODE {
     ode! {
-        name: "one_cmt_oral",
+        name: "one_cmt_acetaminophen_oral",
         params: [ka, ke, v],
         states: [gut, central],
         outputs: [plasma],
         routes: [
-            bolus(oral) -> gut,
+            bolus(acetaminophen_oral) -> gut,
         ],
         diffeq: |x, _p, _t, dx, _cov| {
             dx[gut] = -ka * x[gut];
@@ -87,12 +87,12 @@ fn ode_one_cpt() -> ODE {
 /// Two-compartment model with infusion
 fn ode_two_cpt() -> ODE {
     ode! {
-        name: "two_cmt_iv",
+        name: "two_cmt_acetaminophen_iv",
         params: [cl, v, vp, q],
         states: [central, peripheral],
         outputs: [plasma],
         routes: [
-            infusion(iv) -> central,
+            infusion(acetaminophen_iv) -> central,
         ],
[…]
