+++ benches/runtime_matrix.rs
@@ -1,0 +1,409 @@
+use criterion::{criterion_group, criterion_main, Criterion};
+
+#[cfg(all(
+    feature = "dsl-jit",
+    feature = "dsl-aot",
+    feature = "dsl-aot-load",
+    feature = "dsl-wasm"
+))]
+mod self_contained {
+    use std::error::Error;
+    use std::io;
+    use std::path::PathBuf;
+
+    use pharmsol::dsl::{
+        self, CompiledRuntimeModel, CompiledWasmModule, RuntimeCompilationTarget,
+        RuntimePredictions, WasmCompileCache,
+    };
+    use pharmsol::{Subject, SubjectBuilderExt};
+    use tempfile::{tempdir, TempDir};
+
+    const ODE_SOURCE: &str = r#"
+name = one_cmt_acetaminophen_oral_iv
+kind = ode
+
+params = ka, cl, v, tlag, f_oral
+covariates = wt@linear
+states = depot, central
+derived = cl_i, ke
+outputs = plasma
+
+bolus(acetaminophen_oral) -> depot
+infusion(acetaminophen_iv) -> central
+
+lag(acetaminophen_oral) = tlag
+fa(acetaminophen_oral) = f_oral
+
+cl_i = cl * pow(wt / 70.0, 0.75)
+ke = cl_i / v
+
+dx(depot) = -ka * depot
+dx(central) = ka * depot - ke * central
+
+out(plasma) = central / v ~ continuous()
+"#;
+
+    const ANALYTICAL_SOURCE: &str = r#"
+name = one_cmt_acetaminophen_oral
+kind = analytical
+
+params = ka, ke, v, tlag, f_oral
+states = depot, central
+outputs = plasma
+
+bolus(acetaminophen_oral) -> depot
+
+lag(acetaminophen_oral) = tlag
+fa(acetaminophen_oral) = f_oral
+
+structure = one_compartment_with_absorption
+
+out(plasma) = central / v ~ continuous()
+"#;
+
+    const SDE_SOURCE: &str = r#"
+name = acetaminophen_iv_sde
+kind = sde
+
+params = ka, ke0, kcp, kpc, vol, ske
+covariates = wt@locf
+states = depot, central, peripheral, ke_latent
+particles = 16
+outputs = plasma
+
+bolus(acetaminophen_iv) -> depot
+
+init(ke_latent) = ke0
+
+dx(depot) = -ka * depot
+dx(central) = ka * depot - (ke_latent + kcp) * central + kpc * peripheral
+dx(peripheral) = kcp * central - kpc * peripheral
+dx(ke_latent) = -ke_latent + ke0
+
+noise(ke_latent) = ske
+
+out(plasma) = central / (vol * wt) ~ continuous()
+"#;
+
+    const SDE_PARTICLE_COUNT: usize = 16;
+
+    #[derive(Debug, Clone, Copy, PartialEq, Eq)]
+    pub enum CorpusCase {
+        Ode,
+        Analytical,
+        Sde,
+    }
+
+    impl CorpusCase {
+        pub fn label(self) -> &'static str {
+            match self {
+                Self::Ode => "dsl-ode-one_cmt_acetaminophen_oral_iv",
+                Self::Analytical => "dsl-analytical-one_cmt_acetaminophen_oral",
+                Self::Sde => "dsl-sde-acetaminophen_iv_sde",
+            }
+        }
+
+        fn model_name(self) -> &'static str {
+            match self {
+                Self::Ode => "one_cmt_acetaminophen_oral_iv",
+                Self::Analytical => "one_cmt_acetaminophen_oral",
+                Self::Sde => "acetaminophen_iv_sde",
+            }
+        }
+
+        fn source(self) -> &'static str {
[…]
