+++ src/simulator/likelihood/matrix.rs
@@ -18,19 +18,28 @@ use super::progress::ProgressTracker;
 /// important when dealing with many observations or extreme parameter values that could
 /// cause the regular likelihood to underflow to zero.
 ///
+/// `support_points` must already be a dense matrix in model order. If the
+/// incoming columns are in an external named order, validate that order once
+/// with [`crate::ParameterOrder`] and reorder before calling this function.
+///
 /// # Parameters
 /// - `equation`: The equation to use for simulation
 /// - `subjects`: The subject data
 /// - `support_points`: The support points to evaluate (rows = support points, cols = parameters)
 /// - `error_models`: The error models to use (observation-based sigma)
-/// - `progress`: Whether to display a progress bar during computation`
+/// - `progress`: Whether to display a progress bar during computation
 ///
 /// # Returns
 /// A 2D array of log-likelihoods with shape (n_subjects, n_support_points)
 ///
 /// # Example
 /// ```ignore
-/// use pharmsol::prelude::simulator::log_likelihood_matrix;
+/// use ndarray::array;
+/// use pharmsol::{ParameterOrder, prelude::simulator::log_likelihood_matrix};
+///
+/// let order = ParameterOrder::with_model(&equation, ["ka", "ke"])?;
+/// let support_points_in_source_order = array![[0.1, 0.3], [0.2, 0.4]];
+/// let support_points = order.matrix(support_points_in_source_order)?;
 ///
 /// let log_liks = log_likelihood_matrix(
 ///     &equation,
@@ -138,4 +147,92 @@ pub fn psi(
 
     // Exponentiate to get likelihoods (may underflow to 0 for extreme values)
     Ok(log_psi_matrix.mapv(f64::exp))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::log_likelihood_matrix;
+    use crate::data::builder::SubjectBuilderExt;
+    use crate::data::error_model::{AssayErrorModel, ErrorPoly};
+    use crate::{fa, lag, metadata, AssayErrorModels, Data, ModelKind, ODE, ParameterOrder};
+    use ndarray::array;
+
+    fn likelihood_named_order_ode() -> ODE {
+        ODE::new(
+            |_x, _p, _t, _dx, _b, _rateiv, _cov| {},
+            |_p, _t, _cov| lag! {},
+            |_p, _t, _cov| fa! {},
+            |_p, _t, _cov, x| {
+                x[0] = 0.0;
+            },
+            |_x, p, _t, _cov, y| {
+                y[0] = p[0] - p[1];
+            },
+        )
+        .with_nstates(1)
+        .with_ndrugs(0)
+        .with_nout(1)
+        .with_metadata(
+            metadata::new("likelihood_named_order")
+                .kind(ModelKind::Ode)
+                .parameters(["v", "ke"])
+                .states(["central"])
+                .outputs(["cp"]),
+        )
+        .expect("attach metadata")
+    }
+
+    fn likelihood_error_models() -> AssayErrorModels {
+        AssayErrorModels::empty()
+            .add(
+                0,
+                AssayErrorModel::additive(ErrorPoly::new(0.0, 1.0, 0.0, 0.0), 0.0),
+            )
+            .expect("add error model")
+    }
+
+    #[test]
+    fn parameter_order_feeds_likelihood_matrix_once() {
+        let equation = likelihood_named_order_ode();
+        let data = Data::from(
+            crate::Subject::builder("likelihood-named-order")
+                .observation(0.0, 9.5, "cp")
+                .build(),
+        );
+        let error_models = likelihood_error_models();
+        let manual_support_points = array![[10.0, 0.5], [20.0, 0.7]];
+        let source_order_support_points = array![[0.5, 10.0], [0.7, 20.0]];
[…]

[…]
