### 2026-05-12 e18048d
Update CI
To run new benchmarks behind feature flag
