+++ .github/workflows/base_benchmarks.yml
@@ -28,4 +28,4 @@ jobs:
           --err \
           --adapter rust_criterion \
           --github-actions '${{ secrets.GITHUB_TOKEN }}' \
-          cargo bench
+          cargo bench --all-features
