### 2026-05-12 6a2867e
feat: analytical models now allow covariates to replace primary parameters, detect missing parameters and proper report
