+++ src/simulator/equation/analytical/three_compartment_cl_models.rs
@@ -132,10 +132,10 @@ mod tests {
         .with_ndrugs(3);
 
         let op_ode = ode
-            .estimate_predictions(&subject, &[0.1, 3.0, 2.0, 1.0, 3.0, 4.0])
+            .estimate_predictions(&subject, &crate::Parameters::dense([0.1, 3.0, 2.0, 1.0, 3.0, 4.0]))
             .unwrap();
         let op_analytical = analytical
-            .estimate_predictions(&subject, &[0.1, 3.0, 2.0, 1.0, 3.0, 4.0])
+            .estimate_predictions(&subject, &crate::Parameters::dense([0.1, 3.0, 2.0, 1.0, 3.0, 4.0]))
             .unwrap();
 
         let pred_ode = &op_ode.flat_predictions()[..];
@@ -194,10 +194,10 @@ mod tests {
         );
 
         let op_ode = ode
-            .estimate_predictions(&subject, &[1.0, 0.1, 3.0, 2.0, 1.0, 3.0, 4.0])
+            .estimate_predictions(&subject, &crate::Parameters::dense([1.0, 0.1, 3.0, 2.0, 1.0, 3.0, 4.0]))
             .unwrap();
         let op_analytical = analytical
-            .estimate_predictions(&subject, &[1.0, 0.1, 3.0, 2.0, 1.0, 3.0, 4.0])
+            .estimate_predictions(&subject, &crate::Parameters::dense([1.0, 0.1, 3.0, 2.0, 1.0, 3.0, 4.0]))
             .unwrap();
 
         let pred_ode = &op_ode.flat_predictions()[..];
