+++ src/error/mod.rs
@@ -7,9 +7,9 @@ use crate::data::error_model::ErrorModelError;
 #[cfg(not(all(target_arch = "wasm32", target_os = "unknown")))]
 use crate::data::row::DataError;
 
+use crate::parameters::ParameterError;
 #[cfg(not(all(target_arch = "wasm32", target_os = "unknown")))]
 use crate::CovariateError;
-use crate::parameters::ParameterError;
 
 #[cfg(not(all(target_arch = "wasm32", target_os = "unknown")))]
 use ndarray::ShapeError;
