+++ pharmsol-dsl/src/semantic.rs
@@ -9,8 +9,7 @@ use crate::diagnostic::{
 };
 use crate::ir::*;
 use crate::name_match::{
-    common_prefix_len, edit_distance, is_high_confidence_match,
-    is_single_adjacent_transposition,
+    common_prefix_len, edit_distance, is_high_confidence_match, is_single_adjacent_transposition,
 };
 use crate::{ModelKind, NUMERIC_OUTPUT_PREFIX, NUMERIC_ROUTE_PREFIX, RATE_FUNCTION_NAME};
 
@@ -675,10 +674,7 @@ impl<'a> Analyzer<'a> {
                         .replacement_suggestion(
                             ident.span,
                             format!("{}_derived", ident.text),
-                            format!(
-                                "rename this derive target to `{}_derived`",
-                                ident.text
-                            ),
+                            format!("rename this derive target to `{}_derived`", ident.text),
                             Applicability::MaybeIncorrect,
                         )
                         .apply(SemanticError::new(
@@ -2854,9 +2850,9 @@ mod tests {
         assert!(matches!(ke_symbol.ty, SymbolType::Scalar(ValueType::Real)));
     }
 
-        #[test]
-        fn analytical_model_accepts_straight_line_required_derived_assignment() {
-                let src = r#"
+    #[test]
+    fn analytical_model_accepts_straight_line_required_derived_assignment() {
+        let src = r#"
 model analytical_ok {
     kind analytical
     parameters { ka, ke0, v }
@@ -2874,17 +2870,17 @@ model analytical_ok {
 }
 "#;
 
-                let model = parse_model(src).expect("model parses");
-                let typed = analyze_model(&model).expect("model analyzes");
-                assert!(matches!(
-                        typed.analytical.as_ref().map(|value| value.structure),
-                        Some(AnalyticalKernel::OneCompartmentWithAbsorption)
-                ));
-        }
+        let model = parse_model(src).expect("model parses");
+        let typed = analyze_model(&model).expect("model analyzes");
+        assert!(matches!(
+            typed.analytical.as_ref().map(|value| value.structure),
+            Some(AnalyticalKernel::OneCompartmentWithAbsorption)
+        ));
+    }
 
-        #[test]
-        fn analytical_model_accepts_required_derived_assignment_across_if_else() {
-                let src = r#"
+    #[test]
+    fn analytical_model_accepts_required_derived_assignment_across_if_else() {
+        let src = r#"
 model analytical_ok {
     kind analytical
     parameters { ka, ke0, v }
@@ -2906,13 +2902,13 @@ model analytical_ok {
 }
 "#;
 
-                let model = parse_model(src).expect("model parses");
-                analyze_model(&model).expect("model analyzes");
-        }
+        let model = parse_model(src).expect("model parses");
+        analyze_model(&model).expect("model analyzes");
+    }
 
-        #[test]
-        fn analytical_model_accepts_loop_updates_after_initial_derived_assignment() {
-                let src = r#"
+    #[test]
+    fn analytical_model_accepts_loop_updates_after_initial_derived_assignment() {
+        let src = r#"
 model analytical_ok {
     kind analytical
     parameters { ka, ke0, v }
@@ -2933,13 +2929,13 @@ model analytical_ok {
 }
 "#;
 
-                let model = parse_model(src).expect("model parses");
-                analyze_model(&model).expect("model analyzes");
-        }
+        let model = parse_model(src).expect("model parses");
+        analyze_model(&model).expect("model analyzes");
+    }
 
-        #[test]
-        fn analytical_model_rejects_missing_required_structure_name_across_params_and_derived() {
-                let src = r#"
+    #[test]
[…]
