### 2026-05-13 c769ac8
chore: remove parallel Parameter definition, clippy+fmt
