+++ tests/ode_optimizations.rs
@@ -14,22 +14,154 @@ use pharmsol::*;
 const REL_TOL: f64 = 1e-2; // 1% relative tolerance for most comparisons
 const ABS_TOL: f64 = 1e-6;
 
+fn parameters_for_analytical(
+    label: &str,
+    analytical: &equation::Analytical,
+    param_names: &[&str],
+    params: &[f64],
+) -> Parameters {
+    assert_eq!(
+        param_names.len(),
+        params.len(),
+        "{label}: expected {} parameter value(s), got {}",
+        param_names.len(),
+        params.len()
+    );
+
+    Parameters::with_model(
+        analytical,
+        param_names.iter().copied().zip(params.iter().copied()),
+    )
+    .unwrap_or_else(|error| panic!("{label}: analytical parameters should validate: {error}"))
+}
+
+fn parameters_for_ode(
+    label: &str,
+    ode: &equation::ODE,
+    param_names: &[&str],
+    params: &[f64],
+) -> Parameters {
+    assert_eq!(
+        param_names.len(),
+        params.len(),
+        "{label}: expected {} parameter value(s), got {}",
+        param_names.len(),
+        params.len()
+    );
+
+    Parameters::with_model(ode, param_names.iter().copied().zip(params.iter().copied()))
+        .unwrap_or_else(|error| panic!("{label}: ODE parameters should validate: {error}"))
+}
+
+fn with_one_compartment_analytical_metadata(
+    analytical: equation::Analytical,
+    model_name: &str,
+) -> equation::Analytical {
+    analytical
+        .with_ndrugs(1)
+        .with_metadata(
+            equation::metadata::new(model_name)
+                .kind(equation::ModelKind::Analytical)
+                .parameters(["ke", "v"])
+                .states(["central"])
+                .outputs(["cp"])
+                .routes([
+                    equation::Route::bolus("iv_bolus").to_state("central"),
+                    equation::Route::infusion("iv").to_state("central"),
+                ])
+                .analytical_kernel(equation::AnalyticalKernel::OneCompartment),
+        )
+        .expect("one-compartment analytical metadata should validate")
+}
+
+fn with_one_compartment_ode_metadata(ode: equation::ODE, model_name: &str) -> equation::ODE {
+    ode.with_ndrugs(1)
+        .with_metadata(
+            equation::metadata::new(model_name)
+                .parameters(["ke", "v"])
+                .states(["central"])
+                .outputs(["cp"])
+                .routes([
+                    equation::Route::bolus("iv_bolus")
+                        .to_state("central")
+                        .expect_explicit_input(),
+                    equation::Route::infusion("iv")
+                        .to_state("central")
+                        .expect_explicit_input(),
+                ]),
+        )
+        .expect("one-compartment ODE metadata should validate")
+}
+
+fn with_absorption_analytical_metadata(
+    analytical: equation::Analytical,
+    model_name: &str,
+) -> equation::Analytical {
+    analytical
+        .with_ndrugs(1)
+        .with_metadata(
+            equation::metadata::new(model_name)
+                .kind(equation::ModelKind::Analytical)
+                .parameters(["ka", "ke", "v"])
+                .states(["gut", "central"])
[…]

[…]
