+++ benches/likelihood.rs
@@ -8,9 +8,9 @@ use std::time::Duration;
 
 fn benchmark_equation() -> ODE {
     equation::ODE::new(
-        |x, p, _t, dx, _b, _rateiv, _cov| {
+        |x, p, _t, dx, b, _rateiv, _cov| {
             fetch_params!(p, ka, ke, _tlag, _v);
-            dx[0] = -ka * x[0];
+            dx[0] = -ka * x[0] + b[0];
             dx[1] = ka * x[0] - ke * x[1];
         },
         |p, _t, _cov| {
