+++ src/simulator/equation/analytical/one_compartment_models.rs
@@ -93,10 +93,10 @@ mod tests {
         .with_nout(1);
 
         let op_ode = ode
-            .estimate_predictions(&subject, &crate::Parameters::dense([0.1, 1.0]))
+            .estimate_predictions(&subject, &crate::parameters::dense([0.1, 1.0]))
             .unwrap();
         let op_analytical = analytical
-            .estimate_predictions(&subject, &crate::Parameters::dense([0.1, 1.0]))
+            .estimate_predictions(&subject, &crate::parameters::dense([0.1, 1.0]))
             .unwrap();
 
         let pred_ode = &op_ode.flat_predictions()[..];
@@ -150,10 +150,10 @@ mod tests {
         .with_nout(1);
 
         let op_ode = ode
-            .estimate_predictions(&subject, &crate::Parameters::dense([1.0, 0.1, 1.0]))
+            .estimate_predictions(&subject, &crate::parameters::dense([1.0, 0.1, 1.0]))
             .unwrap();
         let op_analytical = analytical
-            .estimate_predictions(&subject, &crate::Parameters::dense([1.0, 0.1, 1.0]))
+            .estimate_predictions(&subject, &crate::parameters::dense([1.0, 0.1, 1.0]))
             .unwrap();
 
         let pred_ode = &op_ode.flat_predictions()[..];
