+++ src/simulator/equation/analytical/two_compartment_cl_models.rs
@@ -113,10 +113,10 @@ mod tests {
         .with_ndrugs(2);
 
         let op_ode = ode
-            .estimate_predictions(&subject, &crate::Parameters::dense([0.1, 3.0, 1.0, 3.0]))
+            .estimate_predictions(&subject, &crate::parameters::dense([0.1, 3.0, 1.0, 3.0]))
             .unwrap();
         let op_analytical = analytical
-            .estimate_predictions(&subject, &crate::Parameters::dense([0.1, 3.0, 1.0, 3.0]))
+            .estimate_predictions(&subject, &crate::parameters::dense([0.1, 3.0, 1.0, 3.0]))
             .unwrap();
 
         let pred_ode = &op_ode.flat_predictions()[..];
@@ -174,13 +174,13 @@ mod tests {
         let op_ode = ode
             .estimate_predictions(
                 &subject,
-                &crate::Parameters::dense([1.0, 0.1, 3.0, 1.0, 3.0]),
+                &crate::parameters::dense([1.0, 0.1, 3.0, 1.0, 3.0]),
             )
             .unwrap();
         let op_analytical = analytical
             .estimate_predictions(
                 &subject,
-                &crate::Parameters::dense([1.0, 0.1, 3.0, 1.0, 3.0]),
+                &crate::parameters::dense([1.0, 0.1, 3.0, 1.0, 3.0]),
             )
             .unwrap();
 
