+++ benches/common/mod.rs
@@ -1,0 +1,685 @@
+//! Shared bench fixtures for the native + DSL benchmark suites.
+//!
+//! This module centralizes the model definitions, subjects, parameter vectors,
+//! and error-model fixtures that every backend variant in `benches/native_matrix.rs`
+//! and `benches/dsl_matrix.rs` reuses. The goal is to keep all backends
+//! (handwritten, macro, dsl-jit, dsl-aot, dsl-wasm) measuring **the same**
+//! model, subject, and parameters so the only thing that differs between
+//! cells is the engine itself.
+//!
+//! Two workloads are exposed:
+//! - [`Workload::Short`]: 1-compartment, single 100 mg PO bolus at t=0,
+//!   9 observations spread across 12 hours.
+//! - [`Workload::Repeat`]: 2-compartment, 100 mg IV bolus q12h × 10
+//!   (doses at t=0,12,…,108), 14 observations spread across 120 hours.
+//!
+//! For each workload three solver kinds (ODE, Analytical, SDE) and three
+//! authoring/backend flavors (handwritten, macro, DSL-source) are provided.
+//!
+//! Some helpers (`SDE_PARTICLES`, `matrix_data`, `support_points`) are
+//! deliberately public — both bench binaries consume them, but cargo compiles
+//! each bench separately so unused helpers would trigger dead-code lints in
+//! one binary or the other. We silence those with `#![allow(dead_code)]`.
+
+#![allow(dead_code)]
+
+use ndarray::Array2;
+use pharmsol::prelude::*;
+use pharmsol::simulator::equation::analytical::{
+    one_compartment_with_absorption, two_compartments,
+};
+use pharmsol::{
+    equation::{self, Route},
+    Analytical, ResidualErrorModel, ResidualErrorModels, ODE, SDE,
+};
+
+/// Build the `ModelMetadata` for `(workload, kind)`. Handwritten factories
+/// attach this so they resolve string route labels (`"po"`/`"iv"`) and output
+/// labels (`"plasma"`) the same way the `ode!` / `analytical!` / `sde!`
+/// macros and the DSL do.
+fn model_metadata(workload: Workload, kind: SolverKind) -> equation::ModelMetadata {
+    let name = match (workload, kind) {
+        (Workload::Short, SolverKind::Ode) => "bench_one_cpt_po_ode",
+        (Workload::Short, SolverKind::Analytical) => "bench_one_cpt_po_analytical",
+        (Workload::Short, SolverKind::Sde) => "bench_one_cpt_po_sde",
+        (Workload::Repeat, SolverKind::Ode) => "bench_two_cpt_iv_ode",
+        (Workload::Repeat, SolverKind::Analytical) => "bench_two_cpt_iv_analytical",
+        (Workload::Repeat, SolverKind::Sde) => "bench_two_cpt_iv_sde",
+    };
+    match workload {
+        Workload::Short => {
+            let params: &[&str] = match kind {
+                SolverKind::Sde => &["ka", "ke", "v", "sigma_ke"],
+                _ => &["ka", "ke", "v"],
+            };
+            equation::metadata::new(name)
+                .parameters(params.iter().copied())
+                .states(["depot", "central"])
+                .outputs(["plasma"])
+                .route(
+                    Route::bolus("po")
+                        .to_state("depot")
+                        .inject_input_to_destination(),
+                )
+        }
+        Workload::Repeat => {
+            let params: &[&str] = match kind {
+                SolverKind::Sde => &["ke", "kcp", "kpc", "v", "sigma_ke"],
+                _ => &["ke", "kcp", "kpc", "v"],
+            };
+            equation::metadata::new(name)
[…]
