+++ benches/dsl_matrix.rs
@@ -1,0 +1,355 @@
+//! DSL backend benchmark matrix (feature-gated).
+//!
+//! Covers the three DSL backends — JIT, native AoT, WASM — across the same
+//! two workloads and three solver kinds as `native_matrix.rs`.
+//!
+//! Coverage notes:
+//! - Predictions: every (workload, kind, backend) cell.
+//! - Log-likelihood + likelihood-matrix: **ODE only** across all three
+//!   backends. The DSL ODE runtime model is `RuntimeOdeModel = NativeOdeModel`
+//!   (see `src/dsl/runtime.rs:117`), and `NativeOdeModel` implements both the
+//!   `Equation` and `Cache` traits. `NativeAnalyticalModel` and
+//!   `NativeSdeModel` do not, so analytical/SDE backends remain
+//!   predictions-only on this bench surface.
+//! - Hot/cold cache: ODE only, via `.disable_cache()` on the extracted
+//!   `NativeOdeModel`. Analytical/SDE skip the cache axis.
+//! - Compile-time: one bench per (workload, kind, backend) that measures the
+//!   `source -> CompiledRuntimeModel` pipeline end-to-end.
+//!
+//! Bench id format:
+//! - `dsl/compile`               → `{workload}/{kind}/{backend}`
+//! - `dsl/predictions`           → `{workload}/{kind}/{backend}/{cache}` (ODE)
+//!                                 `{workload}/{kind}/{backend}` (Ana/SDE)
+//! - `dsl/log-likelihood`        → `{workload}/ode/{backend}/{cache}`
+//! - `dsl/likelihood-matrix`     → `{workload}/ode/{backend}`
+
+use std::hint::black_box;
+use std::path::PathBuf;
+use std::time::Duration;
+
+use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, SamplingMode};
+use tempfile::TempDir;
+
+use pharmsol::dsl::{
+    compile_module_source_to_runtime, CompiledRuntimeModel, NativeAotCompileOptions, NativeOdeModel,
+    RuntimeCompilationTarget,
+};
+use pharmsol::prelude::*;
+use pharmsol::Cache;
+
+mod common;
+use common::{
+    assay_error_models, dsl_model_name, dsl_source, matrix_data, params, subject_for_likelihood,
+    subject_for_predictions, support_points, SolverKind, Workload,
+};
+
+const MATRIX_N_SUBJECTS: usize = 32;
+const MATRIX_N_SUPPORT: usize = 64;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+enum Backend {
+    Jit,
+    Aot,
+    Wasm,
+}
+
+impl Backend {
+    fn label(self) -> &'static str {
+        match self {
+            Self::Jit => "dsl-jit",
+            Self::Aot => "dsl-aot",
+            Self::Wasm => "dsl-wasm",
+        }
+    }
+
+    fn all() -> [Backend; 3] {
+        [Backend::Jit, Backend::Aot, Backend::Wasm]
+    }
+}
+
+#[derive(Debug, Clone, Copy)]
+enum CacheState {
+    Hot,
+    Cold,
+}
+
+impl CacheState {
+    fn label(self) -> &'static str {
+        match self {
+            Self::Hot => "hot",
+            Self::Cold => "cold",
+        }
+    }
+
+    fn all() -> [CacheState; 2] {
+        [CacheState::Hot, CacheState::Cold]
+    }
+}
+
+/// Per-bench AoT workspace. One `TempDir` for the whole bench binary; each
+/// compile call uses a fresh subdirectory so artifacts don't clash.
+struct AotWorkspace {
+    root: TempDir,
+    counter: std::cell::Cell<usize>,
+}
+
[…]
