+++ src/dsl/native.rs
@@ -24,11 +24,11 @@ use crate::{
     data::error_model::AssayErrorModels,
     data::{Covariates, Infusion, InputLabel, OutputLabel},
     simulator::{
-        cache::{PredictionCache, DEFAULT_CACHE_SIZE},
+        cache::{PredictionCache, SdeLikelihoodCache, DEFAULT_CACHE_SIZE},
         equation::{
             ode::{closure_helpers::PMProblem, ExplicitRkTableau, OdeSolver, SdirkTableau},
             sde::simulate_sde_event_with,
-            EqnKind, Equation, EquationPriv, EquationTypes,
+            EqnKind, Equation, EquationPriv, EquationTypes, Predictions,
         },
         likelihood::{Prediction, SubjectPredictions},
         Fa, Lag, M, T, V,
@@ -745,12 +745,14 @@ pub struct NativeOdeModel {
 #[derive(Clone, Debug)]
 pub struct NativeAnalyticalModel {
     shared: Arc<SharedNativeModel>,
+    cache: Option<PredictionCache>,
 }
 
 #[derive(Clone, Debug)]
 pub struct NativeSdeModel {
     shared: Arc<SharedNativeModel>,
     nparticles: usize,
+    cache: Option<SdeLikelihoodCache>,
 }
 
 #[derive(Clone, Debug)]
@@ -1224,6 +1226,7 @@ impl NativeAnalyticalModel {
     pub(crate) fn new(info: NativeModelInfo, artifact: impl RuntimeArtifact + 'static) -> Self {
         Self {
             shared: Arc::new(SharedNativeModel::new(info, artifact)),
+            cache: Some(PredictionCache::new(DEFAULT_CACHE_SIZE)),
         }
     }
 
@@ -1377,12 +1380,189 @@ impl NativeAnalyticalModel {
     }
 }
 
+#[inline(always)]
+fn runtime_analytical_predictions(
+    model: &NativeAnalyticalModel,
+    subject: &Subject,
+    support_point: &[f64],
+) -> Result<SubjectPredictions, PharmsolError> {
+    if let Some(cache) = &model.cache {
+        let key = (
+            subject.hash(),
+            crate::simulator::equation::spphash(support_point),
+        );
+        if let Some(cached) = cache.get(&key) {
+            return Ok(cached);
+        }
+
+        let result = model.estimate_predictions(subject, support_point)?;
+        cache.insert(key, result.clone());
+        Ok(result)
+    } else {
+        model.estimate_predictions(subject, support_point)
+    }
+}
+
+impl crate::simulator::equation::Cache for NativeAnalyticalModel {
+    fn with_cache_capacity(mut self, size: u64) -> Self {
+        self.cache = Some(PredictionCache::new(size));
+        self
+    }
+
+    fn enable_cache(mut self) -> Self {
+        self.cache = Some(PredictionCache::new(DEFAULT_CACHE_SIZE));
+        self
+    }
+
+    fn clear_cache(&self) {
+        if let Some(cache) = &self.cache {
+            cache.invalidate_all();
+        }
+    }
+
+    fn disable_cache(mut self) -> Self {
+        self.cache = None;
+        self
+    }
+}
+
+impl EquationTypes for NativeAnalyticalModel {
+    type S = V;
+    type P = SubjectPredictions;
+}
+
+impl EquationPriv for NativeAnalyticalModel {
+    fn lag(&self) -> &Lag {
+        &(runtime_no_lag as Lag)
+    }
+
+    fn fa(&self) -> &Fa {
+        &(runtime_no_fa as Fa)
+    }
+
+    fn get_nstates(&self) -> usize {
+        self.shared.info.state_len
+    }
+
[…]
