### 2026-05-14 ead8c82
chore: Refactor benchmarks (#258)
* chore: Refactor benchmarks
