+++ benches/native_matrix.rs
@@ -1,0 +1,570 @@
+//! Native bench matrix: handwritten + macro models, all solvers, both workloads.
+//! Default features only — this is the CI regression signal. DSL backends are in `dsl_matrix.rs`.
+//!
+//! IDs:
+//! - `native/predictions`       → `{workload}/{solver}/{authoring}/{cache}`
+//! - `native/log-likelihood`    → `{workload}/{solver}/{authoring}/{cache}`
+//! - `native/likelihood-matrix` → `{workload}/{solver}/{authoring}`
+
+use std::hint::black_box;
+use std::time::Duration;
+
+use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, SamplingMode};
+use pharmsol::prelude::*;
+use pharmsol::Cache;
+
+mod common;
+use common::{
+    assay_error_models, handwritten_analytical, handwritten_ode, handwritten_sde, macro_analytical,
+    macro_ode, macro_sde, matrix_data, params, subject_for_likelihood, subject_for_predictions,
+    support_points, SolverKind, Workload,
+};
+
+const MATRIX_N_SUBJECTS: usize = 32;
+const MATRIX_N_SUPPORT: usize = 64;
+
+#[derive(Debug, Clone, Copy)]
+enum Authoring {
+    Handwritten,
+    Macro,
+}
+
+impl Authoring {
+    fn label(self) -> &'static str {
+        match self {
+            Self::Handwritten => "handwritten",
+            Self::Macro => "macro",
+        }
+    }
+
+    fn all() -> [Authoring; 2] {
+        [Authoring::Handwritten, Authoring::Macro]
+    }
+}
+
+#[derive(Debug, Clone, Copy)]
+enum CacheState {
+    Hot,
+    Cold,
+}
+
+impl CacheState {
+    fn label(self) -> &'static str {
+        match self {
+            Self::Hot => "hot",
+            Self::Cold => "cold",
+        }
+    }
+
+    fn all() -> [CacheState; 2] {
+        [CacheState::Hot, CacheState::Cold]
+    }
+}
+
+fn id(workload: Workload, kind: SolverKind, authoring: Authoring, cache: CacheState) -> String {
+    format!(
+        "{}/{}/{}/{}",
+        workload.label(),
+        kind.label(),
+        authoring.label(),
+        cache.label()
+    )
+}
+
+fn predictions_group(c: &mut Criterion) {
+    let mut group = c.benchmark_group("native/predictions");
+    group.sampling_mode(SamplingMode::Flat);
+
+    for workload in Workload::all() {
+        let subject = subject_for_predictions(workload);
+        for kind in SolverKind::all() {
+            let theta = params(workload, kind);
+            for authoring in Authoring::all() {
+                for cache in CacheState::all() {
+                    let bench_id =
+                        BenchmarkId::from_parameter(id(workload, kind, authoring, cache));
+                    group.bench_function(bench_id, |b| match (kind, authoring, cache) {
+                        (SolverKind::Ode, Authoring::Handwritten, CacheState::Hot) => {
+                            let model = handwritten_ode(workload);
+                            b.iter(|| {
+                                black_box(
+                                    model
+                                        .estimate_predictions(
+                                            black_box(&subject),
+                                            black_box(&theta),
+                                        )
+                                        .unwrap(),
+                                )
+                            });
+                        }
+                        (SolverKind::Ode, Authoring::Handwritten, CacheState::Cold) => {
+                            let model = handwritten_ode(workload).disable_cache();
+                            b.iter(|| {
+                                black_box(
+                                    model
+                                        .estimate_predictions(
+                                            black_box(&subject),
+                                            black_box(&theta),
+                                        )
+                                        .unwrap(),
+                                )
+                            });
+                        }
[…]
