+++ .github/workflows/pr_benchmarks.yml
@@ -29,4 +29,4 @@ jobs:
           --err \
           --adapter rust_criterion \
           --github-actions '${{ secrets.GITHUB_TOKEN }}' \
-          cargo bench
+          cargo bench --all-features
