### 2026-05-14 56e9a23
refactor: Rename support points to parameters (#251)
Parameters are now named. Analytic structures are parameter-order agnostic and there is proper error propagation.
---------

Co-authored-by: Julián D. Otálvaro <juliandavid347@gmail.com>
