+++ pharmsol-dsl/src/ir.rs
@@ -1,5 +1,11 @@
+use std::error::Error;
+use std::fmt;
+
 use serde::{Deserialize, Serialize};
 
+use crate::name_match::{
+    common_prefix_len, edit_distance, is_high_confidence_match, is_single_adjacent_transposition,
+};
 use crate::{ModelKind, RouteKind, Span};
 
 pub type SymbolId = usize;
@@ -207,6 +213,44 @@ impl AnalyticalKernel {
         }
     }
 
+    pub fn name(self) -> &'static str {
+        match self {
+            Self::OneCompartment => "one_compartment",
+            Self::OneCompartmentCl => "one_compartment_cl",
+            Self::OneCompartmentClWithAbsorption => "one_compartment_cl_with_absorption",
+            Self::OneCompartmentWithAbsorption => "one_compartment_with_absorption",
+            Self::TwoCompartments => "two_compartments",
+            Self::TwoCompartmentsCl => "two_compartments_cl",
+            Self::TwoCompartmentsClWithAbsorption => "two_compartments_cl_with_absorption",
+            Self::TwoCompartmentsWithAbsorption => "two_compartments_with_absorption",
+            Self::ThreeCompartments => "three_compartments",
+            Self::ThreeCompartmentsCl => "three_compartments_cl",
+            Self::ThreeCompartmentsClWithAbsorption => "three_compartments_cl_with_absorption",
+            Self::ThreeCompartmentsWithAbsorption => "three_compartments_with_absorption",
+        }
+    }
+
+    pub fn required_parameter_names(self) -> &'static [&'static str] {
+        match self {
+            Self::OneCompartment => &["ke"],
+            Self::OneCompartmentCl => &["cl", "v"],
+            Self::OneCompartmentClWithAbsorption => &["ka", "cl", "v"],
+            Self::OneCompartmentWithAbsorption => &["ka", "ke"],
+            Self::TwoCompartments => &["ke", "kcp", "kpc"],
+            Self::TwoCompartmentsCl => &["cl", "q", "vc", "vp"],
+            Self::TwoCompartmentsClWithAbsorption => &["ka", "cl", "q", "vc", "vp"],
+            Self::TwoCompartmentsWithAbsorption => &["ke", "ka", "kcp", "kpc"],
+            Self::ThreeCompartments => &["k10", "k12", "k13", "k21", "k31"],
+            Self::ThreeCompartmentsCl => &["cl", "q2", "q3", "vc", "v2", "v3"],
+            Self::ThreeCompartmentsClWithAbsorption => &["ka", "cl", "q2", "q3", "vc", "v2", "v3"],
+            Self::ThreeCompartmentsWithAbsorption => &["ka", "k10", "k12", "k13", "k21", "k31"],
+        }
+    }
+
+    pub fn required_parameter_count(self) -> usize {
+        self.required_parameter_names().len()
+    }
+
     pub fn state_count(self) -> usize {
         match self {
             Self::OneCompartment | Self::OneCompartmentCl => 1,
@@ -216,6 +260,251 @@ impl AnalyticalKernel {
             Self::ThreeCompartments | Self::ThreeCompartmentsCl => 3,
             Self::ThreeCompartmentsClWithAbsorption | Self::ThreeCompartmentsWithAbsorption => 4,
         }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum AnalyticalStructureInputSource {
+    Primary,
+    Derived,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct AnalyticalStructureInputBinding {
[…]
