+++ benches/native_matrix.rs
@@ -11,13 +11,13 @@ use std::time::Duration;
 
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, SamplingMode};
 use pharmsol::prelude::*;
-use pharmsol::Cache;
+use pharmsol::{Analytical, Cache, Parameters, ODE, SDE};
 
 mod common;
 use common::{
     assay_error_models, handwritten_analytical, handwritten_ode, handwritten_sde, macro_analytical,
-    macro_ode, macro_sde, matrix_data, params, subject_for_likelihood, subject_for_predictions,
-    support_points, SolverKind, Workload,
+    macro_ode, macro_sde, matrix_data, named_params, subject_for_likelihood,
+    subject_for_predictions, support_points, SolverKind, Workload,
 };
 
 const MATRIX_N_SUBJECTS: usize = 32;
@@ -71,6 +71,21 @@ fn id(workload: Workload, kind: SolverKind, authoring: Authoring, cache: CacheSt
     )
 }
 
+fn ode_parameters(model: &ODE, workload: Workload) -> Parameters {
+    Parameters::with_model(model, named_params(workload, SolverKind::Ode))
+        .expect("native ODE bench parameters should validate")
+}
+
+fn analytical_parameters(model: &Analytical, workload: Workload) -> Parameters {
+    Parameters::with_model(model, named_params(workload, SolverKind::Analytical))
+        .expect("native analytical bench parameters should validate")
+}
+
+fn sde_parameters(model: &SDE, workload: Workload) -> Parameters {
+    Parameters::with_model(model, named_params(workload, SolverKind::Sde))
+        .expect("native SDE bench parameters should validate")
+}
+
 fn predictions_group(c: &mut Criterion) {
     let mut group = c.benchmark_group("native/predictions");
     group.sampling_mode(SamplingMode::Flat);
@@ -78,7 +93,6 @@ fn predictions_group(c: &mut Criterion) {
     for workload in Workload::all() {
         let subject = subject_for_predictions(workload);
         for kind in SolverKind::all() {
-            let theta = params(workload, kind);
             for authoring in Authoring::all() {
                 for cache in CacheState::all() {
                     let bench_id =
@@ -86,6 +100,7 @@ fn predictions_group(c: &mut Criterion) {
                     group.bench_function(bench_id, |b| match (kind, authoring, cache) {
                         (SolverKind::Ode, Authoring::Handwritten, CacheState::Hot) => {
                             let model = handwritten_ode(workload);
+                            let theta = ode_parameters(&model, workload);
                             b.iter(|| {
                                 black_box(
                                     model
@@ -99,6 +114,7 @@ fn predictions_group(c: &mut Criterion) {
                         }
                         (SolverKind::Ode, Authoring::Handwritten, CacheState::Cold) => {
                             let model = handwritten_ode(workload).disable_cache();
+                            let theta = ode_parameters(&model, workload);
                             b.iter(|| {
                                 black_box(
                                     model
@@ -112,6 +128,7 @@ fn predictions_group(c: &mut Criterion) {
                         }
                         (SolverKind::Ode, Authoring::Macro, CacheState::Hot) => {
                             let model = macro_ode(workload);
+                            let theta = ode_parameters(&model, workload);
                             b.iter(|| {
                                 black_box(
                                     model
@@ -125,6 +142,7 @@ fn predictions_group(c: &mut Criterion) {
                         }
                         (SolverKind::Ode, Authoring::Macro, CacheState::Cold) => {
                             let model = macro_ode(workload).disable_cache();
+                            let theta = ode_parameters(&model, workload);
                             b.iter(|| {
                                 black_box(
                                     model
@@ -138,6 +156,7 @@ fn predictions_group(c: &mut Criterion) {
[…]
