+++ pharmsol-dsl/Cargo.toml
@@ -6,7 +6,8 @@ authors = ["Julián D. Otálvaro <juliandavid347@gmail.com>", "Markus Hovd"]
 description = "Backend-neutral frontend crate for the pharmsol model DSL."
 license = "GPL-3.0"
 readme = "README.md"
-publish = false
+repository = "https://github.com/LAPKB/pharmsol"
+publish = true
 
 [lib]
 bench = false
