+++ Cargo.toml
@@ -2,14 +2,28 @@
 members = [".", "pharmsol-dsl", "pharmsol-macros"]
 resolver = "2"
 
-[package]
-name = "pharmsol"
+[workspace.package]
 version = "0.27.0"
 edition = "2021"
-autobenches = false
-authors = ["Julián D. Otálvaro <juliandavid347@gmail.com>", "Markus Hovd"]
-description = "Rust library for solving analytic and ode-defined pharmacometric models."
 license = "GPL-3.0"
+authors = ["Julián D. Otálvaro <juliandavid347@gmail.com>", "Markus Hovd"]
+repository = "https://github.com/LAPKB/pharmsol"
+
+[workspace.dependencies]
+pharmsol-dsl = { path = "pharmsol-dsl", version = "0.27.0" }
+pharmsol-macros = { path = "pharmsol-macros", version = "0.27.0" }
+serde = { version = "1.0.228", features = ["derive"] }
+serde_json = "1.0.149"
+
+[package]
+name = "pharmsol"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+autobenches = false
+description = "Rust library for solving analytic and ode-defined pharmacometric models."
 documentation = "https://lapkb.github.io/pharmsol/"
 
 [features]
@@ -29,15 +43,15 @@ dsl-wasm = ["dsl-wasm-compile", "dep:wasmtime"]
 wasm = ["dsl-wasm"]
 
 [dependencies]
-pharmsol-dsl = { path = "pharmsol-dsl", version = "0.1.0" }
-pharmsol-macros = { path = "pharmsol-macros", version = "0.1.1" }
+pharmsol-dsl = { workspace = true }
+pharmsol-macros = { workspace = true }
 libloading = { version = "0.9.0", optional = true, features = [] }
 cranelift = { version = "0.131.0", optional = true }
 cranelift-jit = { version = "0.131.0", optional = true }
 cranelift-module = { version = "0.131.0", optional = true }
 cranelift-native = { version = "0.131.0", optional = true }
-serde = { version = "1.0.228", features = ["derive"] }
-serde_json = "1.0.149"
+serde = { workspace = true }
+serde_json = { workspace = true }
 thiserror = "2.0.18"
 tracing = "0.1.44"
 wasm-encoder = { version = "0.247.0", optional = true }
