+++ pharmsol-macros/Cargo.toml
@@ -1,17 +1,18 @@
 [package]
 name = "pharmsol-macros"
-version = "0.1.1"
-edition = "2021"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
 description = "Procedural macros for pharmsol. Not intended for direct use."
-license = "GPL-3.0"
-repository = "https://github.com/LAPKB/pharmsol"
 readme = "README.md"
 
 [lib]
 proc-macro = true
 
 [dependencies]
-pharmsol-dsl = { path = "../pharmsol-dsl" }
+pharmsol-dsl = { workspace = true }
 proc-macro2 = "1.0.106"
 quote = "1.0.45"
 syn = { version = "2.0.117", features = ["full", "visit", "visit-mut"] }
