+++ pharmsol-dsl/Cargo.toml
@@ -1,17 +1,17 @@
 [package]
 name = "pharmsol-dsl"
-version = "0.1.0"
-edition = "2021"
-authors = ["Julián D. Otálvaro <juliandavid347@gmail.com>", "Markus Hovd"]
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
 description = "Backend-neutral frontend crate for the pharmsol model DSL."
-license = "GPL-3.0"
 readme = "README.md"
-repository = "https://github.com/LAPKB/pharmsol"
 publish = true
 
 [lib]
 bench = false
 
 [dependencies]
-serde = { version = "1.0.228", features = ["derive"] }
-serde_json = "1.0.149"
+serde = { workspace = true }
+serde_json = { workspace = true }
