### 2026-05-14 8df3664
chore: Use cargo workspaces (#260)
* Update Cargo.toml

* Update workspace

* Ensure packages are released together
