### 2026-05-14 ff4d607
chore: release v0.27.0 (#261)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
