+++ pharmsol-dsl/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.27.0](https://github.com/LAPKB/pharmsol/compare/pharmsol-macros-v0.1.0...pharmsol-macros-v0.27.0) - 2026-05-14
+## [0.27.0](https://github.com/LAPKB/pharmsol/releases/tag/pharmsol-dsl-v0.27.0) - 2026-05-14
 
 ### Added
 
@@ -16,17 +16,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Other
 
 - Use cargo workspaces ([#260](https://github.com/LAPKB/pharmsol/pull/260))
-- release ([#259](https://github.com/LAPKB/pharmsol/pull/259))
-- Rename support points to parameters ([#251](https://github.com/LAPKB/pharmsol/pull/251))
-- add missing macros ([#253](https://github.com/LAPKB/pharmsol/pull/253))
-
-## [0.1.1](https://github.com/LAPKB/pharmsol/compare/pharmsol-macros-v0.1.0...pharmsol-macros-v0.1.1) - 2026-05-14
-
-### Added
-
-- Domain Specific Language (DSL) using JIT or AOT ([#252](https://github.com/LAPKB/pharmsol/pull/252))
-
-### Other
-
 - Rename support points to parameters ([#251](https://github.com/LAPKB/pharmsol/pull/251))
 - add missing macros ([#253](https://github.com/LAPKB/pharmsol/pull/253))
