### 2026-05-14 71769ca
feature: Metadata contract
