+++ src/dsl/native.rs
@@ -16,12 +16,12 @@ use cranelift_jit::JITModule;
 use libloading::Library;
 use pharmsol_dsl::execution::KernelRole;
 use pharmsol_dsl::{
-    AnalyticalKernel, AnalyticalStructureInputKind, AnalyticalStructureInputPlan, RouteKind,
-    NUMERIC_OUTPUT_PREFIX, NUMERIC_ROUTE_PREFIX,
+    AnalyticalKernel, AnalyticalStructureInputKind, AnalyticalStructureInputPlan, ModelKind,
+    RouteKind, NUMERIC_OUTPUT_PREFIX, NUMERIC_ROUTE_PREFIX,
 };
 
 pub use super::model_info::{
-    NativeCovariateInfo, NativeModelInfo, NativeOutputInfo, NativeRouteInfo,
+    NativeCovariateInfo, NativeModelInfo, NativeOutputInfo, NativeRouteInfo, NativeStateInfo,
 };
 use crate::{
     data::error_model::AssayErrorModels,
@@ -39,7 +39,7 @@ use crate::{
         likelihood::{Prediction, SubjectPredictions},
         Fa, Lag, M, T, V,
     },
-    Event, Observation, Occasion, Parameters, PharmsolError, Subject,
+    Event, Observation, Occasion, Parameters, PharmsolError, Subject, ValidatedModelMetadata,
 };
 
 pub type DenseKernelFn = unsafe extern "C" fn(
@@ -274,8 +274,207 @@ impl RuntimeArtifact for NativeExecutionArtifact {
 #[derive(Clone, Debug)]
 struct SharedNativeModel {
     info: Arc<NativeModelInfo>,
+    metadata: Arc<ValidatedModelMetadata>,
     route_semantics: Arc<RouteInputSemantics>,
     artifact: Arc<dyn RuntimeArtifact>,
+}
+
+fn compiled_metadata_error(info: &NativeModelInfo, detail: impl Into<String>) -> PharmsolError {
+    PharmsolError::OtherError(format!(
+        "compiled model `{}` has invalid runtime metadata: {}",
+        info.name,
+        detail.into()
+    ))
+}
+
+fn sorted_dense_metadata<'a, T>(
+    info: &NativeModelInfo,
+    domain: &str,
+    expected_len: usize,
+    entries: &'a [T],
+    index_of: impl Fn(&T) -> usize,
+) -> Result<Vec<&'a T>, PharmsolError> {
+    if entries.len() != expected_len {
+        return Err(compiled_metadata_error(
+            info,
+            format!(
+                "expected {expected_len} {domain} entr{} but found {}",
+                if expected_len == 1 { "y" } else { "ies" },
+                entries.len()
+            ),
+        ));
+    }
+
+    let mut sorted = entries.iter().collect::<Vec<_>>();
+    sorted.sort_by_key(|entry| index_of(entry));
+    for (expected, entry) in sorted.iter().enumerate() {
+        let found = index_of(entry);
+        if found != expected {
+            return Err(compiled_metadata_error(
+                info,
+                format!(
+                    "{domain} metadata must use dense 0-based indices; expected {expected}, found {found}"
+                ),
+            ));
+        }
+    }
+
+    Ok(sorted)
+}
+
+fn runtime_model_metadata(info: &NativeModelInfo) -> Result<ValidatedModelMetadata, PharmsolError> {
+    let states = sorted_dense_metadata(info, "state", info.state_len, &info.states, |state| {
+        state.offset
+    })?;
+    let state_names = states
+        .iter()
+        .map(|state| state.name.clone())
+        .collect::<Vec<_>>();
+
+    let covariates = sorted_dense_metadata(
+        info,
+        "covariate",
+        info.covariates.len(),
[…]
