### 2026-05-14 4a7e67f
fix: addressing some issues
