+++ src/dsl/native.rs
@@ -322,10 +322,104 @@ fn sorted_dense_metadata<'a, T>(
     Ok(sorted)
 }
 
+fn sorted_state_metadata<'a>(
+    info: &'a NativeModelInfo,
+) -> Result<Vec<&'a NativeStateInfo>, PharmsolError> {
+    if info.state_len == 0 {
+        if info.states.is_empty() {
+            return Ok(Vec::new());
+        }
+        return Err(compiled_metadata_error(
+            info,
+            format!(
+                "expected no state metadata for an empty state buffer, found {} declaration(s)",
+                info.states.len()
+            ),
+        ));
+    }
+
+    if info.states.is_empty() {
+        return Err(compiled_metadata_error(
+            info,
+            format!(
+                "expected state metadata for {} state slot(s), found none",
+                info.state_len
+            ),
+        ));
+    }
+
+    let mut states = info.states.iter().collect::<Vec<_>>();
+    states.sort_by_key(|state| state.offset);
+
+    if states[0].offset != 0 {
+        return Err(compiled_metadata_error(
+            info,
+            format!(
+                "state metadata must start at offset 0; first declaration starts at {}",
+                states[0].offset
+            ),
+        ));
+    }
+
+    for window in states.windows(2) {
+        let current = window[0];
+        let next = window[1];
+        if next.offset <= current.offset {
+            return Err(compiled_metadata_error(
+                info,
+                format!(
+                    "state metadata offsets must be strictly increasing; saw {} followed by {}",
+                    current.offset, next.offset
+                ),
+            ));
+        }
+    }
+
+    let last_offset = states.last().expect("non-empty states").offset;
+    if last_offset >= info.state_len {
+        return Err(compiled_metadata_error(
+            info,
+            format!(
+                "state metadata offset {} is out of range for state buffer length {}",
+                last_offset, info.state_len
+            ),
+        ));
+    }
+
+    Ok(states)
+}
+
+fn state_declaration_for_offset<'a>(
+    info: &NativeModelInfo,
+    states: &[&'a NativeStateInfo],
+    offset: usize,
+) -> Result<(usize, &'a NativeStateInfo), PharmsolError> {
+    if offset >= info.state_len {
+        return Err(compiled_metadata_error(
+            info,
+            format!(
+                "state offset {} is out of range for state buffer length {}",
+                offset, info.state_len
+            ),
+        ));
+    }
+
+    let declaration_index = match states.binary_search_by_key(&offset, |state| state.offset) {
+        Ok(index) => index,
+        Err(0) => {
+            return Err(compiled_metadata_error(
+                info,
+                format!("state offset {} precedes the first declared state", offset),
+            ));
+        }
+        Err(index) => index - 1,
+    };
+
+    Ok((declaration_index, states[declaration_index]))
+}
+
 fn runtime_model_metadata(info: &NativeModelInfo) -> Result<ValidatedModelMetadata, PharmsolError> {
-    let states = sorted_dense_metadata(info, "state", info.state_len, &info.states, |state| {
-        state.offset
-    })?;
+    let states = sorted_state_metadata(info)?;
     let state_names = states
         .iter()
         .map(|state| state.name.clone())
@@ -373,18 +467,9 @@ fn runtime_model_metadata(info: &NativeModelInfo) -> Result<ValidatedModelMetada
     }
 
     for route in &routes {
-        let destination = state_names
-            .get(route.destination_offset)
-            .ok_or_else(|| {
-                compiled_metadata_error(
-                    info,
-                    format!(
-                        "route `{}` targets out-of-range state offset {}",
-                        route.name, route.destination_offset
-                    ),
-                )
-            })?
-            .clone();
[…]
