### 2026-05-14 1932cc1
chore: New error variant
To reduce use of OtherError
