+++ src/dsl/native.rs
@@ -279,14 +279,6 @@ struct SharedNativeModel {
     artifact: Arc<dyn RuntimeArtifact>,
 }
 
-fn compiled_metadata_error(info: &NativeModelInfo, detail: impl Into<String>) -> PharmsolError {
-    PharmsolError::OtherError(format!(
-        "compiled model `{}` has invalid runtime metadata: {}",
-        info.name,
-        detail.into()
-    ))
-}
-
 fn sorted_dense_metadata<'a, T>(
     info: &NativeModelInfo,
     domain: &str,
@@ -295,14 +287,14 @@ fn sorted_dense_metadata<'a, T>(
     index_of: impl Fn(&T) -> usize,
 ) -> Result<Vec<&'a T>, PharmsolError> {
     if entries.len() != expected_len {
-        return Err(compiled_metadata_error(
-            info,
-            format!(
+        return Err(PharmsolError::InvalidMetadata {
+            model: info.name.clone(),
+            detail: format!(
                 "expected {expected_len} {domain} entr{} but found {}",
                 if expected_len == 1 { "y" } else { "ies" },
                 entries.len()
             ),
-        ));
+        });
     }
 
     let mut sorted = entries.iter().collect::<Vec<_>>();
@@ -310,12 +302,12 @@ fn sorted_dense_metadata<'a, T>(
     for (expected, entry) in sorted.iter().enumerate() {
         let found = index_of(entry);
         if found != expected {
-            return Err(compiled_metadata_error(
-                info,
-                format!(
+            return Err(PharmsolError::InvalidMetadata {
+                model: info.name.clone(),
+                detail: format!(
                     "{domain} metadata must use dense 0-based indices; expected {expected}, found {found}"
                 ),
-            ));
+            });
         }
     }
 
@@ -329,61 +321,61 @@ fn sorted_state_metadata<'a>(
         if info.states.is_empty() {
             return Ok(Vec::new());
         }
-        return Err(compiled_metadata_error(
-            info,
-            format!(
+        return Err(PharmsolError::InvalidMetadata {
+            model: info.name.clone(),
+            detail: format!(
                 "expected no state metadata for an empty state buffer, found {} declaration(s)",
                 info.states.len()
             ),
-        ));
+        });
     }
 
     if info.states.is_empty() {
-        return Err(compiled_metadata_error(
-            info,
-            format!(
+        return Err(PharmsolError::InvalidMetadata {
+            model: info.name.clone(),
+            detail: format!(
                 "expected state metadata for {} state slot(s), found none",
                 info.state_len
             ),
-        ));
+        });
     }
 
     let mut states = info.states.iter().collect::<Vec<_>>();
     states.sort_by_key(|state| state.offset);
 
     if states[0].offset != 0 {
-        return Err(compiled_metadata_error(
-            info,
-            format!(
+        return Err(PharmsolError::InvalidMetadata {
+            model: info.name.clone(),
+            detail: format!(
                 "state metadata must start at offset 0; first declaration starts at {}",
                 states[0].offset
             ),
-        ));
+        });
     }
 
     for window in states.windows(2) {
         let current = window[0];
         let next = window[1];
         if next.offset <= current.offset {
-            return Err(compiled_metadata_error(
-                info,
-                format!(
+            return Err(PharmsolError::InvalidMetadata {
+                model: info.name.clone(),
+                detail: format!(
                     "state metadata offsets must be strictly increasing; saw {} followed by {}",
                     current.offset, next.offset
                 ),
-            ));
+            });
         }
     }
 
     let last_offset = states.last().expect("non-empty states").offset;
     if last_offset >= info.state_len {
-        return Err(compiled_metadata_error(
-            info,
[…]
