### 2026-05-14 8029aa4
feature: Metadata contract (#262)
* feature: Metadata contract

* fix: addressing some issues

* chore: New error variant

To reduce use of OtherError

* fix: c->C it is always C...

---------

Co-authored-by: Markus <markushh@uio.no>
