### 2026-05-15 #10580 @editorialbot
[REVIEW]: pharmsol: A high-performance Rust library for pharmacokinetic/pharmacodynamic modeling and simulation
**Submitting author:** @mhovd (<a href="http://orcid.org/0000-0002-6077-0934">Markus Hovd</a>)
**Repository:** https://github.com/LAPKB/pharmsol
**Branch with paper.md** (empty if default branch): joss
**Version:** pharmsol-v0.28.4
