+++ src/dsl/native.rs
@@ -109,8 +109,12 @@ impl std::fmt::Debug for NativeArtifactOwner {
     }
 }
 
-pub struct NativeExecutionArtifact {
-    pub model_name: String,
+/// Table of compiled native kernels indexed by [`KernelRole`].
+///
+/// `outputs` is the only required kernel; every other role is optional and
+/// absent when the source model does not declare it.
+#[derive(Clone, Copy)]
+pub struct NativeKernels {
     pub derive: Option<DenseKernelFn>,
     pub dynamics: Option<DenseKernelFn>,
     pub outputs: DenseKernelFn,
@@ -119,6 +123,49 @@ pub struct NativeExecutionArtifact {
     pub diffusion: Option<DenseKernelFn>,
     pub route_lag: Option<DenseKernelFn>,
     pub route_bioavailability: Option<DenseKernelFn>,
+}
+
+impl NativeKernels {
+    /// Look up the kernel registered for `role`, treating `Outputs` as always
+    /// present and `Analytical` as never having a dense kernel.
+    pub(crate) fn get(&self, role: KernelRole) -> Option<DenseKernelFn> {
+        match role {
+            KernelRole::Derive => self.derive,
+            KernelRole::Dynamics => self.dynamics,
+            KernelRole::Outputs => Some(self.outputs),
+            KernelRole::Init => self.init,
+            KernelRole::Drift => self.drift,
+            KernelRole::Diffusion => self.diffusion,
+            KernelRole::RouteLag => self.route_lag,
+            KernelRole::RouteBioavailability => self.route_bioavailability,
+            KernelRole::Analytical => None,
+        }
+    }
+
+    pub(crate) fn has(&self, role: KernelRole) -> bool {
+        self.get(role).is_some()
+    }
+}
+
+impl std::fmt::Debug for NativeKernels {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        let as_ptr = |k: Option<DenseKernelFn>| k.map(|ptr| ptr as *const ());
+        f.debug_struct("NativeKernels")
+            .field("derive", &as_ptr(self.derive))
+            .field("dynamics", &as_ptr(self.dynamics))
+            .field("outputs", &(self.outputs as *const ()))
+            .field("init", &as_ptr(self.init))
+            .field("drift", &as_ptr(self.drift))
+            .field("diffusion", &as_ptr(self.diffusion))
+            .field("route_lag", &as_ptr(self.route_lag))
+            .field("route_bioavailability", &as_ptr(self.route_bioavailability))
+            .finish()
+    }
+}
+
+pub struct NativeExecutionArtifact {
+    pub model_name: String,
+    pub kernels: NativeKernels,
     _owner: Option<NativeArtifactOwner>,
 }
 
@@ -129,74 +176,34 @@ impl std::fmt::Debug for NativeExecutionArtifact {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("NativeExecutionArtifact")
             .field("model_name", &self.model_name)
-            .field("derive", &self.derive.map(|ptr| ptr as *const ()))
-            .field("dynamics", &self.dynamics.map(|ptr| ptr as *const ()))
-            .field("outputs", &(self.outputs as *const ()))
-            .field("init", &self.init.map(|ptr| ptr as *const ()))
-            .field("drift", &self.drift.map(|ptr| ptr as *const ()))
-            .field("diffusion", &self.diffusion.map(|ptr| ptr as *const ()))
-            .field("route_lag", &self.route_lag.map(|ptr| ptr as *const ()))
-            .field(
-                "route_bioavailability",
-                &self.route_bioavailability.map(|ptr| ptr as *const ()),
-            )
[…]
