+++ src/dsl/wasm.rs
@@ -2,15 +2,13 @@ use std::ops::Range;
 use std::path::Path;
 use std::sync::Mutex;
 
-#[allow(unused)]
-use crate::dsl::model_info;
 use wasmtime::{Engine, Instance, Linker, Memory, Module, Store, TypedFunc};
 
 use super::compiled_backend_abi::{
-    decode_compiled_model_info, CompiledKernelAvailability, ALLOC_F64_BUFFER_SYMBOL,
-    API_VERSION_SYMBOL, DERIVE_SYMBOL, DIFFUSION_SYMBOL, DRIFT_SYMBOL, DYNAMICS_SYMBOL,
-    FREE_F64_BUFFER_SYMBOL, INIT_SYMBOL, MODEL_INFO_JSON_LEN_SYMBOL, MODEL_INFO_JSON_PTR_SYMBOL,
-    OUTPUTS_SYMBOL, ROUTE_BIOAVAILABILITY_SYMBOL, ROUTE_LAG_SYMBOL,
+    decode_compiled_model_info, kernel_output_len, CompiledKernelAvailability,
+    ALLOC_F64_BUFFER_SYMBOL, API_VERSION_SYMBOL, DERIVE_SYMBOL, DIFFUSION_SYMBOL, DRIFT_SYMBOL,
+    DYNAMICS_SYMBOL, FREE_F64_BUFFER_SYMBOL, INIT_SYMBOL, MODEL_INFO_JSON_LEN_SYMBOL,
+    MODEL_INFO_JSON_PTR_SYMBOL, OUTPUTS_SYMBOL, ROUTE_BIOAVAILABILITY_SYMBOL, ROUTE_LAG_SYMBOL,
 };
 use super::native::{KernelSession, NativeModelInfo, RuntimeArtifact, RuntimeBackend};
 use super::wasm_compile::{WasmError, WASM_API_VERSION};
@@ -20,52 +18,11 @@ use super::wasm_direct_emitter::{
 use crate::PharmsolError;
 use pharmsol_dsl::execution::KernelRole;
 
-#[derive(Clone, Copy, Debug, Default)]
-struct WasmKernelAvailability {
-    derive: bool,
-    dynamics: bool,
-    outputs: bool,
-    init: bool,
-    drift: bool,
-    diffusion: bool,
-    route_lag: bool,
-    route_bioavailability: bool,
-}
-
-impl WasmKernelAvailability {
-    fn has(self, role: KernelRole) -> bool {
-        match role {
-            KernelRole::Derive => self.derive,
-            KernelRole::Dynamics => self.dynamics,
-            KernelRole::Outputs => self.outputs,
-            KernelRole::Init => self.init,
-            KernelRole::Drift => self.drift,
-            KernelRole::Diffusion => self.diffusion,
-            KernelRole::RouteLag => self.route_lag,
-            KernelRole::RouteBioavailability => self.route_bioavailability,
-            KernelRole::Analytical => false,
-        }
-    }
-
-    fn compiled(self) -> CompiledKernelAvailability {
-        CompiledKernelAvailability {
-            derive: self.derive,
-            dynamics: self.dynamics,
-            outputs: self.outputs,
-            init: self.init,
-            drift: self.drift,
-            diffusion: self.diffusion,
-            route_lag: self.route_lag,
-            route_bioavailability: self.route_bioavailability,
-        }
-    }
-}
-
 pub(crate) struct WasmExecutionArtifact {
     info: NativeModelInfo,
     engine: Engine,
     module: Module,
-    kernels: WasmKernelAvailability,
+    kernels: CompiledKernelAvailability,
     session_pool: Mutex<Vec<WasmKernelSession>>,
 }
 
@@ -120,7 +77,7 @@ impl WasmKernelSession {
[…]
