### 2026-06-02 b0297f3
fix: Added mutable to lookup in example
