### 2026-06-02 5a5065a
fix: update wasm.rs tests to use kernels.* fields and re-export NativeStateInfo
