### 2026-06-02 b89ae7c
fix: Cargo fmt
