### 2026-06-06 #10580 @axiomcura
Great paper overall - I have finished my review. Below are a few additional comments. In particular, there is one major comment I would strongly encourage the authors to address:

The authors state:

> “As a fully open-source solution, pharmsol empowers users to inspect, modify, and extend the simulation capabilities without licensing constraints.”

and

> “pharmsol is open-source with a permissive license, allowing developers to inspect, extend, and integrate the library into new or existing solutions.”

However, according to the repo's licensing, the project is currently distributed under the GPL-3.0 license, which is not considered a permissive license and does impose licensing constraints.

For example, if another software project links against or incorporates pharmsol, the resulting combined work must also be distributed under the GPL-3.0 license. Consequently, this may limit integration into proprietary software, commercially closed-source tools, or even projects under permissive licenses such as MIT unless the combined project is also relicensed under GPL-3.0.

If this licensing model is intentional, I would recommend revising the manuscript statements to more accurately reflect the implications of the GPL-3.0 license. Alternatively, if the intention is to maximize downstream adoption and integration flexibility, the authors may wish to consider adopting a more permissive license such as MIT, BSD-2-Clause, BSD-3-Clause, or Apache-2.0.
