### 2026-06-17 1246a51
Expanding the documentation for the macro-based entrypoints and the Optimize module. #272
