### 2026-06-17 bf66058
rework traits
