+++ src/simulator/backends/ode/mod.rs
@@ -11,18 +11,19 @@ pub(crate) mod closure_helpers {
 }
 
 use crate::{
-    data::{Covariates, Infusion},
+    core::{ModelInfo, Solver},
+    data::Covariates,
     error_model::AssayErrorModels,
     prelude::simulator::SubjectPredictions,
     simulator::{DiffEq, Fa, Init, Lag, Neqs, Out, M, V},
-    Event, Observation, Parameters, PharmsolError, Subject,
+    Event, PharmsolError, Subject,
 };
 
-use super::parameters_hash;
+use crate::simulator::backends::parameters_hash;
 use crate::simulator::cache::{
-    BoundErrorModelCache, PredictionCache, DEFAULT_BOUND_ERROR_MODEL_CACHE_SIZE, DEFAULT_CACHE_SIZE,
+    BoundErrorModelCache, PredictionCache, DEFAULT_CACHE_SIZE,
 };
-use crate::simulator::equation::Predictions;
+use crate::core::Predictions;
 use closure::PMProblem;
 use diffsol::{
     error::OdeSolverError, ode_solver::method::OdeSolverMethod, NalgebraContext, OdeBuilder,
@@ -32,10 +33,8 @@ use nalgebra::DVector;
 use pharmsol_dsl::ModelKind;
 use thiserror::Error;
 
-use super::{
-    EqnKind, Equation, EquationPriv, EquationTypes, ModelMetadata, ModelMetadataError, State,
-    ValidatedModelMetadata,
-};
+use crate::core::metadata::{ModelMetadata, ModelMetadataError, ValidatedModelMetadata};
+use crate::core::State;
 
 const RTOL: f64 = 1e-4;
 const ATOL: f64 = 1e-4;
@@ -97,58 +96,47 @@ pub enum OdeMetadataError {
 
 #[derive(Clone, Debug)]
 pub struct ODE {
+    core: crate::core::ModelCore<PredictionCache>,
     diffeq: DiffEq,
     lag: Lag,
     fa: Fa,
     init: Init,
     out: Out,
-    neqs: Neqs,
     solver: OdeSolver,
     rtol: f64,
     atol: f64,
-    metadata: Option<ValidatedModelMetadata>,
-    cache: Option<PredictionCache>,
-    error_model_cache: Option<BoundErrorModelCache>,
 }
 
 impl ODE {
     pub fn new(diffeq: DiffEq, lag: Lag, fa: Fa, init: Init, out: Out) -> Self {
         Self {
+            core: crate::core::ModelCore::new(Some(PredictionCache::new(DEFAULT_CACHE_SIZE))),
             diffeq,
             lag,
             fa,
             init,
             out,
-            neqs: Neqs::default(),
             solver: OdeSolver::default(),
             rtol: RTOL,
             atol: ATOL,
-            metadata: None,
-            cache: Some(PredictionCache::new(DEFAULT_CACHE_SIZE)),
-            error_model_cache: Some(BoundErrorModelCache::new(
-                DEFAULT_BOUND_ERROR_MODEL_CACHE_SIZE,
-            )),
         }
     }
 
     /// Set the number of state variables (ODE compartments).
     pub fn with_nstates(mut self, nstates: usize) -> Self {
-        self.neqs.nstates = nstates;
-        self.invalidate_metadata();
+        self.core = self.core.with_nstates(nstates);
         self
     }
 
     /// Set the number of drug inputs (size of bolus[] and rateiv[]).
     pub fn with_ndrugs(mut self, ndrugs: usize) -> Self {
-        self.neqs.ndrugs = ndrugs;
-        self.invalidate_metadata();
+        self.core = self.core.with_ndrugs(ndrugs);
         self
     }
 
[…]
