+++ src/simulator/backends/analytical/three_compartment_models.rs
@@ -252,8 +252,8 @@ pub fn pm_three_compartments_with_absorption(
 #[cfg(test)]
 mod tests {
     use super::super::tests::SubjectInfo;
-    use crate::core::Simulate;
     use super::{three_compartments, three_compartments_with_absorption};
+    use crate::core::Simulate;
     use crate::*;
     use approx::assert_relative_eq;
 
