+++ src/simulator/backends/analytical/two_compartment_models.rs
@@ -118,8 +118,8 @@ pub fn pm_two_compartments_with_absorption(x: &V, p: &V, t: T, rateiv: &V, cov: 
 #[cfg(test)]
 mod tests {
     use super::super::tests::SubjectInfo;
-    use crate::core::Simulate;
     use super::{two_compartments, two_compartments_with_absorption};
+    use crate::core::Simulate;
     use crate::*;
     use approx::assert_relative_eq;
 
