+++ src/dsl/native.rs
@@ -29,13 +29,13 @@ use crate::{
     data::error_model::AssayErrorModels,
     data::{Covariates, Infusion, InputLabel, OutputLabel},
     simulator::{
-        cache::{
-            BoundErrorModelCache, PredictionCache, SdeLikelihoodCache,
-            DEFAULT_BOUND_ERROR_MODEL_CACHE_SIZE, DEFAULT_CACHE_SIZE,
-        },
         backends::{
             ode::{closure_helpers::PMProblem, ExplicitRkTableau, OdeSolver, SdirkTableau},
             sde::simulate_sde_event_with,
+        },
+        cache::{
+            BoundErrorModelCache, PredictionCache, SdeLikelihoodCache,
+            DEFAULT_BOUND_ERROR_MODEL_CACHE_SIZE, DEFAULT_CACHE_SIZE,
         },
         likelihood::{Prediction, SubjectPredictions},
         Fa, Lag, M, T, V,
@@ -441,8 +441,7 @@ fn runtime_model_metadata(info: &NativeModelInfo) -> Result<ValidatedModelMetada
         .kind(info.kind)
         .parameters(info.parameters.iter().cloned())
         .covariates(covariates.into_iter().map(|covariate| {
-            let mut declaration =
-                crate::core::metadata::Covariate::new(covariate.name.clone());
+            let mut declaration = crate::core::metadata::Covariate::new(covariate.name.clone());
             if let Some(interpolation) = covariate.interpolation {
                 declaration = declaration.with_interpolation(interpolation);
             }
@@ -479,12 +478,8 @@ fn runtime_model_metadata(info: &NativeModelInfo) -> Result<ValidatedModelMetada
         let kind = route.kind.unwrap_or(RouteKind::Bolus);
 
         let mut declaration = match kind {
-            RouteKind::Bolus => {
-                crate::core::metadata::Route::bolus(route.name.clone())
-            }
-            RouteKind::Infusion => {
-                crate::core::metadata::Route::infusion(route.name.clone())
-            }
+            RouteKind::Bolus => crate::core::metadata::Route::bolus(route.name.clone()),
+            RouteKind::Infusion => crate::core::metadata::Route::infusion(route.name.clone()),
         }
         .to_state(destination);
 
@@ -1413,11 +1408,15 @@ fn runtime_ode_predictions(
     }
 }
 
-
 impl Solver for NativeOdeModel {
     type State = V;
 
-    fn initial_state(&self, _params: &[f64], _covariates: &Covariates, _occasion_index: usize) -> V {
+    fn initial_state(
+        &self,
+        _params: &[f64],
+        _covariates: &Covariates,
+        _occasion_index: usize,
+    ) -> V {
         V::zeros(self.shared.info.state_len, NalgebraContext)
     }
 
@@ -1427,36 +1426,64 @@ impl Solver for NativeOdeModel {
 }
 
 impl ModelInfo for NativeOdeModel {
-    fn nstates(&self) -> usize { self.shared.info.state_len }
-    fn ndrugs(&self) -> usize { self.shared.info.route_len }
-    fn nout(&self) -> usize { self.shared.info.output_len }
-    fn metadata(&self) -> Option<&ValidatedModelMetadata> { Some(self.shared.metadata()) }
-    fn lag(&self) -> &Lag { &(runtime_no_lag as Lag) }
-    fn fa(&self) -> &Fa { &(runtime_no_fa as Fa) }
+    fn nstates(&self) -> usize {
+        self.shared.info.state_len
+    }
+    fn ndrugs(&self) -> usize {
+        self.shared.info.route_len
+    }
+    fn nout(&self) -> usize {
[…]
