### 2026-06-18 605cfe7
chore: Apply suggestions from code review
Mostly related to doc links

Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>
