### 2026-06-18 356cf2f
Trying to fix compilation on Ubuntu
