### 2026-06-18 a225aa8
answering copilot's comments
