+++ src/optimize/mod.rs
@@ -4,7 +4,7 @@
 //!
 //! - [`effect`] — Find the maximum effect (`E2`) for dual-site PD models
 //!   via Nelder‑Mead optimization in log‑space.
-//! - [`parameters`] — Nelder‑Mead parameter refinement for an [`Equation`]
+//! - [`parameters`] — Nelder‑Mead parameter refinement for a [`Simulate`] model
 //!   against a [`Data`] set and [`AssayErrorModels`].
 
 pub mod effect;
