### 2026-06-18 c46ce7d
force SDEs to use the standard loop
