+++ src/simulator/backends/sde/mod.rs
@@ -2,19 +2,19 @@ mod em;
 
 use diffsol::{NalgebraContext, Vector};
 use nalgebra::DVector;
-use ndarray::{concatenate, Array2, Axis};
+use ndarray::Array2;
 use pharmsol_dsl::ModelKind;
 use rand::{rng, RngExt};
 use rayon::prelude::*;
 use thiserror::Error;
 
-use crate::core::{ModelInfo, Simulate, Solver};
+use crate::core::{ModelInfo, Simulate};
 use crate::{
     data::{Covariates, Infusion},
     error_model::AssayErrorModels,
     prelude::simulator::Prediction,
     simulator::{Diffusion, Drift, Fa, Init, Lag, Neqs, Out, V},
-    Event, Observation, PharmsolError, Subject,
+    Observation, PharmsolError, Subject,
 };
 
 use crate::simulator::backends::parameters_hash;
@@ -236,7 +236,7 @@ impl crate::core::PredictionsContainer for Array2<Prediction> {
     }
 
     fn push(&mut self, pred: Prediction) {
-        let col = Array2::from_shape_vec((self.nrows(), 1), vec![pred]).unwrap();
+        let col = Array2::from_shape_vec((self.nrows(), 1), vec![pred; self.nrows()]).unwrap();
         *self = ndarray::concatenate(ndarray::Axis(1), &[self.view(), col.view()]).unwrap();
     }
 
@@ -403,9 +403,15 @@ impl crate::core::Solver for SDE {
         Ok(())
     }
 
+    fn process_bolus(&self, state: &mut Self::State, input: usize, amount: f64) {
+        if !self.injected_bolus_mappings.apply(state, input, amount) {
+            state.add_bolus(input, amount);
+        }
+    }
+
     fn process_observation(
         &self,
-        x: &Self::State,
+        x: &mut Self::State,
         parameters: &[f64],
         observation: &Observation,
         error_models: Option<&AssayErrorModels>,
@@ -430,15 +436,16 @@ impl crate::core::Solver for SDE {
             *p = observation.to_prediction(y[outeq], x[i].as_slice().to_vec());
         });
 
-        // Resampling and likelihood computation
+        // Resampling — mutate state to concentrate particles on high-likelihood regions
         let lik = if let Some(em) = error_models {
             let q: Vec<f64> = preds
                 .iter()
                 .map(|p| p.log_likelihood(em).map(f64::exp).unwrap_or(0.0))
                 .collect();
             let sum_q: f64 = q.iter().sum();
-            // Note: resampling is skipped here because state is borrowed.
-            // Full resampling happens in simulate_subject.
+            let w: Vec<f64> = q.iter().map(|qi| qi / sum_q).collect();
+            let indices = sysresample(&w);
+            *x = indices.iter().map(|&i| x[i].clone()).collect();
             Some(sum_q / nparticles as f64)
         } else {
             None
@@ -542,103 +549,12 @@ impl crate::core::Simulate for SDE {
         params: &[f64],
         error_models: Option<&AssayErrorModels>,
     ) -> Result<(Self::Predictions, Option<f64>), PharmsolError> {
-        let bound_em = match error_models {
-            Some(em) => Some(crate::core::simulate::bind_error_models_inner(self, em)?),
-            None => None,
-        };
-
-        let mut output =
-            Array2::<Prediction>::from_shape_fn((self.nparticles, 0), |_| Prediction::default());
-        let mut likelihood = Vec::new();
-
-        for occasion in subject.occasions() {
-            let covariates = occasion.covariates();
-            let events = self.resolve_events(occasion, params, covariates)?;
-            let mut state = self.initial_state(params, covariates, occasion.index());
[…]
