+++ .github/workflows/build.yml
@@ -34,7 +34,7 @@ jobs:
             native_aot_smoke_target: ""
     steps:
       - name: Checkout code
-        uses: actions/checkout@v6
+        uses: actions/checkout@v7
 
       - name: Install Rust toolchain
         uses: dtolnay/rust-toolchain@stable
