### 2026-06-29 d96bd1c
feat: Use quick_cache for caching
