+++ .github/workflows/build.yml
@@ -58,7 +58,7 @@ jobs:
           cache-dependency-path: tests/browser-e2e/package-lock.json
 
       - name: Setup cache
-        uses: actions/cache@v5
+        uses: actions/cache@v6
         with:
           path: |
             ~/.cargo/registry
