### 2026-07-06 4653382
feat: Use quick_cache for caching (#280)
