+++ Cargo.toml
@@ -68,7 +68,7 @@ rayon = "1.12.0"
 statrs = "0.18.0"
 argmin = "0.11.0"
 argmin-math = "0.5.1"
-moka = { version = "0.12.15", features = ["sync"] }
+quick_cache = "0.7.0"
 wasmtime = { version = "44.0.0", optional = true }
 
 [dev-dependencies]
