### 2026-07-07 282aeb3
Adding explanations to the examples #272 (#275)
Co-authored-by: Markus Hovd <markushh@uio.no>
