### 2026-07-07 385d89a
docs: Improve documentation for macros (#276)
* Expanding the documentation for the macro-based entrypoints and the Optimize module. #272

* chore: Apply suggestions from code review

Mostly related to doc links

Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>

---------

Co-authored-by: Markus Hovd <markushh@uio.no>
Co-authored-by: Copilot Autofix powered by AI <175728472+Copilot@users.noreply.github.com>
