### 2026-07-07 58c5086
chore: Bump version number to 0.28.0
