### 2026-07-07 8fda5f2
chore: release v0.27.2 (#278)
* chore: release v0.27.2

* chore: Bump version number to 0.28.0

---------

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Markus <66058642+mhovd@users.noreply.github.com>
