### 2026-07-08 f1765c3
fix: Fix data expansion
Now uses each subjects maximal time, instead of the global max time.

TAD is now applied after the last dose, not the last observation.
