### 2026-07-08 f6eb3da
chore: Remove dead argument
Also update data expansion to use the time per subject occasion, not subject as a whole
