### 2026-07-08 667af88
chore: Add tests for expand
