### 2026-07-08 68fdeee
chore: Use diffsol 0.15.1
