+++ src/simulator/equation/analytical/mod.rs
@@ -62,7 +62,8 @@ pub struct Analytical {
 pub(crate) fn compact_public_vector(vector: &V) -> V {
     V::from_vec(
         vector.as_slice().get(1..).unwrap_or(&[]).to_vec(),
-        NalgebraContext,
+        NalgebraContext::new(),
+
     )
 }
 
@@ -71,7 +72,7 @@ pub(crate) fn pad_public_vector(vector: &V) -> V {
     let mut padded = Vec::with_capacity(vector.len() + 1);
     padded.push(0.0);
     padded.extend(vector.as_slice().iter().copied());
-    V::from_vec(padded, NalgebraContext)
+    V::from_vec(padded, NalgebraContext::new())
 }
 
 #[inline(always)]
@@ -328,8 +329,8 @@ impl EquationPriv for Analytical {
 
         // 2) March over each sub-interval
         let mut current_t = ts[0];
-        let mut parameters_v = V::from_vec(parameters.to_vec(), NalgebraContext);
-        let mut rateiv = V::zeros(self.get_ndrugs(), NalgebraContext);
+        let mut parameters_v = V::from_vec(parameters.to_vec(), NalgebraContext::new());
+        let mut rateiv = V::zeros(self.get_ndrugs(), NalgebraContext::new());
 
         for &next_t in &ts[1..] {
             // prepare parameters and infusion rate for [current_t .. next_t]
@@ -379,11 +380,11 @@ impl EquationPriv for Analytical {
         likelihood: &mut Vec<f64>,
         output: &mut Self::P,
     ) -> Result<(), PharmsolError> {
-        let mut y = V::zeros(self.get_nouteqs(), NalgebraContext);
+        let mut y = V::zeros(self.get_nouteqs(), NalgebraContext::new());
         let out = &self.out;
         (out)(
             x,
-            &V::from_vec(parameters.to_vec(), NalgebraContext),
+            &V::from_vec(parameters.to_vec(), NalgebraContext::new()),
             observation.time(),
             covariates,
             &mut y,
@@ -409,10 +410,10 @@ impl EquationPriv for Analytical {
         occasion_index: usize,
     ) -> V {
         let init = &self.init;
-        let mut x = V::zeros(self.get_nstates(), NalgebraContext);
+        let mut x = V::zeros(self.get_nstates(), NalgebraContext::new());
         if occasion_index == 0 {
             (init)(
-                &V::from_vec(parameters.to_vec(), NalgebraContext),
+                &V::from_vec(parameters.to_vec(), NalgebraContext::new()),
                 0.0,
                 covariates,
                 &mut x,
@@ -769,17 +770,17 @@ pub(crate) mod tests {
         compact_rateiv: Vec<f64>,
     ) {
         let covariates = Covariates::new();
-        let compact_x = V::from_vec(compact_x, NalgebraContext);
-        let params = V::from_vec(params, NalgebraContext);
-        let compact_rateiv = V::from_vec(compact_rateiv, NalgebraContext);
+        let compact_x = V::from_vec(compact_x, NalgebraContext::new());
+        let params = V::from_vec(params, NalgebraContext::new());
+        let compact_rateiv = V::from_vec(compact_rateiv, NalgebraContext::new());
 
         let mut padded_x = vec![1234.0];
         padded_x.extend(compact_x.as_slice().iter().copied());
-        let padded_x = V::from_vec(padded_x, NalgebraContext);
+        let padded_x = V::from_vec(padded_x, NalgebraContext::new());
 
         let mut padded_rateiv = vec![5678.0];
         padded_rateiv.extend(compact_rateiv.as_slice().iter().copied());
[…]
