### 2026-07-10 f628d20
fix: Avoid infinite loops in data expansion
Expansion can not be made at a finer grid than every microsecond, or equivalent time-units
