### 2026-07-13 af1338a
paper: Specify license constraints
