### 2026-07-13 d3a704d
paper: Add Pmetrics citation count
