### 2026-07-13 d97f156
fix: Fix data expansion (#282)
* fix: Fix data expansion

Now uses each subjects maximal time, instead of the global max time.

TAD is now applied after the last dose, not the last observation.

* chore: Remove dead argument

Also update data expansion to use the time per subject occasion, not subject as a whole

* chore: Add tests for expand

* fix: Avoid infinite loops in data expansion

Expansion can not be made at a finer grid than every microsecond, or equivalent time-units
