+++ src/data/structs.rs
@@ -143,8 +143,11 @@ impl Data {
     ///
     /// # Arguments
     ///
-    /// * `idelta` - Time interval between added observations
-    /// * `tad` - Additional time to add after the last observation
+    /// * `idelta` - Time interval between added observations. Times are handled
+    ///   with microsecond resolution, so `idelta` must be at least `5e-7`
+    ///   (0.5 µs, which rounds up to 1 µs); any smaller positive value rounds
+    ///   to zero and the dataset is returned unchanged.
+    /// * `tad` - Additional time to add after the last dose (time after dose)
     ///
     /// # Returns
     ///
@@ -154,20 +157,13 @@ impl Data {
             return self.clone();
         }
 
-        // Determine the last time across all subjects and occasions
-        let last_time = self
-            .subjects
-            .iter()
-            .flat_map(|subject| &subject.occasions)
-            .flat_map(|occasion| &occasion.events)
-            .filter_map(|event| match event {
-                Event::Observation(observation) => Some(observation.time()),
-                Event::Infusion(infusion) => Some(infusion.time() + infusion.duration()),
-                _ => None,
-            })
-            .max_by(|a, b| a.partial_cmp(b).unwrap())
-            .unwrap_or(0.0)
-            + tad;
+        // Work in integer microseconds so the grid always makes progress.
+        // A sub-microsecond `idelta` would otherwise round back to the previous
+        // time on every iteration and spin forever.
+        let step_us = (idelta * 1e6).round() as u64;
+        if step_us == 0 {
+            return self.clone();
+        }
 
         // Collect unique output equations more efficiently
         let outeq_values = self.get_output_equations();
@@ -181,7 +177,22 @@ impl Data {
                     .occasions
                     .iter()
                     .map(|occasion| {
-                        let old_events = occasion.process_events(None, true);
+                        let old_events = occasion.process_events(None);
+
+                        // Determine the last dose time for this occasion
+                        let last_time = occasion
+                            .events
+                            .iter()
+                            .filter_map(|event| match event {
+                                Event::Bolus(bolus) => Some(bolus.time()),
+                                Event::Infusion(infusion) => {
+                                    Some(infusion.time() + infusion.duration())
+                                }
+                                _ => None,
+                            })
+                            .max_by(|a, b| a.partial_cmp(b).unwrap())
+                            .unwrap_or(0.0)
+                            + tad;
 
                         // Create a set of existing (time, outeq) pairs for fast lookup
                         let existing_obs: std::collections::HashSet<(u64, OutputLabel)> =
@@ -197,11 +208,13 @@ impl Data {
                                 })
                                 .collect();
 
-                        // Generate new observation times
+                        // Generate new observation times, stepping in integer
+                        // microseconds to guarantee forward progress.
                         let mut new_events = Vec::new();
-                        let mut time = 0.0;
-                        while time < last_time {
-                            let time_key = (time * 1e6).round() as u64;
+                        let last_time_us = (last_time * 1e6).round() as u64;
+                        let mut time_key = 0u64;
+                        while time_key <= last_time_us {
+                            let time = time_key as f64 / 1e6;
 
[…]
