### 2026-07-14 cd669c7
chore: Add more informative error messages
