+++ src/simulator/equation/ode/mod.rs
@@ -334,108 +334,120 @@ fn _simulate_subject_dense(
     let mut y_out = V::zeros(ode.get_nouteqs(), NalgebraContext);
 
     for occasion in subject.occasions() {
-        let covariates = occasion.covariates();
-        let events = ode.resolve_occasion_events(occasion, parameters, covariates)?;
+        // Run one occasion in a closure so any error can be tagged with the
+        // subject and support point in a single place below.
+        let occasion_result: Result<(), PharmsolError> = (|| {
+            let covariates = occasion.covariates();
+            let events = ode.resolve_occasion_events(occasion, parameters, covariates)?;
 
-        let problem = OdeBuilder::<M>::new()
-            .atol(vec![ode.atol])
-            .rtol(ode.rtol)
-            .t0(occasion.initial_time())
-            .h0(1e-3)
-            .p(parameters_vec.clone())
-            .build_from_eqn(PMProblem::with_params_v(
-                move |x, p, t, dx, bolus, rateiv, cov| {
-                    (ode.diffeq)(x, p, t, dx, bolus, rateiv, cov);
-                },
-                nstates,
-                ndrugs,
-                parameters_v.clone(),
-                covariates,
-                events.iter().filter_map(|event| match event {
-                    Event::Infusion(infusion) => Some(infusion),
-                    _ => None,
-                }),
-                ode.initial_state(parameters, covariates, occasion.index()),
-            )?)?;
+            let problem = OdeBuilder::<M>::new()
+                .atol(vec![ode.atol])
+                .rtol(ode.rtol)
+                .t0(occasion.initial_time())
+                .h0(1e-3)
+                .p(parameters_vec.clone())
+                .build_from_eqn(PMProblem::with_params_v(
+                    move |x, p, t, dx, bolus, rateiv, cov| {
+                        (ode.diffeq)(x, p, t, dx, bolus, rateiv, cov);
+                    },
+                    nstates,
+                    ndrugs,
+                    parameters_v.clone(),
+                    covariates,
+                    events.iter().filter_map(|event| match event {
+                        Event::Infusion(infusion) => Some(infusion),
+                        _ => None,
+                    }),
+                    ode.initial_state(parameters, covariates, occasion.index()),
+                )?)?;
 
-        match &ode.solver {
-            OdeSolver::Bdf => {
-                let mut solver = problem.bdf::<diffsol::NalgebraLU<f64>>()?;
-                ODE::run_events(
-                    ode,
-                    &mut solver,
-                    &events,
-                    &parameters_v,
-                    covariates,
-                    bound_error_models,
-                    &mut bolus_v,
-                    &zero_bolus,
-                    &zero_rateiv,
-                    &mut state_with_bolus,
-                    &mut state_without_bolus,
-                    &mut y_out,
-                    &mut likelihood,
-                    &mut output,
-                )?;
+            match &ode.solver {
+                OdeSolver::Bdf => {
+                    let mut solver = problem.bdf::<diffsol::NalgebraLU<f64>>()?;
+                    ODE::run_events(
+                        ode,
+                        &mut solver,
+                        &events,
+                        &parameters_v,
+                        covariates,
+                        bound_error_models,
+                        &mut bolus_v,
+                        &zero_bolus,
+                        &zero_rateiv,
+                        &mut state_with_bolus,
+                        &mut state_without_bolus,
+                        &mut y_out,
+                        &mut likelihood,
+                        &mut output,
+                    )?;
+                }
+                OdeSolver::ExplicitRk(ExplicitRkTableau::Tsit45) => {
+                    let mut solver = problem.tsit45()?;
+                    ODE::run_events(
+                        ode,
+                        &mut solver,
+                        &events,
+                        &parameters_v,
[…]

[…]
