### 2026-07-14 dda9842
Errors for available parameters
