### 2026-07-15 13e1d8f
Likelihood extraction
