+++ examples/one_compartment.rs
@@ -1,8 +1,7 @@
 //! Compares analytical and ODE solutions for a one-compartment IV model.
 //!
-//! Both the `analytical!` and `ode!` macros define the same model. Predictions
-//! and log-likelihoods are computed with each and printed side by side to
-//! confirm they match.
+//! Both the `analytical!` and `ode!` macros define the same model. Their
+//! noiseless predictions are printed side by side to confirm that they match.
 
 fn main() -> Result<(), pharmsol::PharmsolError> {
     use pharmsol::{prelude::*, Parameters};
@@ -50,13 +49,6 @@ fn main() -> Result<(), pharmsol::PharmsolError> {
         .missing_observation(12.0, "cp")
         .build();
 
-    // Define the assay error models once by label and reuse them across both
-    // equations.
-    let ems = AssayErrorModels::new().add(
-        "cp",
-        AssayErrorModel::additive(ErrorPoly::new(0.0, 0.05, 0.0, 0.0), 0.0),
-    )?;
-
     // Define the parameter values for the simulations
     let ke = 1.022; // Elimination rate constant
     let v = 194.0; // Volume of distribution
@@ -65,27 +57,13 @@ fn main() -> Result<(), pharmsol::PharmsolError> {
     let ode_parameters =
         Parameters::with_model(&ode, [("ke", ke), ("v", v)]).expect("valid named parameters");
 
-    // Compute likelihoods and predictions for both models
-    let analytical_likelihoods =
-        analytical.estimate_log_likelihood(&subject, &analytical_parameters, &ems)?;
-
     let analytical_predictions =
         analytical.estimate_predictions(&subject, &analytical_parameters)?;
-
-    let ode_likelihoods = ode.estimate_log_likelihood(&subject, &ode_parameters, &ems)?;
-
     let ode_predictions = ode.estimate_predictions(&subject, &ode_parameters)?;
 
-    // Print comparison table
+    // Structural prediction parity is a simulation-engine invariant.
     println!("\n┌───────────┬─────────────────┬─────────────────┬─────────────────────┐");
     println!("│           │   Analytical    │       ODE       │     Difference      │");
-    println!("├───────────┼─────────────────┼─────────────────┼─────────────────────┤");
-    println!(
-        "│ Log-Likeli│ {:>15.6} │ {:>15.6} │ {:>19.2e} │",
-        analytical_likelihoods,
-        ode_likelihoods,
-        analytical_likelihoods - ode_likelihoods
-    );
     println!("├───────────┼─────────────────┼─────────────────┼─────────────────────┤");
     println!("│   Time    │   Prediction    │   Prediction    │                     │");
     println!("├───────────┼─────────────────┼─────────────────┼─────────────────────┤");
