+++ src/data/parser/pmetrics.rs
@@ -53,7 +53,7 @@ use std::str::FromStr;
 /// # Expected columns
 ///
 /// The canonical columns are `ID`, `TIME`, `EVID`, `DOSE`, `DUR`, `ADDL`,
-/// `II`, `INPUT`, `OUT`, `OUTEQ`, `CENS`, and optional `C0..C3` error
+/// `II`, `INPUT`, `OUT`, `OUTEQ`, `CENS`, and optional `C0..C3` observation
 /// coefficients.
 ///
 /// All other numeric columns are treated as covariates.
@@ -134,16 +134,16 @@ struct Row {
     /// Censoring output
     #[serde(default, deserialize_with = "deserialize_option_censor")]
     cens: Option<Censor>,
-    /// First element of the error polynomial
+    /// C0 observation coefficient
     #[serde(deserialize_with = "deserialize_option_f64")]
     c0: Option<f64>,
-    /// Second element of the error polynomial
+    /// C1 observation coefficient
     #[serde(deserialize_with = "deserialize_option_f64")]
     c1: Option<f64>,
-    /// Third element of the error polynomial
+    /// C2 observation coefficient
     #[serde(deserialize_with = "deserialize_option_f64")]
     c2: Option<f64>,
-    /// Fourth element of the error polynomial
+    /// C3 observation coefficient
     #[serde(deserialize_with = "deserialize_option_f64")]
     c3: Option<f64>,
     /// All other columns are covariates
@@ -331,8 +331,8 @@ impl Data {
                             };
                             let (c0, c1, c2, c3) = obs
                                 .errorpoly()
-                                .map(|poly| {
-                                    let (c0, c1, c2, c3) = poly.coefficients();
+                                .map(|errorpoly| {
+                                    let (c0, c1, c2, c3) = errorpoly.coefficients();
                                     (
                                         c0.to_string(),
                                         c1.to_string(),
@@ -537,6 +537,68 @@ mod tests {
         assert_eq!(second.get(10), Some("-1"));
         assert_eq!(second.get(11), Some("."));
         assert_eq!(second.get(14), Some("."));
+    }
+
+    #[test]
+    fn pmetrics_roundtrip_preserves_exact_observation_metadata() {
+        let subject = Subject::builder("roundtrip")
+            .observation_with_error(
+                0.25,
+                2.5,
+                "cp",
+                ErrorPoly::new(0.1, 0.2, 0.3, 0.4),
+                Censor::BLOQ,
+            )
+            .censored_observation(1.5, 3.5, "effect", Censor::ALOQ)
+            .build();
+        let data = Data::new(vec![subject]);
+        let file = NamedTempFile::new().unwrap();
+        data.write_pmetrics(file.as_file()).unwrap();
+
+        let reparsed = read_pmetrics(file.path().display().to_string()).unwrap();
+        let events = reparsed.subjects()[0].occasions()[0].events();
+        let observations: Vec<_> = events
+            .iter()
+            .filter_map(|event| match event {
+                Event::Observation(observation) => Some(observation),
+                _ => None,
+            })
+            .collect();
+
+        assert_eq!(observations.len(), 2);
+        assert_eq!(observations[0].time(), 0.25);
+        assert_eq!(observations[0].value(), Some(2.5));
+        assert_eq!(observations[0].outeq().as_str(), "cp");
+        assert_eq!(observations[0].censoring(), Censor::BLOQ);
+        assert_eq!(
+            observations[0].errorpoly().unwrap().coefficients(),
[…]
