### 2026-07-19 9f44189
chore: Use diffsol 0.15.1 (#283)
* chore: Use diffsol 0.15.1

* fmt
