### 2026-07-20 c8bbb9f
chore: Optimize event processing
Refactors lag-time and bioavailability application in the event-processing path and adds test coverage for ADDL expansion ordering. Investigated as part of issue #292.
