+++ pharmsol-macros/Cargo.toml
@@ -15,4 +15,4 @@ proc-macro = true
 pharmsol-dsl = { workspace = true }
 proc-macro2 = "1.0.106"
 quote = "1.0.45"
-syn = { version = "2.0.117", features = ["full", "visit", "visit-mut"] }
+syn = { version = "3.0.2", features = ["full", "visit", "visit-mut"] }
