### 2026-07-20 0bdb87f
chore: Add more informative error messages (#290)
* chore: Add more informative error messages

* Errors for available parameters

* More errors
