### 2026-07-20 e06c9c6
chore: Optimize event processing (#294)
* chore: Optimize event processing

Refactors lag-time and bioavailability application in the event-processing path and adds test coverage for ADDL expansion ordering. Investigated as part of issue #292.

* Delete testing artifact

---------

Co-authored-by: Julian Otalvaro <juliandavid347@gmail.com>
