+++ src/data/structs.rs
@@ -1,6 +1,6 @@
 use crate::{
     data::*,
-    simulator::{Fa, Lag},
+    simulator::{Fa, Lag, V},
     Censor,
 };
 use serde::{Deserialize, Serialize};
@@ -609,37 +609,58 @@ impl Occasion {
     }
 
     fn add_lagtime(&mut self, reorder: Option<(&Fa, &Lag, &[f64], &Covariates)>) {
-        if let Some((_, fn_lag, parameters, covariates)) = reorder {
-            let parameters = nalgebra::DVector::from_vec(parameters.to_vec());
-            for event in self.events.iter_mut() {
-                let time = event.time();
-                if let Event::Bolus(bolus) = event {
-                    let lagtime = fn_lag(&parameters.clone().into(), time, covariates);
-                    if let Some(input) = bolus.input_index() {
-                        if let Some(l) = lagtime.get(&input) {
-                            *bolus.mut_time() += l;
-                        }
-                    }
+        let Some((_, fn_lag, parameters, covariates)) = reorder else {
+            // No model context: events are already time-sorted from construction.
+            return;
+        };
+
+        // Build the parameter vector once and reuse it for every dose.
+        let parameters: V = nalgebra::DVector::from_vec(parameters.to_vec()).into();
+        let mut shifted = false;
+
+        for event in self.events.iter_mut() {
+            // Lag time delays boluses only; infusions are never lagged.
+            let Event::Bolus(bolus) = event else {
+                continue;
+            };
+            let Some(input) = bolus.input_index() else {
+                continue;
+            };
+            let lagtime = fn_lag(&parameters, bolus.time(), covariates);
+            if let Some(&l) = lagtime.get(&input) {
+                if l != 0.0 {
+                    *bolus.mut_time() += l;
+                    shifted = true;
                 }
             }
         }
-        self.sort();
+
+        // Re-sort only when a lag actually moved an event; the events were
+        // already sorted at construction time, so an unchanged pass stays sorted.
+        if shifted {
+            self.sort();
+        }
     }
 
     fn add_bioavailability(&mut self, reorder: Option<(&Fa, &Lag, &[f64], &Covariates)>) {
-        // If lagtime is empty, return early
-        if let Some((fn_fa, _, parameters, covariates)) = reorder {
-            let parameters = nalgebra::DVector::from_vec(parameters.to_vec());
-            for event in self.events.iter_mut() {
-                let time = event.time();
-                if let Event::Bolus(bolus) = event {
-                    let fa = fn_fa(&parameters.clone().into(), time, covariates);
-                    if let Some(input) = bolus.input_index() {
-                        if let Some(f) = fa.get(&input) {
-                            bolus.set_amount(bolus.amount() * f);
-                        }
-                    }
-                }
+        let Some((fn_fa, _, parameters, covariates)) = reorder else {
+            return;
+        };
+
+        // Build the parameter vector once and reuse it for every dose.
+        let parameters: V = nalgebra::DVector::from_vec(parameters.to_vec()).into();
+
+        for event in self.events.iter_mut() {
+            // Bioavailability scales bolus amounts only.
+            let Event::Bolus(bolus) = event else {
+                continue;
+            };
+            let Some(input) = bolus.input_index() else {
+                continue;
+            };
[…]
