### 2026-07-20 d15e379
chore: release v0.28.2 (#293)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
