+++ Cargo.toml
@@ -3,15 +3,15 @@ members = [".", "pharmsol-dsl", "pharmsol-macros"]
 resolver = "2"
 
 [workspace.package]
-version = "0.28.1"
+version = "0.28.2"
 edition = "2021"
 license = "GPL-3.0"
 authors = ["Julián D. Otálvaro <juliandavid347@gmail.com>", "Markus Hovd"]
 repository = "https://github.com/LAPKB/pharmsol"
 
 [workspace.dependencies]
-pharmsol-dsl = { path = "pharmsol-dsl", version = "0.28.1" }
-pharmsol-macros = { path = "pharmsol-macros", version = "0.28.1" }
+pharmsol-dsl = { path = "pharmsol-dsl", version = "0.28.2" }
+pharmsol-macros = { path = "pharmsol-macros", version = "0.28.2" }
 serde = { version = "1.0.228", features = ["derive"] }
 serde_json = "1.0.149"
 
