### 2026-07-21 aee97bb
feat: Harden DSL
