### 2026-07-21 d032c95
feat: Use plain language for DSL types
