+++ pharmsol-dsl/src/execution.rs
@@ -1,19 +1,30 @@
+//! Ready-to-run model compiled from an analyzed model.
+//!
+//! An [`ExecutionModel`] carries everything a simulation backend needs:
+//! resolved symbols, dense buffer layouts, and the model functions
+//! (dynamics, outputs, init, and friends) as straight-line programs.
+
 use std::collections::BTreeMap;
 use std::fmt;
 use std::sync::Arc;
 
 use crate::{
-    AnalyticalKernel, ConstValue, CovariateInterpolation, Diagnostic, DiagnosticPhase,
-    DiagnosticReport, MathIntrinsic, ModelKind, RouteKind, RoutePropertyKind, Span, Symbol,
-    SymbolId, SymbolKind, SymbolType, TypedAssignTargetKind, TypedBinaryOp, TypedCall, TypedExpr,
-    TypedExprKind, TypedModel, TypedModule, TypedRangeExpr, TypedStatePlace, TypedStatementBlock,
-    TypedStmt, TypedStmtKind, TypedUnaryOp, ValueType, DSL_LOWERING_GENERIC,
+    AnalyticalKernel, AnalyzedAssignTargetKind, AnalyzedBinaryOp, AnalyzedCall, AnalyzedExpr,
+    AnalyzedExprKind, AnalyzedModel, AnalyzedModule, AnalyzedRangeExpr, AnalyzedStatePlace,
+    AnalyzedStatementBlock, AnalyzedStmt, AnalyzedStmtKind, AnalyzedUnaryOp, ConstValue,
+    CovariateInterpolation, Diagnostic, DiagnosticPhase, DiagnosticReport, MathFunction, ModelKind,
+    RouteKind, RoutePropertyKind, Span, Symbol, SymbolId, SymbolKind, SymbolType, ValueType,
+    DSL_COMPILE_GENERIC,
 };
 
-pub fn lower_typed_module(module: &TypedModule) -> Result<ExecutionModule, LoweringError> {
+/// Compiles every model in an analyzed module into its ready-to-run form.
+///
+/// This is the final pipeline stage after [`parse_module`](crate::parse_module)
+/// and [`analyze_module`](crate::analyze_module).
+pub fn compile_analyzed_module(module: &AnalyzedModule) -> Result<ExecutionModule, CompileError> {
     let mut models = Vec::with_capacity(module.models.len());
     for model in &module.models {
-        models.push(lower_typed_model(model)?);
+        models.push(compile_analyzed_model(model)?);
     }
     Ok(ExecutionModule {
         models,
@@ -21,8 +32,12 @@ pub fn lower_typed_module(module: &TypedModule) -> Result<ExecutionModule, Lower
     })
 }
 
-pub fn lower_typed_model(model: &TypedModel) -> Result<ExecutionModel, LoweringError> {
-    ExecutionLowerer::new(model)?.lower()
+/// Compiles an analyzed model into its ready-to-run [`ExecutionModel`].
+///
+/// This is the final pipeline stage after [`parse_model`](crate::parse_model)
+/// and [`analyze_model`](crate::analyze_model).
+pub fn compile_analyzed_model(model: &AnalyzedModel) -> Result<ExecutionModel, CompileError> {
+    ModelCompiler::new(model)?.compile()
 }
 
 #[derive(Debug, Clone, PartialEq)]
@@ -36,14 +51,14 @@ pub struct ExecutionModel {
     pub name: String,
     pub kind: ModelKind,
     pub metadata: ExecutionMetadata,
-    pub abi: ExecutionAbi,
-    pub kernels: Vec<ExecutionKernel>,
+    pub layout: ExecutionLayout,
+    pub functions: Vec<ModelFunction>,
     pub span: Span,
 }
 
[…]
