### 2026-07-21 027fd4d
Fix rebase integration
