### 2026-07-22 b05d6fa
feat: Rename DSL types and add additional safeguards (#302)
* feat: Harden DSL

* feat: Use plain language for DSL types

* Fix failing test
