+++ src/dsl/jit.rs
@@ -7,20 +7,20 @@ use cranelift::prelude::*;
 use cranelift_jit::{JITBuilder, JITModule};
 use cranelift_module::{Linkage, Module};
 
-pub use super::native::DenseKernelFn;
+pub use super::native::CompiledModelFunction;
 use super::native::{
     CompiledNativeModel, NativeAnalyticalModel, NativeExecutionArtifact, NativeModelInfo,
     NativeOdeModel, NativeSdeModel,
 };
 use pharmsol_dsl::execution::{
     ExecutionBlock, ExecutionCall, ExecutionExpr, ExecutionExprKind, ExecutionForStmt,
-    ExecutionIfStmt, ExecutionKernel, ExecutionLoad, ExecutionModel, ExecutionProgram,
-    ExecutionStateRef, ExecutionStmt, ExecutionStmtKind, ExecutionTarget, ExecutionTargetKind,
-    KernelImplementation, KernelRole,
+    ExecutionIfStmt, ExecutionLoad, ExecutionModel, ExecutionProgram, ExecutionStateRef,
+    ExecutionStmt, ExecutionStmtKind, ExecutionTarget, ExecutionTargetKind, FunctionBody,
+    ModelFunction, ModelFunctionKind,
 };
 use pharmsol_dsl::{
-    ConstValue, Diagnostic, DiagnosticPhase, DiagnosticReport, MathIntrinsic, ModelKind, Span,
-    TypedBinaryOp, TypedUnaryOp, ValueType, DSL_BACKEND_GENERIC,
+    AnalyzedBinaryOp, AnalyzedUnaryOp, ConstValue, Diagnostic, DiagnosticPhase, DiagnosticReport,
+    MathFunction, ModelKind, Span, ValueType, DSL_BACKEND_GENERIC,
 };
 
 mod externs {
@@ -141,8 +141,8 @@ impl std::fmt::Display for JitCompileError {
         let span = self.diagnostic.primary_span();
         write!(
             f,
-            "{} at bytes {}..{}",
-            self.diagnostic.message, span.start, span.end
+            "error[{}]: {} (at bytes {}..{})",
+            self.diagnostic.code, self.diagnostic.message, span.start, span.end
         )
     }
 }
@@ -190,7 +190,7 @@ struct ExternRefs {
 }
 
 #[derive(Clone, Copy)]
-struct KernelArgs {
+struct FunctionArgs {
     _time: Value,
     states: Value,
     params: Value,
@@ -208,7 +208,7 @@ struct LocalBinding {
 
 struct EmitEnv<'a> {
     _ptr_ty: Type,
-    args: KernelArgs,
+    args: FunctionArgs,
     externs: ExternRefs,
     locals: &'a BTreeMap<usize, LocalBinding>,
 }
@@ -219,9 +219,9 @@ struct LoweredValue {
     ty: ValueType,
 }
 
-/// Compile one lowered execution model into a reusable JIT kernel artifact.
+/// Compile one compiled execution model into a reusable JIT function artifact.
 ///
-/// This builds the raw Cranelift-compiled kernel bundle for all roles present in
+/// This builds the raw Cranelift-compiled function bundle for all roles present in
 /// the model. Most callers should use [`compile_execution_model_to_jit`] instead.
 pub fn compile_execution_artifact(
     model: &ExecutionModel,
@@ -261,78 +261,78 @@ pub fn compile_execution_artifact(
     let mut ctx = module.make_context();
     let mut builder_context = FunctionBuilderContext::new();
 
-    let derive = compile_role_kernel(
+    let derive = compile_role_function(
         &mut module,
         &mut ctx,
         &mut builder_context,
         ptr_ty,
         externs,
         model,
-        KernelRole::Derive,
[…]
