+++ pharmsol-dsl/Cargo.toml
@@ -15,3 +15,4 @@ bench = false
 [dependencies]
 serde = { workspace = true }
 serde_json = { workspace = true }
+thiserror = { workspace = true }

[…]
