### 2026-07-23 a64cd8d
refactor: Hide handwritten models from public API
Massive refactor
