+++ src/dsl/jit.rs
@@ -1342,10 +1342,10 @@ pub fn compile_sde_model_to_jit(model: &ExecutionModel) -> Result<JitSdeModel, J
 mod tests {
     use super::*;
     use crate::equation::ode::{ExplicitRkTableau, OdeSolver};
+    use crate::equation::{ODE, SDE};
     use crate::simulator::equation::analytical::one_compartment_with_absorption;
     use crate::simulator::equation::{Equation, Predictions as PredictionTrait};
     use crate::test_fixtures::STRUCTURED_BLOCK_CORPUS;
-    use crate::equation::{ODE, SDE};
     use crate::{equation, Parameters, Subject, SubjectBuilderExt};
     use approx::assert_relative_eq;
     use diffsol::Vector;
