+++ src/dsl/jit.rs
@@ -1453,14 +1453,14 @@ out(cp) = central / v ~ continuous()
             .build();
 
         let reference_subject = Subject::builder("ode")
-            .bolus(0.0, 120.0, 0)
-            .infusion(6.0, 60.0, 0, 2.0)
-            .observation(0.5, 0.0, 0)
-            .observation(1.0, 0.0, 0)
-            .observation(2.0, 0.0, 0)
-            .observation(6.0, 0.0, 0)
-            .observation(7.0, 0.0, 0)
-            .observation(9.0, 0.0, 0)
+            .bolus(0.0, 120.0, "oral")
+            .infusion(6.0, 60.0, "iv", 2.0)
+            .observation(0.5, 0.0, "cp")
+            .observation(1.0, 0.0, "cp")
+            .observation(2.0, 0.0, "cp")
+            .observation(6.0, 0.0, "cp")
+            .observation(7.0, 0.0, "cp")
+            .observation(9.0, 0.0, "cp")
             .build();
 
         let support = Parameters::with_model(
@@ -1489,7 +1489,22 @@ out(cp) = central / v ~ continuous()
         .with_nstates(2)
         .with_ndrugs(1)
         .with_nout(1)
-        .with_solver(OdeSolver::ExplicitRk(ExplicitRkTableau::Tsit45));
+        .with_solver(OdeSolver::ExplicitRk(ExplicitRkTableau::Tsit45))
+        .with_metadata(
+            equation::metadata::new("shared_authoring")
+                .parameters(["ka", "ke", "v"])
+                .states(["depot", "central"])
+                .outputs(["cp"])
+                .routes([
+                    equation::Route::bolus("oral")
+                        .to_state("depot")
+                        .expect_explicit_input(),
+                    equation::Route::infusion("iv")
+                        .to_state("central")
+                        .expect_explicit_input(),
+                ]),
+        )
+        .expect("reference ode metadata should validate");
 
         let reference_predictions = reference
             .estimate_predictions(&reference_subject, &support)
@@ -1620,14 +1635,14 @@ out(cp) = central / v ~ continuous()
 
         let reference_subject = Subject::builder("ode")
             .covariate("wt", 0.0, 70.0)
-            .bolus(0.0, 120.0, 0)
-            .infusion(6.0, 60.0, 1, 2.0)
-            .missing_observation(0.5, 0)
-            .missing_observation(1.0, 0)
-            .missing_observation(2.0, 0)
-            .missing_observation(6.0, 0)
-            .missing_observation(7.0, 0)
-            .missing_observation(9.0, 0)
+            .bolus(0.0, 120.0, "oral")
+            .infusion(6.0, 60.0, "iv", 2.0)
+            .missing_observation(0.5, "cp")
+            .missing_observation(1.0, "cp")
+            .missing_observation(2.0, "cp")
+            .missing_observation(6.0, "cp")
+            .missing_observation(7.0, "cp")
+            .missing_observation(9.0, "cp")
             .build();
 
         let support = Parameters::with_model(
@@ -1684,7 +1699,25 @@ out(cp) = central / v ~ continuous()
         .with_nstates(2)
         .with_ndrugs(2)
         .with_nout(1)
-        .with_solver(OdeSolver::ExplicitRk(ExplicitRkTableau::Tsit45));
+        .with_solver(OdeSolver::ExplicitRk(ExplicitRkTableau::Tsit45))
+        .with_metadata(
+            equation::metadata::new("one_cmt_oral_iv")
+                .parameters(["ka", "cl", "v", "tlag", "f_oral"])
+                .states(["depot", "central"])
+                .outputs(["cp"])
+                .routes([
+                    equation::Route::bolus("oral")
[…]
