### 2026-07-28 f76d85d
feat: Support time in the DSL
Time can now be referred to, accessible using either `t` or `time`
