### 2026-07-28 caef213
feat: Support time in the DSL (#310)
* feat: Support time in the DSL

Time can now be referred to, accessible using either `t` or `time`

* Improve test coverage
