### 2026-07-29 fbc4993
increase tolerance for some specific tests
