+++ src/data/row.rs
@@ -62,7 +62,7 @@ use thiserror::Error;
 /// let obs = DataRow::builder("pt1", 1.0)
 ///     .evid(0)
 ///     .out(25.5)
-///     .output("cp")
+///     .outeq("cp")
 ///     .build();
 ///
 /// let dose = DataRow::builder("pt1", 0.0)
@@ -74,7 +74,7 @@ use thiserror::Error;
 ///     .build();
 ///
 /// let events = dose.into_events().unwrap();
-/// assert_eq!(obs.output.as_ref().map(|label| label.as_str()), Some("cp"));
+/// assert_eq!(obs.outeq.as_ref().map(|label| label.as_str()), Some("cp"));
 /// assert_eq!(events.len(), 11);
 /// ```
 #[derive(Debug, Clone, Default)]
@@ -98,16 +98,16 @@ pub struct DataRow {
     /// Observed value (for EVID=0)
     pub out: Option<f64>,
     /// Output label
-    pub output: Option<OutputLabel>,
+    pub outeq: Option<OutputLabel>,
     /// Censoring indicator
     pub cens: Option<Censor>,
-    /// Pmetrics C0 transport coefficient
+    /// Pmetrics ErrorPoly C0 coefficient
     pub c0: Option<f64>,
-    /// Pmetrics C1 transport coefficient
+    /// Pmetrics ErrorPoly C1 coefficient
     pub c1: Option<f64>,
-    /// Pmetrics C2 transport coefficient
+    /// Pmetrics ErrorPoly C2 coefficient
     pub c2: Option<f64>,
-    /// Pmetrics C3 transport coefficient
+    /// Pmetrics ErrorPoly C3 coefficient
     pub c3: Option<f64>,
     /// Covariate values at this time point
     pub covariates: HashMap<String, f64>,
@@ -197,7 +197,7 @@ impl DataRow {
             0 => {
                 // Observation event
                 let outeq =
-                    self.output
+                    self.outeq
                         .clone()
                         .ok_or_else(|| DataError::MissingObservationOuteq {
                             id: self.id.clone(),
@@ -318,7 +318,7 @@ impl DataRow {
 /// let row = DataRow::builder("patient_001", 1.5)
 ///     .evid(0)
 ///     .out(25.5)
-///     .output("cp")
+///     .outeq("cp")
 ///     .cens(Censor::None)
 ///     .covariate("weight", 70.0)
 ///     .covariate("age", 45.0)
@@ -416,8 +416,8 @@ impl DataRowBuilder {
     /// Required for EVID=0 observation rows.
     /// The provided value is preserved as the public label until downstream
     /// model resolution.
-    pub fn output(mut self, outeq: impl ToString) -> Self {
-        self.row.output = Some(OutputLabel::new(outeq));
+    pub fn outeq(mut self, outeq: impl ToString) -> Self {
+        self.row.outeq = Some(OutputLabel::new(outeq));
         self
     }
 
@@ -427,7 +427,7 @@ impl DataRowBuilder {
         self
     }
 
-    /// Set ErrorPoly C0-C3 data.
+    /// Set ErrorPoly coefficients.
     pub fn error_poly(mut self, c0: f64, c1: f64, c2: f64, c3: f64) -> Self {
         self.row.c0 = Some(c0);
         self.row.c1 = Some(c1);
@@ -477,11 +477,11 @@ impl DataRowBuilder {
 ///     DataRow::builder("pt1", 0.0)
 ///         .evid(1).dose(100.0).input("iv").build(),
 ///     DataRow::builder("pt1", 1.0)
-///         .evid(0).out(50.0).output("cp").build(),
[…]
