### 2026-07-29 96fad71
feat: Add static or dynamic stepsize
