### 2026-07-30 8f7fd29
Implement canonical pmetrics CSV v1
