### 2026-07-31 4f98650
Correct canonical Pmetrics CSV compatibility
