+++ src/data/covariate.rs
@@ -31,9 +31,6 @@ struct CovariateSegment {
     from: f64,
     to: Option<f64>,
     method: Interpolation,
-    /// Exact source value. Older serde payloads reconstruct it from `method`.
-    #[serde(default)]
-    observation: Option<f64>,
 }
 
 impl CovariateSegment {
@@ -44,13 +41,8 @@ impl CovariateSegment {
     /// * `from` - Start time of the segment
     /// * `to` - End time of the segment (None for unbounded)
     /// * `method` - Interpolation method to use within this segment
-    pub(crate) fn new(from: f64, to: Option<f64>, method: Interpolation, observation: f64) -> Self {
-        CovariateSegment {
-            from,
-            to,
-            method,
-            observation: Some(observation),
-        }
+    pub(crate) fn new(from: f64, to: Option<f64>, method: Interpolation) -> Self {
+        CovariateSegment { from, to, method }
     }
 
     /// Get the original observation time (same as 'from' for observation-based segments)
@@ -58,12 +50,19 @@ impl CovariateSegment {
         self.from
     }
 
+    /// Whether this zero-width segment retains an exact source observation.
+    fn is_observation_marker(&self) -> bool {
+        self.to
+            .is_some_and(|to| to.to_bits() == self.from.to_bits())
+            && matches!(self.method, Interpolation::CarryForward { .. })
+    }
+
     /// Get the original observation value
     fn value(&self) -> f64 {
-        self.observation.unwrap_or(match self.method {
+        match self.method {
             Interpolation::Linear { slope, intercept } => slope * self.from + intercept,
             Interpolation::CarryForward { value } => value,
-        })
+        }
     }
 
     /// Interpolate the covariate value at a specific time within this segment
@@ -122,10 +121,20 @@ impl Covariate {
         let mut observations: Vec<(f64, f64)> = self
             .segments
             .iter()
+            .filter(|segment| segment.is_observation_marker())
             .map(|segment| (segment.time(), segment.value()))
             .collect();
 
-        // Remove duplicates and sort by time
+        // Older serialized covariates have no marker segments. Reconstruct their
+        // observations from the existing interpolation representation.
+        if observations.is_empty() {
+            observations = self
+                .segments
+                .iter()
+                .map(|segment| (segment.time(), segment.value()))
+                .collect();
+        }
+
         observations.sort_by(|a, b| a.0.total_cmp(&b.0));
         observations.dedup_by(|a, b| a.0 == b.0);
         observations
@@ -139,7 +148,6 @@ impl Covariate {
         if let Some(existing_segment) = self.segments.iter_mut().find(|seg| seg.time() == time) {
             // Update the existing observation's value
             existing_segment.method = Interpolation::CarryForward { value };
-            existing_segment.observation = Some(value);
             self.build_segments();
             return;
         }
@@ -149,7 +157,6 @@ impl Covariate {
             time,
             Some(time),
             Interpolation::CarryForward { value },
-            value,
         ));
 
         // Rebuild all segments
@@ -200,6 +207,16 @@ impl Covariate {
             let next_obs = observations.get(i + 1);
             let to_time = next_obs.map(|next| next.0);
 
+            // Retain the exact source pair without changing the serialized
+            // CovariateSegment layout. Zero-width markers never interpolate.
+            self.segments.push(CovariateSegment::new(
+                current_obs.0,
+                Some(current_obs.0),
+                Interpolation::CarryForward {
+                    value: current_obs.1,
+                },
+            ));
+
             if self.fixed {
[…]
