+++ src/data/parser/pmetrics.rs
@@ -8,18 +8,18 @@
 //! numeric values such as `1` are preserved as numeric-looking labels.
 
 use crate::{data::*, PharmsolError};
-use csv::{ReaderBuilder, StringRecord, Terminator, WriterBuilder};
+use csv::{ReaderBuilder, StringRecord, WriterBuilder};
 use serde::de::{MapAccess, Visitor};
 use serde::{de, Deserialize, Deserializer, Serialize};
-use std::collections::{BTreeMap, HashMap, HashSet};
+use std::collections::{HashMap, HashSet};
 use std::fmt;
 use std::fs::File;
-use std::io::{Cursor, Read, Write};
+use std::io::{Cursor, Read};
 use std::str::FromStr;
 
 use crate::data::row::{build_data, DataError, DataRow};
 
-const CORE_HEADERS: [&str; 15] = [
+pub(super) const CORE_HEADERS: [&str; 15] = [
     "ID", "EVID", "TIME", "DUR", "DOSE", "ADDL", "II", "INPUT", "OUT", "OUTEQ", "CENS", "C0", "C1",
     "C2", "C3",
 ];
@@ -108,7 +108,14 @@ pub fn read_pmetrics_csv_v1<R: Read>(mut reader: R) -> Result<Data, DataError> {
         ));
     }
 
-    read_pmetrics_reader(Cursor::new(bytes), true)
+    let data = read_pmetrics_reader(Cursor::new(bytes.as_slice()), true)?;
+    let canonical = data.to_pmetrics_csv_v1()?;
+    if canonical != bytes {
+        return Err(DataError::InvalidCanonicalFormat(
+            "input bytes are not canonical pmetrics-csv.v1".to_string(),
+        ));
+    }
+    Ok(data)
 }
 
 fn read_pmetrics_reader<R: Read>(reader: R, canonical: bool) -> Result<Data, DataError> {
@@ -216,7 +223,7 @@ fn validate_canonical_headers(headers: &StringRecord) -> Result<(), DataError> {
     Ok(())
 }
 
-fn validate_covariate_header(header: &str) -> Result<(), DataError> {
+pub(super) fn validate_covariate_header(header: &str) -> Result<(), DataError> {
     let base = header.strip_suffix('!').unwrap_or(header);
     if base.is_empty()
         || base.ends_with('!')
@@ -230,6 +237,17 @@ fn validate_covariate_header(header: &str) -> Result<(), DataError> {
         )));
     }
     Ok(())
+}
+
+fn ensure_finite(value: f64, field: &str, id: &str) -> Result<(), DataError> {
+    if value.is_finite() {
+        Ok(())
+    } else {
+        Err(DataError::NonFiniteValue {
+            field: field.to_string(),
+            id: id.to_string(),
+        })
+    }
 }
 
 /// One row from a Pmetrics file after serde deserialization.
@@ -597,327 +615,124 @@ where
     deserializer.deserialize_map(CovsVisitor)
 }
 
-#[derive(Debug)]
-struct CanonicalCovariate {
-    name: String,
-    header: String,
-}
-
-#[derive(Debug)]
-struct CanonicalRow {
-    time: f64,
-    rank: u8,
-    sequence: usize,
-    fields: Vec<String>,
-}
-
-fn ensure_finite(value: f64, field: &str, id: &str) -> Result<(), DataError> {
-    if value.is_finite() {
-        Ok(())
-    } else {
-        Err(DataError::NonFiniteValue {
-            field: field.to_string(),
-            id: id.to_string(),
-        })
-    }
-}
-
-fn ensure_label(label: &str, field: &str, id: &str) -> Result<(), DataError> {
[…]
