### 2026-07-31 36042e9
Preserve legacy covariate execution semantics
