+++ src/data/parser/pmetrics_csv_tests.rs
@@ -1,11 +1,9 @@
-use super::pmetrics::read_pmetrics_csv_v1;
 use super::{read_pmetrics, DataError};
 use crate::{Censor, Data, ErrorPoly, Event, Subject, SubjectBuilderExt};
 use std::collections::BTreeMap;
-use std::io::Cursor;
 use tempfile::NamedTempFile;
 
-const GOLDEN: &[u8] = include_bytes!("../../tests/data/pmetrics_csv_v1.csv");
+const GOLDEN: &[u8] = include_bytes!("../../tests/data/pmetrics_csv.csv");
 const CORE_HEADER: &str = "ID,EVID,TIME,DUR,DOSE,ADDL,II,INPUT,OUT,OUTEQ,CENS,C0,C1,C2,C3\n";
 
 fn fixture_data() -> Data {
@@ -13,7 +11,7 @@ fn fixture_data() -> Data {
         .infusion(0.0, 25.0, "1", 2.0)
         .observation(0.0, 4.5, "central")
         .covariate("wt", 1.0, 60.0)
-        .observation(2.0, -99.0, "neg")
+        .observation(2.0, -98.5, "neg")
         .build();
 
     let mut ten = Subject::builder("10")
@@ -41,7 +39,7 @@ fn fixture_data() -> Data {
         assert!(occasion.covariates_mut().set_covariate_fixed("age", true));
     }
 
-    // Deliberately unsorted subjects prove canonical subject ordering.
+    // Deliberately unsorted to check deterministic subject ordering.
     Data::new(vec![alpha, ten])
 }
 
@@ -137,72 +135,65 @@ fn assert_data_equivalent(left: &Data, right: &Data) {
 }
 
 #[test]
-fn canonical_bytes_match_golden_and_are_deterministic() {
+fn bytes_include_the_complete_dataset_and_are_deterministic() {
     let data = fixture_data();
-    let first = data.to_pmetrics_csv_v1().unwrap();
-    let second = data.to_pmetrics_csv_v1().unwrap();
+    let first = data.as_bytes().unwrap();
+    let second = data.as_bytes().unwrap();
     assert_eq!(first, GOLDEN);
     assert_eq!(first, second);
+    assert!(std::str::from_utf8(&first).unwrap().contains(",4,"));
+    assert!(std::str::from_utf8(&first).unwrap().contains(",2,"));
 }
 
 #[test]
-fn canonical_data_round_trip_is_scientifically_equivalent() {
+fn bytes_round_trip_the_complete_dataset() {
     let data = fixture_data();
-    let bytes = data.to_pmetrics_csv_v1().unwrap();
-    let parsed = read_pmetrics_csv_v1(Cursor::new(bytes)).unwrap();
+    let bytes = data.as_bytes().unwrap();
+    let parsed = Data::from_bytes(&bytes).unwrap();
     assert_data_equivalent(&data, &parsed);
+    assert_eq!(parsed.as_bytes().unwrap(), bytes);
 }
 
 #[test]
-fn canonical_bytes_reencode_exactly() {
-    let parsed = read_pmetrics_csv_v1(Cursor::new(GOLDEN)).unwrap();
-    assert_eq!(parsed.to_pmetrics_csv_v1().unwrap(), GOLDEN);
+fn write_pmetrics_writes_as_bytes() {
+    let data = fixture_data();
+    let expected = data.as_bytes().unwrap();
+    let file = NamedTempFile::new().unwrap();
+    data.write_pmetrics(file.as_file()).unwrap();
+    assert_eq!(std::fs::read(file.path()).unwrap(), expected);
 }
 
 #[test]
-fn alternate_missing_and_numeric_spellings_are_rejected() {
[…]
