+++ src/data/parser/pmetrics_csv.rs
@@ -1,20 +1,19 @@
-//! Canonical `pmetrics-csv.v1` writer implementation.
+//! Pmetrics CSV byte encoder.
 
 use super::pmetrics::{validate_covariate_header, CORE_HEADERS};
 use crate::data::row::DataError;
 use crate::data::{Censor, Data, Event};
 use csv::{Terminator, WriterBuilder};
 use std::collections::{BTreeMap, HashMap};
-use std::io::Write;
 
 #[derive(Debug)]
-struct CanonicalCovariate {
+struct CsvCovariate {
     name: String,
     header: String,
 }
 
 #[derive(Debug)]
-struct CanonicalRow {
+struct CsvRow {
     time: f64,
     rank: u8,
     sequence: usize,
@@ -34,7 +33,7 @@ fn ensure_finite(value: f64, field: &str, id: &str) -> Result<(), DataError> {
 
 fn ensure_label(label: &str, field: &str, id: &str) -> Result<(), DataError> {
     if label.is_empty() || label == "." || label == "NA" || label.contains('\r') {
-        Err(DataError::InvalidCanonicalFormat(format!(
+        Err(DataError::InvalidPmetricsData(format!(
             "{field} label `{label}` for {id} is reserved as missing"
         )))
     } else {
@@ -42,7 +41,7 @@ fn ensure_label(label: &str, field: &str, id: &str) -> Result<(), DataError> {
     }
 }
 
-fn collect_covariate_schema(data: &Data) -> Result<Vec<CanonicalCovariate>, DataError> {
+fn collect_covariate_schema(data: &Data) -> Result<Vec<CsvCovariate>, DataError> {
     let mut fixed_by_name = BTreeMap::<String, bool>::new();
     let mut folded_names = HashMap::<String, String>::new();
 
@@ -50,21 +49,21 @@ fn collect_covariate_schema(data: &Data) -> Result<Vec<CanonicalCovariate>, Data
         for occasion in subject.occasions() {
             for (key, covariate) in occasion.covariates().covariates() {
                 if key != covariate.name() {
-                    return Err(DataError::InvalidCanonicalFormat(format!(
+                    return Err(DataError::InvalidPmetricsData(format!(
                         "covariate key `{key}` does not match name `{}`",
                         covariate.name()
                     )));
                 }
                 validate_covariate_header(&key)?;
                 if key.ends_with('!') {
-                    return Err(DataError::InvalidCanonicalFormat(format!(
+                    return Err(DataError::InvalidPmetricsData(format!(
                         "covariate name `{key}` reserves trailing ! for fixed covariates"
                     )));
                 }
                 let folded = key.to_lowercase();
                 if let Some(existing) = folded_names.insert(folded, key.clone()) {
                     if existing != key {
-                        return Err(DataError::InvalidCanonicalFormat(format!(
+                        return Err(DataError::InvalidPmetricsData(format!(
                             "covariate names `{existing}` and `{key}` are ambiguous"
                         )));
                     }
@@ -77,7 +76,7 @@ fn collect_covariate_schema(data: &Data) -> Result<Vec<CanonicalCovariate>, Data
                     });
                 }
                 if covariate.observations().is_empty() {
-                    return Err(DataError::InvalidCanonicalFormat(format!(
+                    return Err(DataError::InvalidPmetricsData(format!(
                         "covariate `{key}` for subject `{}` occasion {} has no observations",
                         subject.id(),
                         occasion.index()
@@ -85,8 +84,8 @@ fn collect_covariate_schema(data: &Data) -> Result<Vec<CanonicalCovariate>, Data
                 }
                 if let Some(existing) = fixed_by_name.insert(key.clone(), covariate.fixed()) {
[…]
