+++ src/lib.rs
@@ -206,7 +206,7 @@ pub mod prelude {
     pub mod data {
         pub use crate::data::{
             error_model::{AssayErrorModel, AssayErrorModels},
-            parser::{read_pmetrics, read_pmetrics_csv_v1, DataError, DataRow, DataRowBuilder},
+            parser::{read_pmetrics, DataError, DataRow, DataRowBuilder},
             residual_error::{ResidualErrorModel, ResidualErrorModels},
             Covariates, Data, Event, Occasion, Subject,
         };
