+++ src/data/parser/pmetrics_csv.rs
@@ -3,8 +3,11 @@
 use super::pmetrics::{validate_covariate_header, CORE_HEADERS};
 use crate::data::row::DataError;
 use crate::data::{Censor, Data, Event};
+use crate::PharmsolError;
 use csv::{Terminator, WriterBuilder};
-use std::collections::{BTreeMap, HashMap};
+use std::collections::BTreeMap;
+use std::fs::File;
+use std::io::Write;
 
 #[derive(Debug)]
 struct CsvCovariate {
@@ -15,8 +18,6 @@ struct CsvCovariate {
 #[derive(Debug)]
 struct CsvRow {
     time: f64,
-    rank: u8,
-    sequence: usize,
     fields: Vec<String>,
 }
 
@@ -43,7 +44,6 @@ fn ensure_label(label: &str, field: &str, id: &str) -> Result<(), DataError> {
 
 fn collect_covariate_schema(data: &Data) -> Result<Vec<CsvCovariate>, DataError> {
     let mut fixed_by_name = BTreeMap::<String, bool>::new();
-    let mut folded_names = HashMap::<String, String>::new();
 
     for subject in data.subjects() {
         for occasion in subject.occasions() {
@@ -60,13 +60,10 @@ fn collect_covariate_schema(data: &Data) -> Result<Vec<CsvCovariate>, DataError>
                         "covariate name `{key}` reserves trailing ! for fixed covariates"
                     )));
                 }
-                let folded = key.to_lowercase();
-                if let Some(existing) = folded_names.insert(folded, key.clone()) {
-                    if existing != key {
-                        return Err(DataError::InvalidPmetricsData(format!(
-                            "covariate names `{existing}` and `{key}` are ambiguous"
-                        )));
-                    }
+                if key != key.to_lowercase() {
+                    return Err(DataError::InvalidPmetricsData(format!(
+                        "covariate name `{key}` must be lowercase for Pmetrics export"
+                    )));
                 }
                 if covariate.has_legacy_unmarked_linear_segments() {
                     return Err(DataError::LegacyLinearCovariate {
@@ -93,7 +90,7 @@ fn collect_covariate_schema(data: &Data) -> Result<Vec<CsvCovariate>, DataError>
         }
     }
 
-    let mut schema = fixed_by_name
+    Ok(fixed_by_name
         .into_iter()
         .map(|(name, fixed)| CsvCovariate {
             header: if fixed {
@@ -103,9 +100,7 @@ fn collect_covariate_schema(data: &Data) -> Result<Vec<CsvCovariate>, DataError>
             },
             name,
         })
-        .collect::<Vec<_>>();
-    schema.sort_by(|left, right| left.header.cmp(&right.header));
-    Ok(schema)
+        .collect())
 }
 
 fn empty_row(id: &str, evid: i32, time: f64, covariate_count: usize) -> Vec<String> {
@@ -138,10 +133,19 @@ fn event_row(id: &str, event: &Event, covariate_count: usize) -> Result<Vec<Stri
 
     match event {
         Event::Observation(observation) => {
-            if let Some(value) = observation.value() {
-                ensure_finite(value, "OUT", id)?;
-                fields[8] = value.to_string();
-            }
+            fields[8] = match observation.value() {
+                Some(value) => {
+                    ensure_finite(value, "OUT", id)?;
+                    if value == -99.0 {
+                        return Err(DataError::InvalidPmetricsData(format!(
+                            "observation OUT=-99 for {id} at time {} is reserved for missing data",
+                            observation.time()
+                        )));
+                    }
+                    value.to_string()
+                }
+                None => "-99".to_string(),
+            };
[…]
