### 2026-07-31 4b272f6
Reject ambiguous Pmetrics rows
