+++ src/data/parser/pmetrics_csv_tests.rs
@@ -273,6 +273,66 @@ fn later_occasion_starts_with_a_real_reset_dose() {
 }
 
 #[test]
+fn mixed_dose_types_at_reset_time_are_not_exportable() {
+    let input = concat!(
+        "ID,EVID,TIME,DUR,DOSE,ADDL,II,INPUT,OUT,OUTEQ,CENS,C0,C1,C2,C3\n",
+        "s,1,0,0,1,.,.,iv,.,.,.,.,.,.,.\n",
+        "s,4,0,2,2,.,.,iv,.,.,.,.,.,.,.\n",
+        "s,1,0,0,3,.,.,oral,.,.,.,.,.,.,.\n"
+    );
+    let data = Data::from_bytes(input.as_bytes()).unwrap();
+    assert!(matches!(
+        data.as_bytes(),
+        Err(DataError::InvalidPmetricsData(message))
+            if message.contains("both bolus and infusion doses at time 0")
+    ));
+}
+
+#[test]
+fn same_time_bolus_doses_keep_the_reset_order() {
+    let input = concat!(
+        "ID,EVID,TIME,DUR,DOSE,ADDL,II,INPUT,OUT,OUTEQ,CENS,C0,C1,C2,C3\n",
+        "s,1,0,0,1,.,.,iv,.,.,.,.,.,.,.\n",
+        "s,4,0,0,2,.,.,iv,.,.,.,.,.,.,.\n",
+        "s,1,0,0,3,.,.,oral,.,.,.,.,.,.,.\n"
+    );
+    let rows = records(
+        &Data::from_bytes(input.as_bytes())
+            .unwrap()
+            .as_bytes()
+            .unwrap(),
+    );
+    assert_eq!(rows[1].get(1), Some("4"));
+    assert_eq!(rows[1].get(4), Some("2"));
+    assert_eq!(rows[1].get(7), Some("iv"));
+    assert_eq!(rows[2].get(1), Some("1"));
+    assert_eq!(rows[2].get(4), Some("3"));
+    assert_eq!(rows[2].get(7), Some("oral"));
+}
+
+#[test]
+fn same_time_infusion_doses_keep_the_reset_order() {
+    let input = concat!(
+        "ID,EVID,TIME,DUR,DOSE,ADDL,II,INPUT,OUT,OUTEQ,CENS,C0,C1,C2,C3\n",
+        "s,1,0,0,1,.,.,iv,.,.,.,.,.,.,.\n",
+        "s,4,0,2,2,.,.,iv,.,.,.,.,.,.,.\n",
+        "s,1,0,3,3,.,.,peripheral,.,.,.,.,.,.,.\n"
+    );
+    let rows = records(
+        &Data::from_bytes(input.as_bytes())
+            .unwrap()
+            .as_bytes()
+            .unwrap(),
+    );
+    assert_eq!(rows[1].get(1), Some("4"));
+    assert_eq!(rows[1].get(3), Some("2"));
+    assert_eq!(rows[1].get(7), Some("iv"));
+    assert_eq!(rows[2].get(1), Some("1"));
+    assert_eq!(rows[2].get(3), Some("3"));
+    assert_eq!(rows[2].get(7), Some("peripheral"));
+}
+
+#[test]
 fn later_occasion_without_an_initial_dose_is_not_exportable() {
     let data = Data::new(vec![Subject::builder("s")
         .bolus(0.0, 1.0, "iv")
@@ -350,6 +410,56 @@ fn positive_addl_reset_round_trips_as_individual_doses() {
 
     let reparsed = Data::from_bytes(&bytes).unwrap();
     assert_data_equivalent(&data, &reparsed);
+}
+
+#[test]
[…]
