+++ src/data/parser/pmetrics_csv_tests.rs
@@ -1,11 +1,27 @@
+use super::pmetrics::CORE_HEADERS;
 use super::{read_pmetrics, DataError};
 use crate::{Censor, Covariate, Data, ErrorPoly, Event, Subject, SubjectBuilderExt};
 use csv::StringRecord;
 use std::collections::BTreeMap;
 use tempfile::NamedTempFile;
 
-const GOLDEN: &[u8] = include_bytes!("../../tests/data/pmetrics_csv.csv");
-const CORE_HEADER: &str = "ID,EVID,TIME,DUR,DOSE,ADDL,II,INPUT,OUT,OUTEQ,CENS,C0,C1,C2,C3\n";
+const GOLDEN: &[u8] = concat!(
+    "ID,EVID,TIME,DUR,DOSE,ADDL,II,INPUT,OUT,OUTEQ,CENS,C0,C1,C2,C3,age!,crcl,wt\n",
+    "10,0,0,.,.,.,.,.,1.25,cp,1,0.1,0.2,0.3,0.4,40,.,70\n",
+    "10,1,0,0,100,.,.,iv,.,.,.,.,.,.,.,40,.,70\n",
+    "10,0,1,.,.,.,.,.,-99,2,0,.,.,.,.,.,.,.\n",
+    "10,1,2,4,50,.,.,1,.,.,.,.,.,.,.,41,.,72\n",
+    "10,4,0,0,80,.,.,oral,.,.,.,.,.,.,.,42,.,.\n",
+    "10,0,3,.,.,.,.,.,9,1,-1,.,.,.,.,.,80,.\n",
+    "alpha,0,0,.,.,.,.,.,4.5,central,0,.,.,.,.,.,.,60\n",
+    "alpha,1,0,2,25,.,.,1,.,.,.,.,.,.,.,.,.,60\n",
+    "alpha,0,2,.,.,.,.,.,-98.5,neg,0,.,.,.,.,.,.,.\n",
+)
+.as_bytes();
+
+fn core_header() -> String {
+    format!("{}\n", CORE_HEADERS.join(","))
+}
 
 fn fixture_data() -> Data {
     let alpha = Subject::builder("alpha")
@@ -144,20 +160,20 @@ fn assert_data_equivalent(left: &Data, right: &Data) {
 #[test]
 fn representable_data_round_trips_with_exact_covariate_values() {
     let data = fixture_data();
-    let first = data.as_bytes().unwrap();
-    let second = data.as_bytes().unwrap();
+    let first = data.to_pmetrics_csv_bytes().unwrap();
+    let second = data.to_pmetrics_csv_bytes().unwrap();
     assert_eq!(first, GOLDEN);
     assert_eq!(first, second);
 
-    let parsed = Data::from_bytes(&first).unwrap();
+    let parsed = Data::from_pmetrics_csv_bytes(&first).unwrap();
     assert_data_equivalent(&data, &parsed);
-    assert_eq!(parsed.as_bytes().unwrap(), first);
+    assert_eq!(parsed.to_pmetrics_csv_bytes().unwrap(), first);
 }
 
 #[test]
 fn write_pmetrics_writes_as_bytes() {
     let data = fixture_data();
-    let expected = data.as_bytes().unwrap();
+    let expected = data.to_pmetrics_csv_bytes().unwrap();
     let file = NamedTempFile::new().unwrap();
     data.write_pmetrics(file.as_file()).unwrap();
     assert_eq!(std::fs::read(file.path()).unwrap(), expected);
@@ -171,14 +187,14 @@ fn existing_pmetrics_files_remain_readable() {
     );
     let bytes = std::fs::read(&path).unwrap();
[…]
