### 2026-08-01 30f636f
Centralize Pmetrics CSV schema and validation
