+++ src/data/parser/pmetrics.rs
@@ -10,17 +10,88 @@
 use crate::data::*;
 use csv::{ReaderBuilder, StringRecord};
 use serde::de::{MapAccess, Visitor};
-use serde::{de, Deserialize, Deserializer, Serialize};
+use serde::{de, Deserialize, Deserializer};
 use std::collections::{HashMap, HashSet};
 use std::fmt;
 use std::str::FromStr;
 
 use crate::data::row::{build_data, DataError, DataRow};
 
-pub(super) const CORE_HEADERS: [&str; 15] = [
-    "ID", "EVID", "TIME", "DUR", "DOSE", "ADDL", "II", "INPUT", "OUT", "OUTEQ", "CENS", "C0", "C1",
-    "C2", "C3",
-];
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[repr(usize)]
+pub(super) enum CoreColumn {
+    Id,
+    Evid,
+    Time,
+    Dur,
+    Dose,
+    Addl,
+    Ii,
+    Input,
+    Out,
+    Outeq,
+    Cens,
+    C0,
+    C1,
+    C2,
+    C3,
+}
+
+impl CoreColumn {
+    pub(super) const ALL: [Self; 15] = [
+        Self::Id,
+        Self::Evid,
+        Self::Time,
+        Self::Dur,
+        Self::Dose,
+        Self::Addl,
+        Self::Ii,
+        Self::Input,
+        Self::Out,
+        Self::Outeq,
+        Self::Cens,
+        Self::C0,
+        Self::C1,
+        Self::C2,
+        Self::C3,
+    ];
+    const REQUIRED: [Self; 3] = [Self::Id, Self::Evid, Self::Time];
+    pub(super) const COUNT: usize = Self::ALL.len();
+
+    pub(super) const fn index(self) -> usize {
+        self as usize
+    }
+
+    pub(super) const fn header(self) -> &'static str {
+        match self {
+            Self::Id => "ID",
+            Self::Evid => "EVID",
+            Self::Time => "TIME",
+            Self::Dur => "DUR",
+            Self::Dose => "DOSE",
+            Self::Addl => "ADDL",
+            Self::Ii => "II",
+            Self::Input => "INPUT",
+            Self::Out => "OUT",
+            Self::Outeq => "OUTEQ",
+            Self::Cens => "CENS",
+            Self::C0 => "C0",
+            Self::C1 => "C1",
+            Self::C2 => "C2",
+            Self::C3 => "C3",
+        }
+    }
+
+    fn from_header(header: &str) -> Option<Self> {
+        Self::ALL
+            .into_iter()
+            .find(|column| column.header().eq_ignore_ascii_case(header))
+    }
+}
+
+pub(super) fn core_headers() -> impl ExactSizeIterator<Item = &'static str> {
+    CoreColumn::ALL.into_iter().map(CoreColumn::header)
+}
 
 /// Read a Pmetrics CSV file into [`Data`].
 ///
@@ -64,9 +135,9 @@ pub(super) const CORE_HEADERS: [&str; 15] = [
 ///
 /// # Expected columns
 ///
-/// The core columns are `ID`, `TIME`, `EVID`, `DOSE`, `DUR`, `ADDL`,
-/// `II`, `INPUT`, `OUT`, `OUTEQ`, `CENS`, and optional `C0..C3` error
-/// coefficients.
+/// `ID`, `EVID`, and `TIME` are required. The remaining core columns are
+/// `DOSE`, `DUR`, `ADDL`, `II`, `INPUT`, `OUT`, `OUTEQ`, `CENS`, and optional
+/// `C0..C3` error coefficients.
 ///
 /// All other numeric columns are treated as covariates.
 ///
[…]
