+++ src/data/parser/pmetrics_csv.rs
@@ -1,26 +1,15 @@
 //! Pmetrics CSV byte encoder.
 
-use super::pmetrics::{normalize_covariate_name, validate_covariate_header, CORE_HEADERS};
+use super::pmetrics::{
+    core_headers, normalize_covariate_name, validate_covariate_header, CoreColumn,
+};
 use crate::data::row::DataError;
 use crate::data::{Censor, Data, Event, Occasion, Subject};
 use crate::PharmsolError;
 use csv::{Terminator, WriterBuilder};
 use std::collections::BTreeMap;
 use std::fs::File;
-use std::io::Write;
-
-mod column {
-    pub const ID: usize = 0;
-    pub const EVID: usize = 1;
-    pub const TIME: usize = 2;
-    pub const DUR: usize = 3;
-    pub const DOSE: usize = 4;
-    pub const INPUT: usize = 7;
-    pub const OUT: usize = 8;
-    pub const OUTEQ: usize = 9;
-    pub const CENS: usize = 10;
-    pub const C0: usize = 11;
-}
+use std::io::{Seek, SeekFrom, Write};
 
 #[derive(Clone, Copy)]
 enum PmetricsEvid {
@@ -139,10 +128,10 @@ fn collect_covariate_schema(data: &Data) -> Result<Vec<PmetricsCovariateColumn>,
 }
 
 fn empty_row(id: &str, evid: PmetricsEvid, time: f64, covariate_count: usize) -> Vec<String> {
-    let mut fields = vec![".".to_string(); CORE_HEADERS.len() + covariate_count];
-    fields[column::ID] = id.to_string();
-    fields[column::EVID] = evid.as_str().to_string();
-    fields[column::TIME] = time.to_string();
+    let mut fields = vec![".".to_string(); CoreColumn::COUNT + covariate_count];
+    fields[CoreColumn::Id.index()] = id.to_string();
+    fields[CoreColumn::Evid.index()] = evid.as_str().to_string();
+    fields[CoreColumn::Time.index()] = time.to_string();
     fields
 }
 
@@ -156,7 +145,7 @@ fn event_row(id: &str, event: &Event, covariate_count: usize) -> Result<Vec<Stri
 
     match event {
         Event::Observation(observation) => {
-            fields[column::OUT] = match observation.value() {
+            fields[CoreColumn::Out.index()] = match observation.value() {
                 Some(value) => {
                     ensure_finite(value, "OUT", id)?;
                     if value == -99.0 {
@@ -171,8 +160,8 @@ fn event_row(id: &str, event: &Event, covariate_count: usize) -> Result<Vec<Stri
             };
             let outeq = observation.outeq().to_string();
             ensure_label(&outeq, "OUTEQ", id)?;
-            fields[column::OUTEQ] = outeq;
-            fields[column::CENS] = match observation.censoring() {
+            fields[CoreColumn::Outeq.index()] = outeq;
+            fields[CoreColumn::Cens.index()] = match observation.censoring() {
                 Censor::None => "0",
                 Censor::BLOQ => "1",
                 Censor::ALOQ => "-1",
@@ -190,7 +179,7 @@ fn event_row(id: &str, event: &Event, covariate_count: usize) -> Result<Vec<Stri
                 .enumerate()
                 {
                     ensure_finite(value, &format!("C{offset}"), id)?;
-                    fields[column::C0 + offset] = value.to_string();
+                    fields[CoreColumn::C0.index() + offset] = value.to_string();
                 }
             }
         }
@@ -198,9 +187,9 @@ fn event_row(id: &str, event: &Event, covariate_count: usize) -> Result<Vec<Stri
             ensure_finite(bolus.amount(), "DOSE", id)?;
[…]

[…]
