### 2026-08-02 3139a73
answering some comments
