### 2026-08-03 15ae0b0
paper: Address editor comments
Line 46: "with under" -> "and released under"?
Line 52: "a flexible" -> "with a flexible"
Case protected the Neely reference to Pmetrics
Added DOI to mrgsolve reference
