+++ joss/paper.bib
@@ -11,7 +11,7 @@
 @Article{pmetrics,
   author    = {Neely, Michael N. and van Guilder, Michael G. and Yamada, Walter M. and Schumitzky, Alan and Jelliffe, Roger W.},
   journal   = {Therapeutic Drug Monitoring},
-  title     = {Accurate Detection of Outliers and Subpopulations With Pmetrics, a Nonparametric and Parametric Pharmacometric Modeling and Simulation Package for R},
+  title     = {Accurate Detection of Outliers and Subpopulations With {Pmetrics}, a Nonparametric and Parametric Pharmacometric Modeling and Simulation Package for {R}},
   year      = {2012},
   issn      = {0163-4356},
   month     = aug,
@@ -104,6 +104,7 @@
   title = {mrgsolve: Simulate from ODE-Based Models},
   author = {Kyle T Baron},
   year = {2025},
+  doi = {10.32614/cran.package.mrgsolve},
   note = {R package version 1.6.1},
   url = {https://mrgsolve.org/docs/},
 }
