### 2026-08-03 df8eb82
chore: release v0.28.4 (#314)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
