+++ Cargo.toml
@@ -47,30 +47,31 @@ wasm = ["dsl-wasm"]
 pharmsol-dsl = { workspace = true }
 pharmsol-macros = { workspace = true }
 libloading = { version = "0.9.0", optional = true, features = [] }
-cranelift = { version = "0.131.0", optional = true }
-cranelift-jit = { version = "0.131.0", optional = true }
-cranelift-module = { version = "0.131.0", optional = true }
-cranelift-native = { version = "0.131.0", optional = true }
+cranelift = { version = "0.134.3", optional = true }
+cranelift-jit = { version = "0.134.3", optional = true }
+cranelift-module = { version = "0.134.3", optional = true }
+cranelift-native = { version = "0.134.3", optional = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 thiserror = { workspace = true }
 tracing = "0.1.44"
-wasm-encoder = { version = "0.247.0", optional = true }
+wasm-encoder = { version = "0.255.0", optional = true }
 
 [target.'cfg(not(all(target_arch = "wasm32", target_os = "unknown")))'.dependencies]
 ahash = "0.8.12"
 csv = "1.4.0"
-diffsol = "=0.15.1"
+diffsol = "=0.16.1"
+diffsol-la = "0.1.0"
 nalgebra = "0.35.0"
 ndarray = { version = "0.17.2", features = ["rayon"] }
 rand = "0.10.1"
 rand_distr = "0.6.0"
 rayon = "1.12.0"
-statrs = "0.18.0"
+statrs = "0.19.0"
 argmin = "0.11.0"
 argmin-math = "0.5.1"
 quick_cache = "0.7.0"
-wasmtime = { version = "44.0.0", optional = true }
+wasmtime = { version = "47.0.3", optional = true }
 
 [dev-dependencies]
 criterion = { version = "0.8.2", features = ["html_reports"] }
