### 2026-08-04 0a548af
feat: Allow macro use in transitive dependencies
If pharmsol is not imported directly, the ode! macro did not resolve like it was supposed to, as it was hardlocked.
