### 2026-08-04 29f18f3
refactor(macros): split monolithic lib.rs into modules
