+++ pharmsol-macros/src/lib.rs
@@ -3,3378 +3,18 @@
 //! This crate is not intended to be used directly. Use the re-exports from the
 //! `pharmsol` crate instead.
 
-use pharmsol_dsl::{
-    AnalyticalKernel as ResolverAnalyticalKernel, AnalyticalStructureInputKind,
-    AnalyticalStructureInputPlan, AnalyticalStructureInputSource,
-};
+mod analysis;
+mod bindings;
+mod crate_path;
+mod expand;
+mod input;
+mod kernel;
+mod symbols;
+mod validate;
+
 use proc_macro::TokenStream;
-use proc_macro2::{Group, Spacing, Span, TokenStream as TokenStream2, TokenTree};
-use proc_macro_crate::{crate_name, FoundCrate};
-use quote::{quote, ToTokens};
-use std::collections::{HashMap, HashSet};
-use syn::{
-    ext::IdentExt,
-    parse::{Parse, ParseStream, Parser},
-    punctuated::Punctuated,
-    token,
-    visit::Visit,
-    visit_mut::VisitMut,
-    Expr, ExprClosure, Ident, Lit, LitInt, LitStr, Pat, Path, PathArguments, Stmt, Token,
-};
 
-// ---------------------------------------------------------------------------
-// Crate path resolution
-// ---------------------------------------------------------------------------
-//
-// Every `quote!` block in this file emits the placeholder path `::pharmsol`.
-// Just before a macro returns, `rewrite_crate_paths` replaces that placeholder
-// with the path resolved for the current invocation, so nested macro calls
-// (`::pharmsol::fetch_cov!`) are rewritten alongside plain type paths.
-//
-// A leading `::` only resolves against the extern prelude, which is populated
-// from direct dependencies, so it breaks for crates that receive `pharmsol`
-// transitively. The `macro_rules!` wrappers in `pharmsol` therefore forward
-// `$crate` through the `@pharmsol_crate(...)` marker below, which resolves
-// correctly no matter how the caller reached the crate.
-
-mod kw {
-    syn::custom_keyword!(pharmsol_crate);
-}
-
-/// Consumes the `@pharmsol_crate($crate)` prefix injected by the `macro_rules!`
-/// wrappers in `pharmsol`.
-fn parse_crate_marker(input: ParseStream) -> syn::Result<Option<TokenStream2>> {
-    if !(input.peek(Token![@]) && input.peek2(kw::pharmsol_crate)) {
-        return Ok(None);
-    }
-
-    input.parse::<Token![@]>()?;
-    input.parse::<kw::pharmsol_crate>()?;
-    let marker;
-    syn::parenthesized!(marker in input);
-    let path = marker.parse::<TokenStream2>()?;
-    if input.peek(Token![,]) {
-        input.parse::<Token![,]>()?;
-    }
-
-    Ok(Some(path))
-}
-
-/// Resolves the crate path for one macro invocation.
-///
-/// An explicit `crate: "..."` key wins, then the `$crate` forwarded by the
-/// wrapper macros, then detection from the caller's `Cargo.toml`.
-fn resolve_crate_path(
-    explicit: Option<LitStr>,
-    forwarded: Option<TokenStream2>,
-) -> syn::Result<TokenStream2> {
-    if let Some(literal) = explicit {
-        let path: Path = literal.parse()?;
-        if let Some(segment) = path
-            .segments
-            .iter()
-            .find(|segment| !matches!(segment.arguments, PathArguments::None))
-        {
-            return Err(syn::Error::new_spanned(
-                segment,
-                "`crate` must be a plain module path without generic arguments",
-            ));
-        }
-        return Ok(absolute_crate_path(path));
-    }
-
[…]
