+++ pharmsol-macros/src/validate/derive.rs
@@ -1,0 +1,321 @@
+//! Validation of the `derived` / `derive` contract for built-in analytical
+//! structures.
+
+use pharmsol_dsl::{
+    AnalyticalKernel as ResolverAnalyticalKernel, AnalyticalStructureInputKind,
+    AnalyticalStructureInputPlan, AnalyticalStructureInputSource,
+};
+use proc_macro2::Span;
+use std::collections::HashSet;
+use syn::{visit::Visit, Expr, ExprClosure, Ident, Pat, Stmt};
+
+use crate::analysis::generated_ident;
+use crate::bindings::generate_supported_input_aliases;
+
+fn analytical_error_span<'a>(names: &'a [Ident], target: &str) -> Option<&'a Ident> {
+    names.iter().find(|ident| *ident == target)
+}
+
+/// Resolves how the kernel's required inputs are sourced from `params` and
+/// `derived`, translating resolver errors into spanned diagnostics.
+pub(crate) fn validate_analytical_structure_inputs(
+    structure: &Ident,
+    function: ResolverAnalyticalKernel,
+    params: &[Ident],
+    derived: &[Ident],
+) -> syn::Result<AnalyticalStructureInputPlan> {
+    let primary_names = params.iter().map(Ident::to_string).collect::<Vec<_>>();
+    let derived_names = derived.iter().map(Ident::to_string).collect::<Vec<_>>();
+    AnalyticalStructureInputPlan::for_kernel(function, &primary_names, &derived_names).map_err(
+        |error| match error {
+            pharmsol_dsl::AnalyticalStructureInputError::DuplicatePrimary { name } => {
+                let span = analytical_error_span(params, &name).unwrap_or(structure);
+                syn::Error::new_spanned(span, format!("duplicate primary parameter `{name}`"))
+            }
+            pharmsol_dsl::AnalyticalStructureInputError::DuplicateDerived { name } => {
+                let span = analytical_error_span(derived, &name).unwrap_or(structure);
+                syn::Error::new_spanned(span, format!("duplicate derived parameter `{name}`"))
+            }
+            pharmsol_dsl::AnalyticalStructureInputError::ConflictingName { name } => {
+                let span = analytical_error_span(derived, &name)
+                    .or_else(|| analytical_error_span(params, &name))
+                    .unwrap_or(structure);
+                syn::Error::new_spanned(
+                    span,
+                    format!("`{name}` is declared in both `params` and `derived`"),
+                )
+            }
+            pharmsol_dsl::AnalyticalStructureInputError::MissingRequiredName {
+                structure,
+                name,
+                suggestion,
+            } => {
+                let message = if let Some(candidate) = suggestion {
+                    format!(
+                        "analytical structure `{structure}` requires `{name}`; did you mean `{candidate}`? declare it in `params: [...]` or `derived: [...]`"
+                    )
+                } else {
+                    format!(
+                        "analytical structure `{structure}` requires `{name}`; declare it in `params: [...]` or `derived: [...]`"
+                    )
+                };
+                syn::Error::new_spanned(structure, message)
+            }
+        },
+    )
+}
+
+#[derive(Clone)]
+struct DeriveValidationContext {
+    params: HashSet<String>,
+    covariates: HashSet<String>,
+    derived: HashSet<String>,
+}
+
+impl DeriveValidationContext {
+    fn new(params: &[Ident], covariates: &[Ident], derived: &[Ident]) -> Self {
+        Self {
+            params: params.iter().map(Ident::to_string).collect(),
+            covariates: covariates.iter().map(Ident::to_string).collect(),
+            derived: derived.iter().map(Ident::to_string).collect(),
+        }
+    }
+
+    fn invalid_target_error(&self, ident: &Ident) -> syn::Error {
[…]

[…]
