### 2026-08-05 0220695
force the solver to stop at the infusions end time
