### 2026-08-05 c5730b1
the previous implementation was a bit fragile. trying a different approach
