+++ src/simulator/equation/ode/closure.rs
@@ -3,7 +3,10 @@ use diffsol::{
     ConstantOp, LinearOp, MatrixCommon, NalgebraContext, NalgebraMat, NonLinearOp,
     NonLinearOpJacobian, OdeEquations, OdeEquationsRef, Op, UnitCallable, Vector,
 };
-use std::{cell::RefCell, cmp::Ordering};
+use std::{
+    cell::{Cell, RefCell},
+    cmp::Ordering,
+};
 type M = NalgebraMat<f64>;
 type V = <M as MatrixCommon>::V;
 type C = <M as MatrixCommon>::C;
@@ -37,7 +40,43 @@ impl InfusionTrack {
         }
     }
 
-    fn rate_at(&self, time: f64) -> f64 {
+    fn rate_at(&self, time: f64, left_continuity_time: Option<f64>) -> f64 {
+        if let Some(left_continuity_time) = left_continuity_time {
+            if left_continuity_time == time {
+                return self.rate_at_left(time);
+            }
+        }
+
+        self.rate_at_right(time)
+    }
+
+    fn rate_at_left(&self, time: f64) -> f64 {
+        if self.event_times.is_empty() {
+            return 0.0;
+        }
+
+        let event_search = self
+            .event_times
+            .binary_search_by(|probe| probe.partial_cmp(&time).unwrap_or(Ordering::Less));
+
+        match event_search {
+            Ok(mut idx) => {
+                while idx > 0 && self.event_times[idx - 1] == time {
+                    idx -= 1;
+                }
+
+                if idx == 0 {
+                    0.0
+                } else {
+                    self.cumulative_rates[idx - 1]
+                }
+            }
+            Err(0) => 0.0,
+            Err(idx) => self.cumulative_rates[idx - 1],
+        }
+    }
+
+    fn rate_at_right(&self, time: f64) -> f64 {
         if self.event_times.is_empty() {
             return 0.0;
         }
@@ -63,6 +102,8 @@ impl InfusionTrack {
 #[derive(Debug, Clone, Default)]
 struct InfusionSchedule {
     tracks: Vec<InfusionTrack>,
+    discontinuity_times: Vec<f64>,
+    left_continuity_time: Cell<Option<f64>>,
 }
 
 impl InfusionSchedule {
@@ -71,11 +112,16 @@ impl InfusionSchedule {
         I: IntoIterator<Item = &'a Infusion>,
     {
         if ndrugs == 0 {
-            return Ok(Self { tracks: Vec::new() });
+            return Ok(Self {
+                tracks: Vec::new(),
+                discontinuity_times: Vec::new(),
+                left_continuity_time: Cell::new(None),
+            });
         }
 
         let mut per_input: Vec<Vec<(f64, f64)>> = vec![Vec::new(); ndrugs];
         let mut saw_infusion = false;
+        let mut discontinuity_times = Vec::new();
         for infusion in infusions {
             saw_infusion = true;
             if infusion.duration() <= 0.0 {
@@ -90,12 +136,22 @@ impl InfusionSchedule {
             }
 
             let rate = infusion.amount() / infusion.duration();
+            let end = infusion.time() + infusion.duration();
+
             per_input[input].push((infusion.time(), rate));
-            per_input[input].push((infusion.time() + infusion.duration(), -rate));
+            per_input[input].push((end, -rate));
+            discontinuity_times.push(end);
         }
 
+        discontinuity_times.sort_by(|a, b| a.total_cmp(b));
+        discontinuity_times.dedup_by(|a, b| a == b);
+
         if !saw_infusion {
-            return Ok(Self { tracks: Vec::new() });
+            return Ok(Self {
+                tracks: Vec::new(),
+                discontinuity_times,
[…]
