+++ src/simulator/equation/ode/mod.rs
@@ -544,28 +544,6 @@ impl EquationPriv for ODE {
 }
 
 impl ODE {
-    fn infusion_end_times(events: &[Event]) -> Vec<f64> {
-        let mut stops: Vec<f64> = events
-            .iter()
-            .filter_map(|event| {
-                let infusion = match event {
-                    Event::Infusion(infusion) => infusion,
-                    _ => return None,
-                };
-                let start = infusion.time();
-                let end = start + infusion.duration();
-                if end.is_finite() && end > start {
-                    Some(end)
-                } else {
-                    None
-                }
-            })
-            .collect();
-        stops.sort_by(|a, b| a.total_cmp(b));
-        stops.dedup_by(|a, b| a == b);
-        stops
-    }
-
     /// Generic event-loop runner, parameterized over the concrete solver type.
     #[allow(clippy::too_many_arguments)]
     fn run_events<'a, F, S>(
@@ -588,8 +566,8 @@ impl ODE {
         F: Fn(&V, &V, f64, &mut V, &V, &V, &Covariates) + 'a,
         S: OdeSolverMethod<'a, PMProblem<'a, F>>,
     {
-        const STOP_TIME_EPS: f64 = 1e-12;
-        let infusion_stop_times = Self::infusion_end_times(events);
+        let infusion_end_times = solver.problem().eqn.infusion_end_times();
+        let mut infusion_end_cursor = 0usize;
         let mut index = 0usize;
         while index < events.len() {
             let event = &events[index];
@@ -671,23 +649,44 @@ impl ODE {
 
             // Advance to the next event time if it exists
             if let Some(next_event) = next_event {
-                while (next_event.time() - solver.state().t) > STOP_TIME_EPS {
-                    let mut stop_time = next_event.time();
-                    let current_t = solver.state().t;
-                    for stop in &infusion_stop_times {
-                        if stop > &(current_t + STOP_TIME_EPS)
-                            && stop < &(stop_time - STOP_TIME_EPS)
-                        {
-                            stop_time = *stop;
-                            break;
-                        }
+                let next_event_time = next_event.time();
+                while next_event_time > solver.state().t {
+                    while infusion_end_cursor < infusion_end_times.len()
+                        && infusion_end_times[infusion_end_cursor] <= solver.state().t
+                    {
+                        infusion_end_cursor += 1;
                     }
+
+                    let (stop_time, is_infusion_end) =
+                        if let Some(stop_time) = infusion_end_times.get(infusion_end_cursor) {
+                            if *stop_time <= next_event_time {
+                                infusion_end_cursor += 1;
+                                (*stop_time, true)
+                            } else {
+                                (next_event_time, false)
+                            }
+                        } else {
+                            (next_event_time, false)
+                        };
+
+                    solver
+                        .problem()
+                        .eqn
+                        .set_left_continuity_time(if is_infusion_end {
+                            Some(stop_time)
+                        } else {
+                            None
+                        });
 
                     match solver.set_stop_time(stop_time) {
                         Ok(_) => loop {
                             match solver.step() {
                                 Ok(OdeSolverStopReason::InternalTimestep) => continue,
                                 Ok(OdeSolverStopReason::TstopReached) => {
+                                    solver.problem().eqn.set_left_continuity_time(None);
+                                    if is_infusion_end {
+                                        let _ = solver.state_mut();
+                                    }
[…]
