### 2026-08-05 b00477f
the previous implementation was a bit fragile. trying a different approach
