+++ src/simulator/equation/ode/mod.rs
@@ -328,6 +328,8 @@ fn _simulate_subject_dense(
     let zero_bolus = V::zeros(ndrugs, NalgebraContext::new());
     let zero_rateiv = V::zeros(ndrugs, NalgebraContext::new());
     let mut bolus_v = V::zeros(ndrugs, NalgebraContext::new());
+    // Scratch for refreshing the solver's derivative at infusion boundaries.
+    let mut dy_scratch = V::zeros(nstates, NalgebraContext::new());
     let parameters_vec = parameters.to_vec();
     let parameters_v: V = DVector::from_vec(parameters_vec.clone()).into();
 
@@ -376,6 +378,7 @@ fn _simulate_subject_dense(
                         &zero_rateiv,
                         &mut state_with_bolus,
                         &mut state_without_bolus,
+                        &mut dy_scratch,
                         &mut y_out,
                         &mut likelihood,
                         &mut output,
@@ -395,6 +398,7 @@ fn _simulate_subject_dense(
                         &zero_rateiv,
                         &mut state_with_bolus,
                         &mut state_without_bolus,
+                        &mut dy_scratch,
                         &mut y_out,
                         &mut likelihood,
                         &mut output,
@@ -414,6 +418,7 @@ fn _simulate_subject_dense(
                         &zero_rateiv,
                         &mut state_with_bolus,
                         &mut state_without_bolus,
+                        &mut dy_scratch,
                         &mut y_out,
                         &mut likelihood,
                         &mut output,
@@ -433,6 +438,7 @@ fn _simulate_subject_dense(
                         &zero_rateiv,
                         &mut state_with_bolus,
                         &mut state_without_bolus,
+                        &mut dy_scratch,
                         &mut y_out,
                         &mut likelihood,
                         &mut output,
@@ -543,6 +549,46 @@ impl EquationPriv for ODE {
     }
 }
 
+/// Restart the solver after an infusion boundary (a RHS discontinuity).
+///
+/// The multi-step history and the internal Jacobian were built for the
+/// pre-boundary RHS, so the first step into the new segment must not reuse
+/// them as-is:
+/// - `set_state` forces diffsol to recompute its BDF coefficients and
+///   reinitialize the internal Jacobian for the current state;
+/// - the stored derivative is refreshed against the post-boundary
+///   (right-continuous) RHS so a first-order restart predicts with the new
+///   dynamics instead of the pre-boundary ones;
+/// - `state_mut` marks the state as modified so the next step restarts the
+///   multi-step method at first order.
+///
+/// The first step is additionally capped just below `next_stop_time` (the
+/// stop the next segment is integrating toward): a restart step landing
+/// exactly on the next discontinuity inherits a Jacobian built for the
+/// pre-boundary order and can make the nonlinear solve oscillate, while a
+/// step that ends just before it converges cleanly (the error controller
+/// takes over from there).
+fn reinitialize_at_boundary<'a, F, S>(solver: &mut S, dy_scratch: &mut V, next_stop_time: f64)
+where
+    F: Fn(&V, &V, f64, &mut V, &V, &V, &Covariates) + 'a,
+    S: OdeSolverMethod<'a, PMProblem<'a, F>>,
+{
+    let state = solver.state_clone();
+    solver.set_state(state);
+
+    let t = solver.state().t;
+    {
+        let y = solver.state().y;
+        solver
+            .problem()
+            .eqn
+            .refresh_state_derivative(t, y, dy_scratch);
+    }
+    let state = solver.state_mut();
+    state.dy.copy_from(dy_scratch);
+    *state.h = (*state.h).min(0.9 * (next_stop_time - t).max(0.0));
+}
+
[…]
