### 2026-08-05 bd2d690
added some tests and covered a edge case scenario
