+++ src/simulator/equation/ode/mod.rs
@@ -561,14 +561,7 @@ impl EquationPriv for ODE {
 ///   dynamics instead of the pre-boundary ones;
 /// - `state_mut` marks the state as modified so the next step restarts the
 ///   multi-step method at first order.
-///
-/// The first step is additionally capped just below `next_stop_time` (the
-/// stop the next segment is integrating toward): a restart step landing
-/// exactly on the next discontinuity inherits a Jacobian built for the
-/// pre-boundary order and can make the nonlinear solve oscillate, while a
-/// step that ends just before it converges cleanly (the error controller
-/// takes over from there).
-fn reinitialize_at_boundary<'a, F, S>(solver: &mut S, dy_scratch: &mut V, next_stop_time: f64)
+fn reinitialize_at_boundary<'a, F, S>(solver: &mut S, dy_scratch: &mut V)
 where
     F: Fn(&V, &V, f64, &mut V, &V, &V, &Covariates) + 'a,
     S: OdeSolverMethod<'a, PMProblem<'a, F>>,
@@ -586,7 +579,6 @@ where
     }
     let state = solver.state_mut();
     state.dy.copy_from(dy_scratch);
-    *state.h = (*state.h).min(0.9 * (next_stop_time - t).max(0.0));
 }
 
 impl ODE {
@@ -618,8 +610,9 @@ impl ODE {
         let mut index = 0usize;
         // Set when the previous stop was an infusion boundary: the solver must
         // be restarted before the first step of the next segment (the RHS is
-        // discontinuous at the boundary). Deferred until the next stop time is
-        // known so the first step can be capped below it.
+        // discontinuous at the boundary). Deferred until `set_stop_time`
+        // succeeds so a stop that is already reached does not trigger a
+        // restart for a zero-length segment.
         let mut pending_reinit = false;
         while index < events.len() {
             let event = &events[index];
@@ -722,11 +715,6 @@ impl ODE {
                         (next_event_time, false)
                     };
 
-                    if pending_reinit {
-                        reinitialize_at_boundary(solver, dy_scratch, stop_time);
-                        pending_reinit = false;
-                    }
-
                     solver
                         .problem()
                         .eqn
@@ -737,28 +725,36 @@ impl ODE {
                         });
 
                     match solver.set_stop_time(stop_time) {
-                        Ok(_) => loop {
-                            match solver.step() {
-                                Ok(OdeSolverStopReason::InternalTimestep) => continue,
-                                Ok(OdeSolverStopReason::TstopReached) => {
-                                    solver.problem().eqn.set_left_continuity_time(None);
-                                    if is_infusion_boundary {
-                                        pending_reinit = true;
+                        Ok(_) => {
+                            if pending_reinit {
+                                reinitialize_at_boundary(solver, dy_scratch);
+                                pending_reinit = false;
+                            }
+                            loop {
+                                match solver.step() {
+                                    Ok(OdeSolverStopReason::InternalTimestep) => continue,
+                                    Ok(OdeSolverStopReason::TstopReached) => {
+                                        solver.problem().eqn.set_left_continuity_time(None);
+                                        if is_infusion_boundary {
+                                            pending_reinit = true;
+                                        }
+                                        break;
                                     }
-                                    break;
[…]
