### 2026-08-05 b7a71da
apply same fix to the jit workflow so tests pass
