+++ pharmsol-macros/src/lib.rs
@@ -76,7 +76,7 @@ struct OdeRouteDecl {
     destination: Ident,
 }
 
-#[derive(Clone, Copy)]
+#[derive(Clone, Copy, PartialEq, Eq, Hash)]
 enum OdeRouteKind {
     Bolus,
     Infusion,
@@ -1461,6 +1461,26 @@ fn ode_route_input_bindings(routes: &[OdeRouteDecl]) -> Vec<(SymbolicIndex, usiz
         .collect()
 }
 
+/// Bolus-route input bindings only, numbered in declaration order.
+///
+/// Lag and bioavailability are bolus-only route properties, so property
+/// closures resolve route names against bolus inputs. When a bolus and an
+/// infusion share a label (one drug given by both routes), this keeps the
+/// generated `const` bindings and numeric rewrites unambiguous.
+fn bolus_route_input_bindings(routes: &[OdeRouteDecl]) -> Vec<(SymbolicIndex, usize)> {
+    let mut next_bolus_index = 0usize;
+
+    routes
+        .iter()
+        .filter(|route| matches!(route.kind, OdeRouteKind::Bolus))
+        .map(|route| {
+            let index = next_bolus_index;
+            next_bolus_index += 1;
+            (route.input.clone(), index)
+        })
+        .collect()
+}
+
 fn dense_index_len(bindings: &[(SymbolicIndex, usize)]) -> usize {
     bindings
         .iter()
@@ -2119,6 +2139,10 @@ fn validate_route_property_kinds(
     for route in routes {
         if property_routes.contains(&route.input.name())
             && matches!(route.kind, OdeRouteKind::Infusion)
+            && !routes.iter().any(|other| {
+                matches!(other.kind, OdeRouteKind::Bolus)
+                    && other.input.name() == route.input.name()
+            })
         {
             return Err(syn::Error::new_spanned(
                 &route.input,
@@ -2240,12 +2264,15 @@ fn expand_route_metadata(
                     quote! { ::pharmsol::equation::Route::infusion(stringify!(#input)) }
                 }
             };
-            let lag_flag = if lag_routes.contains(&route_name) {
+            // Lag and bioavailability are bolus-only; when a bolus and an
+            // infusion share a label, the property binds to the bolus route.
+            let bolus_route = matches!(route.kind, OdeRouteKind::Bolus);
+            let lag_flag = if bolus_route && lag_routes.contains(&route_name) {
                 quote! { .with_lag() }
             } else {
                 quote! {}
             };
-            let fa_flag = if fa_routes.contains(&route_name) {
+            let fa_flag = if bolus_route && fa_routes.contains(&route_name) {
                 quote! { .with_bioavailability() }
             } else {
                 quote! {}
@@ -2281,12 +2308,15 @@ fn expand_analytical_route_metadata(
                     quote! { ::pharmsol::equation::Route::infusion(stringify!(#input)) }
                 }
             };
-            let lag_flag = if lag_routes.contains(&route_name) {
+            // Lag and bioavailability are bolus-only; when a bolus and an
+            // infusion share a label, the property binds to the bolus route.
+            let bolus_route = matches!(route.kind, OdeRouteKind::Bolus);
+            let lag_flag = if bolus_route && lag_routes.contains(&route_name) {
                 quote! { .with_lag() }
             } else {
                 quote! {}
             };
-            let fa_flag = if fa_routes.contains(&route_name) {
+            let fa_flag = if bolus_route && fa_routes.contains(&route_name) {
                 quote! { .with_bioavailability() }
             } else {
                 quote! {}
@@ -2321,12 +2351,15 @@ fn expand_sde_route_metadata(
                     quote! { ::pharmsol::equation::Route::infusion(stringify!(#input)) }
[…]
