### 2026-08-06 ac95881
same fix for the model{}-DSL representation
