+++ src/simulator/equation/metadata.rs
@@ -155,8 +155,15 @@ impl ValidatedModelMetadata {
     }
 
     /// Declared route (input) labels, e.g. for error messages.
+    ///
+    /// A label shared by a bolus and an infusion appears once.
     pub fn route_labels(&self) -> Vec<&str> {
-        self.routes.iter().map(ValidatedRoute::name).collect()
+        let mut seen = std::collections::HashSet::with_capacity(self.routes.len());
+        self.routes
+            .iter()
+            .filter(|route| seen.insert(route.name()))
+            .map(ValidatedRoute::name)
+            .collect()
     }
 
     /// Get the number of dense execution input slots needed for routes.
@@ -200,8 +207,20 @@ impl ValidatedModelMetadata {
     }
 
     /// Look up a route by public name and return its declaration-order index.
+    ///
+    /// Legacy name-only lookup: when a bolus and an infusion share a label,
+    /// this returns the first matching declaration. Prefer
+    /// [`Self::route_declaration_index_by_kind`] when the route kind is known.
     pub fn route_declaration_index(&self, name: &str) -> Option<usize> {
         self.routes.iter().position(|route| route.name() == name)
+    }
+
+    /// Look up a route by public name and kind, returning its
+    /// declaration-order index.
+    pub fn route_declaration_index_by_kind(&self, name: &str, kind: RouteKind) -> Option<usize> {
+        self.routes
+            .iter()
+            .position(|route| route.kind() == kind && route.name() == name)
     }
 
     /// Look up an output by public name and return its dense output index.
@@ -241,25 +260,6 @@ impl ValidatedModelMetadata {
             })
     }
 
-    /// Kind-agnostic route label lookup.
-    ///
-    /// Canonical `model {}` routes carry no kind and default to bolus
-    /// metadata, yet remain usable as either a bolus or an infusion input;
-    /// the runtime falls back to this lookup when no route of the requested
-    /// kind matches the label.
-    pub fn route_for_label_any_kind(&self, label: &str) -> Option<&ValidatedRoute> {
-        self.routes
-            .iter()
-            .find(|route| route.name() == label)
-            .or_else(|| {
-                if !is_bare_numeric_label(label) {
-                    return None;
-                }
-                let aliased = format!("{NUMERIC_ROUTE_PREFIX}{label}");
-                self.routes.iter().find(|route| route.name() == aliased)
-            })
-    }
-
     /// Resolve a public output label from data to its dense output index.
     ///
     /// Uses the same resolution order as [`Self::route_for_label`]: exact
@@ -288,8 +288,19 @@ impl ValidatedModelMetadata {
         self.state_index(name).map(|index| &self.states[index])
     }
 
+    /// Look up a route by public name.
+    ///
+    /// Legacy name-only lookup: when a bolus and an infusion share a label,
+    /// this returns the first matching declaration. Prefer [`Self::route_by_kind`]
+    /// when the route kind is known.
     pub fn route(&self, name: &str) -> Option<&ValidatedRoute> {
         self.route_declaration_index(name)
+            .map(|index| &self.routes[index])
+    }
+
+    /// Look up a route by public name and kind.
+    pub fn route_by_kind(&self, name: &str, kind: RouteKind) -> Option<&ValidatedRoute> {
+        self.route_declaration_index_by_kind(name, kind)
             .map(|index| &self.routes[index])
     }
 
@@ -802,7 +813,7 @@ impl Route {
 }
 
 /// Returns `true` for labels consisting only of ASCII digits (`"0"`, `"12"`).
-fn is_bare_numeric_label(label: &str) -> bool {
+pub(crate) fn is_bare_numeric_label(label: &str) -> bool {
     !label.is_empty() && label.chars().all(|ch| ch.is_ascii_digit())
 }
 
[…]
