+++ pharmsol-dsl/src/syntax.rs
@@ -117,7 +117,7 @@ pub struct RoutesBlock {
     pub span: Span,
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Serialize, Deserialize)]
 pub enum RouteKind {
     Bolus,
     Infusion,
