### 2026-08-06 a7445c2
test: kindless canonical routes accept both input kinds across backends
