### 2026-08-06 c879e36
add tests to check for metadata perservation and runtime compatibility
