### 2026-08-06 553ac23
fix: Allow boluses and infusions to go to the same comparment (#326)
* removing restrictions

* DSL path

* same fix for the model{}-DSL representation

* corrections

* test: kindless canonical routes accept both input kinds across backends

* add tests to check for metadata perservation and runtime compatibility
