### 2026-08-06 4b047c1
fix: Force the solver to stop at the infusions end time (#324)
* force the solver to stop at the infusions end time

* fmt

* fix

* the previous implementation was a bit fragile. trying a different approach

* the previous implementation was a bit fragile. trying a different approach

* added some tests and covered a edge case scenario

* apply same fix to the jit workflow so tests pass
