+++ src/dsl/native.rs
@@ -1185,6 +1185,8 @@ impl NativeOdeModel {
         self.shared.validate_support_point(support_point)?;
         let mut output = SubjectPredictions::default();
         let support_vector: V = DVector::from_vec(support_point.to_vec()).into();
+        // Scratch for refreshing the solver's derivative at infusion boundaries.
+        let mut dy_scratch = V::zeros(self.shared.info.state_len, NalgebraContext::new());
 
         for occasion in subject.occasions() {
             let mut events = self.shared.resolve_events(occasion)?;
@@ -1305,6 +1307,7 @@ impl NativeOdeModel {
                         support_point,
                         occasion.covariates(),
                         infusions.as_slice(),
+                        &mut dy_scratch,
                         &mut output,
                         &session,
                         &function_error,
@@ -1339,6 +1342,7 @@ impl NativeOdeModel {
         support_point: &[f64],
         covariates: &Covariates,
         infusions: &[Infusion],
+        dy_scratch: &mut V,
         output: &mut SubjectPredictions,
         session: &RefCell<Box<dyn FunctionSession + '_>>,
         function_error: &RefCell<Option<PharmsolError>>,
@@ -1347,7 +1351,25 @@ impl NativeOdeModel {
         F: Fn(&V, &V, f64, &mut V, &V, &V, &Covariates) + 'a,
         S: OdeSolverMethod<'a, PMProblem<'a, F>>,
     {
-        for (index, event) in events.iter().enumerate() {
+        // Mirror the closure-based ODE event loop: stop at every infusion
+        // start and end boundary in addition to subject events, using the
+        // left-continuous rate while integrating toward a boundary and the
+        // right-continuous rate after reaching it. This keeps the JIT
+        // implementation numerically consistent with the reference [`ODE`]
+        // path (see `ode::run_events`).
+        let infusion_boundary_times = solver.problem().eqn.infusion_boundary_times();
+        let mut infusion_boundary_cursor = 0usize;
+        let mut index = 0usize;
+        // Set when the previous stop was an infusion boundary: the solver must
+        // be restarted before the first step of the next segment (the RHS is
+        // discontinuous at the boundary). Deferred until `set_stop_time`
+        // succeeds so a stop that is already reached does not trigger a
+        // restart for a zero-length segment.
+        let mut pending_reinit = false;
+        while index < events.len() {
+            let event = &events[index];
+            let next_event = events.get(index + 1);
+
             match event {
                 Event::Bolus(bolus) => {
                     let input = bolus.input_index().ok_or_else(|| {
@@ -1379,44 +1401,113 @@ impl NativeOdeModel {
                 }
             }
 
-            if let Some(next_event) = events.get(index + 1) {
-                if event.time() == next_event.time() {
-                    continue;
-                }
+            // Advance to the next event time if it exists.
+            if let Some(next_event) = next_event {
+                let next_event_time = next_event.time();
+                while next_event_time > solver.state().t {
+                    while infusion_boundary_cursor < infusion_boundary_times.len()
+                        && infusion_boundary_times[infusion_boundary_cursor] <= solver.state().t
+                    {
+                        infusion_boundary_cursor += 1;
+                    }
 
-                match solver.set_stop_time(next_event.time()) {
-                    Ok(_) => loop {
-                        match solver.step() {
-                            Ok(_) if function_error.borrow().is_some() => {
-                                return Err(function_error.borrow_mut().take().unwrap());
+                    let (stop_time, is_infusion_boundary) = if let Some(stop_time) =
+                        infusion_boundary_times.get(infusion_boundary_cursor)
+                    {
+                        if *stop_time <= next_event_time {
+                            infusion_boundary_cursor += 1;
+                            (*stop_time, true)
+                        } else {
+                            (next_event_time, false)
+                        }
[…]
