### 2026-08-06 8b45226
feat: Allow macro use in transitive dependencies (#322)
If pharmsol is not imported directly, the ode! macro did not resolve like it was supposed to, as it was hardlocked.

Co-authored-by: Julian Otalvaro <juliandavid347@gmail.com>
