+++ pharmsol-macros/Cargo.toml
@@ -13,6 +13,7 @@ proc-macro = true
 
 [dependencies]
 pharmsol-dsl = { workspace = true }
+proc-macro-crate = "3.4.0"
 proc-macro2 = "1.0.106"
 quote = "1.0.45"
 syn = { version = "3.0.2", features = ["full", "visit", "visit-mut"] }
