+++ pharmsol-macros/CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.28.5](https://github.com/LAPKB/pharmsol/compare/pharmsol-macros-v0.28.4...pharmsol-macros-v0.28.5) - 2026-08-06
+
+### Added
+
+- Allow macro use in transitive dependencies ([#322](https://github.com/LAPKB/pharmsol/pull/322))
+
+### Fixed
+
+- Allow boluses and infusions to go to the same comparment ([#326](https://github.com/LAPKB/pharmsol/pull/326))
+
 ## [0.28.3](https://github.com/LAPKB/pharmsol/compare/pharmsol-macros-v0.28.2...pharmsol-macros-v0.28.3) - 2026-07-28
 
 ### Added
