### 2026-08-06 48b2ce4
Address copilot review
Fixes https://github.com/LAPKB/pharmsol/pull/323#discussion_r3728559833
