### 2026-08-06 377d5ec
chore: release v0.28.5 (#327)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
