### 2026-08-06 f5a6b5d
increased tolerance for stop times
