### 2026-08-07 e8957c0
chore: Bump dependencies (#321)
Co-authored-by: Julian Otalvaro <juliandavid347@gmail.com>
