### 2026-08-07 1a92216
fix: Stop time relative tolerance (#328)
* fix: accept solver stops within a relative tolerance of the current time

diffsol reports StopTimeAtCurrentTime not only for a stop exactly at the
current state time but also when its internal state time landed a few ULPs
past the requested stop. Dense output grids built with floating-point
accumulation (e.g. t += 0.05 over 0..24) routinely place requested times a
few ULPs away from event times, and 0.28.5's event loop turned those stops
into a hard error ("Stop time is at the current state time"), breaking
legitimate dense predictions near events.

R
[…]
