### 2026-08-07 28728f6
multi-line and nested blocks + more testing
