+++ pharmsol-dsl/src/authoring.rs
@@ -103,14 +103,54 @@ impl<'a> AuthoringParser<'a> {
 
     fn parse_module(mut self) -> Result<Module, ParseError> {
         let mut offset = 0;
-        for segment in self.src.split_inclusive('\n') {
+        let mut pending: Option<String> = None;
+        let mut pending_start = 0;
+        let mut brace_depth = 0i32;
+        let mut segments = self.src.split_inclusive('\n').peekable();
+        while let Some(segment) = segments.next() {
             let line = segment.strip_suffix('\n').unwrap_or(segment);
-            self.parse_line(line, offset)?;
+            let code = &line[..line.find('#').unwrap_or(line.len())];
+            let (opens, closes) = brace_counts(code);
+            let starts_with_else = starts_with_keyword(code.trim_start(), "else");
+            if pending.is_none()
+                && brace_depth == 0
+                && opens == 0
+                && closes == 0
+                && !starts_with_else
+            {
+                self.parse_line(line, offset)?;
+                offset += segment.len();
+                continue;
+            }
+
+            if pending.is_none() {
+                pending = Some(String::new());
+                pending_start = offset;
+            }
+            if let Some(text) = &mut pending {
+                text.push_str(code);
+                text.push('\n');
+            }
+            brace_depth += opens - closes;
+
+            if brace_depth <= 0 {
+                let next_is_else = segments.peek().is_some_and(|next| {
+                    let next_line = next.strip_suffix('\n').unwrap_or(next);
+                    let next_code = &next_line[..next_line.find('#').unwrap_or(next_line.len())];
+                    starts_with_keyword(next_code.trim_start(), "else")
+                });
+                if !next_is_else {
+                    let text = pending.take().unwrap();
+                    self.parse_line(text.trim_end_matches('\n'), pending_start)?;
+                    pending_start = 0;
+                    brace_depth = 0;
+                }
+            }
             offset += segment.len();
         }
 
-        if !self.src.is_empty() && !self.src.ends_with('\n') && offset < self.src.len() {
-            self.parse_line(&self.src[offset..], offset)?;
+        if let Some(text) = pending {
+            self.parse_line(text.trim_end_matches('\n'), pending_start)?;
         }
 
         self.validate_declared_outputs_assigned()?;
@@ -320,6 +360,12 @@ impl<'a> AuthoringParser<'a> {
         let span = Span::new(line_offset + leading, line_offset + trailing);
         self.note_span(span);
 
+        if starts_with_keyword(trimmed, "if") {
+            let stmt = self.parse_if_statement(trimmed, span.start, span)?;
+            self.derive_statements.push(stmt);
+            return Ok(());
+        }
+
         if let Some(arrow) = find_top_level_arrow(trimmed) {
             return self.parse_route_line(trimmed, arrow, span.start, span);
         }
@@ -489,6 +535,162 @@ impl<'a> AuthoringParser<'a> {
             self.derive_statements.push(stmt);
         }
         Ok(())
+    }
+
+    fn parse_if_statement(
+        &mut self,
+        text: &str,
+        abs_start: usize,
+        span: Span,
+    ) -> Result<Stmt, ParseError> {
+        let rest = &text[2..];
+        let rest_leading = rest.len() - rest.trim_start().len();
+        let rest = &rest[rest_leading..];
+        let rest_abs = abs_start + 2 + rest_leading;
+        if !rest.starts_with('(') {
[…]
