+++ pharmsol-dsl/tests/dsl_authoring_edge_cases.rs
@@ -794,3 +794,290 @@ out(cp) = central / v ~ continuous()
         rendered
     );
 }
+
+fn collect_if_branch_lens(
+    statements: &[pharmsol_dsl::execution::ExecutionStmt],
+    out: &mut Vec<(usize, Option<usize>)>,
+) {
+    use pharmsol_dsl::execution::ExecutionStmtKind;
+    for statement in statements {
+        match &statement.kind {
+            ExecutionStmtKind::If(if_stmt) => {
+                out.push((
+                    if_stmt.then_branch.len(),
+                    if_stmt.else_branch.as_ref().map(|branch| branch.len()),
+                ));
+                collect_if_branch_lens(&if_stmt.then_branch, out);
+                if let Some(else_branch) = &if_stmt.else_branch {
+                    collect_if_branch_lens(else_branch, out);
+                }
+            }
+            ExecutionStmtKind::For(for_stmt) => {
+                collect_if_branch_lens(&for_stmt.body, out);
+            }
+            _ => {}
+        }
+    }
+}
+
+#[test]
+fn statement_level_if_with_braces_parses_and_lowers() {
+    let src = r#"
+name = braced_if_statement
+kind = ode
+states = central
+params = ke
+ddt(central) = -ke * central
+if (ke > 0.5) {
+    x = 1
+} else {
+    x = 2
+}
+"#;
+    let model = parse_model(src).expect("braced if authoring model parses");
+    let analyzed = analyze_model(&model).expect("braced if model analyzes");
+    let compiled =
+        compile_analyzed_model(&analyzed).expect("braced if model compiles");
+
+    let mut shapes = Vec::new();
+    let function = compiled
+        .function(pharmsol_dsl::execution::ModelFunctionKind::Derive)
+        .expect("dynamics function");
+    if let pharmsol_dsl::execution::FunctionBody::Statements(program) = &function.body {
+        collect_if_branch_lens(&program.body.statements, &mut shapes);
+    }
+    assert_eq!(shapes, vec![(1, Some(1))], "one if with one stmt per branch");
+}
+
+#[test]
+fn else_if_chain_lowers_to_nested_if() {
+    let src = r#"
+name = else_if_chain
+kind = ode
+states = central
+params = ke
+ddt(central) = -ke * central
+if (ke > 1) {
+    x = 1
+} else if (ke > 0.5) {
+    x = 2
+} else {
+    x = 3
+}
+"#;
+    let model = parse_model(src).expect("else-if chain parses");
+    let analyzed = analyze_model(&model).expect("else-if chain analyzes");
+    let compiled =
+        compile_analyzed_model(&analyzed).expect("else-if chain compiles");
+
+    let mut shapes = Vec::new();
+    let function = compiled
+        .function(pharmsol_dsl::execution::ModelFunctionKind::Derive)
+        .expect("dynamics function");
+    if let pharmsol_dsl::execution::FunctionBody::Statements(program) = &function.body {
+        collect_if_branch_lens(&program.body.statements, &mut shapes);
+    }
+    // outer if: then 1, else 1 (nested if); nested if: then 1, else 1
+    assert_eq!(
+        shapes,
+        vec![(1, Some(1)), (1, Some(1))],
+        "outer if nests an inner if in its else branch"
+    );
+}
+
+#[test]
+fn single_line_braced_if_expression_lowers() {
+    let src = r#"
+name = braced_if_expression
+kind = ode
+states = central
+params = ke, v
+ddt(central) = -ke * central
+out(cp) = if (ke > 0.5) { central / v } else { central / (2 * v) }
+"#;
+    let model = parse_model(src).expect("braced if-expression parses");
[…]
