### 2026-08-07 5f10fbd
if/else statement fixes: trailing-token check, trivia before else, comment offsets
