+++ pharmsol-dsl/src/authoring.rs
@@ -106,8 +106,11 @@ impl<'a> AuthoringParser<'a> {
         let mut pending: Option<String> = None;
         let mut pending_start = 0;
         let mut brace_depth = 0i32;
-        let mut segments = self.src.split_inclusive('\n').peekable();
-        while let Some(segment) = segments.next() {
+        let segments: Vec<&str> = self.src.split_inclusive('\n').collect();
+        let mut index = 0;
+        while index < segments.len() {
+            let segment = segments[index];
+            index += 1;
             let line = segment.strip_suffix('\n').unwrap_or(segment);
             let code = &line[..line.find('#').unwrap_or(line.len())];
             let (opens, closes) = brace_counts(code);
@@ -128,17 +131,24 @@ impl<'a> AuthoringParser<'a> {
                 pending_start = offset;
             }
             if let Some(text) = &mut pending {
-                text.push_str(code);
+                text.push_str(&mask_comments(line));
                 text.push('\n');
             }
             brace_depth += opens - closes;
 
             if brace_depth <= 0 {
-                let next_is_else = segments.peek().is_some_and(|next| {
-                    let next_line = next.strip_suffix('\n').unwrap_or(next);
-                    let next_code = &next_line[..next_line.find('#').unwrap_or(next_line.len())];
-                    starts_with_keyword(next_code.trim_start(), "else")
-                });
+                let next_is_else = segments[index..]
+                    .iter()
+                    .find(|next| {
+                        let next_line = next.strip_suffix('\n').unwrap_or(next);
+                        let next_code = &next_line[..next_line.find('#').unwrap_or(next_line.len())];
+                        !next_code.trim().is_empty()
+                    })
+                    .is_some_and(|next| {
+                        let next_line = next.strip_suffix('\n').unwrap_or(next);
+                        let next_code = &next_line[..next_line.find('#').unwrap_or(next_line.len())];
+                        starts_with_keyword(next_code.trim_start(), "else")
+                    });
                 if !next_is_else {
                     let text = pending.take().unwrap();
                     self.parse_line(text.trim_end_matches('\n'), pending_start)?;
@@ -586,15 +596,26 @@ impl<'a> AuthoringParser<'a> {
             let else_rest_leading = else_rest.len() - else_rest.trim_start().len();
             let else_rest_trimmed = &else_rest[else_rest_leading..];
             let else_rest_abs = after_then_abs + 4 + else_rest_leading;
-            if starts_with_keyword(else_rest_trimmed, "if") {
+            let (stmts, else_consumed) = if starts_with_keyword(else_rest_trimmed, "if") {
                 let nested_span = Span::new(else_rest_abs, else_rest_abs + else_rest_trimmed.len());
                 let nested =
                     self.parse_if_statement(else_rest_trimmed, else_rest_abs, nested_span)?;
-                Some(vec![nested])
+                (vec![nested], else_rest_trimmed.len())
             } else {
-                let (stmts, _consumed) = self.parse_if_body(else_rest_trimmed, else_rest_abs)?;
-                Some(stmts)
+                let (stmts, consumed) = self.parse_if_body(else_rest_trimmed, else_rest_abs)?;
+                (stmts, consumed)
+            };
+            let trailing_src = &else_rest_trimmed[else_consumed..];
+            let trailing = trailing_src.trim_start();
+            let trailing_leading = trailing_src.len() - trailing.len();
+            let trailing_abs = else_rest_abs + else_consumed + trailing_leading;
+            if !trailing.is_empty() {
+                return Err(ParseError::new(
+                    "unexpected tokens after `if`/`else` statement",
[…]
