+++ pharmsol-dsl/tests/dsl_authoring_edge_cases.rs
@@ -1081,3 +1081,232 @@ out(cp) = if (ke > 0.5) { 1 } else { 2 } + 3
         rendered
     );
 }
+
+#[test]
+fn else_on_own_line_or_after_trivia_lines_parses() {
+    let variants = [
+        "} else {",
+        "}\nelse {",
+        "}\n\nelse {",
+        "}\n# explanation\nelse {",
+    ];
+    for (i, variant) in variants.iter().enumerate() {
+        let src = format!(
+            r#"
+name = trivia_else_{i}
+kind = ode
+states = central
+params = ke
+ddt(central) = -ke * central
+if (ke > 0.5) {{
+    x = 1
+{variant}
+    x = 2
+}}
+"#
+        );
+        let model = parse_model(&src).unwrap_or_else(|err| {
+            panic!("variant {variant:?} failed: {}", err.render(&src))
+        });
+        let analyzed = analyze_model(&model).expect("model analyzes");
+        let compiled = compile_analyzed_model(&analyzed).expect("model compiles");
+
+        let mut shapes = Vec::new();
+        let function = compiled
+            .function(pharmsol_dsl::execution::ModelFunctionKind::Derive)
+            .expect("dynamics function");
+        if let pharmsol_dsl::execution::FunctionBody::Statements(program) = &function.body {
+            collect_if_branch_lens(&program.body.statements, &mut shapes);
+        }
+        assert_eq!(shapes, vec![(1, Some(1))], "variant {variant:?}");
+    }
+}
+
+#[test]
+fn else_if_chain_with_blank_and_comment_lines_parses() {
+    let src = r#"
+name = chain_with_trivia
+kind = ode
+states = central
+params = ke
+ddt(central) = -ke * central
+if (ke > 1) {
+    x = 1
+}
+
+else if (ke > 0.5) {
+    x = 2
+}
+
+# a comment sitting between the chain links
+else {
+    x = 3
+}
+"#;
+    let model = parse_model(src).expect("else-if chain with trivia parses");
+    let analyzed = analyze_model(&model).expect("model analyzes");
+    let compiled = compile_analyzed_model(&analyzed).expect("model compiles");
+
+    let mut shapes = Vec::new();
+    let function = compiled
+        .function(pharmsol_dsl::execution::ModelFunctionKind::Derive)
+        .expect("dynamics function");
+    if let pharmsol_dsl::execution::FunctionBody::Statements(program) = &function.body {
+        collect_if_branch_lens(&program.body.statements, &mut shapes);
+    }
+    assert_eq!(
+        shapes,
+        vec![(1, Some(1)), (1, Some(1))],
+        "outer if nests an inner if in its else branch"
+    );
+}
+
+#[test]
+fn nested_statement_if_else_across_lines_parses() {
+    let variants = [
+        "} else {",
+        "}\nelse {",
+        "}\n\nelse {",
+        "}\n    # explanation\n    else {",
+    ];
+    for (i, variant) in variants.iter().enumerate() {
+        let src = format!(
+            r#"
+name = nested_if_else_{i}
+kind = ode
+states = central
+params = ke
+ddt(central) = -ke * central
+if (ke > 0.5) {{
+    if (ke > 1) {{
+        x = 1
+    {variant}
+        x = 2
+    }}
+}}
+"#
+        );
+        let model = parse_model(&src).unwrap_or_else(|err| {
+            panic!("variant {variant:?} failed: {}", err.render(&src))
+        });
+        let analyzed = analyze_model(&model).expect("model analyzes");
+        let compiled = compile_analyzed_model(&analyzed).expect("model compiles");
+
+        let mut shapes = Vec::new();
+        let function = compiled
+            .function(pharmsol_dsl::execution::ModelFunctionKind::Derive)
[…]
