### 2026-08-07 4e743b3
regression test: legacy unbraced if-expression RHS still parses and lowers
