### 2026-08-07 cece776
reject then-branch trailing tokens in if-expressions; cap statement-if nesting depth
