### 2026-08-07 3a027b4
end-to-end tests: multi-assignment statement-if branches; definite-assignment failure without else
