+++ pharmsol-dsl/tests/dsl_authoring_edge_cases.rs
@@ -836,8 +836,7 @@ if (ke > 0.5) {
 "#;
     let model = parse_model(src).expect("braced if authoring model parses");
     let analyzed = analyze_model(&model).expect("braced if model analyzes");
-    let compiled =
-        compile_analyzed_model(&analyzed).expect("braced if model compiles");
+    let compiled = compile_analyzed_model(&analyzed).expect("braced if model compiles");
 
     let mut shapes = Vec::new();
     let function = compiled
@@ -846,7 +845,11 @@ if (ke > 0.5) {
     if let pharmsol_dsl::execution::FunctionBody::Statements(program) = &function.body {
         collect_if_branch_lens(&program.body.statements, &mut shapes);
     }
-    assert_eq!(shapes, vec![(1, Some(1))], "one if with one stmt per branch");
+    assert_eq!(
+        shapes,
+        vec![(1, Some(1))],
+        "one if with one stmt per branch"
+    );
 }
 
 #[test]
@@ -867,8 +870,7 @@ if (ke > 1) {
 "#;
     let model = parse_model(src).expect("else-if chain parses");
     let analyzed = analyze_model(&model).expect("else-if chain analyzes");
-    let compiled =
-        compile_analyzed_model(&analyzed).expect("else-if chain compiles");
+    let compiled = compile_analyzed_model(&analyzed).expect("else-if chain compiles");
 
     let mut shapes = Vec::new();
     let function = compiled
@@ -897,8 +899,7 @@ out(cp) = if (ke > 0.5) { central / v } else { central / (2 * v) }
 "#;
     let model = parse_model(src).expect("braced if-expression parses");
     let analyzed = analyze_model(&model).expect("braced if-expression analyzes");
-    let compiled =
-        compile_analyzed_model(&analyzed).expect("braced if-expression compiles");
+    let compiled = compile_analyzed_model(&analyzed).expect("braced if-expression compiles");
 
     let mut shapes = Vec::new();
     let function = compiled
@@ -962,8 +963,7 @@ if (ke > 0.5) { x = 1 } else { x = 2 }
 "#;
     let model = parse_model(src).expect("single-line statement if parses");
     let analyzed = analyze_model(&model).expect("single-line statement if analyzes");
-    let compiled =
-        compile_analyzed_model(&analyzed).expect("single-line statement if compiles");
+    let compiled = compile_analyzed_model(&analyzed).expect("single-line statement if compiles");
 
     let mut shapes = Vec::new();
     let function = compiled
@@ -991,8 +991,7 @@ out(cp) = if (ke > 0.5) {
 "#;
     let model = parse_model(src).expect("multi-line if-expression parses");
     let analyzed = analyze_model(&model).expect("multi-line if-expression analyzes");
-    let compiled =
-        compile_analyzed_model(&analyzed).expect("multi-line if-expression compiles");
+    let compiled = compile_analyzed_model(&analyzed).expect("multi-line if-expression compiles");
 
     let mut shapes = Vec::new();
     let function = compiled
@@ -1021,8 +1020,7 @@ if (ke > 0.5) {
 "#;
     let model = parse_model(src).expect("commented if block parses");
     let analyzed = analyze_model(&model).expect("commented if block analyzes");
-    let compiled =
-        compile_analyzed_model(&analyzed).expect("commented if block compiles");
+    let compiled = compile_analyzed_model(&analyzed).expect("commented if block compiles");
 
     let mut shapes = Vec::new();
     let function = compiled
@@ -1046,8 +1044,7 @@ out(cp) = if (ka > 1) { if (kb > 0.5) { 1 } else { 2 } } else { 3 }
 "#;
     let model = parse_model(src).expect("nested braced if-expression parses");
[…]
