### 2026-08-08 aab4e35
update tests to the new likelihood API
