### 2026-08-10 6ca2771
Feat/if braces (#330)
* multi-line and nested blocks + more testing

* if/else statement fixes: trailing-token check, trivia before else, comment offsets

* regression test: legacy unbraced if-expression RHS still parses and lowers

* reject then-branch trailing tokens in if-expressions; cap statement-if nesting depth

* end-to-end tests: multi-assignment statement-if branches; definite-assignment failure without else

* fmt
