### 2026-08-10 2c68341
refactor: Update prediction structure
