### 2026-08-10 775b750
refactor: Refactor monolith to modules (#323)
* feat: Allow macro use in transitive dependencies

If pharmsol is not imported directly, the ode! macro did not resolve like it was supposed to, as it was hardlocked.

* refactor(macros): split monolithic lib.rs into modules

* Address copilot review

Fixes https://github.com/LAPKB/pharmsol/pull/323#discussion_r3728559833

---------

Co-authored-by: Julian Otalvaro <juliandavid347@gmail.com>
