+++ pharmsol-macros/src/analysis.rs
@@ -1,0 +1,241 @@
+//! Inspection and rewriting of user-written closure bodies.
+
+use proc_macro2::Span;
+use quote::quote;
+use std::collections::{HashMap, HashSet};
+use syn::{
+    parse::Parser, punctuated::Punctuated, visit::Visit, visit_mut::VisitMut, Expr, ExprClosure,
+    Ident, Lit, LitInt, Pat, Token,
+};
+
+use crate::symbols::{RoutePropertyEntry, SymbolicIndex};
+
+fn param_name(pat: &Pat) -> String {
+    match pat {
+        Pat::Ident(p) => p.ident.to_string(),
+        _ => String::new(),
+    }
+}
+
+pub(crate) fn closure_param_names(c: &ExprClosure) -> Vec<String> {
+    c.inputs.iter().map(param_name).collect()
+}
+
+pub(crate) fn closure_param_ident(c: &ExprClosure, index: usize) -> Option<Ident> {
+    c.inputs.get(index).and_then(|pat| match pat {
+        Pat::Ident(pat_ident) => Some(pat_ident.ident.clone()),
+        _ => None,
+    })
+}
+
+pub(crate) fn generated_ident(name: &str) -> Ident {
+    Ident::new(name, Span::call_site())
+}
+
+/// Which names a closure body mentions, so only the referenced bindings are
+/// materialized in the expansion.
+#[derive(Default)]
+pub(crate) struct ClosureBodyUsage {
+    idents: HashSet<String>,
+    indexed_idents: HashSet<String>,
+    assigned_indexed_idents: HashSet<String>,
+    contains_macro: bool,
+}
+
+impl ClosureBodyUsage {
+    pub(crate) fn analyze(expr: &Expr) -> Self {
+        let mut usage = Self::default();
+        usage.visit_expr(expr);
+        usage
+    }
+
+    pub(crate) fn uses(&self, ident: &Ident) -> bool {
+        self.contains_macro || self.idents.contains(&ident.to_string())
+    }
+
+    pub(crate) fn mentions(&self, ident: &Ident) -> bool {
+        self.idents.contains(&ident.to_string())
+    }
+
+    pub(crate) fn indexes(&self, ident: &Ident) -> bool {
+        self.indexed_idents.contains(&ident.to_string())
+    }
+
+    pub(crate) fn assigns_index(&self, ident: &Ident) -> bool {
+        self.assigned_indexed_idents.contains(&ident.to_string())
+    }
+}
+
+impl<'ast> Visit<'ast> for ClosureBodyUsage {
+    fn visit_expr_path(&mut self, expr_path: &'ast syn::ExprPath) {
+        if expr_path.qself.is_none()
+            && expr_path.path.leading_colon.is_none()
+            && expr_path.path.segments.len() == 1
+        {
+            self.idents
+                .insert(expr_path.path.segments[0].ident.to_string());
+        }
+
+        syn::visit::visit_expr_path(self, expr_path);
+    }
+
+    fn visit_expr_macro(&mut self, expr_macro: &'ast syn::ExprMacro) {
+        self.contains_macro = true;
+        syn::visit::visit_expr_macro(self, expr_macro);
+    }
+
+    fn visit_stmt_macro(&mut self, stmt_macro: &'ast syn::StmtMacro) {
+        self.contains_macro = true;
+        syn::visit::visit_stmt_macro(self, stmt_macro);
+    }
+
+    fn visit_expr_index(&mut self, expr_index: &'ast syn::ExprIndex) {
+        if let Expr::Path(expr_path) = expr_index.expr.as_ref() {
+            if expr_path.qself.is_none()
+                && expr_path.path.leading_colon.is_none()
+                && expr_path.path.segments.len() == 1
[…]
