+++ pharmsol-macros/src/expand/analytical.rs
@@ -1,0 +1,522 @@
+//! `analytical!` expansion.
+
+use pharmsol_dsl::{AnalyticalStructureInputKind, AnalyticalStructureInputSource};
+use proc_macro2::{Span, TokenStream as TokenStream2};
+use quote::quote;
+use std::collections::HashSet;
+use syn::{ExprClosure, Ident};
+
+use super::{covariate_metadata, empty_init, empty_route_map};
+use crate::analysis::{
+    closure_param_ident, generated_ident, IndexRewriteTarget, NumericLabelRewriter,
+};
+use crate::bindings::{
+    generate_covariate_bindings, generate_derived_bindings, generate_index_consts,
+    generate_mapped_index_consts, generate_parameter_bindings, generate_supported_input_aliases,
+};
+use crate::crate_path::rewrite_crate_paths;
+use crate::input::AnalyticalInput;
+use crate::kernel::resolve_analytical_structure;
+use crate::symbols::{
+    bolus_route_input_bindings, dense_index_len, ode_route_input_bindings, symbolic_index_bindings,
+    symbolic_numeric_binding_map, OdeRouteDecl, OdeRouteKind, SymbolicIndex,
+};
+use crate::validate::{
+    extract_route_property_routes, validate_analytical_structure_inputs,
+    validate_route_property_kinds,
+};
+
+const MACRO_LABEL: &str = "built-in `analytical!`";
+
+pub(crate) fn expand(input: AnalyticalInput) -> syn::Result<TokenStream2> {
+    let route_bindings = ode_route_input_bindings(&input.routes);
+    let bolus_route_bindings = bolus_route_input_bindings(&input.routes);
+    let function_spec = resolve_analytical_structure(&input.structure)?;
+    let projection = validate_analytical_structure_inputs(
+        &input.structure,
+        function_spec.function,
+        &input.params,
+        &input.derived,
+    )?;
+
+    let lag_routes = route_property_routes(input.lag.as_ref(), "lag", &input.routes)?;
+    let fa_routes = route_property_routes(input.fa.as_ref(), "fa", &input.routes)?;
+
+    let derive = expand_derive(
+        input.derive.as_ref(),
+        &input.params,
+        &input.covariates,
+        &input.derived,
+    )?;
+    let eq = expand_runtime(&function_spec.runtime_path, projection.kind());
+
+    let out = expand_out(
+        &input.out,
+        &input.params,
+        &input.derived,
+        &input.covariates,
+        &input.states,
+        &input.outputs,
+    )?;
+
+    let lag = match input.lag.as_ref() {
+        Some(closure) => expand_route_map(
+            "lag",
+            closure,
+            &input.params,
+            &input.derived,
+            &input.covariates,
+            &bolus_route_bindings,
+        )?,
+        None => empty_route_map(),
+    };
+
+    let fa = match input.fa.as_ref() {
+        Some(closure) => expand_route_map(
+            "fa",
+            closure,
+            &input.params,
+            &input.derived,
+            &input.covariates,
+            &bolus_route_bindings,
+        )?,
+        None => empty_route_map(),
+    };
+
+    let init = match input.init.as_ref() {
+        Some(closure) => expand_init(
+            closure,
+            &input.params,
+            &input.derived,
+            &input.covariates,
+            &input.states,
+        )?,
+        None => empty_init(),
+    };
+
+    let nstates = input.states.len();
+    let ndrugs = dense_index_len(&route_bindings);
+    let nout = input.outputs.len();
+    let name = &input.name;
+    let params = &input.params;
+    let states = &input.states;
+    let outputs = &input.outputs;
[…]

[…]
