### 2026-08-10 1ffbf38
fix(parser): allow addl<0 when evid=4, module organization
