### 2026-08-11 b15910b
refactor: Remove WASM
