+++ src/dsl/compiled_backend_abi.rs
@@ -3,122 +3,32 @@ use serde::{Deserialize, Serialize};
 use super::model_info::NativeModelInfo;
 use pharmsol_dsl::execution::{ExecutionModel, ModelFunctionKind};
 
-#[cfg(any(
-    test,
-    feature = "dsl-aot",
-    feature = "dsl-aot-load",
-    feature = "dsl-wasm",
-    feature = "dsl-wasm-compile"
-))]
+#[cfg(any(test, feature = "dsl-aot", feature = "dsl-aot-load"))]
 pub const API_VERSION_SYMBOL: &str = "pharmsol_dsl_api_version";
-#[cfg(any(
-    test,
-    feature = "dsl-aot",
-    feature = "dsl-aot-load",
-    feature = "dsl-wasm",
-    feature = "dsl-wasm-compile"
-))]
+#[cfg(any(test, feature = "dsl-aot", feature = "dsl-aot-load"))]
 pub const MODEL_INFO_JSON_PTR_SYMBOL: &str = "pharmsol_dsl_model_info_json_ptr";
-#[cfg(any(
-    test,
-    feature = "dsl-aot",
-    feature = "dsl-aot-load",
-    feature = "dsl-wasm",
-    feature = "dsl-wasm-compile"
-))]
+#[cfg(any(test, feature = "dsl-aot", feature = "dsl-aot-load"))]
 pub const MODEL_INFO_JSON_LEN_SYMBOL: &str = "pharmsol_dsl_model_info_json_len";
-#[cfg(any(
-    test,
-    feature = "dsl-aot",
-    feature = "dsl-aot-load",
-    feature = "dsl-wasm",
-    feature = "dsl-wasm-compile"
-))]
+#[cfg(any(test, feature = "dsl-aot", feature = "dsl-aot-load"))]
 pub const DERIVE_SYMBOL: &str = "pharmsol_dsl_kernel_derive";
-#[cfg(any(
-    test,
-    feature = "dsl-aot",
-    feature = "dsl-aot-load",
-    feature = "dsl-wasm",
-    feature = "dsl-wasm-compile"
-))]
+#[cfg(any(test, feature = "dsl-aot", feature = "dsl-aot-load"))]
 pub const DYNAMICS_SYMBOL: &str = "pharmsol_dsl_kernel_dynamics";
-#[cfg(any(
-    test,
-    feature = "dsl-aot",
-    feature = "dsl-aot-load",
-    feature = "dsl-wasm",
-    feature = "dsl-wasm-compile"
-))]
+#[cfg(any(test, feature = "dsl-aot", feature = "dsl-aot-load"))]
 pub const OUTPUTS_SYMBOL: &str = "pharmsol_dsl_kernel_outputs";
-#[cfg(any(
-    test,
-    feature = "dsl-aot",
-    feature = "dsl-aot-load",
-    feature = "dsl-wasm",
-    feature = "dsl-wasm-compile"
-))]
+#[cfg(any(test, feature = "dsl-aot", feature = "dsl-aot-load"))]
 pub const INIT_SYMBOL: &str = "pharmsol_dsl_kernel_init";
-#[cfg(any(
-    test,
-    feature = "dsl-aot",
-    feature = "dsl-aot-load",
-    feature = "dsl-wasm",
-    feature = "dsl-wasm-compile"
-))]
+#[cfg(any(test, feature = "dsl-aot", feature = "dsl-aot-load"))]
 pub const DRIFT_SYMBOL: &str = "pharmsol_dsl_kernel_drift";
-#[cfg(any(
-    test,
-    feature = "dsl-aot",
-    feature = "dsl-aot-load",
-    feature = "dsl-wasm",
[…]
