### 2026-08-11 8776b2b
feat: Refactor Prediction structures
