### 2026-08-12 bb7a6e6
feat: Improve Pmetrics import/export (#313)
* Implement canonical pmetrics CSV v1

* Correct canonical Pmetrics CSV compatibility

* Preserve legacy covariate execution semantics

* Keep canonical CSV fixture LF-only

* Remove Pmetrics CSV document

* cleaning

* Correct Pmetrics CSV export

* Fix Pmetrics CSV edge cases

* Finish Pmetrics CSV fixes

* Reject ambiguous Pmetrics rows

* Address Pmetrics CSV review feedback

* Centralize Pmetrics CSV schema and validation

* answering some comments

* fix(parser): allow addl<0 when evid=4, module organization

* chrore(parser): use i64
