+++ src/data/parser/pmetrics/row.rs
@@ -1,15 +1,8 @@
-//! Row-shaped data ingestion for [`Data`] and [`Subject`] assembly.
+//! Pmetrics-shaped row ingestion for [`Data`] and [`Subject`] assembly.
 //!
-//! Use this module when your source data already looks like rows from a table,
-//! CSV file, database export, or ETL pipeline.
-//!
-//! Choose the ingestion path by source shape:
-//! - Use [`crate::data::builder::SubjectBuilder`] when you want to author a
-//!   schedule directly in Rust.
-//! - Use [`DataRow`] and [`build_data`] when your application already has
-//!   validated row records in memory.
-//! - Use [`crate::data::parser::read_pmetrics`] when the source file already
-//!   follows the Pmetrics column convention.
+//! This module owns the Pmetrics row schema, validation, `ADDL` expansion, and
+//! occasion construction used by the Pmetrics parser. The public row names are
+//! re-exported through their existing paths for API compatibility.
 //!
 //! [`DataRow`] keeps public route and output labels as strings. Labels such as
 //! `"iv"`, `"depot"`, and `"cp"` are preserved through row parsing and later
@@ -36,9 +29,20 @@ use crate::data::*;
 use std::collections::HashMap;
 use thiserror::Error;
 
-/// A format-agnostic representation of one input row.
+fn ensure_finite(value: f64, field: &str, id: &str) -> Result<(), DataError> {
+    if value.is_finite() {
+        Ok(())
+    } else {
+        Err(DataError::NonFiniteValue {
+            field: field.to_string(),
+            id: id.to_string(),
+        })
+    }
+}
+
+/// A common representation of one row-shaped event record.
 ///
-/// [`DataRow`] collects the canonical fields needed to turn one external row
+/// [`DataRow`] collects the common fields needed to turn one external row
 /// into one or more [`Event`] values.
 ///
 /// Build this type from your own column mapping or external schema, then call
@@ -53,6 +57,7 @@ use thiserror::Error;
 /// - `input` and `outeq` preserve the route and output labels from the source data
 /// - `evid`: 0=observation, 1=dose, 4=reset/new occasion
 /// - `addl`: positive=forward in time, negative=backward in time
+/// - a covariate name ending in `!` selects fixed carry-forward interpolation
 ///
 /// # Example
 ///
@@ -84,7 +89,7 @@ pub struct DataRow {
     /// Event time (required)
     pub time: f64,
     /// Event type: 0=observation, 1=dose, 4=reset/new occasion
-    pub evid: i32,
+    pub evid: i64,
     /// Dose amount (for EVID=1)
     pub dose: Option<f64>,
     /// Infusion duration (if > 0, dose is infusion; otherwise bolus)
@@ -134,6 +139,77 @@ impl DataRow {
     /// ```
     pub fn builder(id: impl Into<String>, time: f64) -> DataRowBuilder {
         DataRowBuilder::new(id, time)
+    }
+
+    fn validate(&self) -> Result<(), DataError> {
+        if self.id.is_empty() {
+            return Err(DataError::InvalidDataRow(
+                "subject ID cannot be empty".to_string(),
+            ));
+        }
+        ensure_finite(self.time, "TIME", &self.id)?;
+        for (field, value) in [
+            ("DUR", self.dur),
+            ("DOSE", self.dose),
+            ("II", self.ii),
+            ("OUT", self.out),
+            ("C0", self.c0),
+            ("C1", self.c1),
+            ("C2", self.c2),
+            ("C3", self.c3),
+        ] {
+            if let Some(value) = value {
+                ensure_finite(value, field, &self.id)?;
+            }
+        }
+        for (name, value) in &self.covariates {
[…]
