+++ src/data/parser/pmetrics/mod.rs
@@ -1,0 +1,533 @@
+//! Pmetrics CSV parsing and export helpers.
+//!
+//! This module reads and writes the Pmetrics-style tabular format while keeping
+//! pharmsol's public input and output labels intact.
+//!
+//! `INPUT` and `OUTEQ` values are parsed as labels, not rewritten to dense
+//! indices. Named values such as `iv` and `cp` are preserved exactly, and
+//! numeric values such as `1` are preserved as numeric-looking labels.
+
+use crate::data::*;
+use ::csv::{ReaderBuilder, StringRecord};
+use serde::de::{MapAccess, Visitor};
+use serde::{de, Deserialize, Deserializer};
+use std::collections::{HashMap, HashSet};
+use std::fmt;
+use std::str::FromStr;
+
+mod csv;
+mod row;
+
+#[cfg(test)]
+#[path = "tests.rs"]
+mod csv_tests;
+
+pub use row::{build_data, DataError, DataRow, DataRowBuilder};
+
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[repr(usize)]
+pub(super) enum CoreColumn {
+    Id,
+    Evid,
+    Time,
+    Dur,
+    Dose,
+    Addl,
+    Ii,
+    Input,
+    Out,
+    Outeq,
+    Cens,
+    C0,
+    C1,
+    C2,
+    C3,
+}
+
+impl CoreColumn {
+    pub(super) const ALL: [Self; 15] = [
+        Self::Id,
+        Self::Evid,
+        Self::Time,
+        Self::Dur,
+        Self::Dose,
+        Self::Addl,
+        Self::Ii,
+        Self::Input,
+        Self::Out,
+        Self::Outeq,
+        Self::Cens,
+        Self::C0,
+        Self::C1,
+        Self::C2,
+        Self::C3,
+    ];
+    const REQUIRED: [Self; 3] = [Self::Id, Self::Evid, Self::Time];
+    pub(super) const COUNT: usize = Self::ALL.len();
+
+    pub(super) const fn index(self) -> usize {
+        self as usize
+    }
+
+    pub(super) const fn header(self) -> &'static str {
+        match self {
+            Self::Id => "ID",
+            Self::Evid => "EVID",
+            Self::Time => "TIME",
+            Self::Dur => "DUR",
+            Self::Dose => "DOSE",
+            Self::Addl => "ADDL",
+            Self::Ii => "II",
+            Self::Input => "INPUT",
+            Self::Out => "OUT",
+            Self::Outeq => "OUTEQ",
+            Self::Cens => "CENS",
+            Self::C0 => "C0",
+            Self::C1 => "C1",
+            Self::C2 => "C2",
+            Self::C3 => "C3",
+        }
+    }
+
+    fn from_header(header: &str) -> Option<Self> {
+        Self::ALL
+            .into_iter()
+            .find(|column| column.header().eq_ignore_ascii_case(header))
+    }
+}
+
+pub(super) fn core_headers() -> impl ExactSizeIterator<Item = &'static str> {
+    CoreColumn::ALL.into_iter().map(CoreColumn::header)
+}
+
+/// Read a Pmetrics CSV file into [`Data`].
+///
+/// Use [`read_pmetrics`] when the source file already follows the usual
+/// Pmetrics column convention instead of mapping the file into [`DataRow`]
+/// values yourself.
+///
+/// The parser normalizes header names to lowercase, preserves `INPUT` and
+/// `OUTEQ` as public labels, expands `ADDL` dosing rows through the shared row
+/// ingestion path, and groups rows into occasions using `EVID=4`.
+///
+/// All columns not claimed by the core Pmetrics schema are treated as
[…]
