+++ src/lib.rs
@@ -357,7 +357,7 @@ pub mod prelude {
     pub mod data {
         pub use crate::data::{
             error_model::{AssayErrorModel, AssayErrorModels},
-            parser::{read_pmetrics, DataRow, DataRowBuilder},
+            parser::{read_pmetrics, DataError, DataRow, DataRowBuilder},
             residual_error::{ResidualErrorModel, ResidualErrorModels},
             Covariates, Data, Event, Occasion, Subject,
         };
