### 2026-08-12 88d00c3
ci: Add code coverage
