### 2026-08-12 dff9607
refactor: Remove WASM (#335)
