+++ src/dsl/runtime.rs
@@ -2,7 +2,7 @@
 //!
 //! Use this module when you already know you want an executable model and need
 //! one backend-neutral surface for compile, load, and prediction workflows.
-//! It normalizes the backend-specific JIT, native AoT, and WASM entrypoints so
+//! It normalizes the backend-specific JIT and native AoT entrypoints so
 //! callers can choose a deployment target without rewriting the downstream
 //! prediction code.
 //!
@@ -11,8 +11,6 @@
 //!
 //! - [`super::jit`] for direct in-process JIT compilation.
 //! - [`compile_module_source_to_aot`][crate::dsl::compile_module_source_to_aot] for native artifact export and reload.
-//! - [`compile_module_source_to_wasm_bytes`][crate::dsl::compile_module_source_to_wasm_bytes] and [`load_runtime_wasm_bytes`] for portable WASM bytes,
-//!   browser-loader assets, and host-side WASM loading.
 //!
 //! Main entrypoints:
 //!
@@ -20,8 +18,8 @@
 //!   path.
 //! - [`compile_execution_model_to_runtime`] when you already have an
 //!   [`ExecutionModel`](pharmsol_dsl::ExecutionModel).
-//! - [`load_runtime_artifact`] and [`load_runtime_wasm_bytes`] when the model
-//!   has already been compiled and stored elsewhere.
+//! - [`load_runtime_artifact`] when the model has already been compiled and
+//!   stored elsewhere.
 //! - [`CompiledRuntimeModel::estimate_predictions`] for backend-neutral
 //!   execution against a [`Subject`](crate::Subject).
 //!
@@ -32,9 +30,6 @@
 //! - [`RuntimeCompilationTarget::NativeAot`] emits a native artifact and reloads
 //!   it into the same runtime model shape. Use it when you want reusable native
 //!   artifacts and can control the target platform.
-//! - [`RuntimeCompilationTarget::Wasm`] emits portable WASM bytes and reloads
-//!   them into the host-side runtime adapter. Choose this target when you need a portable
-//!   artifact or browser-aligned deployment story.
 //!
 //! Smallest compile-and-run example:
 //!
@@ -91,18 +86,10 @@ use super::aot::{
 };
 #[cfg(feature = "dsl-jit")]
 use super::jit::{compile_execution_model_to_jit, JitCompileError};
-#[cfg(feature = "dsl-wasm")]
-use super::native::RuntimeArtifact;
 use super::native::{
     CompiledNativeModel, NativeAnalyticalModel, NativeCovariateInfo, NativeModelInfo,
     NativeOdeModel, NativeOutputInfo, NativeRouteInfo, NativeSdeModel, NativeStateInfo,
     RuntimeBackend,
-};
-#[cfg(feature = "dsl-wasm")]
-use super::wasm::{load_wasm_artifact, load_wasm_artifact_bytes};
-#[cfg(feature = "dsl-wasm")]
-use super::wasm_compile::{
-    compile_execution_model_to_wasm_bytes, compile_module_source_to_wasm_bytes, WasmError,
 };
 use crate::{
     simulator::likelihood::{Prediction, SubjectPredictions},
@@ -135,9 +122,6 @@ pub enum RuntimeCompilationTarget {
     /// Export a native artifact and reload it as a runtime model.
     #[cfg(all(feature = "dsl-aot", feature = "dsl-aot-load"))]
     NativeAot(NativeAotCompileOptions),
-    /// Emit WASM bytes and reload them through the host-side WASM runtime.
-    #[cfg(feature = "dsl-wasm")]
-    Wasm,
 }
 
 /// Identifies the on-disk artifact format for [`load_runtime_artifact`].
[…]
