### 2026-08-12 89f34e3
refactor: Update prediction structure (#334)
* refactor: Update prediction structure

* feat: Refactor Prediction structures
