### 2026-08-12 820e7c3
chore: release v0.28.7
