+++ pharmsol-dsl/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.28.7](https://github.com/LAPKB/pharmsol/compare/pharmsol-dsl-v0.28.6...pharmsol-dsl-v0.28.7) - 2026-08-12
+
+### Other
+
+- Remove WASM ([#335](https://github.com/LAPKB/pharmsol/pull/335))
+- Feat/if braces ([#330](https://github.com/LAPKB/pharmsol/pull/330))
+
 ## [0.28.5](https://github.com/LAPKB/pharmsol/compare/pharmsol-dsl-v0.28.4...pharmsol-dsl-v0.28.5) - 2026-08-06
 
 ### Fixed
