### 2026-08-12 48fb367
chore: release v0.28.7 (#333)
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
