+++ src/simulator/equation/sde/paper.rs
@@ -1,0 +1,551 @@
+//! Direct fixed-step simulation used to generate the paper's SDE trajectory figures.
+//!
+//! This module intentionally bypasses the particle-filter workflow. Each entry in
+//! [`PopulationTrajectories::subjects`] is one simulated population subject with
+//! its own fixed `ke0` parameter and complete `X(t)`/`Ke(t)` trajectory.
+
+use std::{fs::File, io, path::Path};
+
+use csv::Writer;
+use nalgebra::DVector;
+use rand::{rngs::StdRng, Rng, RngExt, SeedableRng};
+use rand_distr::{Distribution, StandardNormal};
+use thiserror::Error;
+
+use super::em::EM;
+
+/// Configuration for the paper's two-state stochastic PK simulation.
+#[derive(Clone, Debug, PartialEq)]
+pub struct PaperSimulationConfig {
+    pub seed: u64,
+    pub n_subjects: usize,
+    pub t0: f64,
+    pub tf: f64,
+    pub dt: f64,
+    pub x0: f64,
+    pub volume: f64,
+    pub sigma_x: f64,
+    pub sigma_ke: f64,
+    pub mixture_weight_1: f64,
+    pub mixture_mean_1: f64,
+    pub mixture_sd_1: f64,
+    pub mixture_weight_2: f64,
+    pub mixture_mean_2: f64,
+    pub mixture_sd_2: f64,
+}
+
+impl Default for PaperSimulationConfig {
+    fn default() -> Self {
+        Self {
+            seed: 12_345,
+            n_subjects: 100,
+            t0: 0.0,
+            tf: 1.0,
+            dt: 0.0002,
+            x0: 20.0,
+            volume: 1.0,
+            sigma_x: 0.05,
+            // The manuscript heading says 0.1, while the experiment text says
+            // 0.5. The historical figure appears more compatible with 0.5.
+            sigma_ke: 0.5,
+            mixture_weight_1: 0.5,
+            mixture_mean_1: 0.5,
+            mixture_sd_1: 0.05,
+            mixture_weight_2: 0.5,
+            mixture_mean_2: 1.5,
+            mixture_sd_2: 0.15,
+        }
+    }
+}
+
+/// A complete trajectory for one simulated population subject.
+#[derive(Clone, Debug, PartialEq)]
+pub struct SubjectTrajectory {
+    pub subject: usize,
+    pub mixture_component: usize,
+    pub ke0: f64,
+    pub times: Vec<f64>,
+    pub x: Vec<f64>,
+    pub concentration: Vec<f64>,
+    pub ke: Vec<f64>,
+}
+
+/// Complete trajectories for the simulated population.
+#[derive(Clone, Debug, PartialEq)]
+pub struct PopulationTrajectories {
+    pub subjects: Vec<SubjectTrajectory>,
+}
+
+/// Errors raised by the paper-specific simulation or CSV export.
+#[derive(Debug, Error)]
+pub enum PaperSimulationError {
+    #[error("invalid paper simulation configuration: {0}")]
+    InvalidConfiguration(String),
+    #[error("non-finite value for subject {subject} at step {step}: {state}")]
+    NonFinite {
+        subject: usize,
+        step: usize,
+        state: &'static str,
+    },
+    #[error(transparent)]
+    Io(#[from] io::Error),
+    #[error(transparent)]
+    Csv(#[from] csv::Error),
+}
+
+/// Return the number of fixed Euler--Maruyama updates for an interval.
+pub fn fixed_step_count(t0: f64, tf: f64, dt: f64) -> Result<usize, PaperSimulationError> {
+    if !t0.is_finite() || !tf.is_finite() || !dt.is_finite() {
[…]
