### 2026-08-14 fa384ac
plots 1,2 and 5
