+++ examples/paper_sde_estimation.rs
@@ -1,0 +1,770 @@
+//! Generate the population data and fixed-grid NPMLE distributions for Figures 1, 2, and 5.
+//!
+//! This example deliberately reuses the `population_ke0.csv` produced for
+//! Figures 3 and 4. It keeps the expensive estimation workflow separate from
+//! the trajectory example:
+//!
+//! ```text
+//! cargo run --release --example paper_sde_estimation
+//! /tmp/pharmsol-paper-venv/bin/python paper/plot_population_distributions.py
+//!
+//! The default is the practical reduced fixed-grid run (201 support points,
+//! 256 particles). Add `--manuscript` to request the manuscript's 1000 × 1000
+//! Experiment 2 calculation; it is substantially more expensive with the
+//! fixed-step paper particle filter.
+//! ```
+//!
+//! PMcore was inspected for this branch. Its adaptive NPAG runner expects to
+//! own the model/data likelihood evaluation and is not a dependency of this
+//! paper-only package. This example therefore uses the permitted fixed-grid
+//! NPMLE fallback: the likelihood matrix is computed on a fixed one-dimensional
+//! K0 grid and the mixture weights are optimized by the stable mixture EM
+//! iteration. Experiment 1 uses an exact scalar Gaussian likelihood. Experiment
+//! 2 uses a seeded fixed-step particle filter with the manuscript's reject-and-
+//! hold Ke boundary.
+
+use std::{collections::BTreeMap, env, path::PathBuf};
+
+use csv::{Reader, Writer};
+use rand::{rngs::StdRng, RngExt, SeedableRng};
+use rand_distr::{Distribution, StandardNormal};
+use rayon::prelude::*;
+
+const OBSERVATION_TIMES: [f64; 5] = [0.2, 0.4, 0.6, 0.8, 1.0];
+const X0: f64 = 20.0;
+const VOLUME: f64 = 1.0;
+const SIGMA_X: f64 = 0.05;
+const SIGMA_KE_EXPERIMENT_1: f64 = 0.0;
+const SIGMA_KE_EXPERIMENT_2: f64 = 0.5;
+const PF_DT: f64 = 0.002;
+const K0_LOWER: f64 = 0.3;
+const K0_UPPER: f64 = 3.0;
+const N_SUPPORT_POINTS: usize = 1000;
+const N_PARTICLES: usize = 1000;
+const SIGMA_Y: f64 = 0.5;
+const EXPERIMENT_1_SEED: u64 = 20_260_814;
+const EXPERIMENT_2_SEED: u64 = 20_260_815;
+const FIGURE_5_SAMPLE_SEED: u64 = 20_260_816;
+
+#[derive(Clone, Copy, Debug)]
+struct RunConfig {
+    n_subjects: usize,
+    n_support_points: usize,
+    n_particles: usize,
+    output_dir: &'static str,
+}
+
+impl RunConfig {
+    fn full() -> Self {
+        Self {
+            n_subjects: 100,
+            n_support_points: N_SUPPORT_POINTS,
+            n_particles: N_PARTICLES,
+            output_dir: "paper/output",
+        }
+    }
+
+    fn smoke() -> Self {
+        Self {
+            n_subjects: 10,
+            n_support_points: 50,
+            n_particles: 50,
+            output_dir: "paper/output_estimation_smoke",
+        }
+    }
+
+    fn reduced() -> Self {
+        Self {
+            n_subjects: 100,
+            n_support_points: 201,
+            n_particles: 256,
+            output_dir: "paper/output",
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+struct SubjectObservations {
+    subject: usize,
+    true_ke0: f64,
[…]
