+++ paper/plot_population_distributions.py
@@ -1,0 +1,239 @@
+"""Plot Figures 1, 2, and 5 from paper estimation CSV outputs.
+
+The estimation example owns simulation, likelihoods, and fitted weights. This
+script only reads those numerical outputs and renders empirical/discrete
+population distributions.
+
+Usage:
+    python paper/plot_population_distributions.py
+    python paper/plot_population_distributions.py paper/output
+"""
+
+from __future__ import annotations
+
+import csv
+import math
+import sys
+from bisect import bisect_right
+from collections import defaultdict
+from pathlib import Path
+
+try:
+    import matplotlib.pyplot as plt  # type: ignore[import-not-found]
+except ImportError as exc:
+    raise SystemExit(
+        "matplotlib is required; install it with `python -m pip install matplotlib`"
+    ) from exc
+
+
+DEFAULT_OUTPUT = Path("paper/output")
+OBSERVATION_TIMES = (0.2, 0.4, 0.6, 0.8, 1.0)
+
+
+def read_population(path: Path) -> list[float]:
+    try:
+        with path.open(newline="") as handle:
+            reader = csv.DictReader(handle)
+            values = [float(row["ke0"]) for row in reader]
+    except (OSError, KeyError, TypeError, ValueError, csv.Error) as exc:
+        raise ValueError(f"cannot read population file {path}: {exc}") from exc
+    if not values or any(not math.isfinite(value) for value in values):
+        raise ValueError(f"population file {path} has no finite values")
+    return values
+
+
+def read_distribution(path: Path) -> tuple[list[float], list[float]]:
+    supports: list[float] = []
+    weights: list[float] = []
+    try:
+        with path.open(newline="") as handle:
+            for row in csv.DictReader(handle):
+                supports.append(float(row["ke0"]))
+                weights.append(float(row["weight"]))
+    except (OSError, KeyError, TypeError, ValueError, csv.Error) as exc:
+        raise ValueError(f"cannot read distribution file {path}: {exc}") from exc
+    if not supports or len(supports) != len(weights):
+        raise ValueError(f"distribution file {path} is empty or malformed")
+    if any(not math.isfinite(value) for value in supports + weights) or any(
+        weight < 0 for weight in weights
+    ):
+        raise ValueError(f"distribution file {path} has invalid values")
+    total = sum(weights)
+    if not math.isclose(total, 1.0, rel_tol=0.0, abs_tol=1e-8):
+        raise ValueError(f"distribution weights in {path} sum to {total}, not 1")
+    return supports, weights
+
+
+def read_sample(path: Path) -> list[float]:
+    try:
+        with path.open(newline="") as handle:
+            values = [float(row["ke0"]) for row in csv.DictReader(handle)]
+    except (OSError, KeyError, TypeError, ValueError, csv.Error) as exc:
+        raise ValueError(f"cannot read sample file {path}: {exc}") from exc
+    if not values or any(not math.isfinite(value) for value in values):
+        raise ValueError(f"sample file {path} has no finite values")
+    return values
+
+
+def validate_observations(path: Path, expected_subjects: int = 100) -> None:
+    rows: dict[int, list[float]] = defaultdict(list)
+    try:
+        with path.open(newline="") as handle:
+            for row in csv.DictReader(handle):
+                rows[int(row["subject"])].append(float(row["time"]))
[…]
