### 2026-08-17 #10580 @rkurchin
👋 Hi, congrats on being nearly done! As AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks, some of which may require your attention:
- [ ] please spell out USA in affiliation
- [ ] first paragraph of State of the Field: hyphenate "well-known"
- [ ] second paragraph of Research Impact statement: replace "researching" with "investigating" to avoid duplication within the same sentence
- [ ] in AI usage statement, missing serial comma after "produce tests"
