+++ .github/workflows/coverage.yml
@@ -47,7 +47,7 @@ jobs:
                     --lcov --output-path lcov.info
 
             - name: Upload coverage to Codecov
-              uses: codecov/codecov-action@v5
+              uses: codecov/codecov-action@v7
               with:
                   token: ${{ secrets.CODECOV_TOKEN }}
                   files: lcov.info
