### 2026-08-18 da1ff95
chore: Correct spelling erros and typography
Adding Oxford commas, consistent use of American english, and some missing prepositions.
