### 2026-08-18 #10580 @mhovd
Dear @rkurchin, thank you for following up.

I have addressed your comments in two commits to the `joss` branch, specifically in commits [f91321a](https://github.com/LAPKB/pharmsol/pull/340/commits/f91321ab4adc47eb7cbd4232ee0cb40d403239dc) and [da1ff95](https://github.com/LAPKB/pharmsol/pull/340/commits/da1ff951e5a6246c011d9c5fa11221e334256045).

Please let me know if any further action is required.
