### 2026-08-18 f8c5c22
fix BDF restart after bolus events
