### 2026-08-18 84eb9e6
fix numeric equality in DSL
