### 2026-08-18 68297ae
Improve error message for boolean and numeric comparison
