+++ benches/dsl_matrix.rs
@@ -1,4 +1,4 @@
-//! DSL bench matrix (feature-gated): JIT, native AoT across all workloads + solvers.
+//! DSL bench matrix (feature-gated): JIT across all workloads + solvers.
 //! Mirrors `native_matrix.rs` but compiles models from DSL source.
 //!
 //! IDs:
@@ -8,15 +8,13 @@
 //! - `dsl/likelihood-matrix` → `{workload}/{kind}/{backend}`
 
 use std::hint::black_box;
-use std::path::PathBuf;
 use std::time::Duration;
 
 use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, SamplingMode};
-use tempfile::TempDir;
 
 use pharmsol::dsl::{
-    compile_module_source_to_runtime, CompiledRuntimeModel, NativeAnalyticalModel,
-    NativeAotCompileOptions, NativeOdeModel, NativeSdeModel, RuntimeCompilationTarget,
+    compile_module_source_to_runtime, CompiledRuntimeModel, NativeAnalyticalModel, NativeOdeModel,
+    NativeSdeModel, RuntimeCompilationTarget,
 };
 use pharmsol::prelude::*;
 use pharmsol::{Cache, Parameters};
@@ -31,21 +29,17 @@ const MATRIX_N_SUBJECTS: usize = 32;
 const MATRIX_N_SUPPORT: usize = 64;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[allow(dead_code)] // Aot temporarily disabled in `Backend::all`
 enum Backend {
     Jit,
-    Aot,
 }
 
 impl Backend {
     fn label(self) -> &'static str {
         match self {
             Self::Jit => "dsl-jit",
-            Self::Aot => "dsl-aot",
         }
     }
 
-    // AoT backend temporarily disabled — too slow for the current matrix.
     fn all() -> [Backend; 1] {
         [Backend::Jit]
     }
@@ -70,52 +64,19 @@ impl CacheState {
     }
 }
 
-/// One `TempDir` shared across the bench binary; each compile gets a fresh subdir.
-struct AotWorkspace {
-    root: TempDir,
-    counter: std::cell::Cell<usize>,
-}
-
-impl AotWorkspace {
-    fn new() -> Self {
-        Self {
-            root: tempfile::Builder::new()
-                .prefix("pharmsol-bench-dsl-aot-")
-                .tempdir()
-                .expect("create AoT workspace tempdir"),
-            counter: std::cell::Cell::new(0),
-        }
-    }
-
-    fn fresh(&self, stem: &str) -> PathBuf {
-        let n = self.counter.get();
-        self.counter.set(n + 1);
-        self.root.path().join(format!("{stem}-{n:04}"))
-    }
-}
-
 /// Compile `(workload, kind)` with `backend` and return the full `CompiledRuntimeModel`.
-fn compile_runtime(
-    workload: Workload,
-    kind: SolverKind,
-    backend: Backend,
-    aot: &AotWorkspace,
-) -> CompiledRuntimeModel {
+fn compile_runtime(workload: Workload, kind: SolverKind, backend: Backend) -> CompiledRuntimeModel {
     let source = dsl_source(workload, kind);
     let name = dsl_model_name(workload, kind);
     let target = match backend {
         Backend::Jit => RuntimeCompilationTarget::Jit,
-        Backend::Aot => {
-            let dir = aot.fresh(&format!("{}-{}", workload.label(), kind.label()));
-            RuntimeCompilationTarget::NativeAot(NativeAotCompileOptions::new(dir))
-        }
     };
     compile_module_source_to_runtime(source, Some(name), target, |_, _| {})
[…]
