+++ README.md
@@ -108,18 +108,13 @@ see [docs/analytical-authoring-migration.md](docs/analytical-authoring-migration
 ## DSL and Runtime Targets
 
 If the model needs to be loaded or compiled at runtime, pharmsol also provides a DSL with
-the same broad modeling coverage: ODE, analytical, and SDE authoring. The DSL can target
-an in-process JIT runtime or native ahead-of-time artifacts depending on how you want to
-ship and execute the model.
-
-- `dsl-jit`: compile DSL source into a runtime model inside the current process.
-- `dsl-aot` and `dsl-aot-load`: emit a native artifact and load it later.
+the same broad modeling coverage: ODE, analytical, and SDE authoring. Enable the `dsl-jit`
+feature to compile DSL source into a runtime model inside the current process.
 
 See [examples/dsl_runtime_jit.rs](examples/dsl_runtime_jit.rs) for the in-repo JIT flow and
 [examples/dsl_jit_analytical_covariates.rs](examples/dsl_jit_analytical_covariates.rs) for a
 small analytical covariate example written both as DSL JIT source and as an
 `analytical!` model.
-The companion `pharmsol-examples` crate includes an end-to-end native AOT runtime example.
 
 ## Performance
 
