+++ Cargo.toml
@@ -37,13 +37,10 @@ dsl-jit = [
     "dep:cranelift-module",
     "dep:cranelift-native",
 ]
-dsl-aot = ["dsl-core"]
-dsl-aot-load = ["dsl-core", "dep:libloading"]
 
 [dependencies]
 pharmsol-dsl = { workspace = true }
 pharmsol-macros = { workspace = true }
-libloading = { version = "0.9.0", optional = true, features = [] }
 cranelift = { version = "0.134.3", optional = true }
 cranelift-jit = { version = "0.134.3", optional = true }
 cranelift-module = { version = "0.134.3", optional = true }
@@ -83,4 +80,4 @@ harness = false
 [[bench]]
 name = "dsl_matrix"
 harness = false
-required-features = ["dsl-jit", "dsl-aot", "dsl-aot-load"]
+required-features = ["dsl-jit"]
