### 2026-08-18 #10580 @rkurchin
Congrats to @mhovd and thanks to handling editor @csoneson  and reviewers @Armavica and @axiomcura. JOSS wouldn't exist without your volunteer efforts!
