### 2026-08-18 #10580 @editorialbot
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
