### 2026-08-18 #10580 @mhovd
> Congrats to [@mhovd](https://github.com/mhovd) and thanks to handling editor [@csoneson](https://github.com/csoneson) and reviewers [@Armavica](https://github.com/Armavica) and [@axiomcura](https://github.com/axiomcura). JOSS wouldn't exist without your volunteer efforts!

Thank you very much for accepting our paper, and thank you again to the editors and reviewers for their meaningful suggestions!
